How does a call print at the console?

call_print_type() returns the way a call is deparsed and printed at the console. This is useful when generating documents based on R code. The types of calls are:

"prefix" for calls like foo(), unary operators, -1, and operators with more than 2 arguments like `+`(1, 2, 3) (the latter can be obtained by building calls manually).

"infix" for operators like 1 + 2 or foo$bar.

"special" for function definitions, control-flow calls like if or for, and subscripting calls like foo[] and foo[[]].

call_print_type(call)

Arguments

call
A quoted function call. An error is raised if not a call.

Examples

call_print_type(quote(foo(bar)))

#> [1] "prefix"

call_print_type(quote(foo[[bar]]))

#> [1] "special"

call_print_type(quote(+foo))

#> [1] "prefix"

call_print_type(quote(function() foo))

#> [1] "special"

# When an operator call has an artificial number of arguments, R
# often prints it in prefix form:
call <- call("+", 1, 2, 3)
call

#> `+`(1, 2, 3)

call_print_type(call)

#> [1] "prefix"

# But not always:
call <- call("$", 1, 2, 3)
call

#> 1$2

call_print_type(call)

#> [1] "infix"