call_standardise {rlang}	R Documentation

Standardise a call

Description

Deprecated in rlang 0.4.11 in favour of call_match(). call_standardise() was designed for call wrappers that include an environment like formulas or quosures. The function definition was plucked from that environment. However in practice it is rare to use it with wrapped calls, and then it's easy to forget to supply the environment. For these reasons, we have designed call_match() as a simpler wrapper around match.call().

This is essentially equivalent to base::match.call(), but with experimental handling of primitive functions.

Usage

call_standardise(call, env = caller_env())

Arguments

call

env

Value

A quosure if call is a quosure, a raw call otherwise.

[Package rlang version 1.0.1 Index]