caller_frame {rlang}	R Documentation

Get properties of the current or caller frame

Description

The current frame is the execution context of the function that is currently being evaluated.

The caller frame is the execution context of the function that called the function currently being evaluated.

See the call stack topic for more information.

Usage

caller_frame(n = 1)

caller_fn(n = 1)

current_fn()

Arguments

n
The number of generation to go back.

See Also

caller_env() and current_env()

[Package rlang version 0.2.0.9001 Index]