check_dots_empty0 {rlang}	R Documentation

Check that dots are empty (low level variant)

Description

check_dots_empty0() is a more efficient version of check_dots_empty() with a slightly different interface. Instead of inspecting the current environment for dots, it directly takes .... It is only meant for very low level functions where a couple microseconds make a difference.

Usage

check_dots_empty0(..., call = caller_env())

Arguments

...
Dots which should be empty.

[Package rlang version 1.0.1 Index]