deprecated-cnd {rlang}	R Documentation

Deprecated condition constructors

Description

These functions were deprecated in rlang 0.2.0 to follow the convention that return types are indicated as suffixes. Please use cnd(), error_cnd(), warning_cnd() and message_cnd() instead.

Usage

new_cnd(.type = NULL, ..., .msg = NULL)

cnd_error(.type = NULL, ..., .msg = NULL)

cnd_warning(.type = NULL, ..., .msg = NULL)

cnd_message(.type = NULL, ..., .msg = NULL)

Arguments

.type
The condition subclass.

...
Named data fields stored inside the condition object. These dots are evaluated with explicit splicing.

.msg
A default message to inform the user about the condition when it is signalled.

[Package rlang version 0.2.0.9001 Index]