dots_definitions {rlang}	R Documentation

Capture definition objects

Description

Capture definition objects

Usage

dots_definitions(..., .named = FALSE, .ignore_empty = c("trailing", "none",
  "all"))

Arguments

...
For enexprs(), ensyms() and enquos(), names of arguments to capture without evaluation (including ...). For exprs() and quos(), the expressions to capture unevaluated (including expressions contained in ...).

.named
Whether to ensure all dots are named. Unnamed elements are processed with expr_text() to figure out a default name. If an integer, it is passed to the width argument of expr_text(), if TRUE, the default width is used. See exprs_auto_name().

.ignore_empty
Whether to ignore empty arguments. Can be one of "trailing", "none", "all". If "trailing", only the last argument is ignored if it is empty.

Life cycle

dots_definitions() is experimental. Expect API changes.

[Package rlang version 0.2.0.9001 Index]