dots_list {rlang}	R Documentation

Collect dots tidily

Description

dots_list(...) is equivalent to list(...) but provides tidy dots semantics:

You can splice other lists with the unquote-splice !!! operator.

You can unquote names by using the unquote operator !! on the left-hand side of :=.

We call quasiquotation support in dots tidy dots semantics and functions taking dots with dots_list() tidy dots functions. Quasiquotation is an alternative to do.call() idioms and gives the users of your functions an uniform syntax to supply a variable number of arguments or a variable name.

Usage

dots_list(..., .ignore_empty = c("trailing", "none", "all"))

Arguments

...
Arguments with explicit (dots_list()) or list (dots_splice()) splicing semantics. The contents of spliced arguments are embedded in the returned list.

.ignore_empty
Whether to ignore empty arguments. Can be one of "trailing", "none", "all". If "trailing", only the last argument is ignored if it is empty.

Details

Note that while all tidy eval quoting functions have tidy dots semantics, not all tidy dots functions are quoting functions. dots_list() is for standard functions, not quoting functions.

Value

A list of arguments. This list is always named: unnamed arguments are named with the empty string "".

Life cycle

dots_list() returns named dots, even if no names were supplied. This behaviour is for consistency with dots returned by exprs() or quos() or their capturing variants. However we now question this feature. It makes sense in quoting functions that are building up function calls and where usage of names2() was rather systematic before tidy evaluation. However in the case of standard functions such as created by dots_list(), the cost of allocating an empty character vector of names can be expensive.

For this reason we may break the API in the future by returning NULL names when no arguments were named. Please use dots_list() accordingly (i.e. set_names(dots, names2(dots))).

See Also

exprs() for extracting dots without evaluation.

Examples

# Let's create a function that takes a variable number of arguments:
numeric <- function(...) {
  dots <- dots_list(...)
  num <- as.numeric(dots)
  set_names(num, names(dots))
}
numeric(1, 2, 3)

# The main difference with list(...) is that dots_list(...) enables
# the `!!!` syntax to splice lists:
x <- list(2, 3)
numeric(1, !!! x, 4)

# As well as unquoting of names:
nm <- "yup!"
numeric(!!nm := 1)

# One useful application of splicing is to work around exact and
# partial matching of arguments. Let's create a function taking
# named arguments and dots:
fn <- function(data, ...) {
  dots_list(...)
}

# You normally cannot pass an argument named `data` through the dots
# as it will match `fn`'s `data` argument. The splicing syntax
# provides a workaround:
fn("wrong!", data = letters)  # exact matching of `data`
fn("wrong!", dat = letters)   # partial matching of `data`
fn(some_data, !!! list(data = letters))  # no matching

[Package rlang version 0.1.6.9003 Index]