Get the empty environment

The empty environment is the only one that does not have a parent. It is always used as the tail of an environment chain such as the search path (see search_envs()).

empty_env()

Examples

# Create environments with nothing in scope:
child_env(empty_env())

#> <environment: 0x7f85c88cbe78>