Tidy evaluation in a custom environment

This function is soft-deprecated as of rlang 0.2.0.

eval_tidy_(expr, bottom, top = NULL, env = caller_env())

Arguments

expr
An expression to evaluate.

bottom
The environment containing masking objects if the data mask is one environment deep. The bottom environment if the data mask comprises multiple environment.

top
The last environment of the data mask. If the data mask is only one environment deep, top should be the same as bottom.

env
The environment in which to evaluate expr. This environment is always ignored when evaluating quosures. Quosures are evaluated in their own environment.