Ensure that list of expressions are all named

This gives default names to unnamed elements of a list of expressions (or expression wrappers such as formulas or quosures). exprs_auto_name() deparses the expressions with expr_text() by default. quos_auto_name() deparses with quo_text().

exprs_auto_name(exprs, width = 60L, printer = expr_text)

quos_auto_name(quos, width = 60L)

Arguments

exprs
A list of expressions.

width
Maximum width of names.

printer
A function that takes an expression and converts it to a string. This function must take an expression as first argument and width as second argument.

quos
A list of quosures.