global_handle {rlang}	R Documentation

Register default global handlers

Description

global_handle() sets up a default configuration for error, warning, and message handling. It calls:

global_entrace() to enable rlang errors and warnings globally.

global_prompt_install() to recover from packageNotFoundErrors with a user prompt to install the missing package. Note that at the time of writing (R 4.1), there are only very limited situations where this handler works.

Usage

global_handle(entrace = TRUE, prompt_install = TRUE)

Arguments

entrace
Passed as enable argument to global_entrace().

prompt_install
Passed as enable argument to global_prompt_install().

[Package rlang version 1.0.1 Index]