interrupt {rlang}	R Documentation

Simulate interrupt condition

Description

interrupt() simulates a user interrupt of the kind that is signalled with Ctrl-C. It is currently not possible to create custom interrupt condition objects.

Usage

interrupt()

[Package rlang version 1.0.2 Index]