is_pairlist {rlang}	R Documentation

Is object a node or pairlist?

Description

is_pairlist() checks that x has type pairlist or NULL. NULL is treated as a pairlist because it is the terminating node of pairlists and an empty pairlist is thus the NULL object itself.

is_node() checks that x has type pairlist.

In other words, is_pairlist() tests for the data structure while is_node() tests for the internal type.

Usage

is_pairlist(x)

is_node(x)

Arguments

x
Object to test.

See Also

is_lang() tests for language nodes.

[Package rlang version 0.1.6 Index]