Is an object a quosure or quosure-like?

These predicates test for quosure objects.

is_quosure() tests for a tidyeval quosure. These are one-sided formulas with a quosure class.

is_quosureish() tests for general R quosure objects: quosures, or one-sided formulas.

is_quosure(x)

is_quosureish(x, scoped = NULL)

Arguments

x
An object to test.

scoped
A boolean indicating whether the quosure or formula is scoped, that is, has a valid environment attribute. If NULL, the scope is not inspected.

See also

is_formula() and is_formulaish()

Examples

# Quosures are created with quo():
quo(foo)

#> <quosure: local>
#> ~foo

is_quosure(quo(foo))

#> [1] TRUE

# Formulas look similar to quosures but are not quosures:
is_quosure(~foo)

#> [1] FALSE

# But they are quosureish:
is_quosureish(~foo)

#> [1] TRUE

# Note that two-sided formulas are never quosureish:
is_quosureish(a ~ b)

#> [1] FALSE