is_quosureish {rlang}	R Documentation

Test for or coerce to quosure-like objects

Description

These functions are deprecated as of rlang 0.2.0 because they make the assumption that quosures are a subtype of formula, which we are now considering to be an implementation detail.

Usage

is_quosureish(x, scoped = NULL)

as_quosureish(x, env = caller_env())

Arguments

x
An object to test.

scoped
A boolean indicating whether the quosure is scoped, that is, has a valid environment attribute. If NULL, the scope is not inspected.

env
The environment in which the expression should be evaluated. Only used for symbols and calls. This should typically be the environment in which the expression was created.

[Package rlang version 0.3.4 Index]