last_error {rlang}	R Documentation

Last abort() error

Description

last_error() returns the last error entraced by abort() or global_entrace(). The error is printed with a backtrace in simplified form.

last_trace() is a shortcut to return the backtrace stored in the last error. This backtrace is printed in full form.

Usage

last_error()

last_trace()

See Also

rlang_backtrace_on_error to control what is displayed when an error is thrown.

global_entrace() to enable last_error() logging for all errors.

last_warnings() and last_messages().

[Package rlang version 1.0.2 Index]