modify {rlang}	R Documentation

Modify a vector

Description

This function merges a list of arguments into a vector. It always returns a list.

Usage

modify(.x, ...)

Arguments

.x
A vector to modify.

...
List of elements to merge into .x. Named elements already existing in .x are used as replacements. Elements that have new or no names are inserted at the end. These dots support tidy dots features.

Value

A modified vector upcasted to a list.

Life cycle

modify() is in the qestioning stage. We are still figuring out what vector tools belong in rlang.

Examples

modify(c(1, b = 2, 3), 4, b = "foo")

x <- list(a = 1, b = 2)
y <- list(b = 3, c = 4)
modify(x, splice(y))

[Package rlang version 0.3.4 Index]