names_inform_repair {rlang}	R Documentation

Inform about name repair

Description

Inform about name repair

Usage

names_inform_repair(old, new)

Arguments

old
Original names vector.

new
Repaired names vector.

Muffling and silencing messages

Name repair messages are signaled with inform() and are given the class "rlib_message_name_repair". These messages can be muffled with base::suppressMessages().

Name repair messages can also be silenced with the global option rlib_name_repair_verbosity. This option takes the values:

"verbose": Always verbose.

"quiet": Always quiet.

When set to quiet, the message is not displayed and the condition is not signaled. This is particularly useful for silencing messages during testing when combined with local_options().

[Package rlang version 1.0.1 Index]