new-vector-along-retired {rlang}	R Documentation

Create vectors matching the length of a given vector

Description

These functions are deprecated as of rlang 0.3.0 because they are longer to type than the equivalent rep_along() or rep_named() calls without added clarity.

Usage

new_logical_along(x, names = base::names(x))

new_integer_along(x, names = base::names(x))

new_double_along(x, names = base::names(x))

new_character_along(x, names = base::names(x))

new_complex_along(x, names = base::names(x))

new_raw_along(x, names = base::names(x))

new_list_along(x, names = base::names(x))

Arguments

x
A vector.

names
Names for the new vector.

[Package rlang version 0.4.12 Index]