ns_registry_env {rlang}	R Documentation

Return the namespace registry env

Description

Note that the namespace registry does not behave like a normal environment because the parent is NULL instead of the empty environment. This is exported for expert usage in development tools only.

Usage

ns_registry_env()

[Package rlang version 1.0.4 Index]