parse_quo {rlang}	R Documentation

Parsing variants that return quosures

Description

These functions are in the questioning stage because they don't add value compared to using parse_expr() with new_quosure().

Usage

parse_quo(x, env)

parse_quos(x, env)

Arguments

x
Text containing expressions to parse_expr for parse_expr() and parse_exprs(). Can also be an R connection, for instance to a file. If the supplied connection is not open, it will be automatically closed and destroyed.

env
The environment for the quosures. Depending on the use case, a good default might be the global environment but you might also want to evaluate the R code in an isolated context (perhaps a child of the global environment or of the base environment).

[Package rlang version 0.4.8 Index]