Parse text into a quosure

These functions were soft-deprecated and renamed to parse_quo() and parse_quos() in rlang 0.2.0. This is for consistency with the convention that suffixes indicating return types are not abbreviated.

parse_quosure(x, env = caller_env())

Arguments

x
Text containing expressions to parse_expr for parse_expr() and parse_exprs(). Can also be an R connection, for instance to a file. If the supplied connection is not open, it will be automatically closed and destroyed.

env
The environment for the quosures. Depending on the use case, a good default might be the global environment but you might also want to evaluate the R code in an isolated context (perhaps a child of the global environment or of the base environment).