quo-predicates {rlang}	R Documentation

Is a quosure quoting a symbolic, missing or NULL object?

Description

These functions examine the expression of a quosure with a predicate.

Usage

quo_is_missing(quo)

quo_is_symbol(quo)

quo_is_lang(quo)

quo_is_symbolic(quo)

quo_is_null(quo)

Arguments

quo
A quosure.

Empty quosures

When missing arguments are captured as quosures, either through enquo() or quos(), they are returned as an empty quosure. These quosures contain the missing argument and typically have the empty environment as enclosure.

Examples

quo_is_symbol(quo(sym))
quo_is_symbol(quo(foo(bar)))

# You can create empty quosures by calling quo() without input:
quo <- quo()
quo_is_missing(quo)
is_missing(f_rhs(quo))

[Package rlang version 0.1.6 Index]