raw_deparse_str {rlang}	R Documentation

Serialize a raw vector to a string

Description

This function converts a raw vector to a hexadecimal string, optionally adding a prefix and a suffix. It is roughly equivalent to paste0(prefix, paste(format(x), collapse = ""), suffix) and much faster.

Usage

raw_deparse_str(x, prefix = NULL, suffix = NULL)

Arguments

x
A raw vector.

prefix, suffix
Prefix and suffix strings, or 'NULL.

Value

A string.

Examples

raw_deparse_str(raw())
raw_deparse_str(charToRaw("string"))
raw_deparse_str(raw(10), prefix = "'0x", suffix = "'")

[Package rlang version 1.0.1 Index]