rlang-package {rlang}	R Documentation

rlang: Functions for Base Types and Core R and 'Tidyverse' Features

Description

A toolbox for working with base types, core R features like the condition system, and core 'Tidyverse' features like tidy evaluation.

Author(s)

Maintainer: Lionel Henry lionel@rstudio.com

Authors:

Hadley Wickham hadley@rstudio.com

Other contributors:

mikefc mikefc@coolbutuseless.com (Hash implementation based on Mike's xxhashlite) [copyright holder]

Yann Collet (Author of the embedded xxHash library) [copyright holder]

RStudio [copyright holder]

See Also

Useful links:

https://rlang.r-lib.org

https://github.com/r-lib/rlang

Report bugs at https://github.com/r-lib/rlang/issues

[Package rlang version 1.0.1 Index]