rlang__backtrace_on_error {rlang}	R Documentation

Display backtrace on error

Description

Errors thrown with abort() automatically save a backtrace that can be inspected by calling last_error(). Optionally, you can also display the backtrace alongside the error message by setting the experimental option rlang__backtrace_on_error to one of the following values:

"reminder": Display a reminder that the backtrace can be inspected by calling rlang::last_error().

"branch": Display a simplified backtrace.

"collapse": Display a collapsed backtrace tree.

"full": Display the full backtrace tree.

Promote base errors to rlang errors

Call options(error = rlang:::add_backtrace) to instrument base errors with rlang features. This handler does two things:

It saves the base error as an rlang object. This allows you to call last_error() to print the backtrace or inspect its data.

It prints the backtrace for the current error according to the rlang__backtrace_on_error option.

This experimental handler is not exported because it is meant for interactive use only.

Examples

# Display a simplified backtrace on error for both base and rlang
# errors:

# options(
#   rlang__backtrace_on_error = "branch",
#   error = rlang:::add_backtrace
# )
# stop("foo")

[Package rlang version 0.3.0.1 Index]