Trim top call layers from the evaluation stack

ctxt_stack() can be tricky to use in real code because all intervening frames are returned with the stack, including those at ctxt_stack() own call site. stack_trim() makes it easy to remove layers of intervening calls.

stack_trim(stack, n = 1)

Arguments

stack
An evaluation stack.

n
The number of call frames (not eval frames) to trim off the top of the stack. In other words, the number of layers of intervening frames to trim.

Life cycle

These functions are soft-deprecated and replaced by trace_back().

Examples

# Intervening frames appear on the evaluation stack:
identity(identity(ctxt_stack()))

#> [[1]]
#> <frame 37> (35)
#> expr: identity(ctxt_stack())
#> env:  [local 0xba58fe0]
#> 
#> [[2]]
#> <frame 36> (35)
#> expr: identity(identity(ctxt_stack()))
#> env:  [local 0xba590f8]
#> 
#> [[3]]
#> <frame 35> (34)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x9db04a8]
#> 
#> [[4]]
#> <frame 34> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0xba592b8]
#> 
#> [[5]]
#> <frame 33> (24)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0xba594e8]
#> 
#> [[6]]
#> <frame 32> (24)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0xba59a98]
#> 
#> [[7]]
#> <frame 31> (30)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0xba5a010]
#> 
#> [[8]]
#> <frame 30> (29)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0xb96e3e8]
#> 
#> [[9]]
#> <frame 29> (28)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0xb96e730]
#> 
#> [[10]]
#> <frame 28> (27)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0xb96edf8]
#> 
#> [[11]]
#> <frame 27> (26)
#> expr: try(f, silent = TRUE)
#> env:  [local 0xb96f028]
#> 
#> [[12]]
#> <frame 26> (24)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0xb96f1b0]
#> 
#> [[13]]
#> <frame 25> (24)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0xb96f290]
#> 
#> [[14]]
#> <frame 24> (23)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0xb75e208]
#> 
#> [[15]]
#> <frame 23> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x8123d98]
#> 
#> [[16]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x8da0488]
#> 
#> [[17]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0xbba8e50]
#> 
#> [[18]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x9cfe680]
#> 
#> [[19]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0xa0c06f8]
#> 
#> [[20]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x23a49b0]
#> 
#> [[21]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x3468ee0]
#> 
#> [[22]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x8fea6c8]
#> 
#> [[23]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x6567340]
#> 
#> [[24]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x1d4ab00]
#> 
#> [[25]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x20a8108]
#> 
#> [[26]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors)  <...>
#> env:  [frame 0x39122b8]
#> 
#> [[27]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x39132e8]
#> 
#> [[28]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpD1cB4A/file3a5a45f44243"),  <...>
#> env:  [frame 0x3902fc0]
#> 
#> [[29]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpD1cB4A/file3a5a45f44243"),  <...>
#> env:  [frame 0x3903d50]
#> 
#> [[30]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x38feb28]
#> 
#> [[31]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x38ffce0]
#> 
#> [[32]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x3900028]
#> 
#> [[33]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x38fc5b0]
#> 
#> [[34]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x38fcb60]
#> 
#> [[35]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x38fcea8]
#> 
#> [[36]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x38fd1f0]
#> 
#> [[37]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x38fd960]
#> 
#> [[38]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# stack_trim() will trim the first n layers of calls:
stack_trim(identity(identity(ctxt_stack())))

#> [[1]]
#> <frame 35> (34)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x9db04a8]
#> 
#> [[2]]
#> <frame 34> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0xa89fcc0]
#> 
#> [[3]]
#> <frame 33> (24)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0xa89fef0]
#> 
#> [[4]]
#> <frame 32> (24)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0xa8a04a0]
#> 
#> [[5]]
#> <frame 31> (30)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0xa8a0a18]
#> 
#> [[6]]
#> <frame 30> (29)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0xa8a0d60]
#> 
#> [[7]]
#> <frame 29> (28)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0xa8a10a8]
#> 
#> [[8]]
#> <frame 28> (27)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0xa8a1850]
#> 
#> [[9]]
#> <frame 27> (26)
#> expr: try(f, silent = TRUE)
#> env:  [local 0xa89dd68]
#> 
#> [[10]]
#> <frame 26> (24)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0xa89def0]
#> 
#> [[11]]
#> <frame 25> (24)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0xa89dfd0]
#> 
#> [[12]]
#> <frame 24> (23)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x9a06b28]
#> 
#> [[13]]
#> <frame 23> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x8123d98]
#> 
#> [[14]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x8da0488]
#> 
#> [[15]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0xbba8e50]
#> 
#> [[16]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x9cfe680]
#> 
#> [[17]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0xa0c06f8]
#> 
#> [[18]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x23a49b0]
#> 
#> [[19]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x3468ee0]
#> 
#> [[20]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x8fea6c8]
#> 
#> [[21]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x6567340]
#> 
#> [[22]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x1d4ab00]
#> 
#> [[23]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x20a8108]
#> 
#> [[24]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors)  <...>
#> env:  [frame 0x39122b8]
#> 
#> [[25]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x39132e8]
#> 
#> [[26]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpD1cB4A/file3a5a45f44243"),  <...>
#> env:  [frame 0x3902fc0]
#> 
#> [[27]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpD1cB4A/file3a5a45f44243"),  <...>
#> env:  [frame 0x3903d50]
#> 
#> [[28]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x38feb28]
#> 
#> [[29]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x38ffce0]
#> 
#> [[30]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x3900028]
#> 
#> [[31]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x38fc5b0]
#> 
#> [[32]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x38fcb60]
#> 
#> [[33]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x38fcea8]
#> 
#> [[34]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x38fd1f0]
#> 
#> [[35]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x38fd960]
#> 
#> [[36]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# Note that it also takes care of calls intervening at its own call
# site:
identity(identity(
  stack_trim(identity(identity(ctxt_stack())))
))

#> [[1]]
#> <frame 35> (34)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x9db04a8]
#> 
#> [[2]]
#> <frame 34> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0xad92de8]
#> 
#> [[3]]
#> <frame 33> (24)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0xad93018]
#> 
#> [[4]]
#> <frame 32> (24)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0xad935c8]
#> 
#> [[5]]
#> <frame 31> (30)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0xad8fd10]
#> 
#> [[6]]
#> <frame 30> (29)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0xad90058]
#> 
#> [[7]]
#> <frame 29> (28)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0xad903a0]
#> 
#> [[8]]
#> <frame 28> (27)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0xad90a68]
#> 
#> [[9]]
#> <frame 27> (26)
#> expr: try(f, silent = TRUE)
#> env:  [local 0xad90c98]
#> 
#> [[10]]
#> <frame 26> (24)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0xad90e20]
#> 
#> [[11]]
#> <frame 25> (24)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0xad90f00]
#> 
#> [[12]]
#> <frame 24> (23)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x9e9ba00]
#> 
#> [[13]]
#> <frame 23> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x8123d98]
#> 
#> [[14]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x8da0488]
#> 
#> [[15]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0xbba8e50]
#> 
#> [[16]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x9cfe680]
#> 
#> [[17]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0xa0c06f8]
#> 
#> [[18]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x23a49b0]
#> 
#> [[19]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x3468ee0]
#> 
#> [[20]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x8fea6c8]
#> 
#> [[21]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x6567340]
#> 
#> [[22]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x1d4ab00]
#> 
#> [[23]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x20a8108]
#> 
#> [[24]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors)  <...>
#> env:  [frame 0x39122b8]
#> 
#> [[25]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x39132e8]
#> 
#> [[26]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpD1cB4A/file3a5a45f44243"),  <...>
#> env:  [frame 0x3902fc0]
#> 
#> [[27]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpD1cB4A/file3a5a45f44243"),  <...>
#> env:  [frame 0x3903d50]
#> 
#> [[28]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x38feb28]
#> 
#> [[29]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x38ffce0]
#> 
#> [[30]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x3900028]
#> 
#> [[31]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x38fc5b0]
#> 
#> [[32]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x38fcb60]
#> 
#> [[33]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x38fcea8]
#> 
#> [[34]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x38fd1f0]
#> 
#> [[35]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x38fd960]
#> 
#> [[36]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# It is especially useful when used within a function that needs to
# inspect the evaluation stack but should nonetheless be callable
# within nested calls without side effects:
stack_util <- function() {
  # n = 2 means that two layers of intervening calls should be
  # removed: The layer at ctxt_stack()'s call site (including the
  # stack_trim() call), and the layer at stack_util()'s call.
  stack <- stack_trim(ctxt_stack(), n = 2)
  stack
}
user_fn <- function() {
  # A user calls your stack utility with intervening frames:
  identity(identity(stack_util()))
}
# These intervening frames won't appear in the evaluation stack
identity(user_fn())

#> [[1]]
#> <frame 37> (35)
#> expr: user_fn()
#> env:  [local 0xa1323f0]
#> 
#> [[2]]
#> <frame 36> (35)
#> expr: identity(user_fn())
#> env:  [local 0xa132460]
#> 
#> [[3]]
#> <frame 35> (34)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x9db04a8]
#> 
#> [[4]]
#> <frame 34> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0xa132620]
#> 
#> [[5]]
#> <frame 33> (24)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0xa132850]
#> 
#> [[6]]
#> <frame 32> (24)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0xa132e00]
#> 
#> [[7]]
#> <frame 31> (30)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0xa133378]
#> 
#> [[8]]
#> <frame 30> (29)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0xa1336c0]
#> 
#> [[9]]
#> <frame 29> (28)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0xa12fbd8]
#> 
#> [[10]]
#> <frame 28> (27)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0xa1302d8]
#> 
#> [[11]]
#> <frame 27> (26)
#> expr: try(f, silent = TRUE)
#> env:  [local 0xa130508]
#> 
#> [[12]]
#> <frame 26> (24)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0xa130690]
#> 
#> [[13]]
#> <frame 25> (24)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0xa130770]
#> 
#> [[14]]
#> <frame 24> (23)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0xae501c8]
#> 
#> [[15]]
#> <frame 23> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x8123d98]
#> 
#> [[16]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x8da0488]
#> 
#> [[17]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0xbba8e50]
#> 
#> [[18]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x9cfe680]
#> 
#> [[19]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0xa0c06f8]
#> 
#> [[20]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x23a49b0]
#> 
#> [[21]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x3468ee0]
#> 
#> [[22]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x8fea6c8]
#> 
#> [[23]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x6567340]
#> 
#> [[24]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x1d4ab00]
#> 
#> [[25]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x20a8108]
#> 
#> [[26]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors)  <...>
#> env:  [frame 0x39122b8]
#> 
#> [[27]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x39132e8]
#> 
#> [[28]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpD1cB4A/file3a5a45f44243"),  <...>
#> env:  [frame 0x3902fc0]
#> 
#> [[29]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpD1cB4A/file3a5a45f44243"),  <...>
#> env:  [frame 0x3903d50]
#> 
#> [[30]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x38feb28]
#> 
#> [[31]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x38ffce0]
#> 
#> [[32]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x3900028]
#> 
#> [[33]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x38fc5b0]
#> 
#> [[34]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x38fcb60]
#> 
#> [[35]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x38fcea8]
#> 
#> [[36]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x38fd1f0]
#> 
#> [[37]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x38fd960]
#> 
#> [[38]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"