Promote all errors to rlang errors

with_abort() promotes all errors as if they were thrown with abort(). These errors embed a backtrace. They are particularly suitable to be set as parent errors (see parent argument of abort()).

with_abort(expr)

Arguments

expr
An expression run in a context where errors are promoted to rlang errors.

Details

with_abort() installs a calling handler for errors and rethrows non-rlang errors with abort(). However, error handlers installed within with_abort() have priority. For this reason, you should use tryCatch() and exiting handlers outside with_abort() rather than inside.

Examples

# with_abort() automatically casts simple errors thrown by stop()
# to rlang errors:
fn <- function() stop("Base error")
try(with_abort(fn()))
last_error()

#> <error>
#> message: Base error
#> class:   `rlang_error`
#> backtrace:
#>  ─(function (...) ...
#>  ─pkgdown::build_site(...)
#>  ─pkgdown:::build_site_local(...)
#>  ─pkgdown::build_reference(...)
#>  ─purrr::map(...)
#>  ─pkgdown:::.f(.x[[i]], ...)
#>  ─pkgdown:::data_reference_topic(...)
#>  ─pkgdown:::as_data.tag_examples(...)
#>  ─purrr::pmap_chr(...)
#>  ─pkgdown:::.f(...)
#>  ─evaluate::evaluate(code, env, new_device = TRUE)
#>  ─evaluate:::evaluate_call(...)
#>  ─[ base::eval(...) ] with 1 more call
#>  ─global::fn()
#>  ─base::stop("Base error")
#> Call `summary(rlang::last_error())` to see the full backtrace

# with_abort() is handy for rethrowing low level errors. The
# backtraces are then segmented between the low level and high
# level contexts.
low_level1 <- function() low_level2()
low_level2 <- function() stop("Low level error")

high_level <- function() {
  with_handlers(
    with_abort(low_level1()),
    error = ~ abort("High level error", parent = .x)
  )
}

try(high_level())
last_error()

#> <error>
#> message: High level error
#> class:   `rlang_error`
#> backtrace:
#>  ─base::tryCatch(...)
#>  ─pkgdown::build_site(...)
#>  ─pkgdown:::build_site_local(...)
#>  ─pkgdown::build_reference(...)
#>  ─purrr::map(...)
#>  ─pkgdown:::.f(.x[[i]], ...)
#>  ─pkgdown:::data_reference_topic(...)
#>  ─pkgdown:::as_data.tag_examples(...)
#>  ─purrr::pmap_chr(...)
#>  ─pkgdown:::.f(...)
#>  ─evaluate::evaluate(code, env, new_device = TRUE)
#>  ─evaluate:::evaluate_call(...)
#>  ─[ base::eval(...) ] with 1 more call
#>  ─global::high_level()
#> <error: parent>
#> message: Low level error
#> class:   `rlang_error`
#> backtrace:
#>  ─rlang::with_abort(low_level1())
#>  ─global::low_level1()
#>  ─global::low_level2()
#>  ─base::stop("Low level error")
#> Call `summary(rlang::last_error())` to see the full backtrace

summary(last_error())

#> <error>
#> message: High level error
#> class:   `rlang_error`
#> fields:  `message`, `trace` and `parent`
#> backtrace:
#> █
#> ├─base::tryCatch(...)
#> │ └─base:::tryCatchList(expr, classes, parentenv, handlers)
#> │   ├─base:::tryCatchOne(...)
#> │   │ └─base:::doTryCatch(return(expr), name, parentenv, handler)
#> │   └─base:::tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> │     └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> │       └─base:::doTryCatch(return(expr), name, parentenv, handler)
#> ├─base::withCallingHandlers(...)
#> ├─base::saveRDS(...)
#> ├─base::do.call(...)
#> ├─(function (what, args, quote = FALSE, envir = parent.frame()) ...
#> └─(function (...) ...
#>   └─pkgdown::build_site(...)
#>     └─pkgdown:::build_site_local(...)
#>       └─pkgdown::build_reference(...)
#>         └─purrr::map(...)
#>           └─pkgdown:::.f(.x[[i]], ...)
#>             └─pkgdown:::data_reference_topic(...)
#>               ├─pkgdown:::as_data(...)
#>               └─pkgdown:::as_data.tag_examples(...)
#>                 └─purrr::pmap_chr(...)
#>                   └─pkgdown:::.f(...)
#>                     └─evaluate::evaluate(code, env, new_device = TRUE)
#>                       └─evaluate:::evaluate_call(...)
#>                         ├─evaluate:::timing_fn(...)
#>                         ├─evaluate:::handle(...)
#>                         │ └─base::try(f, silent = TRUE)
#>                         │   └─base::tryCatch(...)
#>                         │     └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>                         │       └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>                         │         └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>                         ├─base::withCallingHandlers(...)
#>                         ├─base::withVisible(eval(expr, envir, enclos))
#>                         └─base::eval(expr, envir, enclos)
#>                           └─base::eval(expr, envir, enclos)
#>                             ├─base::try(high_level())
#>                             │ └─base::tryCatch(...)
#>                             │   └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>                             │     └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>                             │       └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>                             └─global::high_level()
#> <error: parent>
#> message: Low level error
#> class:   `rlang_error`
#> fields:  `message`, `trace`, `parent` and `error`
#> backtrace:
#> █
#> ├─rlang::with_abort(low_level1())
#> │ └─base::withCallingHandlers(...) /Users/lionel/Dropbox/Projects/R/hadley/rlang/R/cnd.R:851:2
#> └─global::low_level1()
#>   └─global::low_level2()
#>     └─base::stop("Low level error")