as_overscope {rlang}	R Documentation

Create an overscope

Description

These functions have been soft-deprecated in rlang 0.2.0. Please use as_data_mask() and new_data_mask() instead. We no longer require the mask to be cleaned up so overscope_clean() does not have a replacement.

Usage

as_overscope(quo, data = NULL)

new_overscope(bottom, top = NULL, enclosure = NULL)

overscope_clean(overscope)

Arguments

quo
A quosure.

data
A data frame or named vector of masking data.

bottom
The environment containing masking objects if the data mask is one environment deep. The bottom environment if the data mask comprises multiple environment.

top
The last environment of the data mask. If the data mask is only one environment deep, top should be the same as bottom.

enclosure
The parent argument of new_data_mask().

overscope
A data mask.

[Package rlang version 0.3.4 Index]