dictionary {rlang}	R Documentation

Create a dictionary

Description

The dictionary class is defunct as of rlang 0.2.0. It was trying to be too general and did not prove useful. Please use as_data_pronoun() or your own pronoun class instead.

Usage

as_dictionary(x, lookup_msg = NULL, read_only = FALSE)

is_dictionary(x)

Arguments

x
An object for which you want to find associated data.

lookup_msg
An error message when your data source is accessed inappropriately (by position rather than name).

read_only
Whether users can replace elements of the dictionary.

[Package rlang version 0.3.4 Index]