Evaluate dots with preliminary splicing

This is a tool for advanced users. It captures dots, processes unquoting and splicing operators, and evaluates them. Unlike dots_list() and dots_splice(), it does not flatten spliced objects. They are merely attributed a spliced class (see splice()). You can process spliced objects manually, perhaps with a custom predicate (see flatten_if()).

dots_values(..., .ignore_empty = c("trailing", "none", "all"))

Arguments

...
Arguments to evaluate and process splicing operators.

.ignore_empty
Whether to ignore empty arguments. Can be one of "trailing", "none", "all". If "trailing", only the last argument is ignored if it is empty.

Examples

dots <- dots_values(!!! list(1))
dots

#> [[1]]
#> [[1]]
#> [1] 1
#> 
#> attr(,"class")
#> [1] "spliced"
#> 

# Flatten the spliced objects:
flatten_if(dots, is_spliced)

#> [[1]]
#> [1] 1
#>