empty_env {rlang}	R Documentation

Get the empty environment

Description

The empty environment is the only one that does not have a parent. It is always used as the tail of an environment chain such as the search path (see search_envs()).

Usage

empty_env()

Examples

# Create environments with nothing in scope:
child_env(empty_env())

[Package rlang version 1.0.1 Index]