env_unbind {rlang}	R Documentation

Remove bindings from an environment

Description

env_unbind() is the complement of env_bind(). Like env_has(), it ignores the parent environments of env by default. Set inherit to TRUE to track down bindings in parent environments.

Usage


env_unbind(env = caller_env(), nms, inherit = FALSE)

Arguments

env
An environment.

nms
A character vector containing the names of the bindings to remove.

inherit
Whether to look for bindings in the parent environments.

Value

The input object env with its associated environment modified in place, invisibly.

Examples


data <- set_names(as.list(letters), letters)
env_bind(environment(), !!! data)
env_has(environment(), letters)

# env_unbind() removes bindings:
env_unbind(environment(), letters)
env_has(environment(), letters)

# With inherit = TRUE, it removes bindings in parent environments
# as well:
parent <- child_env(NULL, foo = "a")
env <- child_env(parent, foo = "b")
env_unbind(env, "foo", inherit = TRUE)
env_has(env, "foo", inherit = TRUE)

[Package rlang version 0.4.3 Index]