Evaluate an expression in an environment

eval_bare() is a lightweight version of the base function base::eval(). It does not accept supplementary data, but it is more efficient and does not clutter the evaluation stack. Technically, eval_bare() is a simple wrapper around the C function Rf_eval().

eval_bare(expr, env = parent.frame())

Arguments

expr
An expression to evaluate.

env
The environment in which to evaluate the expression.

Details

base::eval() inserts two call frames in the stack, the second of which features the envir parameter as frame environment. This may unnecessarily clutter the evaluation stack and it can change evaluation semantics with stack sensitive functions in the case where env is an evaluation environment of a stack frame (see ctxt_stack()). Since the base function eval() creates a new evaluation context with env as frame environment there are actually two contexts with the same evaluation environment on the stack when expr is evaluated. Thus, any command that looks up frames on the stack (stack sensitive functions) may find the parasite frame set up by eval() rather than the original frame targetted by env. As a result, code evaluated with base::eval() does not have the property of stack consistency, and stack sensitive functions like base::return(), base::parent.frame() may return misleading results.

See also

with_env

Examples

# eval_bare() works just like base::eval():
env <- child_env(NULL, foo = "bar")
expr <- quote(foo)
eval_bare(expr, env)

#> [1] "bar"

# To explore the consequences of stack inconsistent semantics, let's
# create a function that evaluates `parent.frame()` deep in the call
# stack, in an environment corresponding to a frame in the middle of
# the stack. For consistency we R's lazy evaluation semantics, we'd
# expect to get the caller of that frame as result:
fn <- function(eval_fn) {
  list(
    returned_env = middle(eval_fn),
    actual_env = get_env()
  )
}
middle <- function(eval_fn) {
  deep(eval_fn, get_env())
}
deep <- function(eval_fn, eval_env) {
  expr <- quote(parent.frame())
  eval_fn(expr, eval_env)
}

# With eval_bare(), we do get the expected environment:
fn(rlang::eval_bare)

#> $returned_env
#> <environment: 0x7f8494a2adf0>
#> 
#> $actual_env
#> <environment: 0x7f8494a2adf0>
#> 

# But that's not the case with base::eval():
fn(base::eval)

#> $returned_env
#> <environment: 0x7f8494d74278>
#> 
#> $actual_env
#> <environment: 0x7f8494d7a860>
#> 

# Another difference of eval_bare() compared to base::eval() is
# that it does not insert parasite frames in the evaluation stack:
get_stack <- quote(identity(ctxt_stack()))
eval_bare(get_stack)

#> [[1]]
#> <frame 33> (31)
#> expr: identity(ctxt_stack())
#> env:  [local 0x7f848bd1ee60]
#> 
#> [[2]]
#> <frame 32> (31)
#> expr: eval_bare(get_stack)
#> env:  [local 0x7f848bd1f4a0]
#> 
#> [[3]]
#> <frame 31> (30)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f8494ed13f0]
#> 
#> [[4]]
#> <frame 30> (20)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f848bf10178]
#> 
#> [[5]]
#> <frame 29> (20)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f848bf0fdc0]
#> 
#> [[6]]
#> <frame 28> (20)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f848e25ca40]
#> 
#> [[7]]
#> <frame 27> (26)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f848b938d78]
#> 
#> [[8]]
#> <frame 26> (25)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f848b938790]
#> 
#> [[9]]
#> <frame 25> (24)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f848bb175f0]
#> 
#> [[10]]
#> <frame 24> (23)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f848e4a1878]
#> 
#> [[11]]
#> <frame 23> (22)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f848e4a2628]
#> 
#> [[12]]
#> <frame 22> (20)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f848e4a2468]
#> 
#> [[13]]
#> <frame 21> (20)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f848e4a2350]
#> 
#> [[14]]
#> <frame 20> (19)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f848e378e28]
#> 
#> [[15]]
#> <frame 19> (18)
#> expr: evaluate::evaluate(text, code_env, new_device = TRUE)
#> env:  [frame 0x7f8494c4e518]
#> 
#> [[16]]
#> <frame 18> (16)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f8494d658d0]
#> 
#> [[17]]
#> <frame 17> (16)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f848ab6d548]
#> 
#> [[18]]
#> <frame 16> (12)
#> expr: data_reference_topic(topic, pkg, path = path, examples = examples,  <...>
#> env:  [frame 0x7f848a7a6e30]
#> 
#> [[19]]
#> <frame 15> (14)
#> expr: stopifnot(is.list(x), is.list(val))
#> env:  [frame 0x7f848a7a6390]
#> 
#> [[20]]
#> <frame 14> (13)
#> expr: utils::modifyList(data, data_template(pkg, depth = depth))
#> env:  [frame 0x7f848a7a6588]
#> 
#> [[21]]
#> <frame 13> (12)
#> expr: render_page(pkg, "reference-topic", data = data_reference_topic(topic,  <...>
#> env:  [frame 0x7f848a6ca608]
#> 
#> [[22]]
#> <frame 12> (11)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f848a7abb10]
#> 
#> [[23]]
#> <frame 11> (10)
#> expr: purrr::map(., build_reference_topic, path, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f848aba6200]
#> 
#> [[24]]
#> <frame 10> (8)
#> expr: function_list[[k]](value)
#> env:  [frame 0x7f848abc34d8]
#> 
#> [[25]]
#> <frame 9> (8)
#> expr: withVisible(function_list[[k]](value))
#> env:  [frame 0x7f848abc33c0]
#> 
#> [[26]]
#> <frame 8> (7)
#> expr: freduce(value, `_function_list`)
#> env:  [frame 0x7f848abd1468]
#> 
#> [[27]]
#> <frame 7> (6)
#> expr: `_fseq`(`_lhs`)
#> env:  [frame 0x7f848abd12a8]
#> 
#> [[28]]
#> <frame 6> (5)
#> expr: eval(expr, envir, enclos)
#> env:  [frame 0x7f848adb4230]
#> 
#> [[29]]
#> <frame 5> (3)
#> expr: eval(quote(`_fseq`(`_lhs`)), env, env)
#> env:  [frame 0x7f848abd1078]
#> 
#> [[30]]
#> <frame 4> (3)
#> expr: withVisible(eval(quote(`_fseq`(`_lhs`)), env, env))
#> env:  [frame 0x7f848abd25f0]
#> 
#> [[31]]
#> <frame 3> (2)
#> expr: pkg$topics %>% purrr::transpose() %>% purrr::map(build_reference_topic,  <...>
#> env:  [frame 0x7f848adb4d78]
#> 
#> [[32]]
#> <frame 2> (1)
#> expr: build_reference(pkg, lazy = FALSE, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f8492ff3f40]
#> 
#> [[33]]
#> <frame 1> (0)
#> expr: pkgdown::build_site("/Users/yoplait/Dropbox/Projects/R/hadley/rlang")
#> env:  [frame 0x7f848fb31d48]
#> 
#> [[34]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

eval(get_stack)

#> [[1]]
#> <frame 34> (31)
#> expr: identity(ctxt_stack())
#> env:  [local 0x7f8493e29cd8]
#> 
#> [[2]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f8494ed13f0]
#> 
#> [[3]]
#> <frame 32> (31)
#> expr: eval(get_stack)
#> env:  [local 0x7f8493e298e8]
#> 
#> [[4]]
#> <frame 31> (30)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f8494ed13f0]
#> 
#> [[5]]
#> <frame 30> (20)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f8493bb9060]
#> 
#> [[6]]
#> <frame 29> (20)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f8493bb8df8]
#> 
#> [[7]]
#> <frame 28> (20)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f8493bba230]
#> 
#> [[8]]
#> <frame 27> (26)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f8493901d10]
#> 
#> [[9]]
#> <frame 26> (25)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f8493901920]
#> 
#> [[10]]
#> <frame 25> (24)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f8493bf3f80]
#> 
#> [[11]]
#> <frame 24> (23)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f848d0e6d48]
#> 
#> [[12]]
#> <frame 23> (22)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f848d0e6ae0]
#> 
#> [[13]]
#> <frame 22> (20)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f848d0e68e8]
#> 
#> [[14]]
#> <frame 21> (20)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f848d2b6f78]
#> 
#> [[15]]
#> <frame 20> (19)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f8493d3a078]
#> 
#> [[16]]
#> <frame 19> (18)
#> expr: evaluate::evaluate(text, code_env, new_device = TRUE)
#> env:  [frame 0x7f8494c4e518]
#> 
#> [[17]]
#> <frame 18> (16)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f8494d658d0]
#> 
#> [[18]]
#> <frame 17> (16)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f848ab6d548]
#> 
#> [[19]]
#> <frame 16> (12)
#> expr: data_reference_topic(topic, pkg, path = path, examples = examples,  <...>
#> env:  [frame 0x7f848a7a6e30]
#> 
#> [[20]]
#> <frame 15> (14)
#> expr: stopifnot(is.list(x), is.list(val))
#> env:  [frame 0x7f848a7a6390]
#> 
#> [[21]]
#> <frame 14> (13)
#> expr: utils::modifyList(data, data_template(pkg, depth = depth))
#> env:  [frame 0x7f848a7a6588]
#> 
#> [[22]]
#> <frame 13> (12)
#> expr: render_page(pkg, "reference-topic", data = data_reference_topic(topic,  <...>
#> env:  [frame 0x7f848a6ca608]
#> 
#> [[23]]
#> <frame 12> (11)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f848a7abb10]
#> 
#> [[24]]
#> <frame 11> (10)
#> expr: purrr::map(., build_reference_topic, path, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f848aba6200]
#> 
#> [[25]]
#> <frame 10> (8)
#> expr: function_list[[k]](value)
#> env:  [frame 0x7f848abc34d8]
#> 
#> [[26]]
#> <frame 9> (8)
#> expr: withVisible(function_list[[k]](value))
#> env:  [frame 0x7f848abc33c0]
#> 
#> [[27]]
#> <frame 8> (7)
#> expr: freduce(value, `_function_list`)
#> env:  [frame 0x7f848abd1468]
#> 
#> [[28]]
#> <frame 7> (6)
#> expr: `_fseq`(`_lhs`)
#> env:  [frame 0x7f848abd12a8]
#> 
#> [[29]]
#> <frame 6> (5)
#> expr: eval(expr, envir, enclos)
#> env:  [frame 0x7f848adb4230]
#> 
#> [[30]]
#> <frame 5> (3)
#> expr: eval(quote(`_fseq`(`_lhs`)), env, env)
#> env:  [frame 0x7f848abd1078]
#> 
#> [[31]]
#> <frame 4> (3)
#> expr: withVisible(eval(quote(`_fseq`(`_lhs`)), env, env))
#> env:  [frame 0x7f848abd25f0]
#> 
#> [[32]]
#> <frame 3> (2)
#> expr: pkg$topics %>% purrr::transpose() %>% purrr::map(build_reference_topic,  <...>
#> env:  [frame 0x7f848adb4d78]
#> 
#> [[33]]
#> <frame 2> (1)
#> expr: build_reference(pkg, lazy = FALSE, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f8492ff3f40]
#> 
#> [[34]]
#> <frame 1> (0)
#> expr: pkgdown::build_site("/Users/yoplait/Dropbox/Projects/R/hadley/rlang")
#> env:  [frame 0x7f848fb31d48]
#> 
#> [[35]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"