Evaluate an expression with quosures and pronoun support

eval_tidy() is a variant of base::eval() that powers the tidy evaluation framework. Like eval() it accepts user data as argument. If supplied, it evaluates its input expr in a data mask. In additon eval_tidy() supports:

Quosures. The expression wrapped in the quosure evaluates in its original context (masked by data if supplied).

Pronouns. If data is supplied, the .env and .data pronouns are installed in the data mask. .env is a reference to the calling environment and .data refers to the data argument. These pronouns lets you be explicit about where to find values and throw errors if you try to access non-existent values.

eval_tidy(expr, data = NULL, env = caller_env())

Arguments

expr
An expression to evaluate.

data
A data frame, or named list or vector. Alternatively, a data mask created with as_data_mask() or new_data_mask().

env
The environment in which to evaluate expr. This environment is always ignored when evaluating quosures. Quosures are evaluated in their own environment.

Life cycle

eval_tidy() is stable.

See also

quasiquotation for the second leg of the tidy evaluation framework.

Examples

# With simple quoted expressions eval_tidy() works the same way as
# eval():
apple <- "apple"
kiwi <- "kiwi"
expr <- quote(paste(apple, kiwi))
expr

#> paste(apple, kiwi)

eval(expr)

#> [1] "apple kiwi"

eval_tidy(expr)

#> [1] "apple kiwi"

# Both accept a data mask as argument:
data <- list(apple = "CARROT", kiwi = "TOMATO")
eval(expr, data)

#> [1] "CARROT TOMATO"

eval_tidy(expr, data)

#> [1] "CARROT TOMATO"

# In addition eval_tidy() has support for quosures:
with_data <- function(data, expr) {
  quo <- enquo(expr)
  eval_tidy(quo, data)
}
with_data(NULL, apple)

#> [1] "apple"

with_data(data, apple)

#> [1] "CARROT"

with_data(data, list(apple, kiwi))

#> [[1]]
#> [1] "CARROT"
#> 
#> [[2]]
#> [1] "TOMATO"
#> 

# Secondly eval_tidy() installs handy pronouns that allows users to
# be explicit about where to find symbols:
with_data(data, .data$apple)

#> [1] "CARROT"

with_data(data, .env$apple)

#> [1] "apple"

# Note that instead of using `.env` it is often equivalent and may
# be preferred to unquote a value. There are two differences. First
# unquoting happens earlier, when the quosure is created. Secondly,
# subsetting `.env` with the `$` operator may be brittle because
# `$` does not look through the parents of the environment.
#
# For instance using `.env$name` in a magrittr pipeline is an
# instance where this poses problem, because the magrittr pipe
# currently (as of v1.5.0) evaluates its operands in a *child* of
# the current environment (this child environment is where it
# defines the pronoun `.`).

# NOT RUN {
  data %>% with_data(!!kiwi)     # "kiwi"
  data %>% with_data(.env$kiwi)  # NULL
# }