eval_tidy_ {rlang}	R Documentation

Tidy evaluation in a custom environment

Description

This function is defunct as of rlang 0.3.0.

Usage

eval_tidy_(expr, bottom, top = NULL, env = caller_env())

Arguments

expr
An expression or quosure to evaluate.

bottom
The environment containing masking objects if the data mask is one environment deep. The bottom environment if the data mask comprises multiple environment.

top
The last environment of the data mask. If the data mask is only one environment deep, top should be the same as bottom.

env
The environment in which to evaluate expr. This environment is not applicable for quosures because they have their own environments.

[Package rlang version 0.3.4 Index]