exprs_auto_name {rlang}	R Documentation

Ensure that all elements of a list of expressions are named

Description

This gives default names to unnamed elements of a list of expressions (or expression wrappers such as formulas or quosures). exprs_auto_name() deparses the expressions with expr_name() by default. quos_auto_name() deparses with quo_name().

Usage

exprs_auto_name(
  exprs,
  ...,
  repair_auto = c("minimal", "unique"),
  repair_quiet = FALSE
)

quos_auto_name(quos)

Arguments

exprs
A list of expressions.

...
These dots are for future extensions and must be empty.

repair_auto
Whether to repair the automatic names. By default, minimal names are returned. See ?vctrs::vec_as_names for information about name repairing.

repair_quiet
Whether to inform user about repaired names.

quos
A list of quosures.

[Package rlang version 1.0.1 Index]