Find the position or distance of a frame on the evaluation stack

The frame position on the stack can be computed by counting frames from the global frame (the bottom of the stack, the default) or from the current frame (the top of the stack).

frame_position(frame, from = c("global", "current"))

Arguments

frame
The environment of a frame. Can be any object with a get_env() method. Note that for frame objects, the position from the global frame is simply frame$pos. Alternatively, frame can be an integer that represents the position on the stack (and is thus returned as is if from is "global".

from
Whether to compute distance from the global frame (the bottom of the evaluation stack), or from the current frame (the top of the evaluation stack).

Details

While this function returns the position of the frame on the evaluation stack, it can safely be called with intervening frames as those will be discarded.

Life cycle

These functions are soft-deprecated and replaced by trace_back().

Examples

fn <- function() g(environment())
g <- function(env) frame_position(env)

# frame_position() returns the position of the frame on the evaluation stack:
fn()

#> Warning: `frame_position()` is soft-deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> Warning: `stack_trim()` is soft-deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> Warning: `ctxt_stack()` is soft-deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> Warning: `global_frame()` is soft-deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 36

identity(identity(fn()))

#> [1] 38

# Note that it trims off intervening calls before counting so you
# can safely nest it within other calls:
g <- function(env) identity(identity(frame_position(env)))
fn()

#> [1] 36

# You can also ask for the position from the current frame rather
# than the global frame:
fn <- function() g(environment())
g <- function(env) h(env)
h <- function(env) frame_position(env, from = "current")
fn()

#> [1] 3