Manipulate or access a call

These functions are soft-deprecated, please use call_modify(), call_standardise(), or call_fn() instead.

lang_modify(.lang, ..., .standardise = FALSE)

lang_standardise(lang)

lang_fn(lang)

lang_name(lang)

lang_args(lang)

lang_args_names(lang)

Arguments

...
Named or unnamed expressions (constants, names or calls) used to modify the call. Use NULL to remove arguments. These dots support tidy dots features.

.standardise
If TRUE, the call is standardised beforehand to match existing unnamed arguments to their argument names. This prevents new named arguments from accidentally replacing original unnamed arguments.

lang, .lang
The call or .call argument of the renamed functions.