scoped_interactive {rlang}	R Documentation

Deprecated scoped_ functions

Description

Deprecated as of rlang 0.4.2. Use local_interactive(), local_options(), or local_bindings() instead.

Usage

scoped_interactive(value = TRUE, frame = caller_env())

scoped_options(..., .frame = caller_env())

scoped_bindings(..., .env = .frame, .frame = caller_env())

Arguments

value
A single TRUE or FALSE. This overrides the return value of is_interactive().

frame
The environment of a running function which defines the scope of the temporary options. When the function returns, the options are reset to their original values.

...
For local_options() and push_options(), named values defining new option values. For peek_options(), strings or character vectors of option names.

.frame
The environment of a running function which defines the scope of the temporary options. When the function returns, the options are reset to their original values.

.env
An environment.

[Package rlang version 1.0.1 Index]