Call stack information

The eval_ and call_ families of functions provide a replacement for the base R functions prefixed with sys. (which are all about the context stack), as well as for parent.frame() (which is the only base R function for querying the call stack). The context stack includes all R-level evaluation contexts. It is linear in terms of execution history but due to lazy evaluation it is potentially nonlinear in terms of call history. The call stack history, on the other hand, is homogenous.

global_frame()

current_frame()

ctxt_frame(n = 1)

call_frame(n = 1, clean = TRUE)

ctxt_depth()

call_depth()

ctxt_stack(n = NULL, trim = 0)

call_stack(n = NULL, clean = TRUE)

Arguments

n
The number of frames to go back in the stack.

clean
Whether to post-process the call stack to clean non-standard frames. If TRUE, suboptimal call-stack entries by base::eval() will be cleaned up: the duplicate frame created by eval() is eliminated.

trim
The number of layers of intervening frames to trim off the stack. See stack_trim() and examples.

Details

ctxt_frame() and call_frame() return a frame object containing the following fields: expr and env (call expression and evaluation environment), pos and caller_pos (position of current frame in the context stack and position of the caller), and fun (function of the current frame). ctxt_stack() and call_stack() return a list of all context or call frames on the stack. Finally, ctxt_depth() and call_depth() report the current context position or the number of calling frames on the stack.

The base R functions take two sorts of arguments to indicate which frame to query: which and n. The n argument is straightforward: it's the number of frames to go down the stack, with n = 1 referring to the current context. The which argument is more complicated and changes meaning for values lower than 1. For the sake of consistency, the rlang functions all take the same kind of argument n. This argument has a single meaning (the number of frames to go down the stack) and cannot be lower than 1.

Note finally that parent.frame(1) corresponds to call_frame(2)$env, as n = 1 always refers to the current frame. This makes the _frame() and _stack() functions consistent: ctxt_frame(2) is the same as ctxt_stack()[[2]]. Also, ctxt_depth() returns one more frame than base::sys.nframe() because it counts the global frame. That is consistent with the _stack() functions which return the global frame as well. This way, call_stack(call_depth()) is the same as global_frame().

Life cycle

These functions are soft-deprecated and replaced by trace_back().

Examples

# Expressions within arguments count as contexts
identity(identity(ctxt_depth())) # returns 2

#> Warning: `ctxt_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 26

# But they are not part of the call stack because arguments are
# evaluated within the calling function (or the global environment
# if called at top level)
identity(identity(call_depth())) # returns 0

#> Warning: `call_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 13

# The context stacks includes all intervening execution frames. The
# call stack doesn't:
f <- function(x) identity(x)
f(f(ctxt_stack()))

#> [[1]]
#> <frame 27> (26)
#> expr: identity(x)
#> env:  [local 0x7fc7653c9b90]
#> 
#> [[2]]
#> <frame 26> (23)
#> expr: f(ctxt_stack())
#> env:  [local 0x7fc7653c9ca8]
#> 
#> [[3]]
#> <frame 25> (24)
#> expr: identity(x)
#> env:  [local 0x7fc7653c9dc0]
#> 
#> [[4]]
#> <frame 24> (23)
#> expr: f(f(ctxt_stack()))
#> env:  [local 0x7fc7653c9ed8]
#> 
#> [[5]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc765f50618]
#> 
#> [[6]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc7653ca098]
#> 
#> [[7]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fc7653ca2c8]
#> 
#> [[8]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fc7653ca878]
#> 
#> [[9]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fc7653cadf0]
#> 
#> [[10]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fc7653cb138]
#> 
#> [[11]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fc7653cb480]
#> 
#> [[12]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fc765a4b238]
#> 
#> [[13]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fc765a4b468]
#> 
#> [[14]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fc765a4b5f0]
#> 
#> [[15]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fc765a4b6d0]
#> 
#> [[16]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fc76586a520]
#> 
#> [[17]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fc7613ba6a8]
#> 
#> [[18]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fc7613e42f8]
#> 
#> [[19]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fc76146e290]
#> 
#> [[20]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fc7621175a0]
#> 
#> [[21]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fc762116ea0]
#> 
#> [[22]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fc762143108]
#> 
#> [[23]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fc761e6cf30]
#> 
#> [[24]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc760640160]
#> 
#> [[25]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc761d2f020]
#> 
#> [[26]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fc761d339f8]
#> 
#> [[27]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fc76175faa8]
#> 
#> [[28]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(f(call_stack()))

#> Warning: `call_stack()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [[1]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc765f50618]
#> 
#> [[2]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fc7655ad608]
#> 
#> [[3]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fc7613ba6a8]
#> 
#> [[4]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fc7613e42f8]
#> 
#> [[5]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fc76146e290]
#> 
#> [[6]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fc7621175a0]
#> 
#> [[7]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fc762143108]
#> 
#> [[8]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fc761e6cf30]
#> 
#> [[9]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc760640160]
#> 
#> [[10]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc761d2f020]
#> 
#> [[11]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fc761d339f8]
#> 
#> [[12]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fc76175faa8]
#> 
#> [[13]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

g <- function(cmd) cmd()
f(g(ctxt_stack))

#> [[1]]
#> <frame 26> (23)
#> expr: g(ctxt_stack)
#> env:  [local 0x7fc765cec388]
#> 
#> [[2]]
#> <frame 25> (24)
#> expr: identity(x)
#> env:  [local 0x7fc765cec270]
#> 
#> [[3]]
#> <frame 24> (23)
#> expr: f(g(ctxt_stack))
#> env:  [local 0x7fc765cec158]
#> 
#> [[4]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc765f50618]
#> 
#> [[5]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc765cebf98]
#> 
#> [[6]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fc765cebd68]
#> 
#> [[7]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fc765ceb7b8]
#> 
#> [[8]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fc765ceb240]
#> 
#> [[9]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fc765cf0e08]
#> 
#> [[10]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fc765cf0ac0]
#> 
#> [[11]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fc765cf03c0]
#> 
#> [[12]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fc765cf0190]
#> 
#> [[13]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fc765cf0008]
#> 
#> [[14]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fc765ceff28]
#> 
#> [[15]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fc765f5b1f8]
#> 
#> [[16]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fc7613ba6a8]
#> 
#> [[17]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fc7613e42f8]
#> 
#> [[18]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fc76146e290]
#> 
#> [[19]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fc7621175a0]
#> 
#> [[20]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fc762116ea0]
#> 
#> [[21]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fc762143108]
#> 
#> [[22]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fc761e6cf30]
#> 
#> [[23]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc760640160]
#> 
#> [[24]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc761d2f020]
#> 
#> [[25]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fc761d339f8]
#> 
#> [[26]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fc76175faa8]
#> 
#> [[27]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(g(call_stack))

#> [[1]]
#> <frame 26> (22)
#> expr: g(call_stack)
#> env:  [local 0x7fc76642cbc8]
#> 
#> [[2]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc765f50618]
#> 
#> [[3]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fc7645dc1b8]
#> 
#> [[4]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fc7613ba6a8]
#> 
#> [[5]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fc7613e42f8]
#> 
#> [[6]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fc76146e290]
#> 
#> [[7]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fc7621175a0]
#> 
#> [[8]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fc762143108]
#> 
#> [[9]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fc761e6cf30]
#> 
#> [[10]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc760640160]
#> 
#> [[11]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc761d2f020]
#> 
#> [[12]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fc761d339f8]
#> 
#> [[13]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fc76175faa8]
#> 
#> [[14]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

# The rlang _stack() functions return a list of frame
# objects. Use purrr::transpose() or index a field with
# purrr::map()'s to extract a particular field from a stack:

# stack <- f(f(call_stack()))
# purrr::map(stack, "env")
# purrr::transpose(stack)$expr

# current_frame() is an alias for ctxt_frame(1)
fn <- function() list(current = current_frame(), first = ctxt_frame(1))
fn()

#> Warning: `current_frame()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> $current
#> <frame 24> (23)
#> expr: fn()
#> env:  [local 0x7fc7637d45b0]
#> 
#> $first
#> <frame 24> (23)
#> expr: fn()
#> env:  [local 0x7fc7637d45b0]
#> 

# While current_frame() is the top of the stack, global_frame() is
# the bottom:
fn <- function() {
  n <- ctxt_depth()
  ctxt_frame(n)
}
identical(fn(), global_frame())

#> [1] TRUE

# ctxt_stack() returns a stack with all intervening frames. You can
# trim layers of intervening frames with the trim argument:
identity(identity(ctxt_stack()))

#> [[1]]
#> <frame 25> (23)
#> expr: identity(ctxt_stack())
#> env:  [local 0x7fc76355a808]
#> 
#> [[2]]
#> <frame 24> (23)
#> expr: identity(identity(ctxt_stack()))
#> env:  [local 0x7fc76355a6f0]
#> 
#> [[3]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc765f50618]
#> 
#> [[4]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc76355a530]
#> 
#> [[5]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fc76355a300]
#> 
#> [[6]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fc763559d18]
#> 
#> [[7]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fc76355d6b0]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fc76355d330]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fc76355cfe8]
#> 
#> [[10]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fc76355c920]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fc76355c6f0]
#> 
#> [[12]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fc76355c568]
#> 
#> [[13]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fc76355c488]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fc76357c260]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fc7613ba6a8]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fc7613e42f8]
#> 
#> [[17]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fc76146e290]
#> 
#> [[18]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fc7621175a0]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fc762116ea0]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fc762143108]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fc761e6cf30]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc760640160]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc761d2f020]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fc761d339f8]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fc76175faa8]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(ctxt_stack(trim = 1)))

#> [[1]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc765f50618]
#> 
#> [[2]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc764a74ca0]
#> 
#> [[3]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fc764a74a70]
#> 
#> [[4]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fc764a78398]
#> 
#> [[5]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fc764a77e20]
#> 
#> [[6]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fc764a77ad8]
#> 
#> [[7]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fc764a77758]
#> 
#> [[8]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fc764a77090]
#> 
#> [[9]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fc764a76df0]
#> 
#> [[10]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fc764a76c68]
#> 
#> [[11]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fc764a76b88]
#> 
#> [[12]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fc764a88bf8]
#> 
#> [[13]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fc7613ba6a8]
#> 
#> [[14]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fc7613e42f8]
#> 
#> [[15]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fc76146e290]
#> 
#> [[16]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fc7621175a0]
#> 
#> [[17]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fc762116ea0]
#> 
#> [[18]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fc762143108]
#> 
#> [[19]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fc761e6cf30]
#> 
#> [[20]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc760640160]
#> 
#> [[21]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc761d2f020]
#> 
#> [[22]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fc761d339f8]
#> 
#> [[23]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fc76175faa8]
#> 
#> [[24]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# ctxt_stack() is called within fn() with intervening frames:
fn <- function(trim) identity(identity(ctxt_stack(trim = trim)))
fn(0)

#> [[1]]
#> <frame 26> (24)
#> expr: identity(ctxt_stack(trim = trim))
#> env:  [local 0x7fc764817f28]
#> 
#> [[2]]
#> <frame 25> (24)
#> expr: identity(identity(ctxt_stack(trim = trim)))
#> env:  [local 0x7fc764817e10]
#> 
#> [[3]]
#> <frame 24> (23)
#> expr: fn(0)
#> env:  [local 0x7fc764817cf8]
#> 
#> [[4]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc765f50618]
#> 
#> [[5]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc764817b38]
#> 
#> [[6]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fc764817908]
#> 
#> [[7]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fc764817358]
#> 
#> [[8]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fc764816de0]
#> 
#> [[9]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fc764816a98]
#> 
#> [[10]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fc764816750]
#> 
#> [[11]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fc76481bf28]
#> 
#> [[12]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fc76481bcf8]
#> 
#> [[13]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fc76481bb70]
#> 
#> [[14]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fc76481ba90]
#> 
#> [[15]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fc764832cc8]
#> 
#> [[16]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fc7613ba6a8]
#> 
#> [[17]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fc7613e42f8]
#> 
#> [[18]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fc76146e290]
#> 
#> [[19]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fc7621175a0]
#> 
#> [[20]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fc762116ea0]
#> 
#> [[21]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fc762143108]
#> 
#> [[22]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fc761e6cf30]
#> 
#> [[23]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc760640160]
#> 
#> [[24]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc761d2f020]
#> 
#> [[25]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fc761d339f8]
#> 
#> [[26]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fc76175faa8]
#> 
#> [[27]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# We can trim the first layer of those:
fn(1)

#> [[1]]
#> <frame 24> (23)
#> expr: fn(1)
#> env:  [local 0x7fc76254e140]
#> 
#> [[2]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc765f50618]
#> 
#> [[3]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc76254df80]
#> 
#> [[4]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fc76254dd50]
#> 
#> [[5]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fc76254d7a0]
#> 
#> [[6]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fc76254d1b8]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fc762550a00]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fc7625506b8]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fc76254ffb8]
#> 
#> [[10]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fc76254fd88]
#> 
#> [[11]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fc76254fc00]
#> 
#> [[12]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fc76254fb20]
#> 
#> [[13]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fc762563ce0]
#> 
#> [[14]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fc7613ba6a8]
#> 
#> [[15]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fc7613e42f8]
#> 
#> [[16]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fc76146e290]
#> 
#> [[17]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fc7621175a0]
#> 
#> [[18]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fc762116ea0]
#> 
#> [[19]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fc762143108]
#> 
#> [[20]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fc761e6cf30]
#> 
#> [[21]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc760640160]
#> 
#> [[22]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc761d2f020]
#> 
#> [[23]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fc761d339f8]
#> 
#> [[24]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fc76175faa8]
#> 
#> [[25]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# The outside intervening frames (at the fn() call site) are still
# returned, but can be trimmed as well:
identity(identity(fn(1)))

#> [[1]]
#> <frame 26> (23)
#> expr: fn(1)
#> env:  [local 0x7fc76226fb18]
#> 
#> [[2]]
#> <frame 25> (23)
#> expr: identity(fn(1))
#> env:  [local 0x7fc76226f9c8]
#> 
#> [[3]]
#> <frame 24> (23)
#> expr: identity(identity(fn(1)))
#> env:  [local 0x7fc76226f8b0]
#> 
#> [[4]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc765f50618]
#> 
#> [[5]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc76226f6f0]
#> 
#> [[6]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fc76226f488]
#> 
#> [[7]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fc76226eed8]
#> 
#> [[8]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fc762274800]
#> 
#> [[9]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fc7622744b8]
#> 
#> [[10]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fc762274170]
#> 
#> [[11]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fc762273a70]
#> 
#> [[12]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fc762273840]
#> 
#> [[13]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fc7622736b8]
#> 
#> [[14]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fc7622735d8]
#> 
#> [[15]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fc76228b680]
#> 
#> [[16]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fc7613ba6a8]
#> 
#> [[17]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fc7613e42f8]
#> 
#> [[18]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fc76146e290]
#> 
#> [[19]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fc7621175a0]
#> 
#> [[20]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fc762116ea0]
#> 
#> [[21]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fc762143108]
#> 
#> [[22]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fc761e6cf30]
#> 
#> [[23]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc760640160]
#> 
#> [[24]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc761d2f020]
#> 
#> [[25]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fc761d339f8]
#> 
#> [[26]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fc76175faa8]
#> 
#> [[27]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(fn(2)))

#> [[1]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc765f50618]
#> 
#> [[2]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc7641bb9a8]
#> 
#> [[3]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fc7641bb778]
#> 
#> [[4]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fc7641bb190]
#> 
#> [[5]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fc7641bac18]
#> 
#> [[6]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fc7641ba8d0]
#> 
#> [[7]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fc7641ba518]
#> 
#> [[8]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fc7641bccf8]
#> 
#> [[9]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fc7641bcac8]
#> 
#> [[10]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fc7641bc940]
#> 
#> [[11]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fc7641bc860]
#> 
#> [[12]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fc7641e1af8]
#> 
#> [[13]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fc7613ba6a8]
#> 
#> [[14]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fc7613e42f8]
#> 
#> [[15]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fc76146e290]
#> 
#> [[16]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fc7621175a0]
#> 
#> [[17]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fc762116ea0]
#> 
#> [[18]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fc762143108]
#> 
#> [[19]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fc761e6cf30]
#> 
#> [[20]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc760640160]
#> 
#> [[21]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc761d2f020]
#> 
#> [[22]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fc761d339f8]
#> 
#> [[23]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fc76175faa8]
#> 
#> [[24]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g <- function(trim) identity(identity(fn(trim)))
g(2)

#> [[1]]
#> <frame 24> (23)
#> expr: g(2)
#> env:  [local 0x7fc762d8e368]
#> 
#> [[2]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc765f50618]
#> 
#> [[3]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc762d8e1a8]
#> 
#> [[4]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fc762d8df78]
#> 
#> [[5]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fc762d8d9c8]
#> 
#> [[6]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fc762d91248]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fc762d90f00]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fc762d90bb8]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fc762d904b8]
#> 
#> [[10]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fc762d90288]
#> 
#> [[11]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fc762d90100]
#> 
#> [[12]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fc762d90020]
#> 
#> [[13]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fc762da6d78]
#> 
#> [[14]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fc7613ba6a8]
#> 
#> [[15]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fc7613e42f8]
#> 
#> [[16]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fc76146e290]
#> 
#> [[17]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fc7621175a0]
#> 
#> [[18]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fc762116ea0]
#> 
#> [[19]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fc762143108]
#> 
#> [[20]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fc761e6cf30]
#> 
#> [[21]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc760640160]
#> 
#> [[22]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc761d2f020]
#> 
#> [[23]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fc761d339f8]
#> 
#> [[24]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fc76175faa8]
#> 
#> [[25]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g(3)

#> [[1]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc765f50618]
#> 
#> [[2]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc7628a6400]
#> 
#> [[3]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fc7628a9f58]
#> 
#> [[4]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fc7628a9938]
#> 
#> [[5]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fc7628a93c0]
#> 
#> [[6]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fc7628a8fd0]
#> 
#> [[7]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fc7628a8be0]
#> 
#> [[8]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fc7628a8390]
#> 
#> [[9]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fc7628abfc8]
#> 
#> [[10]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fc7628abe08]
#> 
#> [[11]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fc7628abcf0]
#> 
#> [[12]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fc7628bebe0]
#> 
#> [[13]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fc7613ba6a8]
#> 
#> [[14]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fc7613e42f8]
#> 
#> [[15]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fc76146e290]
#> 
#> [[16]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fc7621175a0]
#> 
#> [[17]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fc762116ea0]
#> 
#> [[18]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fc762143108]
#> 
#> [[19]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fc761e6cf30]
#> 
#> [[20]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc760640160]
#> 
#> [[21]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc761d2f020]
#> 
#> [[22]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fc761d339f8]
#> 
#> [[23]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fc76175faa8]
#> 
#> [[24]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"