Call stack information

The eval_ and call_ families of functions provide a replacement for the base R functions prefixed with sys. (which are all about the context stack), as well as for parent.frame() (which is the only base R function for querying the call stack). The context stack includes all R-level evaluation contexts. It is linear in terms of execution history but due to lazy evaluation it is potentially nonlinear in terms of call history. The call stack history, on the other hand, is homogenous.

global_frame()

current_frame()

ctxt_frame(n = 1)

call_frame(n = 1, clean = TRUE)

ctxt_depth()

call_depth()

ctxt_stack(n = NULL, trim = 0)

call_stack(n = NULL, clean = TRUE)

Arguments

n
The number of frames to go back in the stack.

clean
Whether to post-process the call stack to clean non-standard frames. If TRUE, suboptimal call-stack entries by base::eval() will be cleaned up: the duplicate frame created by eval() is eliminated.

trim
The number of layers of intervening frames to trim off the stack. See stack_trim() and examples.

Details

ctxt_frame() and call_frame() return a frame object containing the following fields: expr and env (call expression and evaluation environment), pos and caller_pos (position of current frame in the context stack and position of the caller), and fun (function of the current frame). ctxt_stack() and call_stack() return a list of all context or call frames on the stack. Finally, ctxt_depth() and call_depth() report the current context position or the number of calling frames on the stack.

The base R functions take two sorts of arguments to indicate which frame to query: which and n. The n argument is straightforward: it's the number of frames to go down the stack, with n = 1 referring to the current context. The which argument is more complicated and changes meaning for values lower than 1. For the sake of consistency, the rlang functions all take the same kind of argument n. This argument has a single meaning (the number of frames to go down the stack) and cannot be lower than 1.

Note finally that parent.frame(1) corresponds to call_frame(2)$env, as n = 1 always refers to the current frame. This makes the _frame() and _stack() functions consistent: ctxt_frame(2) is the same as ctxt_stack()[[2]]. Also, ctxt_depth() returns one more frame than base::sys.nframe() because it counts the global frame. That is consistent with the _stack() functions which return the global frame as well. This way, call_stack(call_depth()) is the same as global_frame().

Life cycle

These functions are soft-deprecated and replaced by trace_back().

Examples

# Expressions within arguments count as contexts
identity(identity(ctxt_depth())) # returns 2

#> Warning: `ctxt_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 27

# But they are not part of the call stack because arguments are
# evaluated within the calling function (or the global environment
# if called at top level)
identity(identity(call_depth())) # returns 0

#> Warning: `call_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 14

# The context stacks includes all intervening execution frames. The
# call stack doesn't:
f <- function(x) identity(x)
f(f(ctxt_stack()))

#> [[1]]
#> <frame 28> (27)
#> expr: identity(x)
#> env:  [local 0x7fc460fb1450]
#> 
#> [[2]]
#> <frame 27> (24)
#> expr: f(ctxt_stack())
#> env:  [local 0x7fc460fb1338]
#> 
#> [[3]]
#> <frame 26> (25)
#> expr: identity(x)
#> env:  [local 0x7fc460fb11e8]
#> 
#> [[4]]
#> <frame 25> (24)
#> expr: f(f(ctxt_stack()))
#> env:  [local 0x7fc460fb10d0]
#> 
#> [[5]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc461609c08]
#> 
#> [[6]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc460fb0f10]
#> 
#> [[7]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fc460fb0ca8]
#> 
#> [[8]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fc460fb44f0]
#> 
#> [[9]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fc460fb3f08]
#> 
#> [[10]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fc460fb3b18]
#> 
#> [[11]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fc4614f3c40]
#> 
#> [[12]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fc4614f5b60]
#> 
#> [[13]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fc4614f52d8]
#> 
#> [[14]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fc4614f5150]
#> 
#> [[15]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fc4614f5000]
#> 
#> [[16]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fc460fcdc30]
#> 
#> [[17]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fc4636fee58]
#> 
#> [[18]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7fc463705c58]
#> 
#> [[19]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fc4637143d0]
#> 
#> [[20]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fc462a13088]
#> 
#> [[21]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fc4640595f8]
#> 
#> [[22]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fc464058e50]
#> 
#> [[23]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fc46407de48]
#> 
#> [[24]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fc460d9ca30]
#> 
#> [[25]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc462261780]
#> 
#> [[26]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc4628bfe08]
#> 
#> [[27]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fc4628bb0b0]
#> 
#> [[28]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fc46109a5c0]
#> 
#> [[29]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(f(call_stack()))

#> Warning: `call_stack()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [[1]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc461609c08]
#> 
#> [[2]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fc460e9d6a8]
#> 
#> [[3]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fc4636fee58]
#> 
#> [[4]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7fc463705c58]
#> 
#> [[5]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fc4637143d0]
#> 
#> [[6]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fc462a13088]
#> 
#> [[7]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fc4640595f8]
#> 
#> [[8]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fc46407de48]
#> 
#> [[9]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fc460d9ca30]
#> 
#> [[10]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc462261780]
#> 
#> [[11]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc4628bfe08]
#> 
#> [[12]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fc4628bb0b0]
#> 
#> [[13]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fc46109a5c0]
#> 
#> [[14]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

g <- function(cmd) cmd()
f(g(ctxt_stack))

#> [[1]]
#> <frame 27> (24)
#> expr: g(ctxt_stack)
#> env:  [local 0x7fc46216f0b8]
#> 
#> [[2]]
#> <frame 26> (25)
#> expr: identity(x)
#> env:  [local 0x7fc46216efa0]
#> 
#> [[3]]
#> <frame 25> (24)
#> expr: f(g(ctxt_stack))
#> env:  [local 0x7fc46216ee50]
#> 
#> [[4]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc461609c08]
#> 
#> [[5]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc46216ec90]
#> 
#> [[6]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fc46216ea28]
#> 
#> [[7]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fc46216e408]
#> 
#> [[8]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fc462171d68]
#> 
#> [[9]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fc4621719b0]
#> 
#> [[10]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fc4621715f8]
#> 
#> [[11]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fc462170ef8]
#> 
#> [[12]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fc462170c58]
#> 
#> [[13]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fc462170a98]
#> 
#> [[14]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fc4621709b8]
#> 
#> [[15]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fc4621b1be0]
#> 
#> [[16]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fc4636fee58]
#> 
#> [[17]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7fc463705c58]
#> 
#> [[18]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fc4637143d0]
#> 
#> [[19]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fc462a13088]
#> 
#> [[20]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fc4640595f8]
#> 
#> [[21]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fc464058e50]
#> 
#> [[22]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fc46407de48]
#> 
#> [[23]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fc460d9ca30]
#> 
#> [[24]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc462261780]
#> 
#> [[25]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc4628bfe08]
#> 
#> [[26]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fc4628bb0b0]
#> 
#> [[27]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fc46109a5c0]
#> 
#> [[28]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(g(call_stack))

#> [[1]]
#> <frame 27> (23)
#> expr: g(call_stack)
#> env:  [local 0x7fc461369340]
#> 
#> [[2]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc461609c08]
#> 
#> [[3]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fc4613aa770]
#> 
#> [[4]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fc4636fee58]
#> 
#> [[5]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7fc463705c58]
#> 
#> [[6]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fc4637143d0]
#> 
#> [[7]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fc462a13088]
#> 
#> [[8]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fc4640595f8]
#> 
#> [[9]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fc46407de48]
#> 
#> [[10]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fc460d9ca30]
#> 
#> [[11]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc462261780]
#> 
#> [[12]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc4628bfe08]
#> 
#> [[13]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fc4628bb0b0]
#> 
#> [[14]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fc46109a5c0]
#> 
#> [[15]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

# The rlang _stack() functions return a list of frame
# objects. Use purrr::transpose() or index a field with
# purrr::map()'s to extract a particular field from a stack:

# stack <- f(f(call_stack()))
# purrr::map(stack, "env")
# purrr::transpose(stack)$expr

# current_frame() is an alias for ctxt_frame(1)
fn <- function() list(current = current_frame(), first = ctxt_frame(1))
fn()

#> Warning: `current_frame()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> $current
#> <frame 25> (24)
#> expr: fn()
#> env:  [local 0x7fc4610e9af8]
#> 
#> $first
#> <frame 25> (24)
#> expr: fn()
#> env:  [local 0x7fc4610e9af8]
#> 

# While current_frame() is the top of the stack, global_frame() is
# the bottom:
fn <- function() {
  n <- ctxt_depth()
  ctxt_frame(n)
}
identical(fn(), global_frame())

#> [1] TRUE

# ctxt_stack() returns a stack with all intervening frames. You can
# trim layers of intervening frames with the trim argument:
identity(identity(ctxt_stack()))

#> [[1]]
#> <frame 26> (24)
#> expr: identity(ctxt_stack())
#> env:  [local 0x7fc460826000]
#> 
#> [[2]]
#> <frame 25> (24)
#> expr: identity(identity(ctxt_stack()))
#> env:  [local 0x7fc460825ee8]
#> 
#> [[3]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc461609c08]
#> 
#> [[4]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc460825cb8]
#> 
#> [[5]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fc460825a50]
#> 
#> [[6]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fc460825468]
#> 
#> [[7]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fc460824e80]
#> 
#> [[8]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fc460824a20]
#> 
#> [[9]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fc4608246d8]
#> 
#> [[10]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fc460827c10]
#> 
#> [[11]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fc460827858]
#> 
#> [[12]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fc460827660]
#> 
#> [[13]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fc460827580]
#> 
#> [[14]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fc4608392a8]
#> 
#> [[15]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fc4636fee58]
#> 
#> [[16]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7fc463705c58]
#> 
#> [[17]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fc4637143d0]
#> 
#> [[18]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fc462a13088]
#> 
#> [[19]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fc4640595f8]
#> 
#> [[20]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fc464058e50]
#> 
#> [[21]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fc46407de48]
#> 
#> [[22]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fc460d9ca30]
#> 
#> [[23]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc462261780]
#> 
#> [[24]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc4628bfe08]
#> 
#> [[25]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fc4628bb0b0]
#> 
#> [[26]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fc46109a5c0]
#> 
#> [[27]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(ctxt_stack(trim = 1)))

#> [[1]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc461609c08]
#> 
#> [[2]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc46015d730]
#> 
#> [[3]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fc46015d378]
#> 
#> [[4]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fc46015cce8]
#> 
#> [[5]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fc4601610f8]
#> 
#> [[6]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fc460160d40]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fc4601608a8]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fc46015fe98]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fc46015fc30]
#> 
#> [[10]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fc46015fa70]
#> 
#> [[11]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fc46015f958]
#> 
#> [[12]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fc460186c40]
#> 
#> [[13]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fc4636fee58]
#> 
#> [[14]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7fc463705c58]
#> 
#> [[15]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fc4637143d0]
#> 
#> [[16]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fc462a13088]
#> 
#> [[17]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fc4640595f8]
#> 
#> [[18]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fc464058e50]
#> 
#> [[19]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fc46407de48]
#> 
#> [[20]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fc460d9ca30]
#> 
#> [[21]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc462261780]
#> 
#> [[22]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc4628bfe08]
#> 
#> [[23]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fc4628bb0b0]
#> 
#> [[24]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fc46109a5c0]
#> 
#> [[25]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# ctxt_stack() is called within fn() with intervening frames:
fn <- function(trim) identity(identity(ctxt_stack(trim = trim)))
fn(0)

#> [[1]]
#> <frame 27> (25)
#> expr: identity(ctxt_stack(trim = trim))
#> env:  [local 0x7fc463f07870]
#> 
#> [[2]]
#> <frame 26> (25)
#> expr: identity(identity(ctxt_stack(trim = trim)))
#> env:  [local 0x7fc463f07988]
#> 
#> [[3]]
#> <frame 25> (24)
#> expr: fn(0)
#> env:  [local 0x7fc463f07aa0]
#> 
#> [[4]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc461609c08]
#> 
#> [[5]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc463f07c60]
#> 
#> [[6]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fc463f07e90]
#> 
#> [[7]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fc463f08440]
#> 
#> [[8]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fc463f04aa8]
#> 
#> [[9]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fc463f04df0]
#> 
#> [[10]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fc463f05138]
#> 
#> [[11]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fc463f05800]
#> 
#> [[12]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fc463f05a68]
#> 
#> [[13]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fc463f05bf0]
#> 
#> [[14]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fc463f05cd0]
#> 
#> [[15]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fc463f21a80]
#> 
#> [[16]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fc4636fee58]
#> 
#> [[17]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7fc463705c58]
#> 
#> [[18]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fc4637143d0]
#> 
#> [[19]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fc462a13088]
#> 
#> [[20]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fc4640595f8]
#> 
#> [[21]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fc464058e50]
#> 
#> [[22]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fc46407de48]
#> 
#> [[23]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fc460d9ca30]
#> 
#> [[24]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc462261780]
#> 
#> [[25]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc4628bfe08]
#> 
#> [[26]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fc4628bb0b0]
#> 
#> [[27]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fc46109a5c0]
#> 
#> [[28]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# We can trim the first layer of those:
fn(1)

#> [[1]]
#> <frame 25> (24)
#> expr: fn(1)
#> env:  [local 0x7fc46475f840]
#> 
#> [[2]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc461609c08]
#> 
#> [[3]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc46475fa00]
#> 
#> [[4]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fc46475fc30]
#> 
#> [[5]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fc4647601e0]
#> 
#> [[6]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fc464760758]
#> 
#> [[7]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fc46475cb90]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fc46475ced8]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fc46475d5a0]
#> 
#> [[10]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fc46475d7d0]
#> 
#> [[11]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fc46475d958]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fc46475da38]
#> 
#> [[13]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fc46474da00]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fc4636fee58]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7fc463705c58]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fc4637143d0]
#> 
#> [[17]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fc462a13088]
#> 
#> [[18]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fc4640595f8]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fc464058e50]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fc46407de48]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fc460d9ca30]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc462261780]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc4628bfe08]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fc4628bb0b0]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fc46109a5c0]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# The outside intervening frames (at the fn() call site) are still
# returned, but can be trimmed as well:
identity(identity(fn(1)))

#> [[1]]
#> <frame 27> (24)
#> expr: fn(1)
#> env:  [local 0x7fc4673d4070]
#> 
#> [[2]]
#> <frame 26> (24)
#> expr: identity(fn(1))
#> env:  [local 0x7fc4673d0278]
#> 
#> [[3]]
#> <frame 25> (24)
#> expr: identity(identity(fn(1)))
#> env:  [local 0x7fc4673d0390]
#> 
#> [[4]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc461609c08]
#> 
#> [[5]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc4673d0550]
#> 
#> [[6]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fc4673d0780]
#> 
#> [[7]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fc4673d0da0]
#> 
#> [[8]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fc4673d1318]
#> 
#> [[9]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fc4673d1660]
#> 
#> [[10]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fc4673d19a8]
#> 
#> [[11]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fc4673d2070]
#> 
#> [[12]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fc4673cb190]
#> 
#> [[13]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fc4673cb318]
#> 
#> [[14]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fc4673cb3f8]
#> 
#> [[15]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fc4673b78e0]
#> 
#> [[16]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fc4636fee58]
#> 
#> [[17]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7fc463705c58]
#> 
#> [[18]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fc4637143d0]
#> 
#> [[19]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fc462a13088]
#> 
#> [[20]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fc4640595f8]
#> 
#> [[21]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fc464058e50]
#> 
#> [[22]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fc46407de48]
#> 
#> [[23]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fc460d9ca30]
#> 
#> [[24]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc462261780]
#> 
#> [[25]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc4628bfe08]
#> 
#> [[26]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fc4628bb0b0]
#> 
#> [[27]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fc46109a5c0]
#> 
#> [[28]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(fn(2)))

#> [[1]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc461609c08]
#> 
#> [[2]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc4675ef240]
#> 
#> [[3]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fc4675ef4a8]
#> 
#> [[4]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fc4675efa90]
#> 
#> [[5]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fc4675f0008]
#> 
#> [[6]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fc4675f0350]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fc4675f0698]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fc4675f0d60]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fc4675f0f90]
#> 
#> [[10]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fc4675ed208]
#> 
#> [[11]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fc4675ed2e8]
#> 
#> [[12]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fc4675decb8]
#> 
#> [[13]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fc4636fee58]
#> 
#> [[14]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7fc463705c58]
#> 
#> [[15]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fc4637143d0]
#> 
#> [[16]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fc462a13088]
#> 
#> [[17]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fc4640595f8]
#> 
#> [[18]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fc464058e50]
#> 
#> [[19]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fc46407de48]
#> 
#> [[20]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fc460d9ca30]
#> 
#> [[21]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc462261780]
#> 
#> [[22]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc4628bfe08]
#> 
#> [[23]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fc4628bb0b0]
#> 
#> [[24]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fc46109a5c0]
#> 
#> [[25]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g <- function(trim) identity(identity(fn(trim)))
g(2)

#> [[1]]
#> <frame 25> (24)
#> expr: g(2)
#> env:  [local 0x7fc466a9da98]
#> 
#> [[2]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc461609c08]
#> 
#> [[3]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc466a9dc58]
#> 
#> [[4]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fc466a9de88]
#> 
#> [[5]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fc466a9e438]
#> 
#> [[6]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fc466a9aaa0]
#> 
#> [[7]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fc466a9ade8]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fc466a9b130]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fc466a9b7f8]
#> 
#> [[10]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fc466a9ba28]
#> 
#> [[11]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fc466a9bbb0]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fc466a9bc90]
#> 
#> [[13]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fc4669d06e0]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fc4636fee58]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7fc463705c58]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fc4637143d0]
#> 
#> [[17]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fc462a13088]
#> 
#> [[18]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fc4640595f8]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fc464058e50]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fc46407de48]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fc460d9ca30]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc462261780]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc4628bfe08]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fc4628bb0b0]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fc46109a5c0]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g(3)

#> [[1]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc461609c08]
#> 
#> [[2]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fc466bdb9a0]
#> 
#> [[3]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fc466bdbbd0]
#> 
#> [[4]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fc466bd8270]
#> 
#> [[5]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fc466bd87e8]
#> 
#> [[6]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fc466bd8b30]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fc466bd8e78]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fc466bd9540]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fc466bd9770]
#> 
#> [[10]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fc466bd98f8]
#> 
#> [[11]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fc466bd99d8]
#> 
#> [[12]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fc466bc9428]
#> 
#> [[13]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fc4636fee58]
#> 
#> [[14]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7fc463705c58]
#> 
#> [[15]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fc4637143d0]
#> 
#> [[16]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fc462a13088]
#> 
#> [[17]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fc4640595f8]
#> 
#> [[18]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fc464058e50]
#> 
#> [[19]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fc46407de48]
#> 
#> [[20]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fc460d9ca30]
#> 
#> [[21]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc462261780]
#> 
#> [[22]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fc4628bfe08]
#> 
#> [[23]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fc4628bb0b0]
#> 
#> [[24]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fc46109a5c0]
#> 
#> [[25]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"