Call stack information

The eval_ and call_ families of functions provide a replacement for the base R functions prefixed with sys. (which are all about the context stack), as well as for parent.frame() (which is the only base R function for querying the call stack). The context stack includes all R-level evaluation contexts. It is linear in terms of execution history but due to lazy evaluation it is potentially nonlinear in terms of call history. The call stack history, on the other hand, is homogenous.

Usage

global_frame()

current_frame()

ctxt_frame(n = 1)

call_frame(n = 1, clean = TRUE)

ctxt_depth()

call_depth()

ctxt_stack(n = NULL, trim = 0)

call_stack(n = NULL, clean = TRUE)

Arguments

n

The number of frames to go back in the stack.

clean

Whether to post-process the call stack to clean non-standard frames. If TRUE, suboptimal call-stack entries by base::eval() will be cleaned up: the duplicate frame created by eval() is eliminated.

trim

The number of layers of intervening frames to trim off the stack. See stack_trim() and examples.

Details

ctxt_frame() and call_frame() return a frame object containing the following fields: expr and env (call expression and evaluation environment), pos and caller_pos (position of current frame in the context stack and position of the caller), and fun (function of the current frame). ctxt_stack() and call_stack() return a list of all context or call frames on the stack. Finally, ctxt_depth() and call_depth() report the current context position or the number of calling frames on the stack.

The base R functions take two sorts of arguments to indicate which frame to query: which and n. The n argument is straightforward: it's the number of frames to go down the stack, with n = 1 referring to the current context. The which argument is more complicated and changes meaning for values lower than 1. For the sake of consistency, the rlang functions all take the same kind of argument n. This argument has a single meaning (the number of frames to go down the stack) and cannot be lower than 1.

Note finally that parent.frame(1) corresponds to call_frame(2)$env, as n = 1 always refers to the current frame. This makes the _frame() and _stack() functions consistent: ctxt_frame(2) is the same as ctxt_stack()[[2]]. Also, ctxt_depth() returns one more frame than base::sys.nframe() because it counts the global frame. That is consistent with the _stack() functions which return the global frame as well. This way, call_stack(call_depth()) is the same as global_frame().

Life cycle

These functions are soft-deprecated and replaced by trace_back().

Examples

# Expressions within arguments count as contexts
identity(identity(ctxt_depth())) # returns 2
#> Warning: `ctxt_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.
#> [1] 28

# But they are not part of the call stack because arguments are
# evaluated within the calling function (or the global environment
# if called at top level)
identity(identity(call_depth())) # returns 0
#> Warning: `call_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.
#> [1] 15

# The context stacks includes all intervening execution frames. The
# call stack doesn't:
f <- function(x) identity(x)
f(f(ctxt_stack()))
#> [[1]]
#> <frame 29> (28)
#> expr: identity(x)
#> env:  [local 0x561f4c3d97d8]
#> 
#> [[2]]
#> <frame 28> (25)
#> expr: f(ctxt_stack())
#> env:  [local 0x561f4c3d96c0]
#> 
#> [[3]]
#> <frame 27> (26)
#> expr: identity(x)
#> env:  [local 0x561f4c3d95a8]
#> 
#> [[4]]
#> <frame 26> (25)
#> expr: f(f(ctxt_stack()))
#> env:  [local 0x561f4c3d9490]
#> 
#> [[5]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x561f4d18a688]
#> 
#> [[6]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x561f4c3d92d0]
#> 
#> [[7]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x561f4c3d90a0]
#> 
#> [[8]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x561f4c3dc920]
#> 
#> [[9]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x561f4c3dc338]
#> 
#> [[10]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x561f4c3dbff0]
#> 
#> [[11]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x561f4c3dbca8]
#> 
#> [[12]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x561f4c3db5a8]
#> 
#> [[13]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x561f4c3db378]
#> 
#> [[14]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x561f4c3db1f0]
#> 
#> [[15]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x561f4c3db110]
#> 
#> [[16]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x561f4c3f0220]
#> 
#> [[17]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x561f4da65958]
#> 
#> [[18]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x561f4da68170]
#> 
#> [[19]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x561f502d1368]
#> 
#> [[20]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x561f4f91c9d8]
#> 
#> [[21]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x561f4fe42a80]
#> 
#> [[22]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x561f4fe43180]
#> 
#> [[23]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x561f500ed7b8]
#> 
#> [[24]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x561f4cfe43a8]
#> 
#> [[25]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x561f4e9b4760]
#> 
#> [[26]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x561f4bcab008]
#> 
#> [[27]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x561f4bca4fa8]
#> 
#> [[28]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x561f49bd2188]
#> 
#> [[29]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x561f478ac310]
#> 
#> [[30]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     
f(f(call_stack()))
#> Warning: `call_stack()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.
#> [[1]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x561f4d18a688]
#> 
#> [[2]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x561f4c07b660]
#> 
#> [[3]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x561f4da65958]
#> 
#> [[4]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x561f4da68170]
#> 
#> [[5]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x561f502d1368]
#> 
#> [[6]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x561f4f91c9d8]
#> 
#> [[7]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x561f4fe42a80]
#> 
#> [[8]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x561f500ed7b8]
#> 
#> [[9]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x561f4cfe43a8]
#> 
#> [[10]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x561f4e9b4760]
#> 
#> [[11]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x561f4bcab008]
#> 
#> [[12]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x561f4bca4fa8]
#> 
#> [[13]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x561f49bd2188]
#> 
#> [[14]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x561f478ac310]
#> 
#> [[15]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

g <- function(cmd) cmd()
f(g(ctxt_stack))
#> [[1]]
#> <frame 28> (25)
#> expr: g(ctxt_stack)
#> env:  [local 0x561f4bb15860]
#> 
#> [[2]]
#> <frame 27> (26)
#> expr: identity(x)
#> env:  [local 0x561f4bb15748]
#> 
#> [[3]]
#> <frame 26> (25)
#> expr: f(g(ctxt_stack))
#> env:  [local 0x561f4bb15630]
#> 
#> [[4]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x561f4d18a688]
#> 
#> [[5]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x561f4bb15470]
#> 
#> [[6]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x561f4bb15208]
#> 
#> [[7]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x561f4bb14bb0]
#> 
#> [[8]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x561f4bb18388]
#> 
#> [[9]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x561f4bb18008]
#> 
#> [[10]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x561f4bb17cc0]
#> 
#> [[11]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x561f4bb17550]
#> 
#> [[12]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x561f4bb172b0]
#> 
#> [[13]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x561f4bb170b8]
#> 
#> [[14]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x561f4bb16fd8]
#> 
#> [[15]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x561f4bb6a910]
#> 
#> [[16]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x561f4da65958]
#> 
#> [[17]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x561f4da68170]
#> 
#> [[18]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x561f502d1368]
#> 
#> [[19]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x561f4f91c9d8]
#> 
#> [[20]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x561f4fe42a80]
#> 
#> [[21]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x561f4fe43180]
#> 
#> [[22]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x561f500ed7b8]
#> 
#> [[23]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x561f4cfe43a8]
#> 
#> [[24]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x561f4e9b4760]
#> 
#> [[25]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x561f4bcab008]
#> 
#> [[26]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x561f4bca4fa8]
#> 
#> [[27]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x561f49bd2188]
#> 
#> [[28]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x561f478ac310]
#> 
#> [[29]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     
f(g(call_stack))
#> [[1]]
#> <frame 28> (24)
#> expr: g(call_stack)
#> env:  [local 0x561f4b41fb58]
#> 
#> [[2]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x561f4d18a688]
#> 
#> [[3]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x561f4b44f938]
#> 
#> [[4]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x561f4da65958]
#> 
#> [[5]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x561f4da68170]
#> 
#> [[6]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x561f502d1368]
#> 
#> [[7]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x561f4f91c9d8]
#> 
#> [[8]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x561f4fe42a80]
#> 
#> [[9]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x561f500ed7b8]
#> 
#> [[10]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x561f4cfe43a8]
#> 
#> [[11]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x561f4e9b4760]
#> 
#> [[12]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x561f4bcab008]
#> 
#> [[13]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x561f4bca4fa8]
#> 
#> [[14]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x561f49bd2188]
#> 
#> [[15]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x561f478ac310]
#> 
#> [[16]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

# The rlang _stack() functions return a list of frame
# objects. Use purrr::transpose() or index a field with
# purrr::map()'s to extract a particular field from a stack:

# stack <- f(f(call_stack()))
# purrr::map(stack, "env")
# purrr::transpose(stack)$expr

# current_frame() is an alias for ctxt_frame(1)
fn <- function() list(current = current_frame(), first = ctxt_frame(1))
fn()
#> Warning: `current_frame()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.
#> $current
#> <frame 26> (25)
#> expr: fn()
#> env:  [local 0x561f4af66590]
#> 
#> $first
#> <frame 26> (25)
#> expr: fn()
#> env:  [local 0x561f4af66590]
#> 

# While current_frame() is the top of the stack, global_frame() is
# the bottom:
fn <- function() {
  n <- ctxt_depth()
  ctxt_frame(n)
}
identical(fn(), global_frame())
#> [1] TRUE

# ctxt_stack() returns a stack with all intervening frames. You can
# trim layers of intervening frames with the trim argument:
identity(identity(ctxt_stack()))
#> [[1]]
#> <frame 27> (25)
#> expr: identity(ctxt_stack())
#> env:  [local 0x561f49e3b068]
#> 
#> [[2]]
#> <frame 26> (25)
#> expr: identity(identity(ctxt_stack()))
#> env:  [local 0x561f49e3af50]
#> 
#> [[3]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x561f4d18a688]
#> 
#> [[4]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x561f49e3ad20]
#> 
#> [[5]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x561f49e3aa80]
#> 
#> [[6]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x561f49e3a380]
#> 
#> [[7]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x561f49e39dd0]
#> 
#> [[8]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x561f49e3f730]
#> 
#> [[9]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x561f49e3f378]
#> 
#> [[10]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x561f49e3ec40]
#> 
#> [[11]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x561f49e3e8f8]
#> 
#> [[12]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x561f49e3e770]
#> 
#> [[13]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x561f49e3e690]
#> 
#> [[14]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x561f49e50290]
#> 
#> [[15]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x561f4da65958]
#> 
#> [[16]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x561f4da68170]
#> 
#> [[17]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x561f502d1368]
#> 
#> [[18]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x561f4f91c9d8]
#> 
#> [[19]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x561f4fe42a80]
#> 
#> [[20]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x561f4fe43180]
#> 
#> [[21]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x561f500ed7b8]
#> 
#> [[22]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x561f4cfe43a8]
#> 
#> [[23]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x561f4e9b4760]
#> 
#> [[24]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x561f4bcab008]
#> 
#> [[25]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x561f4bca4fa8]
#> 
#> [[26]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x561f49bd2188]
#> 
#> [[27]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x561f478ac310]
#> 
#> [[28]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     
identity(identity(ctxt_stack(trim = 1)))
#> [[1]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x561f4d18a688]
#> 
#> [[2]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x561f483d09d0]
#> 
#> [[3]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x561f483d06c0]
#> 
#> [[4]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x561f483d0068]
#> 
#> [[5]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x561f483cfa48]
#> 
#> [[6]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x561f483d2df8]
#> 
#> [[7]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x561f483d2a40]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x561f483d20d8]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x561f483d1e38]
#> 
#> [[10]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x561f483d1c40]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x561f483d1ab8]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x561f48b8ba50]
#> 
#> [[13]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x561f4da65958]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x561f4da68170]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x561f502d1368]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x561f4f91c9d8]
#> 
#> [[17]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x561f4fe42a80]
#> 
#> [[18]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x561f4fe43180]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x561f500ed7b8]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x561f4cfe43a8]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x561f4e9b4760]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x561f4bcab008]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x561f4bca4fa8]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x561f49bd2188]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x561f478ac310]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# ctxt_stack() is called within fn() with intervening frames:
fn <- function(trim) identity(identity(ctxt_stack(trim = trim)))
fn(0)
#> [[1]]
#> <frame 28> (26)
#> expr: identity(ctxt_stack(trim = trim))
#> env:  [local 0x561f4ff00118]
#> 
#> [[2]]
#> <frame 27> (26)
#> expr: identity(identity(ctxt_stack(trim = trim)))
#> env:  [local 0x561f4ff00230]
#> 
#> [[3]]
#> <frame 26> (25)
#> expr: fn(0)
#> env:  [local 0x561f4ff00348]
#> 
#> [[4]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x561f4d18a688]
#> 
#> [[5]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x561f4ff00508]
#> 
#> [[6]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x561f4ff007e0]
#> 
#> [[7]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x561f4ff00d90]
#> 
#> [[8]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x561f4ff01308]
#> 
#> [[9]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x561f4ff01650]
#> 
#> [[10]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x561f4ff01998]
#> 
#> [[11]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x561f4eecb088]
#> 
#> [[12]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x561f4eecb2b8]
#> 
#> [[13]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x561f4eecb440]
#> 
#> [[14]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x561f4eecb520]
#> 
#> [[15]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x561f4fb24390]
#> 
#> [[16]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x561f4da65958]
#> 
#> [[17]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x561f4da68170]
#> 
#> [[18]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x561f502d1368]
#> 
#> [[19]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x561f4f91c9d8]
#> 
#> [[20]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x561f4fe42a80]
#> 
#> [[21]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x561f4fe43180]
#> 
#> [[22]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x561f500ed7b8]
#> 
#> [[23]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x561f4cfe43a8]
#> 
#> [[24]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x561f4e9b4760]
#> 
#> [[25]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x561f4bcab008]
#> 
#> [[26]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x561f4bca4fa8]
#> 
#> [[27]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x561f49bd2188]
#> 
#> [[28]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x561f478ac310]
#> 
#> [[29]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# We can trim the first layer of those:
fn(1)
#> [[1]]
#> <frame 26> (25)
#> expr: fn(1)
#> env:  [local 0x561f4f7d8390]
#> 
#> [[2]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x561f4d18a688]
#> 
#> [[3]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x561f4f7d8550]
#> 
#> [[4]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x561f4f7d8780]
#> 
#> [[5]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x561f4d88c2e0]
#> 
#> [[6]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x561f4d88c858]
#> 
#> [[7]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x561f4d88cba0]
#> 
#> [[8]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x561f4d88cee8]
#> 
#> [[9]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x561f4d88d5b0]
#> 
#> [[10]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x561f4d88d7e0]
#> 
#> [[11]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x561f4d88d968]
#> 
#> [[12]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x561f4d88da48]
#> 
#> [[13]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x561f4e6e81c0]
#> 
#> [[14]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x561f4da65958]
#> 
#> [[15]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x561f4da68170]
#> 
#> [[16]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x561f502d1368]
#> 
#> [[17]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x561f4f91c9d8]
#> 
#> [[18]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x561f4fe42a80]
#> 
#> [[19]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x561f4fe43180]
#> 
#> [[20]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x561f500ed7b8]
#> 
#> [[21]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x561f4cfe43a8]
#> 
#> [[22]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x561f4e9b4760]
#> 
#> [[23]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x561f4bcab008]
#> 
#> [[24]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x561f4bca4fa8]
#> 
#> [[25]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x561f49bd2188]
#> 
#> [[26]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x561f478ac310]
#> 
#> [[27]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# The outside intervening frames (at the fn() call site) are still
# returned, but can be trimmed as well:
identity(identity(fn(1)))
#> [[1]]
#> <frame 28> (25)
#> expr: fn(1)
#> env:  [local 0x561f4e298ac8]
#> 
#> [[2]]
#> <frame 27> (25)
#> expr: identity(fn(1))
#> env:  [local 0x561f4e298be0]
#> 
#> [[3]]
#> <frame 26> (25)
#> expr: identity(identity(fn(1)))
#> env:  [local 0x561f4e298cf8]
#> 
#> [[4]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x561f4d18a688]
#> 
#> [[5]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x561f4e298eb8]
#> 
#> [[6]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x561f4e2990e8]
#> 
#> [[7]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x561f4e299698]
#> 
#> [[8]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x561f4ff955a0]
#> 
#> [[9]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x561f4ff958e8]
#> 
#> [[10]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x561f4ff95c30]
#> 
#> [[11]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x561f4ff962f8]
#> 
#> [[12]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x561f4ff96528]
#> 
#> [[13]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x561f4ff966b0]
#> 
#> [[14]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x561f4ff96790]
#> 
#> [[15]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x561f4f6b5140]
#> 
#> [[16]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x561f4da65958]
#> 
#> [[17]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x561f4da68170]
#> 
#> [[18]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x561f502d1368]
#> 
#> [[19]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x561f4f91c9d8]
#> 
#> [[20]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x561f4fe42a80]
#> 
#> [[21]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x561f4fe43180]
#> 
#> [[22]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x561f500ed7b8]
#> 
#> [[23]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x561f4cfe43a8]
#> 
#> [[24]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x561f4e9b4760]
#> 
#> [[25]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x561f4bcab008]
#> 
#> [[26]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x561f4bca4fa8]
#> 
#> [[27]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x561f49bd2188]
#> 
#> [[28]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x561f478ac310]
#> 
#> [[29]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     
identity(identity(fn(2)))
#> [[1]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x561f4d18a688]
#> 
#> [[2]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x561f4dfb8610]
#> 
#> [[3]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x561f4dfb8840]
#> 
#> [[4]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x561f4dfb8df0]
#> 
#> [[5]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x561f4dfb9368]
#> 
#> [[6]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x561f4dfb96b0]
#> 
#> [[7]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x561f4dfb5bc8]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x561f4dfb6290]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x561f4dfb64c0]
#> 
#> [[10]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x561f4dfb6648]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x561f4dfb6728]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x561f4e6097a8]
#> 
#> [[13]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x561f4da65958]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x561f4da68170]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x561f502d1368]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x561f4f91c9d8]
#> 
#> [[17]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x561f4fe42a80]
#> 
#> [[18]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x561f4fe43180]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x561f500ed7b8]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x561f4cfe43a8]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x561f4e9b4760]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x561f4bcab008]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x561f4bca4fa8]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x561f49bd2188]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x561f478ac310]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g <- function(trim) identity(identity(fn(trim)))
g(2)
#> [[1]]
#> <frame 26> (25)
#> expr: g(2)
#> env:  [local 0x561f4d589c70]
#> 
#> [[2]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x561f4d18a688]
#> 
#> [[3]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x561f4e408830]
#> 
#> [[4]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x561f4e408a60]
#> 
#> [[5]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x561f4e409010]
#> 
#> [[6]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x561f4e409588]
#> 
#> [[7]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x561f4e4098d0]
#> 
#> [[8]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x561f4e409c18]
#> 
#> [[9]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x561f4e40a2e0]
#> 
#> [[10]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x561f4e40a510]
#> 
#> [[11]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x561f4e406868]
#> 
#> [[12]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x561f4e406948]
#> 
#> [[13]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x561f4d507028]
#> 
#> [[14]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x561f4da65958]
#> 
#> [[15]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x561f4da68170]
#> 
#> [[16]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x561f502d1368]
#> 
#> [[17]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x561f4f91c9d8]
#> 
#> [[18]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x561f4fe42a80]
#> 
#> [[19]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x561f4fe43180]
#> 
#> [[20]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x561f500ed7b8]
#> 
#> [[21]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x561f4cfe43a8]
#> 
#> [[22]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x561f4e9b4760]
#> 
#> [[23]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x561f4bcab008]
#> 
#> [[24]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x561f4bca4fa8]
#> 
#> [[25]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x561f49bd2188]
#> 
#> [[26]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x561f478ac310]
#> 
#> [[27]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     
g(3)
#> [[1]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x561f4d18a688]
#> 
#> [[2]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x561f4e70c708]
#> 
#> [[3]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x561f4e70c970]
#> 
#> [[4]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x561f4e70cfc8]
#> 
#> [[5]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x561f4e70d5e8]
#> 
#> [[6]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x561f4e709b00]
#> 
#> [[7]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x561f4e709e48]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x561f4e70a548]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x561f4e70a778]
#> 
#> [[10]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x561f4e70a900]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x561f4e70a9e0]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x561f503eb558]
#> 
#> [[13]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x561f4da65958]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x561f4da68170]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x561f502d1368]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x561f4f91c9d8]
#> 
#> [[17]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x561f4fe42a80]
#> 
#> [[18]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x561f4fe43180]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x561f500ed7b8]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x561f4cfe43a8]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x561f4e9b4760]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x561f4bcab008]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x561f4bca4fa8]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x561f49bd2188]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x561f478ac310]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"