Call stack information

The eval_ and call_ families of functions provide a replacement for the base R functions prefixed with sys. (which are all about the context stack), as well as for parent.frame() (which is the only base R function for querying the call stack). The context stack includes all R-level evaluation contexts. It is linear in terms of execution history but due to lazy evaluation it is potentially nonlinear in terms of call history. The call stack history, on the other hand, is homogenous.

global_frame()

current_frame()

ctxt_frame(n = 1)

call_frame(n = 1, clean = TRUE)

ctxt_depth()

call_depth()

ctxt_stack(n = NULL, trim = 0)

call_stack(n = NULL, clean = TRUE)

Arguments

n
The number of frames to go back in the stack.

clean
Whether to post-process the call stack to clean non-standard frames. If TRUE, suboptimal call-stack entries by base::eval() will be cleaned up: the duplicate frame created by eval() is eliminated.

trim
The number of layers of intervening frames to trim off the stack. See stack_trim() and examples.

Details

ctxt_frame() and call_frame() return a frame object containing the following fields: expr and env (call expression and evaluation environment), pos and caller_pos (position of current frame in the context stack and position of the caller), and fun (function of the current frame). ctxt_stack() and call_stack() return a list of all context or call frames on the stack. Finally, ctxt_depth() and call_depth() report the current context position or the number of calling frames on the stack.

The base R functions take two sorts of arguments to indicate which frame to query: which and n. The n argument is straightforward: it's the number of frames to go down the stack, with n = 1 referring to the current context. The which argument is more complicated and changes meaning for values lower than 1. For the sake of consistency, the rlang functions all take the same kind of argument n. This argument has a single meaning (the number of frames to go down the stack) and cannot be lower than 1.

Note finally that parent.frame(1) corresponds to call_frame(2)$env, as n = 1 always refers to the current frame. This makes the _frame() and _stack() functions consistent: ctxt_frame(2) is the same as ctxt_stack()[[2]]. Also, ctxt_depth() returns one more frame than base::sys.nframe() because it counts the global frame. That is consistent with the _stack() functions which return the global frame as well. This way, call_stack(call_depth()) is the same as global_frame().

Life cycle

These functions are soft-deprecated and replaced by trace_back().

Examples

# Expressions within arguments count as contexts
identity(identity(ctxt_depth())) # returns 2

#> Warning: `ctxt_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 27

# But they are not part of the call stack because arguments are
# evaluated within the calling function (or the global environment
# if called at top level)
identity(identity(call_depth())) # returns 0

#> Warning: `call_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 14

# The context stacks includes all intervening execution frames. The
# call stack doesn't:
f <- function(x) identity(x)
f(f(ctxt_stack()))

#> [[1]]
#> <frame 28> (27)
#> expr: identity(x)
#> env:  [local 0x7f9198462638]
#> 
#> [[2]]
#> <frame 27> (24)
#> expr: f(ctxt_stack())
#> env:  [local 0x7f91984624e8]
#> 
#> [[3]]
#> <frame 26> (25)
#> expr: identity(x)
#> env:  [local 0x7f91984623d0]
#> 
#> [[4]]
#> <frame 25> (24)
#> expr: f(f(ctxt_stack()))
#> env:  [local 0x7f91984622b8]
#> 
#> [[5]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f9198666c70]
#> 
#> [[6]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f9198462088]
#> 
#> [[7]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f9198461e58]
#> 
#> [[8]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f9198461870]
#> 
#> [[9]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f9198461288]
#> 
#> [[10]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f9198460f40]
#> 
#> [[11]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f9198460bc0]
#> 
#> [[12]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f9198464398]
#> 
#> [[13]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f9198464168]
#> 
#> [[14]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f9198463fe0]
#> 
#> [[15]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f9198463f00]
#> 
#> [[16]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f9198475598]
#> 
#> [[17]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f91997b6590]
#> 
#> [[18]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7f91997b4ec0]
#> 
#> [[19]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f919a2a79d8]
#> 
#> [[20]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f919a07b8a8]
#> 
#> [[21]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f919ba81f38]
#> 
#> [[22]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7f919ba82638]
#> 
#> [[23]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f919bad8020]
#> 
#> [[24]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f9197b96d70]
#> 
#> [[25]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f9196500a60]
#> 
#> [[26]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f91963a81a0]
#> 
#> [[27]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f91963ae670]
#> 
#> [[28]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f919417b0c8]
#> 
#> [[29]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(f(call_stack()))

#> Warning: `call_stack()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [[1]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f9198666c70]
#> 
#> [[2]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f9198a25898]
#> 
#> [[3]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f91997b6590]
#> 
#> [[4]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7f91997b4ec0]
#> 
#> [[5]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f919a2a79d8]
#> 
#> [[6]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f919a07b8a8]
#> 
#> [[7]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f919ba81f38]
#> 
#> [[8]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f919bad8020]
#> 
#> [[9]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f9197b96d70]
#> 
#> [[10]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f9196500a60]
#> 
#> [[11]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f91963a81a0]
#> 
#> [[12]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f91963ae670]
#> 
#> [[13]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f919417b0c8]
#> 
#> [[14]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

g <- function(cmd) cmd()
f(g(ctxt_stack))

#> [[1]]
#> <frame 27> (24)
#> expr: g(ctxt_stack)
#> env:  [local 0x7f9197519640]
#> 
#> [[2]]
#> <frame 26> (25)
#> expr: identity(x)
#> env:  [local 0x7f9197519528]
#> 
#> [[3]]
#> <frame 25> (24)
#> expr: f(g(ctxt_stack))
#> env:  [local 0x7f91975193d8]
#> 
#> [[4]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f9198666c70]
#> 
#> [[5]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f9197519218]
#> 
#> [[6]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f9197518fe8]
#> 
#> [[7]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f9197518a00]
#> 
#> [[8]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f91975183e0]
#> 
#> [[9]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f9197518028]
#> 
#> [[10]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f9197517ca8]
#> 
#> [[11]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f919751b410]
#> 
#> [[12]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f919751b1e0]
#> 
#> [[13]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f919751b020]
#> 
#> [[14]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f919751af40]
#> 
#> [[15]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f91975306b8]
#> 
#> [[16]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f91997b6590]
#> 
#> [[17]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7f91997b4ec0]
#> 
#> [[18]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f919a2a79d8]
#> 
#> [[19]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f919a07b8a8]
#> 
#> [[20]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f919ba81f38]
#> 
#> [[21]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7f919ba82638]
#> 
#> [[22]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f919bad8020]
#> 
#> [[23]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f9197b96d70]
#> 
#> [[24]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f9196500a60]
#> 
#> [[25]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f91963a81a0]
#> 
#> [[26]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f91963ae670]
#> 
#> [[27]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f919417b0c8]
#> 
#> [[28]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(g(call_stack))

#> [[1]]
#> <frame 27> (23)
#> expr: g(call_stack)
#> env:  [local 0x7f919723e028]
#> 
#> [[2]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f9198666c70]
#> 
#> [[3]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f919725e6b8]
#> 
#> [[4]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f91997b6590]
#> 
#> [[5]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7f91997b4ec0]
#> 
#> [[6]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f919a2a79d8]
#> 
#> [[7]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f919a07b8a8]
#> 
#> [[8]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f919ba81f38]
#> 
#> [[9]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f919bad8020]
#> 
#> [[10]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f9197b96d70]
#> 
#> [[11]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f9196500a60]
#> 
#> [[12]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f91963a81a0]
#> 
#> [[13]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f91963ae670]
#> 
#> [[14]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f919417b0c8]
#> 
#> [[15]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

# The rlang _stack() functions return a list of frame
# objects. Use purrr::transpose() or index a field with
# purrr::map()'s to extract a particular field from a stack:

# stack <- f(f(call_stack()))
# purrr::map(stack, "env")
# purrr::transpose(stack)$expr

# current_frame() is an alias for ctxt_frame(1)
fn <- function() list(current = current_frame(), first = ctxt_frame(1))
fn()

#> Warning: `current_frame()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> $current
#> <frame 25> (24)
#> expr: fn()
#> env:  [local 0x7f91989805c8]
#> 
#> $first
#> <frame 25> (24)
#> expr: fn()
#> env:  [local 0x7f91989805c8]
#> 

# While current_frame() is the top of the stack, global_frame() is
# the bottom:
fn <- function() {
  n <- ctxt_depth()
  ctxt_frame(n)
}
identical(fn(), global_frame())

#> [1] TRUE

# ctxt_stack() returns a stack with all intervening frames. You can
# trim layers of intervening frames with the trim argument:
identity(identity(ctxt_stack()))

#> [[1]]
#> <frame 26> (24)
#> expr: identity(ctxt_stack())
#> env:  [local 0x7f9197f99bf8]
#> 
#> [[2]]
#> <frame 25> (24)
#> expr: identity(identity(ctxt_stack()))
#> env:  [local 0x7f9197f99aa8]
#> 
#> [[3]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f9198666c70]
#> 
#> [[4]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f9197f998e8]
#> 
#> [[5]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f9197f99680]
#> 
#> [[6]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f9197f99060]
#> 
#> [[7]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f9197f98a40]
#> 
#> [[8]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f9197f9c608]
#> 
#> [[9]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f9197f9c2c0]
#> 
#> [[10]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f9197f9bb50]
#> 
#> [[11]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f9197f9b8e8]
#> 
#> [[12]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f9197f9b760]
#> 
#> [[13]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f9197f9b680]
#> 
#> [[14]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f9197fbc758]
#> 
#> [[15]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f91997b6590]
#> 
#> [[16]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7f91997b4ec0]
#> 
#> [[17]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f919a2a79d8]
#> 
#> [[18]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f919a07b8a8]
#> 
#> [[19]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f919ba81f38]
#> 
#> [[20]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7f919ba82638]
#> 
#> [[21]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f919bad8020]
#> 
#> [[22]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f9197b96d70]
#> 
#> [[23]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f9196500a60]
#> 
#> [[24]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f91963a81a0]
#> 
#> [[25]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f91963ae670]
#> 
#> [[26]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f919417b0c8]
#> 
#> [[27]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(ctxt_stack(trim = 1)))

#> [[1]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f9198666c70]
#> 
#> [[2]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f9198024e80]
#> 
#> [[3]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f9198024c50]
#> 
#> [[4]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f9198024668]
#> 
#> [[5]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f9198027fc8]
#> 
#> [[6]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f9198027c80]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f9198027938]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f91980271c8]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f9198026f60]
#> 
#> [[10]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f9198026dd8]
#> 
#> [[11]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f9198026cf8]
#> 
#> [[12]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f919803beb0]
#> 
#> [[13]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f91997b6590]
#> 
#> [[14]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7f91997b4ec0]
#> 
#> [[15]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f919a2a79d8]
#> 
#> [[16]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f919a07b8a8]
#> 
#> [[17]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f919ba81f38]
#> 
#> [[18]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7f919ba82638]
#> 
#> [[19]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f919bad8020]
#> 
#> [[20]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f9197b96d70]
#> 
#> [[21]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f9196500a60]
#> 
#> [[22]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f91963a81a0]
#> 
#> [[23]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f91963ae670]
#> 
#> [[24]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f919417b0c8]
#> 
#> [[25]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# ctxt_stack() is called within fn() with intervening frames:
fn <- function(trim) identity(identity(ctxt_stack(trim = trim)))
fn(0)

#> [[1]]
#> <frame 27> (25)
#> expr: identity(ctxt_stack(trim = trim))
#> env:  [local 0x7f9197dad1a8]
#> 
#> [[2]]
#> <frame 26> (25)
#> expr: identity(identity(ctxt_stack(trim = trim)))
#> env:  [local 0x7f9197dad058]
#> 
#> [[3]]
#> <frame 25> (24)
#> expr: fn(0)
#> env:  [local 0x7f9197dacf40]
#> 
#> [[4]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f9198666c70]
#> 
#> [[5]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f9197dacd80]
#> 
#> [[6]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f9197dacb18]
#> 
#> [[7]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f9197db0408]
#> 
#> [[8]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f9197dafe20]
#> 
#> [[9]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f9197dafaa0]
#> 
#> [[10]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f9197daf720]
#> 
#> [[11]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f9197daefb0]
#> 
#> [[12]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f9197daed80]
#> 
#> [[13]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f9197daebf8]
#> 
#> [[14]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f9197daeb18]
#> 
#> [[15]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f9197dc76e8]
#> 
#> [[16]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f91997b6590]
#> 
#> [[17]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7f91997b4ec0]
#> 
#> [[18]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f919a2a79d8]
#> 
#> [[19]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f919a07b8a8]
#> 
#> [[20]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f919ba81f38]
#> 
#> [[21]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7f919ba82638]
#> 
#> [[22]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f919bad8020]
#> 
#> [[23]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f9197b96d70]
#> 
#> [[24]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f9196500a60]
#> 
#> [[25]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f91963a81a0]
#> 
#> [[26]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f91963ae670]
#> 
#> [[27]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f919417b0c8]
#> 
#> [[28]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# We can trim the first layer of those:
fn(1)

#> [[1]]
#> <frame 25> (24)
#> expr: fn(1)
#> env:  [local 0x7f9197a22b08]
#> 
#> [[2]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f9198666c70]
#> 
#> [[3]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f9197a22948]
#> 
#> [[4]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f9197a226a8]
#> 
#> [[5]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f9197a25eb8]
#> 
#> [[6]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f9197a25240]
#> 
#> [[7]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f9197a24ef8]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f9197a24ad0]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f9197a2c200]
#> 
#> [[10]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f9197a2bfd0]
#> 
#> [[11]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f9197a2b7b8]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f9197a2b6d8]
#> 
#> [[13]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f9197a49f28]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f91997b6590]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7f91997b4ec0]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f919a2a79d8]
#> 
#> [[17]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f919a07b8a8]
#> 
#> [[18]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f919ba81f38]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7f919ba82638]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f919bad8020]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f9197b96d70]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f9196500a60]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f91963a81a0]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f91963ae670]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f919417b0c8]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# The outside intervening frames (at the fn() call site) are still
# returned, but can be trimmed as well:
identity(identity(fn(1)))

#> [[1]]
#> <frame 27> (24)
#> expr: fn(1)
#> env:  [local 0x7f91953d4878]
#> 
#> [[2]]
#> <frame 26> (24)
#> expr: identity(fn(1))
#> env:  [local 0x7f91953d8600]
#> 
#> [[3]]
#> <frame 25> (24)
#> expr: identity(identity(fn(1)))
#> env:  [local 0x7f91953d8328]
#> 
#> [[4]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f9198666c70]
#> 
#> [[5]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f91953d7f38]
#> 
#> [[6]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f91953d7a68]
#> 
#> [[7]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f91953d71a8]
#> 
#> [[8]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f91953da050]
#> 
#> [[9]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f91953d9bf0]
#> 
#> [[10]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f91953d97c8]
#> 
#> [[11]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f91953d9090]
#> 
#> [[12]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f91953d8df0]
#> 
#> [[13]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f91953d8c30]
#> 
#> [[14]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f91953d8aa8]
#> 
#> [[15]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f91944d0bc8]
#> 
#> [[16]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f91997b6590]
#> 
#> [[17]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7f91997b4ec0]
#> 
#> [[18]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f919a2a79d8]
#> 
#> [[19]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f919a07b8a8]
#> 
#> [[20]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f919ba81f38]
#> 
#> [[21]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7f919ba82638]
#> 
#> [[22]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f919bad8020]
#> 
#> [[23]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f9197b96d70]
#> 
#> [[24]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f9196500a60]
#> 
#> [[25]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f91963a81a0]
#> 
#> [[26]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f91963ae670]
#> 
#> [[27]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f919417b0c8]
#> 
#> [[28]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(fn(2)))

#> [[1]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f9198666c70]
#> 
#> [[2]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f9193d79d50]
#> 
#> [[3]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f9193d79b20]
#> 
#> [[4]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f9193d7d368]
#> 
#> [[5]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f9193d7cd48]
#> 
#> [[6]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f9193d7ca00]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f9193d7c648]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f9193d7bf48]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f9193d7bc70]
#> 
#> [[10]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f9193d7bae8]
#> 
#> [[11]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f9193d7f8e0]
#> 
#> [[12]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f9194617fc8]
#> 
#> [[13]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f91997b6590]
#> 
#> [[14]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7f91997b4ec0]
#> 
#> [[15]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f919a2a79d8]
#> 
#> [[16]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f919a07b8a8]
#> 
#> [[17]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f919ba81f38]
#> 
#> [[18]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7f919ba82638]
#> 
#> [[19]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f919bad8020]
#> 
#> [[20]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f9197b96d70]
#> 
#> [[21]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f9196500a60]
#> 
#> [[22]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f91963a81a0]
#> 
#> [[23]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f91963ae670]
#> 
#> [[24]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f919417b0c8]
#> 
#> [[25]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g <- function(trim) identity(identity(fn(trim)))
g(2)

#> [[1]]
#> <frame 25> (24)
#> expr: g(2)
#> env:  [local 0x7f919bf521d0]
#> 
#> [[2]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f9198666c70]
#> 
#> [[3]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f919bf52390]
#> 
#> [[4]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f919bf525c0]
#> 
#> [[5]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f919bf52b70]
#> 
#> [[6]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f919bf530e8]
#> 
#> [[7]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f919bf4f520]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f919bf4f868]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f919bf4ff30]
#> 
#> [[10]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f919bf50160]
#> 
#> [[11]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f919bf502e8]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f919bf503c8]
#> 
#> [[13]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f919bf3fd38]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f91997b6590]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7f91997b4ec0]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f919a2a79d8]
#> 
#> [[17]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f919a07b8a8]
#> 
#> [[18]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f919ba81f38]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7f919ba82638]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f919bad8020]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f9197b96d70]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f9196500a60]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f91963a81a0]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f91963ae670]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f919417b0c8]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g(3)

#> [[1]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f9198666c70]
#> 
#> [[2]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f919bd14530]
#> 
#> [[3]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f919bd14760]
#> 
#> [[4]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f919bd10e00]
#> 
#> [[5]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f919bd11378]
#> 
#> [[6]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f919bd116c0]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f919bd11a08]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f919bd12108]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f919bd12338]
#> 
#> [[10]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f919bd124c0]
#> 
#> [[11]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f919bd125a0]
#> 
#> [[12]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f919bd01f80]
#> 
#> [[13]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f91997b6590]
#> 
#> [[14]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7f91997b4ec0]
#> 
#> [[15]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f919a2a79d8]
#> 
#> [[16]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f919a07b8a8]
#> 
#> [[17]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f919ba81f38]
#> 
#> [[18]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7f919ba82638]
#> 
#> [[19]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f919bad8020]
#> 
#> [[20]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f9197b96d70]
#> 
#> [[21]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f9196500a60]
#> 
#> [[22]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f91963a81a0]
#> 
#> [[23]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f91963ae670]
#> 
#> [[24]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f919417b0c8]
#> 
#> [[25]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"