Call stack information

The eval_ and call_ families of functions provide a replacement for the base R functions prefixed with sys. (which are all about the context stack), as well as for parent.frame() (which is the only base R function for querying the call stack). The context stack includes all R-level evaluation contexts. It is linear in terms of execution history but due to lazy evaluation it is potentially nonlinear in terms of call history. The call stack history, on the other hand, is homogenous.

Usage

global_frame()

current_frame()

ctxt_frame(n = 1)

call_frame(n = 1, clean = TRUE)

ctxt_depth()

call_depth()

ctxt_stack(n = NULL, trim = 0)

call_stack(n = NULL, clean = TRUE)

Arguments

n

The number of frames to go back in the stack.

clean

Whether to post-process the call stack to clean non-standard frames. If TRUE, suboptimal call-stack entries by base::eval() will be cleaned up: the duplicate frame created by eval() is eliminated.

trim

The number of layers of intervening frames to trim off the stack. See stack_trim() and examples.

Details

ctxt_frame() and call_frame() return a frame object containing the following fields: expr and env (call expression and evaluation environment), pos and caller_pos (position of current frame in the context stack and position of the caller), and fun (function of the current frame). ctxt_stack() and call_stack() return a list of all context or call frames on the stack. Finally, ctxt_depth() and call_depth() report the current context position or the number of calling frames on the stack.

The base R functions take two sorts of arguments to indicate which frame to query: which and n. The n argument is straightforward: it's the number of frames to go down the stack, with n = 1 referring to the current context. The which argument is more complicated and changes meaning for values lower than 1. For the sake of consistency, the rlang functions all take the same kind of argument n. This argument has a single meaning (the number of frames to go down the stack) and cannot be lower than 1.

Note finally that parent.frame(1) corresponds to call_frame(2)$env, as n = 1 always refers to the current frame. This makes the _frame() and _stack() functions consistent: ctxt_frame(2) is the same as ctxt_stack()[[2]]. Also, ctxt_depth() returns one more frame than base::sys.nframe() because it counts the global frame. That is consistent with the _stack() functions which return the global frame as well. This way, call_stack(call_depth()) is the same as global_frame().

Life cycle

These functions are soft-deprecated and replaced by trace_back().

Examples

# Expressions within arguments count as contexts
identity(identity(ctxt_depth())) # returns 2
#> Warning: `ctxt_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.
#> [1] 28

# But they are not part of the call stack because arguments are
# evaluated within the calling function (or the global environment
# if called at top level)
identity(identity(call_depth())) # returns 0
#> Warning: `call_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.
#> [1] 15

# The context stacks includes all intervening execution frames. The
# call stack doesn't:
f <- function(x) identity(x)
f(f(ctxt_stack()))
#> [[1]]
#> <frame 29> (28)
#> expr: identity(x)
#> env:  [local 0x560b421de838]
#> 
#> [[2]]
#> <frame 28> (25)
#> expr: f(ctxt_stack())
#> env:  [local 0x560b421de720]
#> 
#> [[3]]
#> <frame 27> (26)
#> expr: identity(x)
#> env:  [local 0x560b421de5d0]
#> 
#> [[4]]
#> <frame 26> (25)
#> expr: f(f(ctxt_stack()))
#> env:  [local 0x560b421de4b8]
#> 
#> [[5]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x560b43c047b0]
#> 
#> [[6]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x560b421de2f8]
#> 
#> [[7]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x560b421de0c8]
#> 
#> [[8]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x560b421e1948]
#> 
#> [[9]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x560b421e13d0]
#> 
#> [[10]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x560b421e1088]
#> 
#> [[11]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x560b421e0c28]
#> 
#> [[12]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x560b421e0528]
#> 
#> [[13]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x560b421e02f8]
#> 
#> [[14]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x560b421e0170]
#> 
#> [[15]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x560b421e0090]
#> 
#> [[16]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x560b42ce8cb0]
#> 
#> [[17]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x560b42dddec8]
#> 
#> [[18]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x560b42ddc1a0]
#> 
#> [[19]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x560b42d1bde0]
#> 
#> [[20]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x560b429c3cd8]
#> 
#> [[21]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x560b43af02d8]
#> 
#> [[22]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x560b43af09d8]
#> 
#> [[23]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x560b429a6918]
#> 
#> [[24]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x560b4356b840]
#> 
#> [[25]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x560b4116eb88]
#> 
#> [[26]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x560b3e9b18e0]
#> 
#> [[27]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x560b3e9b78a8]
#> 
#> [[28]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x560b3e89c1b0]
#> 
#> [[29]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x560b3c5cd7a8]
#> 
#> [[30]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     
f(f(call_stack()))
#> Warning: `call_stack()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.
#> [[1]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x560b43c047b0]
#> 
#> [[2]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x560b430b48c8]
#> 
#> [[3]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x560b42dddec8]
#> 
#> [[4]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x560b42ddc1a0]
#> 
#> [[5]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x560b42d1bde0]
#> 
#> [[6]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x560b429c3cd8]
#> 
#> [[7]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x560b43af02d8]
#> 
#> [[8]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x560b429a6918]
#> 
#> [[9]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x560b4356b840]
#> 
#> [[10]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x560b4116eb88]
#> 
#> [[11]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x560b3e9b18e0]
#> 
#> [[12]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x560b3e9b78a8]
#> 
#> [[13]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x560b3e89c1b0]
#> 
#> [[14]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x560b3c5cd7a8]
#> 
#> [[15]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

g <- function(cmd) cmd()
f(g(ctxt_stack))
#> [[1]]
#> <frame 28> (25)
#> expr: g(ctxt_stack)
#> env:  [local 0x560b4109f310]
#> 
#> [[2]]
#> <frame 27> (26)
#> expr: identity(x)
#> env:  [local 0x560b4109f1f8]
#> 
#> [[3]]
#> <frame 26> (25)
#> expr: f(g(ctxt_stack))
#> env:  [local 0x560b4109f0e0]
#> 
#> [[4]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x560b43c047b0]
#> 
#> [[5]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x560b4109ef20]
#> 
#> [[6]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x560b4109ecf0]
#> 
#> [[7]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x560b410a2500]
#> 
#> [[8]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x560b410a1f50]
#> 
#> [[9]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x560b410a1bd0]
#> 
#> [[10]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x560b410a1888]
#> 
#> [[11]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x560b410a11c0]
#> 
#> [[12]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x560b410a0f90]
#> 
#> [[13]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x560b410a0e08]
#> 
#> [[14]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x560b410a0d28]
#> 
#> [[15]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x560b410b5ee0]
#> 
#> [[16]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x560b42dddec8]
#> 
#> [[17]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x560b42ddc1a0]
#> 
#> [[18]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x560b42d1bde0]
#> 
#> [[19]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x560b429c3cd8]
#> 
#> [[20]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x560b43af02d8]
#> 
#> [[21]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x560b43af09d8]
#> 
#> [[22]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x560b429a6918]
#> 
#> [[23]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x560b4356b840]
#> 
#> [[24]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x560b4116eb88]
#> 
#> [[25]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x560b3e9b18e0]
#> 
#> [[26]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x560b3e9b78a8]
#> 
#> [[27]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x560b3e89c1b0]
#> 
#> [[28]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x560b3c5cd7a8]
#> 
#> [[29]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     
f(g(call_stack))
#> [[1]]
#> <frame 28> (24)
#> expr: g(call_stack)
#> env:  [local 0x560b40d62b30]
#> 
#> [[2]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x560b43c047b0]
#> 
#> [[3]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x560b40d86a90]
#> 
#> [[4]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x560b42dddec8]
#> 
#> [[5]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x560b42ddc1a0]
#> 
#> [[6]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x560b42d1bde0]
#> 
#> [[7]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x560b429c3cd8]
#> 
#> [[8]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x560b43af02d8]
#> 
#> [[9]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x560b429a6918]
#> 
#> [[10]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x560b4356b840]
#> 
#> [[11]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x560b4116eb88]
#> 
#> [[12]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x560b3e9b18e0]
#> 
#> [[13]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x560b3e9b78a8]
#> 
#> [[14]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x560b3e89c1b0]
#> 
#> [[15]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x560b3c5cd7a8]
#> 
#> [[16]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

# The rlang _stack() functions return a list of frame
# objects. Use purrr::transpose() or index a field with
# purrr::map()'s to extract a particular field from a stack:

# stack <- f(f(call_stack()))
# purrr::map(stack, "env")
# purrr::transpose(stack)$expr

# current_frame() is an alias for ctxt_frame(1)
fn <- function() list(current = current_frame(), first = ctxt_frame(1))
fn()
#> Warning: `current_frame()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.
#> $current
#> <frame 26> (25)
#> expr: fn()
#> env:  [local 0x560b40892618]
#> 
#> $first
#> <frame 26> (25)
#> expr: fn()
#> env:  [local 0x560b40892618]
#> 

# While current_frame() is the top of the stack, global_frame() is
# the bottom:
fn <- function() {
  n <- ctxt_depth()
  ctxt_frame(n)
}
identical(fn(), global_frame())
#> [1] TRUE

# ctxt_stack() returns a stack with all intervening frames. You can
# trim layers of intervening frames with the trim argument:
identity(identity(ctxt_stack()))
#> [[1]]
#> <frame 27> (25)
#> expr: identity(ctxt_stack())
#> env:  [local 0x560b405f30a8]
#> 
#> [[2]]
#> <frame 26> (25)
#> expr: identity(identity(ctxt_stack()))
#> env:  [local 0x560b405f2f58]
#> 
#> [[3]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x560b43c047b0]
#> 
#> [[4]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x560b405f2d98]
#> 
#> [[5]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x560b405f2b30]
#> 
#> [[6]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x560b405f6340]
#> 
#> [[7]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x560b405f5d20]
#> 
#> [[8]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x560b405f58f8]
#> 
#> [[9]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x560b405f5578]
#> 
#> [[10]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x560b405f4e78]
#> 
#> [[11]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x560b405f4c48]
#> 
#> [[12]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x560b405f4a18]
#> 
#> [[13]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x560b405f8730]
#> 
#> [[14]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x560b4060a858]
#> 
#> [[15]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x560b42dddec8]
#> 
#> [[16]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x560b42ddc1a0]
#> 
#> [[17]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x560b42d1bde0]
#> 
#> [[18]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x560b429c3cd8]
#> 
#> [[19]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x560b43af02d8]
#> 
#> [[20]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x560b43af09d8]
#> 
#> [[21]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x560b429a6918]
#> 
#> [[22]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x560b4356b840]
#> 
#> [[23]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x560b4116eb88]
#> 
#> [[24]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x560b3e9b18e0]
#> 
#> [[25]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x560b3e9b78a8]
#> 
#> [[26]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x560b3e89c1b0]
#> 
#> [[27]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x560b3c5cd7a8]
#> 
#> [[28]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     
identity(identity(ctxt_stack(trim = 1)))
#> [[1]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x560b43c047b0]
#> 
#> [[2]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x560b3ff74f50]
#> 
#> [[3]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x560b3ff74c78]
#> 
#> [[4]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x560b3ff74690]
#> 
#> [[5]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x560b3ff77ea0]
#> 
#> [[6]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x560b3ff77b58]
#> 
#> [[7]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x560b3ff77768]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x560b3ff76fc0]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x560b3ff76d58]
#> 
#> [[10]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x560b3ff76bd0]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x560b3ff76af0]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x560b3ff8e2a0]
#> 
#> [[13]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x560b42dddec8]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x560b42ddc1a0]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x560b42d1bde0]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x560b429c3cd8]
#> 
#> [[17]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x560b43af02d8]
#> 
#> [[18]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x560b43af09d8]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x560b429a6918]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x560b4356b840]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x560b4116eb88]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x560b3e9b18e0]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x560b3e9b78a8]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x560b3e89c1b0]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x560b3c5cd7a8]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# ctxt_stack() is called within fn() with intervening frames:
fn <- function(trim) identity(identity(ctxt_stack(trim = trim)))
fn(0)
#> [[1]]
#> <frame 28> (26)
#> expr: identity(ctxt_stack(trim = trim))
#> env:  [local 0x560b3eb82da0]
#> 
#> [[2]]
#> <frame 27> (26)
#> expr: identity(identity(ctxt_stack(trim = trim)))
#> env:  [local 0x560b3eb82c88]
#> 
#> [[3]]
#> <frame 26> (25)
#> expr: fn(0)
#> env:  [local 0x560b3eb82b70]
#> 
#> [[4]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x560b43c047b0]
#> 
#> [[5]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x560b3eb82978]
#> 
#> [[6]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x560b3eb82710]
#> 
#> [[7]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x560b3eb82128]
#> 
#> [[8]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x560b3eb88b00]
#> 
#> [[9]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x560b3eb88780]
#> 
#> [[10]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x560b3eb88400]
#> 
#> [[11]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x560b3eb87cc8]
#> 
#> [[12]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x560b3eb87a60]
#> 
#> [[13]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x560b3eb878d8]
#> 
#> [[14]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x560b3eb877f8]
#> 
#> [[15]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x560b3ebadfa0]
#> 
#> [[16]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x560b42dddec8]
#> 
#> [[17]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x560b42ddc1a0]
#> 
#> [[18]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x560b42d1bde0]
#> 
#> [[19]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x560b429c3cd8]
#> 
#> [[20]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x560b43af02d8]
#> 
#> [[21]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x560b43af09d8]
#> 
#> [[22]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x560b429a6918]
#> 
#> [[23]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x560b4356b840]
#> 
#> [[24]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x560b4116eb88]
#> 
#> [[25]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x560b3e9b18e0]
#> 
#> [[26]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x560b3e9b78a8]
#> 
#> [[27]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x560b3e89c1b0]
#> 
#> [[28]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x560b3c5cd7a8]
#> 
#> [[29]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# We can trim the first layer of those:
fn(1)
#> [[1]]
#> <frame 26> (25)
#> expr: fn(1)
#> env:  [local 0x560b3d200008]
#> 
#> [[2]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x560b43c047b0]
#> 
#> [[3]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x560b3d1ffe10]
#> 
#> [[4]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x560b3d1ffba8]
#> 
#> [[5]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x560b3d203150]
#> 
#> [[6]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x560b3d2014e0]
#> 
#> [[7]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x560b3d204d28]
#> 
#> [[8]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x560b3d205828]
#> 
#> [[9]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x560b3dbf1e40]
#> 
#> [[10]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x560b3dbf4738]
#> 
#> [[11]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x560b3dbf4508]
#> 
#> [[12]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x560b3dbf3938]
#> 
#> [[13]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x560b3e31ca40]
#> 
#> [[14]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x560b42dddec8]
#> 
#> [[15]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x560b42ddc1a0]
#> 
#> [[16]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x560b42d1bde0]
#> 
#> [[17]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x560b429c3cd8]
#> 
#> [[18]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x560b43af02d8]
#> 
#> [[19]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x560b43af09d8]
#> 
#> [[20]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x560b429a6918]
#> 
#> [[21]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x560b4356b840]
#> 
#> [[22]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x560b4116eb88]
#> 
#> [[23]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x560b3e9b18e0]
#> 
#> [[24]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x560b3e9b78a8]
#> 
#> [[25]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x560b3e89c1b0]
#> 
#> [[26]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x560b3c5cd7a8]
#> 
#> [[27]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# The outside intervening frames (at the fn() call site) are still
# returned, but can be trimmed as well:
identity(identity(fn(1)))
#> [[1]]
#> <frame 28> (25)
#> expr: fn(1)
#> env:  [local 0x560b43c481c0]
#> 
#> [[2]]
#> <frame 27> (25)
#> expr: identity(fn(1))
#> env:  [local 0x560b4256cfb8]
#> 
#> [[3]]
#> <frame 26> (25)
#> expr: identity(identity(fn(1)))
#> env:  [local 0x560b4256d0d0]
#> 
#> [[4]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x560b43c047b0]
#> 
#> [[5]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x560b4256d290]
#> 
#> [[6]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x560b4256d4c0]
#> 
#> [[7]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x560b4256da70]
#> 
#> [[8]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x560b4256dfe8]
#> 
#> [[9]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x560b4256e330]
#> 
#> [[10]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x560b4256e678]
#> 
#> [[11]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x560b4256ed40]
#> 
#> [[12]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x560b4360c840]
#> 
#> [[13]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x560b4360c9c8]
#> 
#> [[14]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x560b4360caa8]
#> 
#> [[15]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x560b42e9fc30]
#> 
#> [[16]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x560b42dddec8]
#> 
#> [[17]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x560b42ddc1a0]
#> 
#> [[18]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x560b42d1bde0]
#> 
#> [[19]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x560b429c3cd8]
#> 
#> [[20]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x560b43af02d8]
#> 
#> [[21]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x560b43af09d8]
#> 
#> [[22]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x560b429a6918]
#> 
#> [[23]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x560b4356b840]
#> 
#> [[24]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x560b4116eb88]
#> 
#> [[25]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x560b3e9b18e0]
#> 
#> [[26]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x560b3e9b78a8]
#> 
#> [[27]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x560b3e89c1b0]
#> 
#> [[28]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x560b3c5cd7a8]
#> 
#> [[29]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     
identity(identity(fn(2)))
#> [[1]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x560b43c047b0]
#> 
#> [[2]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x560b41f65850]
#> 
#> [[3]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x560b41f65a80]
#> 
#> [[4]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x560b43407da0]
#> 
#> [[5]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x560b43408318]
#> 
#> [[6]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x560b43408660]
#> 
#> [[7]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x560b434089a8]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x560b434091f8]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x560b43409428]
#> 
#> [[10]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x560b434095b0]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x560b43409700]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x560b42b55ed0]
#> 
#> [[13]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x560b42dddec8]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x560b42ddc1a0]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x560b42d1bde0]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x560b429c3cd8]
#> 
#> [[17]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x560b43af02d8]
#> 
#> [[18]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x560b43af09d8]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x560b429a6918]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x560b4356b840]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x560b4116eb88]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x560b3e9b18e0]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x560b3e9b78a8]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x560b3e89c1b0]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x560b3c5cd7a8]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g <- function(trim) identity(identity(fn(trim)))
g(2)
#> [[1]]
#> <frame 26> (25)
#> expr: g(2)
#> env:  [local 0x560b44268750]
#> 
#> [[2]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x560b43c047b0]
#> 
#> [[3]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x560b44268948]
#> 
#> [[4]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x560b44268b78]
#> 
#> [[5]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x560b44269128]
#> 
#> [[6]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x560b43403fd0]
#> 
#> [[7]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x560b43404318]
#> 
#> [[8]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x560b43404660]
#> 
#> [[9]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x560b43404d28]
#> 
#> [[10]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x560b43404f58]
#> 
#> [[11]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x560b434050e0]
#> 
#> [[12]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x560b434051c0]
#> 
#> [[13]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x560b421c5270]
#> 
#> [[14]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x560b42dddec8]
#> 
#> [[15]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x560b42ddc1a0]
#> 
#> [[16]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x560b42d1bde0]
#> 
#> [[17]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x560b429c3cd8]
#> 
#> [[18]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x560b43af02d8]
#> 
#> [[19]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x560b43af09d8]
#> 
#> [[20]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x560b429a6918]
#> 
#> [[21]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x560b4356b840]
#> 
#> [[22]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x560b4116eb88]
#> 
#> [[23]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x560b3e9b18e0]
#> 
#> [[24]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x560b3e9b78a8]
#> 
#> [[25]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x560b3e89c1b0]
#> 
#> [[26]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x560b3c5cd7a8]
#> 
#> [[27]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     
g(3)
#> [[1]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x560b43c047b0]
#> 
#> [[2]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x560b4491c8f0]
#> 
#> [[3]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x560b4491cb20]
#> 
#> [[4]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x560b4491d108]
#> 
#> [[5]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x560b4491d680]
#> 
#> [[6]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x560b44919b98]
#> 
#> [[7]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x560b44919ee0]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x560b4491a5a8]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x560b4491a7d8]
#> 
#> [[10]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x560b4491a960]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x560b4491aa40]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x560b41dd83b8]
#> 
#> [[13]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x560b42dddec8]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x560b42ddc1a0]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x560b42d1bde0]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x560b429c3cd8]
#> 
#> [[17]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x560b43af02d8]
#> 
#> [[18]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x560b43af09d8]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x560b429a6918]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x560b4356b840]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x560b4116eb88]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x560b3e9b18e0]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x560b3e9b78a8]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x560b3e89c1b0]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x560b3c5cd7a8]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"