Call stack information

The eval_ and call_ families of functions provide a replacement for the base R functions prefixed with sys. (which are all about the context stack), as well as for parent.frame() (which is the only base R function for querying the call stack). The context stack includes all R-level evaluation contexts. It is linear in terms of execution history but due to lazy evaluation it is potentially nonlinear in terms of call history. The call stack history, on the other hand, is homogenous.

Usage

global_frame()

current_frame()

ctxt_frame(n = 1)

call_frame(n = 1, clean = TRUE)

ctxt_depth()

call_depth()

ctxt_stack(n = NULL, trim = 0)

call_stack(n = NULL, clean = TRUE)

Arguments

n

The number of frames to go back in the stack.

clean

Whether to post-process the call stack to clean non-standard frames. If TRUE, suboptimal call-stack entries by base::eval() will be cleaned up: the duplicate frame created by eval() is eliminated.

trim

The number of layers of intervening frames to trim off the stack. See stack_trim() and examples.

Details

ctxt_frame() and call_frame() return a frame object containing the following fields: expr and env (call expression and evaluation environment), pos and caller_pos (position of current frame in the context stack and position of the caller), and fun (function of the current frame). ctxt_stack() and call_stack() return a list of all context or call frames on the stack. Finally, ctxt_depth() and call_depth() report the current context position or the number of calling frames on the stack.

The base R functions take two sorts of arguments to indicate which frame to query: which and n. The n argument is straightforward: it's the number of frames to go down the stack, with n = 1 referring to the current context. The which argument is more complicated and changes meaning for values lower than 1. For the sake of consistency, the rlang functions all take the same kind of argument n. This argument has a single meaning (the number of frames to go down the stack) and cannot be lower than 1.

Note finally that parent.frame(1) corresponds to call_frame(2)$env, as n = 1 always refers to the current frame. This makes the _frame() and _stack() functions consistent: ctxt_frame(2) is the same as ctxt_stack()[[2]]. Also, ctxt_depth() returns one more frame than base::sys.nframe() because it counts the global frame. That is consistent with the _stack() functions which return the global frame as well. This way, call_stack(call_depth()) is the same as global_frame().

Life cycle

These functions are soft-deprecated and replaced by trace_back().

Examples

# Expressions within arguments count as contexts
identity(identity(ctxt_depth())) # returns 2
#> Warning: `ctxt_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.
#> [1] 28

# But they are not part of the call stack because arguments are
# evaluated within the calling function (or the global environment
# if called at top level)
identity(identity(call_depth())) # returns 0
#> Warning: `call_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.
#> [1] 15

# The context stacks includes all intervening execution frames. The
# call stack doesn't:
f <- function(x) identity(x)
f(f(ctxt_stack()))
#> [[1]]
#> <frame 29> (28)
#> expr: identity(x)
#> env:  [local 0x564600fee2c8]
#> 
#> [[2]]
#> <frame 28> (25)
#> expr: f(ctxt_stack())
#> env:  [local 0x564600fee3e0]
#> 
#> [[3]]
#> <frame 27> (26)
#> expr: identity(x)
#> env:  [local 0x564600fee4f8]
#> 
#> [[4]]
#> <frame 26> (25)
#> expr: f(f(ctxt_stack()))
#> env:  [local 0x564600fee610]
#> 
#> [[5]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5645ffd1f118]
#> 
#> [[6]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x564600fee7d0]
#> 
#> [[7]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x564600feea00]
#> 
#> [[8]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x564600feefb0]
#> 
#> [[9]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x564600fef528]
#> 
#> [[10]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x564600fef870]
#> 
#> [[11]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x5645ff15e848]
#> 
#> [[12]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x5645ff15ef10]
#> 
#> [[13]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x5645ff15f140]
#> 
#> [[14]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x5645ff15f2c8]
#> 
#> [[15]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x5645ff15f3a8]
#> 
#> [[16]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x5645ff95a6f8]
#> 
#> [[17]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x5645fa605f70]
#> 
#> [[18]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x5645fa60bec8]
#> 
#> [[19]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x5645fa687e50]
#> 
#> [[20]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x5645fa8458e0]
#> 
#> [[21]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x5645fb8bdf08]
#> 
#> [[22]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x5645fb8bd7d0]
#> 
#> [[23]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x5645fb8e3098]
#> 
#> [[24]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x5645fabb4b70]
#> 
#> [[25]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x5645fac138c0]
#> 
#> [[26]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x5645fa7fca78]
#> 
#> [[27]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x5645fa808e30]
#> 
#> [[28]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x5645fa88c638]
#> 
#> [[29]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x5645f92fd4a8]
#> 
#> [[30]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     
f(f(call_stack()))
#> Warning: `call_stack()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.
#> [[1]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5645ffd1f118]
#> 
#> [[2]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x5645ff2c88d0]
#> 
#> [[3]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x5645fa605f70]
#> 
#> [[4]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x5645fa60bec8]
#> 
#> [[5]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x5645fa687e50]
#> 
#> [[6]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x5645fa8458e0]
#> 
#> [[7]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x5645fb8bdf08]
#> 
#> [[8]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x5645fb8e3098]
#> 
#> [[9]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x5645fabb4b70]
#> 
#> [[10]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x5645fac138c0]
#> 
#> [[11]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x5645fa7fca78]
#> 
#> [[12]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x5645fa808e30]
#> 
#> [[13]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x5645fa88c638]
#> 
#> [[14]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x5645f92fd4a8]
#> 
#> [[15]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

g <- function(cmd) cmd()
f(g(ctxt_stack))
#> [[1]]
#> <frame 28> (25)
#> expr: g(ctxt_stack)
#> env:  [local 0x5645ff5802a0]
#> 
#> [[2]]
#> <frame 27> (26)
#> expr: identity(x)
#> env:  [local 0x5645ff5803b8]
#> 
#> [[3]]
#> <frame 26> (25)
#> expr: f(g(ctxt_stack))
#> env:  [local 0x5645ff5804d0]
#> 
#> [[4]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5645ffd1f118]
#> 
#> [[5]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5645ff580690]
#> 
#> [[6]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x5645ff5808c0]
#> 
#> [[7]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x5645ff580e70]
#> 
#> [[8]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x5645ff57d5b8]
#> 
#> [[9]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x5645ff57d900]
#> 
#> [[10]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x5645ff57dc48]
#> 
#> [[11]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x5645ff57e310]
#> 
#> [[12]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x5645ff57e540]
#> 
#> [[13]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x5645ff57e6c8]
#> 
#> [[14]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x5645ff57e7a8]
#> 
#> [[15]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x5645ff615d80]
#> 
#> [[16]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x5645fa605f70]
#> 
#> [[17]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x5645fa60bec8]
#> 
#> [[18]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x5645fa687e50]
#> 
#> [[19]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x5645fa8458e0]
#> 
#> [[20]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x5645fb8bdf08]
#> 
#> [[21]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x5645fb8bd7d0]
#> 
#> [[22]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x5645fb8e3098]
#> 
#> [[23]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x5645fabb4b70]
#> 
#> [[24]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x5645fac138c0]
#> 
#> [[25]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x5645fa7fca78]
#> 
#> [[26]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x5645fa808e30]
#> 
#> [[27]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x5645fa88c638]
#> 
#> [[28]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x5645f92fd4a8]
#> 
#> [[29]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     
f(g(call_stack))
#> [[1]]
#> <frame 28> (24)
#> expr: g(call_stack)
#> env:  [local 0x5645ffc7c7f0]
#> 
#> [[2]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5645ffd1f118]
#> 
#> [[3]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x5646017deec8]
#> 
#> [[4]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x5645fa605f70]
#> 
#> [[5]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x5645fa60bec8]
#> 
#> [[6]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x5645fa687e50]
#> 
#> [[7]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x5645fa8458e0]
#> 
#> [[8]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x5645fb8bdf08]
#> 
#> [[9]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x5645fb8e3098]
#> 
#> [[10]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x5645fabb4b70]
#> 
#> [[11]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x5645fac138c0]
#> 
#> [[12]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x5645fa7fca78]
#> 
#> [[13]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x5645fa808e30]
#> 
#> [[14]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x5645fa88c638]
#> 
#> [[15]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x5645f92fd4a8]
#> 
#> [[16]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

# The rlang _stack() functions return a list of frame
# objects. Use purrr::transpose() or index a field with
# purrr::map()'s to extract a particular field from a stack:

# stack <- f(f(call_stack()))
# purrr::map(stack, "env")
# purrr::transpose(stack)$expr

# current_frame() is an alias for ctxt_frame(1)
fn <- function() list(current = current_frame(), first = ctxt_frame(1))
fn()
#> Warning: `current_frame()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.
#> $current
#> <frame 26> (25)
#> expr: fn()
#> env:  [local 0x5646001e7380]
#> 
#> $first
#> <frame 26> (25)
#> expr: fn()
#> env:  [local 0x5646001e7380]
#> 

# While current_frame() is the top of the stack, global_frame() is
# the bottom:
fn <- function() {
  n <- ctxt_depth()
  ctxt_frame(n)
}
identical(fn(), global_frame())
#> [1] TRUE

# ctxt_stack() returns a stack with all intervening frames. You can
# trim layers of intervening frames with the trim argument:
identity(identity(ctxt_stack()))
#> [[1]]
#> <frame 27> (25)
#> expr: identity(ctxt_stack())
#> env:  [local 0x5645ffca4870]
#> 
#> [[2]]
#> <frame 26> (25)
#> expr: identity(identity(ctxt_stack()))
#> env:  [local 0x5645ffca4988]
#> 
#> [[3]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5645ffd1f118]
#> 
#> [[4]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5645ffca4b48]
#> 
#> [[5]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x5645ffca4d78]
#> 
#> [[6]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x5645ffca5328]
#> 
#> [[7]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x5645ffca58a0]
#> 
#> [[8]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x5645ffca5be8]
#> 
#> [[9]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x5645ffca5f30]
#> 
#> [[10]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x5645ffca6630]
#> 
#> [[11]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x5645ffca2a30]
#> 
#> [[12]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x5645ffca2bb8]
#> 
#> [[13]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x5645ffca2c98]
#> 
#> [[14]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x564600d78ee0]
#> 
#> [[15]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x5645fa605f70]
#> 
#> [[16]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x5645fa60bec8]
#> 
#> [[17]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x5645fa687e50]
#> 
#> [[18]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x5645fa8458e0]
#> 
#> [[19]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x5645fb8bdf08]
#> 
#> [[20]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x5645fb8bd7d0]
#> 
#> [[21]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x5645fb8e3098]
#> 
#> [[22]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x5645fabb4b70]
#> 
#> [[23]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x5645fac138c0]
#> 
#> [[24]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x5645fa7fca78]
#> 
#> [[25]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x5645fa808e30]
#> 
#> [[26]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x5645fa88c638]
#> 
#> [[27]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x5645f92fd4a8]
#> 
#> [[28]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     
identity(identity(ctxt_stack(trim = 1)))
#> [[1]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5645ffd1f118]
#> 
#> [[2]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x564600980958]
#> 
#> [[3]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x564600980b88]
#> 
#> [[4]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x564600981138]
#> 
#> [[5]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x5646009816b0]
#> 
#> [[6]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x56460097dbc8]
#> 
#> [[7]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x56460097df10]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x56460097e5d8]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x56460097e808]
#> 
#> [[10]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x56460097e990]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x56460097ea70]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x564600a16400]
#> 
#> [[13]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x5645fa605f70]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x5645fa60bec8]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x5645fa687e50]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x5645fa8458e0]
#> 
#> [[17]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x5645fb8bdf08]
#> 
#> [[18]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x5645fb8bd7d0]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x5645fb8e3098]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x5645fabb4b70]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x5645fac138c0]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x5645fa7fca78]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x5645fa808e30]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x5645fa88c638]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x5645f92fd4a8]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# ctxt_stack() is called within fn() with intervening frames:
fn <- function(trim) identity(identity(ctxt_stack(trim = trim)))
fn(0)
#> [[1]]
#> <frame 28> (26)
#> expr: identity(ctxt_stack(trim = trim))
#> env:  [local 0x5645ffd0a768]
#> 
#> [[2]]
#> <frame 27> (26)
#> expr: identity(identity(ctxt_stack(trim = trim)))
#> env:  [local 0x5645ffd0a880]
#> 
#> [[3]]
#> <frame 26> (25)
#> expr: fn(0)
#> env:  [local 0x5645ffd0a998]
#> 
#> [[4]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5645ffd1f118]
#> 
#> [[5]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5645ffd0ab58]
#> 
#> [[6]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x5645ffd0ad88]
#> 
#> [[7]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x5645ffd0b338]
#> 
#> [[8]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x5645ffd0b8b0]
#> 
#> [[9]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x56460117dba8]
#> 
#> [[10]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x56460117def0]
#> 
#> [[11]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x56460117e5b8]
#> 
#> [[12]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x56460117e7e8]
#> 
#> [[13]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x56460117e970]
#> 
#> [[14]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x56460117ea50]
#> 
#> [[15]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x564600e491e8]
#> 
#> [[16]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x5645fa605f70]
#> 
#> [[17]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x5645fa60bec8]
#> 
#> [[18]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x5645fa687e50]
#> 
#> [[19]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x5645fa8458e0]
#> 
#> [[20]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x5645fb8bdf08]
#> 
#> [[21]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x5645fb8bd7d0]
#> 
#> [[22]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x5645fb8e3098]
#> 
#> [[23]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x5645fabb4b70]
#> 
#> [[24]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x5645fac138c0]
#> 
#> [[25]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x5645fa7fca78]
#> 
#> [[26]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x5645fa808e30]
#> 
#> [[27]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x5645fa88c638]
#> 
#> [[28]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x5645f92fd4a8]
#> 
#> [[29]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# We can trim the first layer of those:
fn(1)
#> [[1]]
#> <frame 26> (25)
#> expr: fn(1)
#> env:  [local 0x5645ffd43118]
#> 
#> [[2]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5645ffd1f118]
#> 
#> [[3]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5645ffd432d8]
#> 
#> [[4]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x5645ffd43508]
#> 
#> [[5]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x5645ffd43ab8]
#> 
#> [[6]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x5645ffd40200]
#> 
#> [[7]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x5645ffd40580]
#> 
#> [[8]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x5645ffd408c8]
#> 
#> [[9]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x5645ffd40f90]
#> 
#> [[10]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x5645ffd411c0]
#> 
#> [[11]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x5645ffd41348]
#> 
#> [[12]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x5645ffd41428]
#> 
#> [[13]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x5645ff193f88]
#> 
#> [[14]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x5645fa605f70]
#> 
#> [[15]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x5645fa60bec8]
#> 
#> [[16]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x5645fa687e50]
#> 
#> [[17]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x5645fa8458e0]
#> 
#> [[18]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x5645fb8bdf08]
#> 
#> [[19]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x5645fb8bd7d0]
#> 
#> [[20]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x5645fb8e3098]
#> 
#> [[21]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x5645fabb4b70]
#> 
#> [[22]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x5645fac138c0]
#> 
#> [[23]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x5645fa7fca78]
#> 
#> [[24]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x5645fa808e30]
#> 
#> [[25]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x5645fa88c638]
#> 
#> [[26]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x5645f92fd4a8]
#> 
#> [[27]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# The outside intervening frames (at the fn() call site) are still
# returned, but can be trimmed as well:
identity(identity(fn(1)))
#> [[1]]
#> <frame 28> (25)
#> expr: fn(1)
#> env:  [local 0x564600e5e848]
#> 
#> [[2]]
#> <frame 27> (25)
#> expr: identity(fn(1))
#> env:  [local 0x564600e5e960]
#> 
#> [[3]]
#> <frame 26> (25)
#> expr: identity(identity(fn(1)))
#> env:  [local 0x564600e5ea78]
#> 
#> [[4]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5645ffd1f118]
#> 
#> [[5]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x564600e5ec38]
#> 
#> [[6]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x564600e5ee68]
#> 
#> [[7]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x564600e5f418]
#> 
#> [[8]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x564600e5f990]
#> 
#> [[9]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x564600e5bea8]
#> 
#> [[10]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x564600e5c260]
#> 
#> [[11]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x564600e5c998]
#> 
#> [[12]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x564600e5cbc8]
#> 
#> [[13]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x564600e5cd50]
#> 
#> [[14]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x564600e5ce30]
#> 
#> [[15]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x5646017ecb48]
#> 
#> [[16]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x5645fa605f70]
#> 
#> [[17]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x5645fa60bec8]
#> 
#> [[18]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x5645fa687e50]
#> 
#> [[19]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x5645fa8458e0]
#> 
#> [[20]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x5645fb8bdf08]
#> 
#> [[21]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x5645fb8bd7d0]
#> 
#> [[22]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x5645fb8e3098]
#> 
#> [[23]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x5645fabb4b70]
#> 
#> [[24]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x5645fac138c0]
#> 
#> [[25]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x5645fa7fca78]
#> 
#> [[26]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x5645fa808e30]
#> 
#> [[27]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x5645fa88c638]
#> 
#> [[28]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x5645f92fd4a8]
#> 
#> [[29]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     
identity(identity(fn(2)))
#> [[1]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5645ffd1f118]
#> 
#> [[2]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5645ff7a1210]
#> 
#> [[3]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x5645ff7a0fe0]
#> 
#> [[4]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x5645ff7a0a30]
#> 
#> [[5]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x5645ff7a0480]
#> 
#> [[6]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x5645ff75dee8]
#> 
#> [[7]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x5645ff75dba0]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x5645ff75d4d8]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x5645ff75d2a8]
#> 
#> [[10]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x5645ff75d120]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x5645ff75d040]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x5645ff59a818]
#> 
#> [[13]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x5645fa605f70]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x5645fa60bec8]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x5645fa687e50]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x5645fa8458e0]
#> 
#> [[17]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x5645fb8bdf08]
#> 
#> [[18]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x5645fb8bd7d0]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x5645fb8e3098]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x5645fabb4b70]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x5645fac138c0]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x5645fa7fca78]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x5645fa808e30]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x5645fa88c638]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x5645f92fd4a8]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g <- function(trim) identity(identity(fn(trim)))
g(2)
#> [[1]]
#> <frame 26> (25)
#> expr: g(2)
#> env:  [local 0x5645fd933d90]
#> 
#> [[2]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5645ffd1f118]
#> 
#> [[3]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5645fd933b98]
#> 
#> [[4]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x5645fd937798]
#> 
#> [[5]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x5645fd937178]
#> 
#> [[6]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x5645fd936bc8]
#> 
#> [[7]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x5645fd936880]
#> 
#> [[8]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x5645fd936538]
#> 
#> [[9]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x5645fd935e38]
#> 
#> [[10]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x5645fd935c08]
#> 
#> [[11]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x5645fd935a80]
#> 
#> [[12]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x5645fd939798]
#> 
#> [[13]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x5645fd9546c8]
#> 
#> [[14]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x5645fa605f70]
#> 
#> [[15]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x5645fa60bec8]
#> 
#> [[16]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x5645fa687e50]
#> 
#> [[17]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x5645fa8458e0]
#> 
#> [[18]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x5645fb8bdf08]
#> 
#> [[19]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x5645fb8bd7d0]
#> 
#> [[20]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x5645fb8e3098]
#> 
#> [[21]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x5645fabb4b70]
#> 
#> [[22]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x5645fac138c0]
#> 
#> [[23]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x5645fa7fca78]
#> 
#> [[24]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x5645fa808e30]
#> 
#> [[25]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x5645fa88c638]
#> 
#> [[26]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x5645f92fd4a8]
#> 
#> [[27]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     
g(3)
#> [[1]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5645ffd1f118]
#> 
#> [[2]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5645fd59cc58]
#> 
#> [[3]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x5645fd59c9b8]
#> 
#> [[4]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x5645fd59c398]
#> 
#> [[5]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x5645fd59fc50]
#> 
#> [[6]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x5645fd59f8d0]
#> 
#> [[7]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x5645fd59f470]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x5645fd59ed70]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x5645fd59eb08]
#> 
#> [[10]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x5645fd59e948]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x5645fd59e868]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x5645fd5b4ec0]
#> 
#> [[13]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x5645fa605f70]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x5645fa60bec8]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x5645fa687e50]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x5645fa8458e0]
#> 
#> [[17]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x5645fb8bdf08]
#> 
#> [[18]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x5645fb8bd7d0]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x5645fb8e3098]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x5645fabb4b70]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x5645fac138c0]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x5645fa7fca78]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x5645fa808e30]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x5645fa88c638]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x5645f92fd4a8]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"