Call stack information

The eval_ and call_ families of functions provide a replacement for the base R functions prefixed with sys. (which are all about the context stack), as well as for parent.frame() (which is the only base R function for querying the call stack). The context stack includes all R-level evaluation contexts. It is linear in terms of execution history but due to lazy evaluation it is potentially nonlinear in terms of call history. The call stack history, on the other hand, is homogenous.

Usage

global_frame()

current_frame()

ctxt_frame(n = 1)

call_frame(n = 1, clean = TRUE)

ctxt_depth()

call_depth()

ctxt_stack(n = NULL, trim = 0)

call_stack(n = NULL, clean = TRUE)

Arguments

n

The number of frames to go back in the stack.

clean

Whether to post-process the call stack to clean non-standard frames. If TRUE, suboptimal call-stack entries by base::eval() will be cleaned up: the duplicate frame created by eval() is eliminated.

trim

The number of layers of intervening frames to trim off the stack. See stack_trim() and examples.

Details

ctxt_frame() and call_frame() return a frame object containing the following fields: expr and env (call expression and evaluation environment), pos and caller_pos (position of current frame in the context stack and position of the caller), and fun (function of the current frame). ctxt_stack() and call_stack() return a list of all context or call frames on the stack. Finally, ctxt_depth() and call_depth() report the current context position or the number of calling frames on the stack.

The base R functions take two sorts of arguments to indicate which frame to query: which and n. The n argument is straightforward: it's the number of frames to go down the stack, with n = 1 referring to the current context. The which argument is more complicated and changes meaning for values lower than 1. For the sake of consistency, the rlang functions all take the same kind of argument n. This argument has a single meaning (the number of frames to go down the stack) and cannot be lower than 1.

Note finally that parent.frame(1) corresponds to call_frame(2)$env, as n = 1 always refers to the current frame. This makes the _frame() and _stack() functions consistent: ctxt_frame(2) is the same as ctxt_stack()[[2]]. Also, ctxt_depth() returns one more frame than base::sys.nframe() because it counts the global frame. That is consistent with the _stack() functions which return the global frame as well. This way, call_stack(call_depth()) is the same as global_frame().

Life cycle

These functions are soft-deprecated and replaced by trace_back().

Examples

# Expressions within arguments count as contexts
identity(identity(ctxt_depth())) # returns 2
#> Warning: `ctxt_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.
#> [1] 28

# But they are not part of the call stack because arguments are
# evaluated within the calling function (or the global environment
# if called at top level)
identity(identity(call_depth())) # returns 0
#> Warning: `call_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.
#> [1] 15

# The context stacks includes all intervening execution frames. The
# call stack doesn't:
f <- function(x) identity(x)
f(f(ctxt_stack()))
#> [[1]]
#> <frame 29> (28)
#> expr: identity(x)
#> env:  [local 0x55c5dc285bc0]
#> 
#> [[2]]
#> <frame 28> (25)
#> expr: f(ctxt_stack())
#> env:  [local 0x55c5dc285aa8]
#> 
#> [[3]]
#> <frame 27> (26)
#> expr: identity(x)
#> env:  [local 0x55c5dc2897c0]
#> 
#> [[4]]
#> <frame 26> (25)
#> expr: f(f(ctxt_stack()))
#> env:  [local 0x55c5dc2896a8]
#> 
#> [[5]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x55c5dc5776d8]
#> 
#> [[6]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x55c5dc289478]
#> 
#> [[7]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x55c5dc289210]
#> 
#> [[8]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x55c5dc288b10]
#> 
#> [[9]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x55c5dc2884f0]
#> 
#> [[10]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x55c5dc288138]
#> 
#> [[11]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x55c5dc287b50]
#> 
#> [[12]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x55c5dc28b210]
#> 
#> [[13]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x55c5dc28afe0]
#> 
#> [[14]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x55c5dc28adb0]
#> 
#> [[15]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x55c5dc28acd0]
#> 
#> [[16]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x55c5dc2a12c8]
#> 
#> [[17]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x55c5dccc25e8]
#> 
#> [[18]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x55c5dccc11c8]
#> 
#> [[19]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x55c5dcce2710]
#> 
#> [[20]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x55c5ddf70928]
#> 
#> [[21]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x55c5dfcf2138]
#> 
#> [[22]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x55c5dfcf1a38]
#> 
#> [[23]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x55c5dfc0ead0]
#> 
#> [[24]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x55c5db4b3cf0]
#> 
#> [[25]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x55c5d9bee980]
#> 
#> [[26]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x55c5db6dd428]
#> 
#> [[27]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x55c5db6d9268]
#> 
#> [[28]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x55c5d9abf6e8]
#> 
#> [[29]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x55c5d84dde10]
#> 
#> [[30]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     
f(f(call_stack()))
#> Warning: `call_stack()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.
#> [[1]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x55c5dc5776d8]
#> 
#> [[2]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x55c5dbc68598]
#> 
#> [[3]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x55c5dccc25e8]
#> 
#> [[4]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x55c5dccc11c8]
#> 
#> [[5]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x55c5dcce2710]
#> 
#> [[6]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x55c5ddf70928]
#> 
#> [[7]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x55c5dfcf2138]
#> 
#> [[8]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x55c5dfc0ead0]
#> 
#> [[9]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x55c5db4b3cf0]
#> 
#> [[10]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x55c5d9bee980]
#> 
#> [[11]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x55c5db6dd428]
#> 
#> [[12]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x55c5db6d9268]
#> 
#> [[13]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x55c5d9abf6e8]
#> 
#> [[14]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x55c5d84dde10]
#> 
#> [[15]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

g <- function(cmd) cmd()
f(g(ctxt_stack))
#> [[1]]
#> <frame 28> (25)
#> expr: g(ctxt_stack)
#> env:  [local 0x55c5db629900]
#> 
#> [[2]]
#> <frame 27> (26)
#> expr: identity(x)
#> env:  [local 0x55c5db6297b0]
#> 
#> [[3]]
#> <frame 26> (25)
#> expr: f(g(ctxt_stack))
#> env:  [local 0x55c5db629660]
#> 
#> [[4]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x55c5dc5776d8]
#> 
#> [[5]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x55c5db6294a0]
#> 
#> [[6]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x55c5db629270]
#> 
#> [[7]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x55c5db628c50]
#> 
#> [[8]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x55c5db628630]
#> 
#> [[9]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x55c5db6282e8]
#> 
#> [[10]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x55c5db627f68]
#> 
#> [[11]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x55c5db62f4a0]
#> 
#> [[12]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x55c5db62f270]
#> 
#> [[13]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x55c5db62f0e8]
#> 
#> [[14]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x55c5db62efd0]
#> 
#> [[15]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x55c5db6508d0]
#> 
#> [[16]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x55c5dccc25e8]
#> 
#> [[17]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x55c5dccc11c8]
#> 
#> [[18]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x55c5dcce2710]
#> 
#> [[19]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x55c5ddf70928]
#> 
#> [[20]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x55c5dfcf2138]
#> 
#> [[21]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x55c5dfcf1a38]
#> 
#> [[22]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x55c5dfc0ead0]
#> 
#> [[23]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x55c5db4b3cf0]
#> 
#> [[24]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x55c5d9bee980]
#> 
#> [[25]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x55c5db6dd428]
#> 
#> [[26]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x55c5db6d9268]
#> 
#> [[27]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x55c5d9abf6e8]
#> 
#> [[28]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x55c5d84dde10]
#> 
#> [[29]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     
f(g(call_stack))
#> [[1]]
#> <frame 28> (24)
#> expr: g(call_stack)
#> env:  [local 0x55c5daf97520]
#> 
#> [[2]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x55c5dc5776d8]
#> 
#> [[3]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x55c5dafac480]
#> 
#> [[4]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x55c5dccc25e8]
#> 
#> [[5]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x55c5dccc11c8]
#> 
#> [[6]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x55c5dcce2710]
#> 
#> [[7]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x55c5ddf70928]
#> 
#> [[8]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x55c5dfcf2138]
#> 
#> [[9]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x55c5dfc0ead0]
#> 
#> [[10]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x55c5db4b3cf0]
#> 
#> [[11]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x55c5d9bee980]
#> 
#> [[12]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x55c5db6dd428]
#> 
#> [[13]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x55c5db6d9268]
#> 
#> [[14]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x55c5d9abf6e8]
#> 
#> [[15]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x55c5d84dde10]
#> 
#> [[16]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

# The rlang _stack() functions return a list of frame
# objects. Use purrr::transpose() or index a field with
# purrr::map()'s to extract a particular field from a stack:

# stack <- f(f(call_stack()))
# purrr::map(stack, "env")
# purrr::transpose(stack)$expr

# current_frame() is an alias for ctxt_frame(1)
fn <- function() list(current = current_frame(), first = ctxt_frame(1))
fn()
#> Warning: `current_frame()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.
#> $current
#> <frame 26> (25)
#> expr: fn()
#> env:  [local 0x55c5d9fa8768]
#> 
#> $first
#> <frame 26> (25)
#> expr: fn()
#> env:  [local 0x55c5d9fa8768]
#> 

# While current_frame() is the top of the stack, global_frame() is
# the bottom:
fn <- function() {
  n <- ctxt_depth()
  ctxt_frame(n)
}
identical(fn(), global_frame())
#> [1] TRUE

# ctxt_stack() returns a stack with all intervening frames. You can
# trim layers of intervening frames with the trim argument:
identity(identity(ctxt_stack()))
#> [[1]]
#> <frame 27> (25)
#> expr: identity(ctxt_stack())
#> env:  [local 0x55c5d9b29548]
#> 
#> [[2]]
#> <frame 26> (25)
#> expr: identity(identity(ctxt_stack()))
#> env:  [local 0x55c5d9b29430]
#> 
#> [[3]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x55c5dc5776d8]
#> 
#> [[4]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x55c5d9b291c8]
#> 
#> [[5]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x55c5d9b28dd8]
#> 
#> [[6]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x55c5d9b2c428]
#> 
#> [[7]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x55c5d9b2e0a8]
#> 
#> [[8]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x55c5d9b2dcb8]
#> 
#> [[9]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x55c5d9b2d900]
#> 
#> [[10]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x55c5d9b2cf98]
#> 
#> [[11]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x55c5d9b2ccf8]
#> 
#> [[12]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x55c5d9b2cb70]
#> 
#> [[13]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x55c5d9b2ca20]
#> 
#> [[14]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x55c5d9b4d3a0]
#> 
#> [[15]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x55c5dccc25e8]
#> 
#> [[16]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x55c5dccc11c8]
#> 
#> [[17]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x55c5dcce2710]
#> 
#> [[18]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x55c5ddf70928]
#> 
#> [[19]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x55c5dfcf2138]
#> 
#> [[20]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x55c5dfcf1a38]
#> 
#> [[21]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x55c5dfc0ead0]
#> 
#> [[22]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x55c5db4b3cf0]
#> 
#> [[23]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x55c5d9bee980]
#> 
#> [[24]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x55c5db6dd428]
#> 
#> [[25]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x55c5db6d9268]
#> 
#> [[26]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x55c5d9abf6e8]
#> 
#> [[27]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x55c5d84dde10]
#> 
#> [[28]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     
identity(identity(ctxt_stack(trim = 1)))
#> [[1]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x55c5dc5776d8]
#> 
#> [[2]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x55c5d812b8c0]
#> 
#> [[3]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x55c5d812b0e0]
#> 
#> [[4]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x55c5d8130b68]
#> 
#> [[5]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x55c5d8136d10]
#> 
#> [[6]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x55c5d8136990]
#> 
#> [[7]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x55c5d8136568]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x55c5d8135a78]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x55c5d813bb10]
#> 
#> [[10]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x55c5d813b950]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x55c5d813b870]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x55c5d82286d0]
#> 
#> [[13]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x55c5dccc25e8]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x55c5dccc11c8]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x55c5dcce2710]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x55c5ddf70928]
#> 
#> [[17]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x55c5dfcf2138]
#> 
#> [[18]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x55c5dfcf1a38]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x55c5dfc0ead0]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x55c5db4b3cf0]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x55c5d9bee980]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x55c5db6dd428]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x55c5db6d9268]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x55c5d9abf6e8]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x55c5d84dde10]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# ctxt_stack() is called within fn() with intervening frames:
fn <- function(trim) identity(identity(ctxt_stack(trim = trim)))
fn(0)
#> [[1]]
#> <frame 28> (26)
#> expr: identity(ctxt_stack(trim = trim))
#> env:  [local 0x55c5df652b68]
#> 
#> [[2]]
#> <frame 27> (26)
#> expr: identity(identity(ctxt_stack(trim = trim)))
#> env:  [local 0x55c5df652c80]
#> 
#> [[3]]
#> <frame 26> (25)
#> expr: fn(0)
#> env:  [local 0x55c5df652d98]
#> 
#> [[4]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x55c5dc5776d8]
#> 
#> [[5]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x55c5df652f58]
#> 
#> [[6]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x55c5df653188]
#> 
#> [[7]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x55c5de8730a8]
#> 
#> [[8]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x55c5de873620]
#> 
#> [[9]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x55c5de873968]
#> 
#> [[10]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x55c5de873cb0]
#> 
#> [[11]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x55c5de874378]
#> 
#> [[12]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x55c5de8745a8]
#> 
#> [[13]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x55c5de874730]
#> 
#> [[14]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x55c5de874810]
#> 
#> [[15]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x55c5de801840]
#> 
#> [[16]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x55c5dccc25e8]
#> 
#> [[17]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x55c5dccc11c8]
#> 
#> [[18]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x55c5dcce2710]
#> 
#> [[19]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x55c5ddf70928]
#> 
#> [[20]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x55c5dfcf2138]
#> 
#> [[21]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x55c5dfcf1a38]
#> 
#> [[22]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x55c5dfc0ead0]
#> 
#> [[23]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x55c5db4b3cf0]
#> 
#> [[24]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x55c5d9bee980]
#> 
#> [[25]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x55c5db6dd428]
#> 
#> [[26]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x55c5db6d9268]
#> 
#> [[27]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x55c5d9abf6e8]
#> 
#> [[28]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x55c5d84dde10]
#> 
#> [[29]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# We can trim the first layer of those:
fn(1)
#> [[1]]
#> <frame 26> (25)
#> expr: fn(1)
#> env:  [local 0x55c5dff52080]
#> 
#> [[2]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x55c5dc5776d8]
#> 
#> [[3]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x55c5dff52240]
#> 
#> [[4]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x55c5dff52470]
#> 
#> [[5]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x55c5dff52a20]
#> 
#> [[6]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x55c5dff52f98]
#> 
#> [[7]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x55c5deee0930]
#> 
#> [[8]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x55c5deee0c78]
#> 
#> [[9]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x55c5deee1340]
#> 
#> [[10]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x55c5deee1570]
#> 
#> [[11]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x55c5deee16f8]
#> 
#> [[12]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x55c5deee17d8]
#> 
#> [[13]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x55c5de62a260]
#> 
#> [[14]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x55c5dccc25e8]
#> 
#> [[15]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x55c5dccc11c8]
#> 
#> [[16]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x55c5dcce2710]
#> 
#> [[17]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x55c5ddf70928]
#> 
#> [[18]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x55c5dfcf2138]
#> 
#> [[19]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x55c5dfcf1a38]
#> 
#> [[20]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x55c5dfc0ead0]
#> 
#> [[21]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x55c5db4b3cf0]
#> 
#> [[22]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x55c5d9bee980]
#> 
#> [[23]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x55c5db6dd428]
#> 
#> [[24]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x55c5db6d9268]
#> 
#> [[25]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x55c5d9abf6e8]
#> 
#> [[26]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x55c5d84dde10]
#> 
#> [[27]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# The outside intervening frames (at the fn() call site) are still
# returned, but can be trimmed as well:
identity(identity(fn(1)))
#> [[1]]
#> <frame 28> (25)
#> expr: fn(1)
#> env:  [local 0x55c5e130ab10]
#> 
#> [[2]]
#> <frame 27> (25)
#> expr: identity(fn(1))
#> env:  [local 0x55c5e130ac28]
#> 
#> [[3]]
#> <frame 26> (25)
#> expr: identity(identity(fn(1)))
#> env:  [local 0x55c5e130ad40]
#> 
#> [[4]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x55c5dc5776d8]
#> 
#> [[5]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x55c5e130af00]
#> 
#> [[6]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x55c5e130b130]
#> 
#> [[7]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x55c5e130b718]
#> 
#> [[8]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x55c5e130bc90]
#> 
#> [[9]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x55c5e130bfd8]
#> 
#> [[10]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x55c5ddc1d350]
#> 
#> [[11]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x55c5ddc1dc80]
#> 
#> [[12]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x55c5ddc1deb0]
#> 
#> [[13]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x55c5ddc1e038]
#> 
#> [[14]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x55c5ddc1e118]
#> 
#> [[15]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x55c5e11d67b0]
#> 
#> [[16]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x55c5dccc25e8]
#> 
#> [[17]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x55c5dccc11c8]
#> 
#> [[18]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x55c5dcce2710]
#> 
#> [[19]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x55c5ddf70928]
#> 
#> [[20]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x55c5dfcf2138]
#> 
#> [[21]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x55c5dfcf1a38]
#> 
#> [[22]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x55c5dfc0ead0]
#> 
#> [[23]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x55c5db4b3cf0]
#> 
#> [[24]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x55c5d9bee980]
#> 
#> [[25]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x55c5db6dd428]
#> 
#> [[26]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x55c5db6d9268]
#> 
#> [[27]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x55c5d9abf6e8]
#> 
#> [[28]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x55c5d84dde10]
#> 
#> [[29]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     
identity(identity(fn(2)))
#> [[1]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x55c5dc5776d8]
#> 
#> [[2]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x55c5de24f740]
#> 
#> [[3]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x55c5de24f970]
#> 
#> [[4]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x55c5de24ff20]
#> 
#> [[5]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x55c5de250498]
#> 
#> [[6]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x55c5de2507e0]
#> 
#> [[7]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x55c5de24ccf8]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x55c5de24d3c0]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x55c5de24d5f0]
#> 
#> [[10]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x55c5de24d778]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x55c5de24d858]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x55c5de40b088]
#> 
#> [[13]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x55c5dccc25e8]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x55c5dccc11c8]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x55c5dcce2710]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x55c5ddf70928]
#> 
#> [[17]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x55c5dfcf2138]
#> 
#> [[18]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x55c5dfcf1a38]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x55c5dfc0ead0]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x55c5db4b3cf0]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x55c5d9bee980]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x55c5db6dd428]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x55c5db6d9268]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x55c5d9abf6e8]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x55c5d84dde10]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g <- function(trim) identity(identity(fn(trim)))
g(2)
#> [[1]]
#> <frame 26> (25)
#> expr: g(2)
#> env:  [local 0x55c5df7f5808]
#> 
#> [[2]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x55c5dc5776d8]
#> 
#> [[3]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x55c5df7f59c8]
#> 
#> [[4]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x55c5df7f5bf8]
#> 
#> [[5]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x55c5df7f61a8]
#> 
#> [[6]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x55c5df7f28f0]
#> 
#> [[7]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x55c5df7f2c38]
#> 
#> [[8]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x55c5df7f2fb8]
#> 
#> [[9]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x55c5df7f3680]
#> 
#> [[10]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x55c5df7f38b0]
#> 
#> [[11]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x55c5df7f3a38]
#> 
#> [[12]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x55c5df7f3b18]
#> 
#> [[13]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x55c5dfb64368]
#> 
#> [[14]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x55c5dccc25e8]
#> 
#> [[15]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x55c5dccc11c8]
#> 
#> [[16]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x55c5dcce2710]
#> 
#> [[17]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x55c5ddf70928]
#> 
#> [[18]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x55c5dfcf2138]
#> 
#> [[19]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x55c5dfcf1a38]
#> 
#> [[20]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x55c5dfc0ead0]
#> 
#> [[21]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x55c5db4b3cf0]
#> 
#> [[22]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x55c5d9bee980]
#> 
#> [[23]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x55c5db6dd428]
#> 
#> [[24]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x55c5db6d9268]
#> 
#> [[25]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x55c5d9abf6e8]
#> 
#> [[26]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x55c5d84dde10]
#> 
#> [[27]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     
g(3)
#> [[1]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x55c5dc5776d8]
#> 
#> [[2]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x55c5dfca1560]
#> 
#> [[3]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x55c5dfca1790]
#> 
#> [[4]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x55c5dfc9df10]
#> 
#> [[5]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x55c5dfc9e488]
#> 
#> [[6]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x55c5dfc9e7d0]
#> 
#> [[7]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x55c5dfc9eb18]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x55c5dfc9f1e0]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x55c5dfc9f410]
#> 
#> [[10]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x55c5dfc9f598]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x55c5dfc9f678]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x55c5df09d838]
#> 
#> [[13]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x55c5dccc25e8]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x55c5dccc11c8]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x55c5dcce2710]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x55c5ddf70928]
#> 
#> [[17]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x55c5dfcf2138]
#> 
#> [[18]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x55c5dfcf1a38]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x55c5dfc0ead0]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x55c5db4b3cf0]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x55c5d9bee980]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x55c5db6dd428]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x55c5db6d9268]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x55c5d9abf6e8]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x55c5d84dde10]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"