Call stack information

The eval_ and call_ families of functions provide a replacement for the base R functions prefixed with sys. (which are all about the context stack), as well as for parent.frame() (which is the only base R function for querying the call stack). The context stack includes all R-level evaluation contexts. It is linear in terms of execution history but due to lazy evaluation it is potentially nonlinear in terms of call history. The call stack history, on the other hand, is homogenous.

global_frame()

current_frame()

ctxt_frame(n = 1)

call_frame(n = 1, clean = TRUE)

ctxt_depth()

call_depth()

ctxt_stack(n = NULL, trim = 0)

call_stack(n = NULL, clean = TRUE)

Arguments

n
The number of frames to go back in the stack.

clean
Whether to post-process the call stack to clean non-standard frames. If TRUE, suboptimal call-stack entries by base::eval() will be cleaned up: the duplicate frame created by eval() is eliminated.

trim
The number of layers of intervening frames to trim off the stack. See stack_trim() and examples.

Details

ctxt_frame() and call_frame() return a frame object containing the following fields: expr and env (call expression and evaluation environment), pos and caller_pos (position of current frame in the context stack and position of the caller), and fun (function of the current frame). ctxt_stack() and call_stack() return a list of all context or call frames on the stack. Finally, ctxt_depth() and call_depth() report the current context position or the number of calling frames on the stack.

The base R functions take two sorts of arguments to indicate which frame to query: which and n. The n argument is straightforward: it's the number of frames to go down the stack, with n = 1 referring to the current context. The which argument is more complicated and changes meaning for values lower than 1. For the sake of consistency, the rlang functions all take the same kind of argument n. This argument has a single meaning (the number of frames to go down the stack) and cannot be lower than 1.

Note finally that parent.frame(1) corresponds to call_frame(2)$env, as n = 1 always refers to the current frame. This makes the _frame() and _stack() functions consistent: ctxt_frame(2) is the same as ctxt_stack()[[2]]. Also, ctxt_depth() returns one more frame than base::sys.nframe() because it counts the global frame. That is consistent with the _stack() functions which return the global frame as well. This way, call_stack(call_depth()) is the same as global_frame().

Life cycle

These functions are soft-deprecated and replaced by trace_back().

Examples

# Expressions within arguments count as contexts
identity(identity(ctxt_depth())) # returns 2

#> Warning: `ctxt_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 25

# But they are not part of the call stack because arguments are
# evaluated within the calling function (or the global environment
# if called at top level)
identity(identity(call_depth())) # returns 0

#> Warning: `call_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 13

# The context stacks includes all intervening execution frames. The
# call stack doesn't:
f <- function(x) identity(x)
f(f(ctxt_stack()))

#> [[1]]
#> <frame 26> (25)
#> expr: identity(x)
#> env:  [local 0x7ff31e7d8630]
#> 
#> [[2]]
#> <frame 25> (22)
#> expr: f(ctxt_stack())
#> env:  [local 0x7ff31e7d8748]
#> 
#> [[3]]
#> <frame 24> (23)
#> expr: identity(x)
#> env:  [local 0x7ff31e7d8860]
#> 
#> [[4]]
#> <frame 23> (22)
#> expr: f(f(ctxt_stack()))
#> env:  [local 0x7ff31e7d8978]
#> 
#> [[5]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff31e1abc38]
#> 
#> [[6]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff31e7d8b38]
#> 
#> [[7]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7ff31e7d8d68]
#> 
#> [[8]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7ff31e7d9318]
#> 
#> [[9]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7ff31e7d98c8]
#> 
#> [[10]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7ff31e7d9c10]
#> 
#> [[11]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7ff31e7d9f58]
#> 
#> [[12]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7ff31e7d6710]
#> 
#> [[13]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7ff31e7d6940]
#> 
#> [[14]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7ff31e7d6ac8]
#> 
#> [[15]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7ff31e7d6ba8]
#> 
#> [[16]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7ff31e7a6e58]
#> 
#> [[17]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7ff3152e0bc8]
#> 
#> [[18]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7ff31530fd20]
#> 
#> [[19]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7ff3143167e8]
#> 
#> [[20]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7ff3184e5220]
#> 
#> [[21]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7ff31859d4c0]
#> 
#> [[22]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7ff3186ab8b0]
#> 
#> [[23]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff31a855f90]
#> 
#> [[24]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff31a872c18]
#> 
#> [[25]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7ff3161e47a0]
#> 
#> [[26]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7ff318135d98]
#> 
#> [[27]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(f(call_stack()))

#> Warning: `call_stack()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [[1]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff31e1abc38]
#> 
#> [[2]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7ff31e451318]
#> 
#> [[3]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7ff3152e0bc8]
#> 
#> [[4]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7ff31530fd20]
#> 
#> [[5]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7ff3143167e8]
#> 
#> [[6]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7ff3184e5220]
#> 
#> [[7]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7ff31859d4c0]
#> 
#> [[8]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7ff3186ab8b0]
#> 
#> [[9]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff31a855f90]
#> 
#> [[10]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff31a872c18]
#> 
#> [[11]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7ff3161e47a0]
#> 
#> [[12]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7ff318135d98]
#> 
#> [[13]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

g <- function(cmd) cmd()
f(g(ctxt_stack))

#> [[1]]
#> <frame 25> (22)
#> expr: g(ctxt_stack)
#> env:  [local 0x7ff31c84ed80]
#> 
#> [[2]]
#> <frame 24> (23)
#> expr: identity(x)
#> env:  [local 0x7ff31c84ee98]
#> 
#> [[3]]
#> <frame 23> (22)
#> expr: f(g(ctxt_stack))
#> env:  [local 0x7ff31c84efb0]
#> 
#> [[4]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff31e1abc38]
#> 
#> [[5]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff31c84f170]
#> 
#> [[6]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7ff31c84f3a0]
#> 
#> [[7]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7ff31a66f640]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7ff31a66fbb8]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7ff31a66ff00]
#> 
#> [[10]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7ff31a670248]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7ff31a670910]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7ff31a670b40]
#> 
#> [[13]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7ff31a670cc8]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7ff31a670da8]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7ff31a398c48]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7ff3152e0bc8]
#> 
#> [[17]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7ff31530fd20]
#> 
#> [[18]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7ff3143167e8]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7ff3184e5220]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7ff31859d4c0]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7ff3186ab8b0]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff31a855f90]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff31a872c18]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7ff3161e47a0]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7ff318135d98]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(g(call_stack))

#> [[1]]
#> <frame 25> (21)
#> expr: g(call_stack)
#> env:  [local 0x7ff31bfdbb88]
#> 
#> [[2]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff31e1abc38]
#> 
#> [[3]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7ff31be0b138]
#> 
#> [[4]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7ff3152e0bc8]
#> 
#> [[5]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7ff31530fd20]
#> 
#> [[6]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7ff3143167e8]
#> 
#> [[7]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7ff3184e5220]
#> 
#> [[8]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7ff31859d4c0]
#> 
#> [[9]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7ff3186ab8b0]
#> 
#> [[10]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff31a855f90]
#> 
#> [[11]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff31a872c18]
#> 
#> [[12]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7ff3161e47a0]
#> 
#> [[13]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7ff318135d98]
#> 
#> [[14]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

# The rlang _stack() functions return a list of frame
# objects. Use purrr::transpose() or index a field with
# purrr::map()'s to extract a particular field from a stack:

# stack <- f(f(call_stack()))
# purrr::map(stack, "env")
# purrr::transpose(stack)$expr

# current_frame() is an alias for ctxt_frame(1)
fn <- function() list(current = current_frame(), first = ctxt_frame(1))
fn()

#> Warning: `current_frame()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> $current
#> <frame 23> (22)
#> expr: fn()
#> env:  [local 0x7ff31a52eef8]
#> 
#> $first
#> <frame 23> (22)
#> expr: fn()
#> env:  [local 0x7ff31a52eef8]
#> 

# While current_frame() is the top of the stack, global_frame() is
# the bottom:
fn <- function() {
  n <- ctxt_depth()
  ctxt_frame(n)
}
identical(fn(), global_frame())

#> [1] TRUE

# ctxt_stack() returns a stack with all intervening frames. You can
# trim layers of intervening frames with the trim argument:
identity(identity(ctxt_stack()))

#> [[1]]
#> <frame 24> (22)
#> expr: identity(ctxt_stack())
#> env:  [local 0x7ff31a25f358]
#> 
#> [[2]]
#> <frame 23> (22)
#> expr: identity(identity(ctxt_stack()))
#> env:  [local 0x7ff31a25f240]
#> 
#> [[3]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff31e1abc38]
#> 
#> [[4]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff31a25f080]
#> 
#> [[5]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7ff31a25ee18]
#> 
#> [[6]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7ff31a25e868]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7ff31a2641c8]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7ff31a263e80]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7ff31a263b38]
#> 
#> [[10]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7ff31a263470]
#> 
#> [[11]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7ff31a263240]
#> 
#> [[12]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7ff31a2630b8]
#> 
#> [[13]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7ff31a262fd8]
#> 
#> [[14]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7ff31a27a238]
#> 
#> [[15]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7ff3152e0bc8]
#> 
#> [[16]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7ff31530fd20]
#> 
#> [[17]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7ff3143167e8]
#> 
#> [[18]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7ff3184e5220]
#> 
#> [[19]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7ff31859d4c0]
#> 
#> [[20]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7ff3186ab8b0]
#> 
#> [[21]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff31a855f90]
#> 
#> [[22]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff31a872c18]
#> 
#> [[23]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7ff3161e47a0]
#> 
#> [[24]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7ff318135d98]
#> 
#> [[25]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(ctxt_stack(trim = 1)))

#> [[1]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff31e1abc38]
#> 
#> [[2]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff31ac3ac68]
#> 
#> [[3]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7ff31ac3aa38]
#> 
#> [[4]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7ff31ac40210]
#> 
#> [[5]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7ff31ac3fad8]
#> 
#> [[6]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7ff31ac3f790]
#> 
#> [[7]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7ff31ac3f410]
#> 
#> [[8]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7ff31ac3eca0]
#> 
#> [[9]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7ff31ac3ea70]
#> 
#> [[10]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7ff31ac3e8e8]
#> 
#> [[11]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7ff31ac3e808]
#> 
#> [[12]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7ff31ac54990]
#> 
#> [[13]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7ff3152e0bc8]
#> 
#> [[14]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7ff31530fd20]
#> 
#> [[15]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7ff3143167e8]
#> 
#> [[16]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7ff3184e5220]
#> 
#> [[17]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7ff31859d4c0]
#> 
#> [[18]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7ff3186ab8b0]
#> 
#> [[19]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff31a855f90]
#> 
#> [[20]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff31a872c18]
#> 
#> [[21]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7ff3161e47a0]
#> 
#> [[22]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7ff318135d98]
#> 
#> [[23]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# ctxt_stack() is called within fn() with intervening frames:
fn <- function(trim) identity(identity(ctxt_stack(trim = trim)))
fn(0)

#> [[1]]
#> <frame 25> (23)
#> expr: identity(ctxt_stack(trim = trim))
#> env:  [local 0x7ff31a96f668]
#> 
#> [[2]]
#> <frame 24> (23)
#> expr: identity(identity(ctxt_stack(trim = trim)))
#> env:  [local 0x7ff31a96f550]
#> 
#> [[3]]
#> <frame 23> (22)
#> expr: fn(0)
#> env:  [local 0x7ff31a96f438]
#> 
#> [[4]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff31e1abc38]
#> 
#> [[5]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff31a96f278]
#> 
#> [[6]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7ff31a96f048]
#> 
#> [[7]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7ff31a96ea98]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7ff31a96e4e8]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7ff31a972078]
#> 
#> [[10]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7ff31a971cc0]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7ff31a9713c8]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7ff31a971198]
#> 
#> [[13]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7ff31a971010]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7ff31a970f30]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7ff31a980d38]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7ff3152e0bc8]
#> 
#> [[17]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7ff31530fd20]
#> 
#> [[18]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7ff3143167e8]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7ff3184e5220]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7ff31859d4c0]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7ff3186ab8b0]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff31a855f90]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff31a872c18]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7ff3161e47a0]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7ff318135d98]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# We can trim the first layer of those:
fn(1)

#> [[1]]
#> <frame 23> (22)
#> expr: fn(1)
#> env:  [local 0x7ff315d89b20]
#> 
#> [[2]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff31e1abc38]
#> 
#> [[3]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff31a01faf8]
#> 
#> [[4]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7ff31a01f350]
#> 
#> [[5]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7ff31a020550]
#> 
#> [[6]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7ff31a029a88]
#> 
#> [[7]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7ff31a028518]
#> 
#> [[8]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7ff31a02fc10]
#> 
#> [[9]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7ff31a02e518]
#> 
#> [[10]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7ff31a02e240]
#> 
#> [[11]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7ff31a031fc8]
#> 
#> [[12]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7ff31a031e08]
#> 
#> [[13]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7ff31a0810e8]
#> 
#> [[14]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7ff3152e0bc8]
#> 
#> [[15]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7ff31530fd20]
#> 
#> [[16]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7ff3143167e8]
#> 
#> [[17]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7ff3184e5220]
#> 
#> [[18]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7ff31859d4c0]
#> 
#> [[19]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7ff3186ab8b0]
#> 
#> [[20]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff31a855f90]
#> 
#> [[21]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff31a872c18]
#> 
#> [[22]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7ff3161e47a0]
#> 
#> [[23]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7ff318135d98]
#> 
#> [[24]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# The outside intervening frames (at the fn() call site) are still
# returned, but can be trimmed as well:
identity(identity(fn(1)))

#> [[1]]
#> <frame 25> (22)
#> expr: fn(1)
#> env:  [local 0x7ff319930fa0]
#> 
#> [[2]]
#> <frame 24> (22)
#> expr: identity(fn(1))
#> env:  [local 0x7ff319930e88]
#> 
#> [[3]]
#> <frame 23> (22)
#> expr: identity(identity(fn(1)))
#> env:  [local 0x7ff319930d70]
#> 
#> [[4]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff31e1abc38]
#> 
#> [[5]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff319930bb0]
#> 
#> [[6]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7ff319930948]
#> 
#> [[7]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7ff3199342a8]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7ff319933cf8]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7ff3199339b0]
#> 
#> [[10]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7ff319933668]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7ff319932f68]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7ff319932d38]
#> 
#> [[13]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7ff31993d630]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7ff31993d390]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7ff31997bc20]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7ff3152e0bc8]
#> 
#> [[17]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7ff31530fd20]
#> 
#> [[18]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7ff3143167e8]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7ff3184e5220]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7ff31859d4c0]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7ff3186ab8b0]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff31a855f90]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff31a872c18]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7ff3161e47a0]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7ff318135d98]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(fn(2)))

#> [[1]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff31e1abc38]
#> 
#> [[2]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff318bc0680]
#> 
#> [[3]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7ff318bc0450]
#> 
#> [[4]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7ff318bbfe30]
#> 
#> [[5]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7ff318bc37c8]
#> 
#> [[6]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7ff318bc3480]
#> 
#> [[7]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7ff318bc3138]
#> 
#> [[8]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7ff318bc2a70]
#> 
#> [[9]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7ff318bc2808]
#> 
#> [[10]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7ff318bc2680]
#> 
#> [[11]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7ff318bc2530]
#> 
#> [[12]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7ff318bdb790]
#> 
#> [[13]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7ff3152e0bc8]
#> 
#> [[14]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7ff31530fd20]
#> 
#> [[15]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7ff3143167e8]
#> 
#> [[16]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7ff3184e5220]
#> 
#> [[17]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7ff31859d4c0]
#> 
#> [[18]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7ff3186ab8b0]
#> 
#> [[19]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff31a855f90]
#> 
#> [[20]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff31a872c18]
#> 
#> [[21]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7ff3161e47a0]
#> 
#> [[22]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7ff318135d98]
#> 
#> [[23]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g <- function(trim) identity(identity(fn(trim)))
g(2)

#> [[1]]
#> <frame 23> (22)
#> expr: g(2)
#> env:  [local 0x7ff317b48cf8]
#> 
#> [[2]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff31e1abc38]
#> 
#> [[3]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff317b4dcb8]
#> 
#> [[4]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7ff317b4cbe0]
#> 
#> [[5]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7ff317b4c4a8]
#> 
#> [[6]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7ff317b4fd98]
#> 
#> [[7]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7ff317b53698]
#> 
#> [[8]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7ff317b532e0]
#> 
#> [[9]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7ff317b52ba8]
#> 
#> [[10]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7ff317b52978]
#> 
#> [[11]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7ff317b527f0]
#> 
#> [[12]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7ff317b52710]
#> 
#> [[13]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7ff317b78ef0]
#> 
#> [[14]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7ff3152e0bc8]
#> 
#> [[15]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7ff31530fd20]
#> 
#> [[16]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7ff3143167e8]
#> 
#> [[17]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7ff3184e5220]
#> 
#> [[18]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7ff31859d4c0]
#> 
#> [[19]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7ff3186ab8b0]
#> 
#> [[20]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff31a855f90]
#> 
#> [[21]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff31a872c18]
#> 
#> [[22]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7ff3161e47a0]
#> 
#> [[23]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7ff318135d98]
#> 
#> [[24]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g(3)

#> [[1]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff31e1abc38]
#> 
#> [[2]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff3152eced8]
#> 
#> [[3]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7ff3152ecca8]
#> 
#> [[4]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7ff3152ec5e0]
#> 
#> [[5]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7ff3152ebf88]
#> 
#> [[6]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7ff3152ef290]
#> 
#> [[7]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7ff3152eeea0]
#> 
#> [[8]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7ff3152ee570]
#> 
#> [[9]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7ff3152ee1f0]
#> 
#> [[10]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7ff3152ee068]
#> 
#> [[11]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7ff3152edf50]
#> 
#> [[12]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7ff3157ac4d0]
#> 
#> [[13]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7ff3152e0bc8]
#> 
#> [[14]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7ff31530fd20]
#> 
#> [[15]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7ff3143167e8]
#> 
#> [[16]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7ff3184e5220]
#> 
#> [[17]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7ff31859d4c0]
#> 
#> [[18]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7ff3186ab8b0]
#> 
#> [[19]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff31a855f90]
#> 
#> [[20]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff31a872c18]
#> 
#> [[21]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7ff3161e47a0]
#> 
#> [[22]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7ff318135d98]
#> 
#> [[23]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"