Call stack information

The eval_ and call_ families of functions provide a replacement for the base R functions prefixed with sys. (which are all about the context stack), as well as for parent.frame() (which is the only base R function for querying the call stack). The context stack includes all R-level evaluation contexts. It is linear in terms of execution history but due to lazy evaluation it is potentially nonlinear in terms of call history. The call stack history, on the other hand, is homogenous.

global_frame()

current_frame()

ctxt_frame(n = 1)

call_frame(n = 1, clean = TRUE)

ctxt_depth()

call_depth()

ctxt_stack(n = NULL, trim = 0)

call_stack(n = NULL, clean = TRUE)

Arguments

n
The number of frames to go back in the stack.

clean
Whether to post-process the call stack to clean non-standard frames. If TRUE, suboptimal call-stack entries by base::eval() will be cleaned up: the duplicate frame created by eval() is eliminated.

trim
The number of layers of intervening frames to trim off the stack. See stack_trim() and examples.

Details

ctxt_frame() and call_frame() return a frame object containing the following fields: expr and env (call expression and evaluation environment), pos and caller_pos (position of current frame in the context stack and position of the caller), and fun (function of the current frame). ctxt_stack() and call_stack() return a list of all context or call frames on the stack. Finally, ctxt_depth() and call_depth() report the current context position or the number of calling frames on the stack.

The base R functions take two sorts of arguments to indicate which frame to query: which and n. The n argument is straightforward: it's the number of frames to go down the stack, with n = 1 referring to the current context. The which argument is more complicated and changes meaning for values lower than 1. For the sake of consistency, the rlang functions all take the same kind of argument n. This argument has a single meaning (the number of frames to go down the stack) and cannot be lower than 1.

Note finally that parent.frame(1) corresponds to call_frame(2)$env, as n = 1 always refers to the current frame. This makes the _frame() and _stack() functions consistent: ctxt_frame(2) is the same as ctxt_stack()[[2]]. Also, ctxt_depth() returns one more frame than base::sys.nframe() because it counts the global frame. That is consistent with the _stack() functions which return the global frame as well. This way, call_stack(call_depth()) is the same as global_frame().

Life cycle

These functions are soft-deprecated and replaced by trace_back().

Examples

# Expressions within arguments count as contexts
identity(identity(ctxt_depth())) # returns 2

#> Warning: `ctxt_depth()` is soft-deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 40

# But they are not part of the call stack because arguments are
# evaluated within the calling function (or the global environment
# if called at top level)
identity(identity(call_depth())) # returns 0

#> Warning: `call_depth()` is soft-deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 14

# The context stacks includes all intervening execution frames. The
# call stack doesn't:
f <- function(x) identity(x)
f(f(ctxt_stack()))

#> [[1]]
#> <frame 41> (40)
#> expr: identity(x)
#> env:  [local 0x466a3d8]
#> 
#> [[2]]
#> <frame 40> (37)
#> expr: f(ctxt_stack())
#> env:  [local 0x4670808]
#> 
#> [[3]]
#> <frame 39> (38)
#> expr: identity(x)
#> env:  [local 0x46706b8]
#> 
#> [[4]]
#> <frame 38> (37)
#> expr: f(f(ctxt_stack()))
#> env:  [local 0x46705a0]
#> 
#> [[5]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x59b4f78]
#> 
#> [[6]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x4678108]
#> 
#> [[7]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x4677ed8]
#> 
#> [[8]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x4677420]
#> 
#> [[9]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x468ce20]
#> 
#> [[10]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x468c758]
#> 
#> [[11]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x468c250]
#> 
#> [[12]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x468f638]
#> 
#> [[13]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x468f2f0]
#> 
#> [[14]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x46978a0]
#> 
#> [[15]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x46977c0]
#> 
#> [[16]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x488d858]
#> 
#> [[17]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x6b4dc28]
#> 
#> [[18]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x6b51f38]
#> 
#> [[19]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x6b53b10]
#> 
#> [[20]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x6b58e28]
#> 
#> [[21]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x6b5fb90]
#> 
#> [[22]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6d16988]
#> 
#> [[23]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6d15fe8]
#> 
#> [[24]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x8fc7060]
#> 
#> [[25]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x8bc3500]
#> 
#> [[26]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x813c4a8]
#> 
#> [[27]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x5f0bc08]
#> 
#> [[28]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x261cc68]
#> 
#> [[29]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x24d8200]
#> 
#> [[30]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x41e3290]
#> 
#> [[31]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x41e4250]
#> 
#> [[32]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpYrEbTp/file3a717d7138e4"),  <...>
#> env:  [frame 0x41d40b0]
#> 
#> [[33]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpYrEbTp/file3a717d7138e4"),  <...>
#> env:  [frame 0x41d4e40]
#> 
#> [[34]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x41cfc18]
#> 
#> [[35]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x41d0dd0]
#> 
#> [[36]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x41d1118]
#> 
#> [[37]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x41cd6a0]
#> 
#> [[38]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x41cdc50]
#> 
#> [[39]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x41cdf98]
#> 
#> [[40]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x41ce2e0]
#> 
#> [[41]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x41cea50]
#> 
#> [[42]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(f(call_stack()))

#> Warning: `call_stack()` is soft-deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [[1]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x59b4f78]
#> 
#> [[2]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7ced8d0]
#> 
#> [[3]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x6b4dc28]
#> 
#> [[4]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x6b58e28]
#> 
#> [[5]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x6b5fb90]
#> 
#> [[6]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6d16988]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x8fc7060]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x8bc3500]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x813c4a8]
#> 
#> [[10]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x5f0bc08]
#> 
#> [[11]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x261cc68]
#> 
#> [[12]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x24d8200]
#> 
#> [[13]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x41e3290]
#> 
#> [[14]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

g <- function(cmd) cmd()
f(g(ctxt_stack))

#> [[1]]
#> <frame 40> (37)
#> expr: g(ctxt_stack)
#> env:  [local 0x71626c8]
#> 
#> [[2]]
#> <frame 39> (38)
#> expr: identity(x)
#> env:  [local 0x8ec5720]
#> 
#> [[3]]
#> <frame 38> (37)
#> expr: f(g(ctxt_stack))
#> env:  [local 0x8ec5838]
#> 
#> [[4]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x59b4f78]
#> 
#> [[5]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x8ec59f8]
#> 
#> [[6]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x8ec5c28]
#> 
#> [[7]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x8ec61d8]
#> 
#> [[8]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x8ec6750]
#> 
#> [[9]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x8ec6a98]
#> 
#> [[10]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x8ec6de0]
#> 
#> [[11]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x8ec74a8]
#> 
#> [[12]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x8f0fd08]
#> 
#> [[13]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x8f0fe90]
#> 
#> [[14]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x8f0ff70]
#> 
#> [[15]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x8193480]
#> 
#> [[16]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x6b4dc28]
#> 
#> [[17]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x6b51f38]
#> 
#> [[18]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x6b53b10]
#> 
#> [[19]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x6b58e28]
#> 
#> [[20]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x6b5fb90]
#> 
#> [[21]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6d16988]
#> 
#> [[22]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6d15fe8]
#> 
#> [[23]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x8fc7060]
#> 
#> [[24]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x8bc3500]
#> 
#> [[25]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x813c4a8]
#> 
#> [[26]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x5f0bc08]
#> 
#> [[27]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x261cc68]
#> 
#> [[28]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x24d8200]
#> 
#> [[29]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x41e3290]
#> 
#> [[30]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x41e4250]
#> 
#> [[31]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpYrEbTp/file3a717d7138e4"),  <...>
#> env:  [frame 0x41d40b0]
#> 
#> [[32]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpYrEbTp/file3a717d7138e4"),  <...>
#> env:  [frame 0x41d4e40]
#> 
#> [[33]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x41cfc18]
#> 
#> [[34]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x41d0dd0]
#> 
#> [[35]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x41d1118]
#> 
#> [[36]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x41cd6a0]
#> 
#> [[37]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x41cdc50]
#> 
#> [[38]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x41cdf98]
#> 
#> [[39]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x41ce2e0]
#> 
#> [[40]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x41cea50]
#> 
#> [[41]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(g(call_stack))

#> [[1]]
#> <frame 40> (36)
#> expr: g(call_stack)
#> env:  [local 0x97df888]
#> 
#> [[2]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x59b4f78]
#> 
#> [[3]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x89326d0]
#> 
#> [[4]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x6b4dc28]
#> 
#> [[5]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x6b58e28]
#> 
#> [[6]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x6b5fb90]
#> 
#> [[7]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6d16988]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x8fc7060]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x8bc3500]
#> 
#> [[10]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x813c4a8]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x5f0bc08]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x261cc68]
#> 
#> [[13]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x24d8200]
#> 
#> [[14]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x41e3290]
#> 
#> [[15]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

# The rlang _stack() functions return a list of frame
# objects. Use purrr::transpose() or index a field with
# purrr::map()'s to extract a particular field from a stack:

# stack <- f(f(call_stack()))
# purrr::map(stack, "env")
# purrr::transpose(stack)$expr

# current_frame() is an alias for ctxt_frame(1)
fn <- function() list(current = current_frame(), first = ctxt_frame(1))
fn()

#> Warning: `current_frame()` is soft-deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> $current
#> <frame 38> (37)
#> expr: fn()
#> env:  [local 0x6e5f4e0]
#> 
#> $first
#> <frame 38> (37)
#> expr: fn()
#> env:  [local 0x6e5f4e0]
#> 

# While current_frame() is the top of the stack, global_frame() is
# the bottom:
fn <- function() {
  n <- ctxt_depth()
  ctxt_frame(n)
}
identical(fn(), global_frame())

#> [1] TRUE

# ctxt_stack() returns a stack with all intervening frames. You can
# trim layers of intervening frames with the trim argument:
identity(identity(ctxt_stack()))

#> [[1]]
#> <frame 39> (37)
#> expr: identity(ctxt_stack())
#> env:  [local 0x6ae8120]
#> 
#> [[2]]
#> <frame 38> (37)
#> expr: identity(identity(ctxt_stack()))
#> env:  [local 0x6ae8008]
#> 
#> [[3]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x59b4f78]
#> 
#> [[4]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x6ae7e10]
#> 
#> [[5]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x6ae7b70]
#> 
#> [[6]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x6ae75c0]
#> 
#> [[7]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x6aeae40]
#> 
#> [[8]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x6aeaa18]
#> 
#> [[9]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x6aea6d0]
#> 
#> [[10]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x6aea008]
#> 
#> [[11]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x6ae9dd8]
#> 
#> [[12]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x6ae9c18]
#> 
#> [[13]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x6ae9ac8]
#> 
#> [[14]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x6b21bf0]
#> 
#> [[15]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x6b4dc28]
#> 
#> [[16]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x6b51f38]
#> 
#> [[17]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x6b53b10]
#> 
#> [[18]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x6b58e28]
#> 
#> [[19]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x6b5fb90]
#> 
#> [[20]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6d16988]
#> 
#> [[21]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6d15fe8]
#> 
#> [[22]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x8fc7060]
#> 
#> [[23]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x8bc3500]
#> 
#> [[24]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x813c4a8]
#> 
#> [[25]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x5f0bc08]
#> 
#> [[26]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x261cc68]
#> 
#> [[27]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x24d8200]
#> 
#> [[28]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x41e3290]
#> 
#> [[29]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x41e4250]
#> 
#> [[30]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpYrEbTp/file3a717d7138e4"),  <...>
#> env:  [frame 0x41d40b0]
#> 
#> [[31]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpYrEbTp/file3a717d7138e4"),  <...>
#> env:  [frame 0x41d4e40]
#> 
#> [[32]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x41cfc18]
#> 
#> [[33]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x41d0dd0]
#> 
#> [[34]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x41d1118]
#> 
#> [[35]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x41cd6a0]
#> 
#> [[36]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x41cdc50]
#> 
#> [[37]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x41cdf98]
#> 
#> [[38]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x41ce2e0]
#> 
#> [[39]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x41cea50]
#> 
#> [[40]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(ctxt_stack(trim = 1)))

#> [[1]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x59b4f78]
#> 
#> [[2]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5a5e080]
#> 
#> [[3]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x5a5de50]
#> 
#> [[4]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x5a634d8]
#> 
#> [[5]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x5a62eb8]
#> 
#> [[6]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x5a62b70]
#> 
#> [[7]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x5a62828]
#> 
#> [[8]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x5a62080]
#> 
#> [[9]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x5a61e50]
#> 
#> [[10]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x5a61c90]
#> 
#> [[11]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x5a61bb0]
#> 
#> [[12]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x5b20778]
#> 
#> [[13]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x6b4dc28]
#> 
#> [[14]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x6b51f38]
#> 
#> [[15]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x6b53b10]
#> 
#> [[16]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x6b58e28]
#> 
#> [[17]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x6b5fb90]
#> 
#> [[18]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6d16988]
#> 
#> [[19]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6d15fe8]
#> 
#> [[20]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x8fc7060]
#> 
#> [[21]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x8bc3500]
#> 
#> [[22]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x813c4a8]
#> 
#> [[23]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x5f0bc08]
#> 
#> [[24]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x261cc68]
#> 
#> [[25]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x24d8200]
#> 
#> [[26]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x41e3290]
#> 
#> [[27]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x41e4250]
#> 
#> [[28]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpYrEbTp/file3a717d7138e4"),  <...>
#> env:  [frame 0x41d40b0]
#> 
#> [[29]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpYrEbTp/file3a717d7138e4"),  <...>
#> env:  [frame 0x41d4e40]
#> 
#> [[30]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x41cfc18]
#> 
#> [[31]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x41d0dd0]
#> 
#> [[32]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x41d1118]
#> 
#> [[33]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x41cd6a0]
#> 
#> [[34]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x41cdc50]
#> 
#> [[35]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x41cdf98]
#> 
#> [[36]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x41ce2e0]
#> 
#> [[37]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x41cea50]
#> 
#> [[38]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# ctxt_stack() is called within fn() with intervening frames:
fn <- function(trim) identity(identity(ctxt_stack(trim = trim)))
fn(0)

#> [[1]]
#> <frame 40> (38)
#> expr: identity(ctxt_stack(trim = trim))
#> env:  [local 0x2f20658]
#> 
#> [[2]]
#> <frame 39> (38)
#> expr: identity(identity(ctxt_stack(trim = trim)))
#> env:  [local 0x2f20540]
#> 
#> [[3]]
#> <frame 38> (37)
#> expr: fn(0)
#> env:  [local 0x2f203b8]
#> 
#> [[4]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x59b4f78]
#> 
#> [[5]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x2f200e0]
#> 
#> [[6]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x2f22c78]
#> 
#> [[7]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x2f22070]
#> 
#> [[8]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x2f27188]
#> 
#> [[9]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x2f26d28]
#> 
#> [[10]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x2f26890]
#> 
#> [[11]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x2f297b0]
#> 
#> [[12]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x2f29580]
#> 
#> [[13]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x2f28b00]
#> 
#> [[14]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x2f28978]
#> 
#> [[15]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x30d86a0]
#> 
#> [[16]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x6b4dc28]
#> 
#> [[17]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x6b51f38]
#> 
#> [[18]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x6b53b10]
#> 
#> [[19]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x6b58e28]
#> 
#> [[20]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x6b5fb90]
#> 
#> [[21]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6d16988]
#> 
#> [[22]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6d15fe8]
#> 
#> [[23]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x8fc7060]
#> 
#> [[24]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x8bc3500]
#> 
#> [[25]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x813c4a8]
#> 
#> [[26]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x5f0bc08]
#> 
#> [[27]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x261cc68]
#> 
#> [[28]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x24d8200]
#> 
#> [[29]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x41e3290]
#> 
#> [[30]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x41e4250]
#> 
#> [[31]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpYrEbTp/file3a717d7138e4"),  <...>
#> env:  [frame 0x41d40b0]
#> 
#> [[32]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpYrEbTp/file3a717d7138e4"),  <...>
#> env:  [frame 0x41d4e40]
#> 
#> [[33]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x41cfc18]
#> 
#> [[34]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x41d0dd0]
#> 
#> [[35]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x41d1118]
#> 
#> [[36]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x41cd6a0]
#> 
#> [[37]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x41cdc50]
#> 
#> [[38]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x41cdf98]
#> 
#> [[39]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x41ce2e0]
#> 
#> [[40]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x41cea50]
#> 
#> [[41]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# We can trim the first layer of those:
fn(1)

#> [[1]]
#> <frame 38> (37)
#> expr: fn(1)
#> env:  [local 0x9282a30]
#> 
#> [[2]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x59b4f78]
#> 
#> [[3]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x9282bf0]
#> 
#> [[4]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x927eff0]
#> 
#> [[5]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x927f5a0]
#> 
#> [[6]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x927fb18]
#> 
#> [[7]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x927fe60]
#> 
#> [[8]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x9280288]
#> 
#> [[9]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x8615f38]
#> 
#> [[10]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x86161a0]
#> 
#> [[11]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x8616328]
#> 
#> [[12]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x8616408]
#> 
#> [[13]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x8e4fad8]
#> 
#> [[14]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x6b4dc28]
#> 
#> [[15]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x6b51f38]
#> 
#> [[16]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x6b53b10]
#> 
#> [[17]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x6b58e28]
#> 
#> [[18]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x6b5fb90]
#> 
#> [[19]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6d16988]
#> 
#> [[20]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6d15fe8]
#> 
#> [[21]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x8fc7060]
#> 
#> [[22]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x8bc3500]
#> 
#> [[23]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x813c4a8]
#> 
#> [[24]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x5f0bc08]
#> 
#> [[25]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x261cc68]
#> 
#> [[26]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x24d8200]
#> 
#> [[27]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x41e3290]
#> 
#> [[28]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x41e4250]
#> 
#> [[29]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpYrEbTp/file3a717d7138e4"),  <...>
#> env:  [frame 0x41d40b0]
#> 
#> [[30]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpYrEbTp/file3a717d7138e4"),  <...>
#> env:  [frame 0x41d4e40]
#> 
#> [[31]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x41cfc18]
#> 
#> [[32]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x41d0dd0]
#> 
#> [[33]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x41d1118]
#> 
#> [[34]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x41cd6a0]
#> 
#> [[35]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x41cdc50]
#> 
#> [[36]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x41cdf98]
#> 
#> [[37]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x41ce2e0]
#> 
#> [[38]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x41cea50]
#> 
#> [[39]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# The outside intervening frames (at the fn() call site) are still
# returned, but can be trimmed as well:
identity(identity(fn(1)))

#> [[1]]
#> <frame 40> (37)
#> expr: fn(1)
#> env:  [local 0x8a604b8]
#> 
#> [[2]]
#> <frame 39> (37)
#> expr: identity(fn(1))
#> env:  [local 0x8a603a0]
#> 
#> [[3]]
#> <frame 38> (37)
#> expr: identity(identity(fn(1)))
#> env:  [local 0x8a60288]
#> 
#> [[4]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x59b4f78]
#> 
#> [[5]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x8a600c8]
#> 
#> [[6]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x954bc58]
#> 
#> [[7]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x954b670]
#> 
#> [[8]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x954b0f8]
#> 
#> [[9]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x954ad78]
#> 
#> [[10]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x954aa30]
#> 
#> [[11]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x954a368]
#> 
#> [[12]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x954a138]
#> 
#> [[13]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x9549fb0]
#> 
#> [[14]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x9549ed0]
#> 
#> [[15]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x8fa6560]
#> 
#> [[16]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x6b4dc28]
#> 
#> [[17]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x6b51f38]
#> 
#> [[18]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x6b53b10]
#> 
#> [[19]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x6b58e28]
#> 
#> [[20]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x6b5fb90]
#> 
#> [[21]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6d16988]
#> 
#> [[22]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6d15fe8]
#> 
#> [[23]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x8fc7060]
#> 
#> [[24]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x8bc3500]
#> 
#> [[25]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x813c4a8]
#> 
#> [[26]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x5f0bc08]
#> 
#> [[27]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x261cc68]
#> 
#> [[28]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x24d8200]
#> 
#> [[29]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x41e3290]
#> 
#> [[30]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x41e4250]
#> 
#> [[31]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpYrEbTp/file3a717d7138e4"),  <...>
#> env:  [frame 0x41d40b0]
#> 
#> [[32]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpYrEbTp/file3a717d7138e4"),  <...>
#> env:  [frame 0x41d4e40]
#> 
#> [[33]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x41cfc18]
#> 
#> [[34]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x41d0dd0]
#> 
#> [[35]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x41d1118]
#> 
#> [[36]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x41cd6a0]
#> 
#> [[37]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x41cdc50]
#> 
#> [[38]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x41cdf98]
#> 
#> [[39]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x41ce2e0]
#> 
#> [[40]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x41cea50]
#> 
#> [[41]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(fn(2)))

#> [[1]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x59b4f78]
#> 
#> [[2]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x8544f78]
#> 
#> [[3]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x8544d10]
#> 
#> [[4]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x8544728]
#> 
#> [[5]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x85441b0]
#> 
#> [[6]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x8543e68]
#> 
#> [[7]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x8549870]
#> 
#> [[8]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x8549170]
#> 
#> [[9]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x8548f40]
#> 
#> [[10]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x8548db8]
#> 
#> [[11]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x8548ca0]
#> 
#> [[12]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x8567538]
#> 
#> [[13]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x6b4dc28]
#> 
#> [[14]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x6b51f38]
#> 
#> [[15]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x6b53b10]
#> 
#> [[16]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x6b58e28]
#> 
#> [[17]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x6b5fb90]
#> 
#> [[18]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6d16988]
#> 
#> [[19]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6d15fe8]
#> 
#> [[20]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x8fc7060]
#> 
#> [[21]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x8bc3500]
#> 
#> [[22]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x813c4a8]
#> 
#> [[23]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x5f0bc08]
#> 
#> [[24]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x261cc68]
#> 
#> [[25]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x24d8200]
#> 
#> [[26]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x41e3290]
#> 
#> [[27]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x41e4250]
#> 
#> [[28]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpYrEbTp/file3a717d7138e4"),  <...>
#> env:  [frame 0x41d40b0]
#> 
#> [[29]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpYrEbTp/file3a717d7138e4"),  <...>
#> env:  [frame 0x41d4e40]
#> 
#> [[30]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x41cfc18]
#> 
#> [[31]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x41d0dd0]
#> 
#> [[32]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x41d1118]
#> 
#> [[33]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x41cd6a0]
#> 
#> [[34]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x41cdc50]
#> 
#> [[35]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x41cdf98]
#> 
#> [[36]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x41ce2e0]
#> 
#> [[37]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x41cea50]
#> 
#> [[38]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g <- function(trim) identity(identity(fn(trim)))
g(2)

#> [[1]]
#> <frame 38> (37)
#> expr: g(2)
#> env:  [local 0x67dc0a0]
#> 
#> [[2]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x59b4f78]
#> 
#> [[3]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x67de298]
#> 
#> [[4]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x67dd038]
#> 
#> [[5]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x67e4180]
#> 
#> [[6]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x67e3b60]
#> 
#> [[7]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x67e2ee8]
#> 
#> [[8]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x67e2b30]
#> 
#> [[9]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x67e5e38]
#> 
#> [[10]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x67e5b98]
#> 
#> [[11]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x67e5a10]
#> 
#> [[12]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x67e5930]
#> 
#> [[13]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x682d9a0]
#> 
#> [[14]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x6b4dc28]
#> 
#> [[15]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x6b51f38]
#> 
#> [[16]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x6b53b10]
#> 
#> [[17]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x6b58e28]
#> 
#> [[18]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x6b5fb90]
#> 
#> [[19]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6d16988]
#> 
#> [[20]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6d15fe8]
#> 
#> [[21]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x8fc7060]
#> 
#> [[22]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x8bc3500]
#> 
#> [[23]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x813c4a8]
#> 
#> [[24]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x5f0bc08]
#> 
#> [[25]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x261cc68]
#> 
#> [[26]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x24d8200]
#> 
#> [[27]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x41e3290]
#> 
#> [[28]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x41e4250]
#> 
#> [[29]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpYrEbTp/file3a717d7138e4"),  <...>
#> env:  [frame 0x41d40b0]
#> 
#> [[30]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpYrEbTp/file3a717d7138e4"),  <...>
#> env:  [frame 0x41d4e40]
#> 
#> [[31]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x41cfc18]
#> 
#> [[32]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x41d0dd0]
#> 
#> [[33]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x41d1118]
#> 
#> [[34]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x41cd6a0]
#> 
#> [[35]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x41cdc50]
#> 
#> [[36]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x41cdf98]
#> 
#> [[37]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x41ce2e0]
#> 
#> [[38]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x41cea50]
#> 
#> [[39]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g(3)

#> [[1]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x59b4f78]
#> 
#> [[2]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x48de1b0]
#> 
#> [[3]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x48ddf10]
#> 
#> [[4]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x48dd8b8]
#> 
#> [[5]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x48dd308]
#> 
#> [[6]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x48dc9d8]
#> 
#> [[7]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x48dc578]
#> 
#> [[8]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x48e0d88]
#> 
#> [[9]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x48e0ab0]
#> 
#> [[10]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x48e08b8]
#> 
#> [[11]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x48e07d8]
#> 
#> [[12]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x4a1d9c8]
#> 
#> [[13]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x6b4dc28]
#> 
#> [[14]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x6b51f38]
#> 
#> [[15]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x6b53b10]
#> 
#> [[16]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x6b58e28]
#> 
#> [[17]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x6b5fb90]
#> 
#> [[18]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6d16988]
#> 
#> [[19]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6d15fe8]
#> 
#> [[20]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x8fc7060]
#> 
#> [[21]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x8bc3500]
#> 
#> [[22]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x813c4a8]
#> 
#> [[23]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x5f0bc08]
#> 
#> [[24]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x261cc68]
#> 
#> [[25]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x24d8200]
#> 
#> [[26]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x41e3290]
#> 
#> [[27]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x41e4250]
#> 
#> [[28]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpYrEbTp/file3a717d7138e4"),  <...>
#> env:  [frame 0x41d40b0]
#> 
#> [[29]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpYrEbTp/file3a717d7138e4"),  <...>
#> env:  [frame 0x41d4e40]
#> 
#> [[30]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x41cfc18]
#> 
#> [[31]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x41d0dd0]
#> 
#> [[32]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x41d1118]
#> 
#> [[33]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x41cd6a0]
#> 
#> [[34]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x41cdc50]
#> 
#> [[35]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x41cdf98]
#> 
#> [[36]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x41ce2e0]
#> 
#> [[37]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x41cea50]
#> 
#> [[38]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"