Call stack information

The eval_ and call_ families of functions provide a replacement for the base R functions prefixed with sys. (which are all about the context stack), as well as for parent.frame() (which is the only base R function for querying the call stack). The context stack includes all R-level evaluation contexts. It is linear in terms of execution history but due to lazy evaluation it is potentially nonlinear in terms of call history. The call stack history, on the other hand, is homogenous.

global_frame()

current_frame()

ctxt_frame(n = 1)

call_frame(n = 1, clean = TRUE)

ctxt_depth()

call_depth()

ctxt_stack(n = NULL, trim = 0)

call_stack(n = NULL, clean = TRUE)

Arguments

n
The number of frames to go back in the stack.

clean
Whether to post-process the call stack to clean non-standard frames. If TRUE, suboptimal call-stack entries by base::eval() will be cleaned up: the duplicate frame created by eval() is eliminated.

trim
The number of layers of intervening frames to trim off the stack. See stack_trim() and examples.

Details

ctxt_frame() and call_frame() return a frame object containing the following fields: expr and env (call expression and evaluation environment), pos and caller_pos (position of current frame in the context stack and position of the caller), and fun (function of the current frame). ctxt_stack() and call_stack() return a list of all context or call frames on the stack. Finally, ctxt_depth() and call_depth() report the current context position or the number of calling frames on the stack.

The base R functions take two sorts of arguments to indicate which frame to query: which and n. The n argument is straightforward: it's the number of frames to go down the stack, with n = 1 referring to the current context. The which argument is more complicated and changes meaning for values lower than 1. For the sake of consistency, the rlang functions all take the same kind of argument n. This argument has a single meaning (the number of frames to go down the stack) and cannot be lower than 1.

Note finally that parent.frame(1) corresponds to call_frame(2)$env, as n = 1 always refers to the current frame. This makes the _frame() and _stack() functions consistent: ctxt_frame(2) is the same as ctxt_stack()[[2]]. Also, ctxt_depth() returns one more frame than base::sys.nframe() because it counts the global frame. That is consistent with the _stack() functions which return the global frame as well. This way, call_stack(call_depth()) is the same as global_frame().

Life cycle

These functions are soft-deprecated and replaced by trace_back().

Examples

# Expressions within arguments count as contexts
identity(identity(ctxt_depth())) # returns 2

#> Warning: `ctxt_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 25

# But they are not part of the call stack because arguments are
# evaluated within the calling function (or the global environment
# if called at top level)
identity(identity(call_depth())) # returns 0

#> Warning: `call_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 13

# The context stacks includes all intervening execution frames. The
# call stack doesn't:
f <- function(x) identity(x)
f(f(ctxt_stack()))

#> [[1]]
#> <frame 26> (25)
#> expr: identity(x)
#> env:  [local 0x7fd44a89b478]
#> 
#> [[2]]
#> <frame 25> (22)
#> expr: f(ctxt_stack())
#> env:  [local 0x7fd44a89b590]
#> 
#> [[3]]
#> <frame 24> (23)
#> expr: identity(x)
#> env:  [local 0x7fd44a89b6a8]
#> 
#> [[4]]
#> <frame 23> (22)
#> expr: f(f(ctxt_stack()))
#> env:  [local 0x7fd44a8978b0]
#> 
#> [[5]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fd44aaa1bb8]
#> 
#> [[6]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fd44a897a70]
#> 
#> [[7]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fd44a897ca0]
#> 
#> [[8]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fd44a898250]
#> 
#> [[9]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fd44a898838]
#> 
#> [[10]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fd44a898b80]
#> 
#> [[11]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fd44a898ec8]
#> 
#> [[12]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fd44a895df0]
#> 
#> [[13]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fd44a8960c8]
#> 
#> [[14]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fd44a8962f8]
#> 
#> [[15]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fd44a8963d8]
#> 
#> [[16]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fd44ac3c670]
#> 
#> [[17]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fd442a5b880]
#> 
#> [[18]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fd442a8a240]
#> 
#> [[19]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fd4435f7d28]
#> 
#> [[20]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fd444ddac20]
#> 
#> [[21]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fd444f49288]
#> 
#> [[22]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fd4468eeba8]
#> 
#> [[23]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fd445275fa8]
#> 
#> [[24]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fd442ff0318]
#> 
#> [[25]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7fd442ea0a40]
#> 
#> [[26]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7fd4448607d0]
#> 
#> [[27]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(f(call_stack()))

#> Warning: `call_stack()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [[1]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fd44aaa1bb8]
#> 
#> [[2]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fd44aa016b0]
#> 
#> [[3]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fd442a5b880]
#> 
#> [[4]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fd442a8a240]
#> 
#> [[5]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fd4435f7d28]
#> 
#> [[6]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fd444ddac20]
#> 
#> [[7]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fd444f49288]
#> 
#> [[8]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fd4468eeba8]
#> 
#> [[9]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fd445275fa8]
#> 
#> [[10]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fd442ff0318]
#> 
#> [[11]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7fd442ea0a40]
#> 
#> [[12]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7fd4448607d0]
#> 
#> [[13]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

g <- function(cmd) cmd()
f(g(ctxt_stack))

#> [[1]]
#> <frame 25> (22)
#> expr: g(ctxt_stack)
#> env:  [local 0x7fd4499489a8]
#> 
#> [[2]]
#> <frame 24> (23)
#> expr: identity(x)
#> env:  [local 0x7fd449948890]
#> 
#> [[3]]
#> <frame 23> (22)
#> expr: f(g(ctxt_stack))
#> env:  [local 0x7fd449948778]
#> 
#> [[4]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fd44aaa1bb8]
#> 
#> [[5]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fd4499485b8]
#> 
#> [[6]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fd449948388]
#> 
#> [[7]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fd4464978e8]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fd446497370]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fd446497028]
#> 
#> [[10]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fd446496ce0]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fd4464965e0]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fd446496180]
#> 
#> [[13]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fd446495ff8]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fd446495f18]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fd44648bd20]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fd442a5b880]
#> 
#> [[17]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fd442a8a240]
#> 
#> [[18]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fd4435f7d28]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fd444ddac20]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fd444f49288]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fd4468eeba8]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fd445275fa8]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fd442ff0318]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7fd442ea0a40]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7fd4448607d0]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(g(call_stack))

#> [[1]]
#> <frame 25> (21)
#> expr: g(call_stack)
#> env:  [local 0x7fd445f4ad20]
#> 
#> [[2]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fd44aaa1bb8]
#> 
#> [[3]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fd445f61158]
#> 
#> [[4]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fd442a5b880]
#> 
#> [[5]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fd442a8a240]
#> 
#> [[6]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fd4435f7d28]
#> 
#> [[7]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fd444ddac20]
#> 
#> [[8]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fd444f49288]
#> 
#> [[9]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fd4468eeba8]
#> 
#> [[10]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fd445275fa8]
#> 
#> [[11]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fd442ff0318]
#> 
#> [[12]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7fd442ea0a40]
#> 
#> [[13]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7fd4448607d0]
#> 
#> [[14]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

# The rlang _stack() functions return a list of frame
# objects. Use purrr::transpose() or index a field with
# purrr::map()'s to extract a particular field from a stack:

# stack <- f(f(call_stack()))
# purrr::map(stack, "env")
# purrr::transpose(stack)$expr

# current_frame() is an alias for ctxt_frame(1)
fn <- function() list(current = current_frame(), first = ctxt_frame(1))
fn()

#> Warning: `current_frame()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> $current
#> <frame 23> (22)
#> expr: fn()
#> env:  [local 0x7fd4462f6a38]
#> 
#> $first
#> <frame 23> (22)
#> expr: fn()
#> env:  [local 0x7fd4462f6a38]
#> 

# While current_frame() is the top of the stack, global_frame() is
# the bottom:
fn <- function() {
  n <- ctxt_depth()
  ctxt_frame(n)
}
identical(fn(), global_frame())

#> [1] TRUE

# ctxt_stack() returns a stack with all intervening frames. You can
# trim layers of intervening frames with the trim argument:
identity(identity(ctxt_stack()))

#> [[1]]
#> <frame 24> (22)
#> expr: identity(ctxt_stack())
#> env:  [local 0x7fd445dce9d0]
#> 
#> [[2]]
#> <frame 23> (22)
#> expr: identity(identity(ctxt_stack()))
#> env:  [local 0x7fd445dce8b8]
#> 
#> [[3]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fd44aaa1bb8]
#> 
#> [[4]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fd445dce688]
#> 
#> [[5]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fd445dce3e8]
#> 
#> [[6]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fd445dcdce8]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fd445dcd770]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fd445dcd428]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fd445dcd0a8]
#> 
#> [[10]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fd445dd4810]
#> 
#> [[11]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fd445dd45e0]
#> 
#> [[12]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fd445dd4458]
#> 
#> [[13]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fd445dd4378]
#> 
#> [[14]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fd445de9b60]
#> 
#> [[15]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fd442a5b880]
#> 
#> [[16]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fd442a8a240]
#> 
#> [[17]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fd4435f7d28]
#> 
#> [[18]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fd444ddac20]
#> 
#> [[19]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fd444f49288]
#> 
#> [[20]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fd4468eeba8]
#> 
#> [[21]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fd445275fa8]
#> 
#> [[22]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fd442ff0318]
#> 
#> [[23]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7fd442ea0a40]
#> 
#> [[24]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7fd4448607d0]
#> 
#> [[25]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(ctxt_stack(trim = 1)))

#> [[1]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fd44aaa1bb8]
#> 
#> [[2]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fd44724af98]
#> 
#> [[3]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fd44724ad30]
#> 
#> [[4]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fd44724a780]
#> 
#> [[5]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fd44724a208]
#> 
#> [[6]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fd44724ddd0]
#> 
#> [[7]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fd44724da18]
#> 
#> [[8]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fd44724cd30]
#> 
#> [[9]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fd44724ca20]
#> 
#> [[10]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fd44724c898]
#> 
#> [[11]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fd44724c7b8]
#> 
#> [[12]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fd447263388]
#> 
#> [[13]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fd442a5b880]
#> 
#> [[14]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fd442a8a240]
#> 
#> [[15]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fd4435f7d28]
#> 
#> [[16]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fd444ddac20]
#> 
#> [[17]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fd444f49288]
#> 
#> [[18]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fd4468eeba8]
#> 
#> [[19]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fd445275fa8]
#> 
#> [[20]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fd442ff0318]
#> 
#> [[21]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7fd442ea0a40]
#> 
#> [[22]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7fd4448607d0]
#> 
#> [[23]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# ctxt_stack() is called within fn() with intervening frames:
fn <- function(trim) identity(identity(ctxt_stack(trim = trim)))
fn(0)

#> [[1]]
#> <frame 25> (23)
#> expr: identity(ctxt_stack(trim = trim))
#> env:  [local 0x7fd44702edd0]
#> 
#> [[2]]
#> <frame 24> (23)
#> expr: identity(identity(ctxt_stack(trim = trim)))
#> env:  [local 0x7fd44702ecb8]
#> 
#> [[3]]
#> <frame 23> (22)
#> expr: fn(0)
#> env:  [local 0x7fd44702eb68]
#> 
#> [[4]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fd44aaa1bb8]
#> 
#> [[5]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fd44702e9a8]
#> 
#> [[6]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fd44702e740]
#> 
#> [[7]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fd44702e190]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fd447031a80]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fd447031738]
#> 
#> [[10]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fd4470313f0]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fd447030c80]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fd447030a50]
#> 
#> [[13]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fd4470308c8]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fd4470307e8]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fd447058e40]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fd442a5b880]
#> 
#> [[17]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fd442a8a240]
#> 
#> [[18]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fd4435f7d28]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fd444ddac20]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fd444f49288]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fd4468eeba8]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fd445275fa8]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fd442ff0318]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7fd442ea0a40]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7fd4448607d0]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# We can trim the first layer of those:
fn(1)

#> [[1]]
#> <frame 23> (22)
#> expr: fn(1)
#> env:  [local 0x7fd446dc76a8]
#> 
#> [[2]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fd44aaa1bb8]
#> 
#> [[3]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fd446dc7478]
#> 
#> [[4]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fd446dc7248]
#> 
#> [[5]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fd446dc6bb8]
#> 
#> [[6]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fd446dcc4a8]
#> 
#> [[7]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fd446dcc160]
#> 
#> [[8]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fd446dcbde0]
#> 
#> [[9]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fd446dcb6a8]
#> 
#> [[10]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fd446dcb440]
#> 
#> [[11]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fd446dcb2b8]
#> 
#> [[12]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fd446dcb1a0]
#> 
#> [[13]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fd446de37f8]
#> 
#> [[14]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fd442a5b880]
#> 
#> [[15]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fd442a8a240]
#> 
#> [[16]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fd4435f7d28]
#> 
#> [[17]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fd444ddac20]
#> 
#> [[18]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fd444f49288]
#> 
#> [[19]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fd4468eeba8]
#> 
#> [[20]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fd445275fa8]
#> 
#> [[21]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fd442ff0318]
#> 
#> [[22]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7fd442ea0a40]
#> 
#> [[23]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7fd4448607d0]
#> 
#> [[24]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# The outside intervening frames (at the fn() call site) are still
# returned, but can be trimmed as well:
identity(identity(fn(1)))

#> [[1]]
#> <frame 25> (22)
#> expr: fn(1)
#> env:  [local 0x7fd4469d2438]
#> 
#> [[2]]
#> <frame 24> (22)
#> expr: identity(fn(1))
#> env:  [local 0x7fd4469d2320]
#> 
#> [[3]]
#> <frame 23> (22)
#> expr: identity(identity(fn(1)))
#> env:  [local 0x7fd4469d8070]
#> 
#> [[4]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fd44aaa1bb8]
#> 
#> [[5]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fd4469d7e78]
#> 
#> [[6]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fd4469d7bd8]
#> 
#> [[7]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fd4469d75b8]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fd4469d6ba8]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fd4469d6630]
#> 
#> [[10]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fd4469d6240]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fd4469d99a8]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fd4469d9740]
#> 
#> [[13]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fd4469d9580]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fd4469d93f8]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fd446a05660]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fd442a5b880]
#> 
#> [[17]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fd442a8a240]
#> 
#> [[18]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fd4435f7d28]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fd444ddac20]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fd444f49288]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fd4468eeba8]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fd445275fa8]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fd442ff0318]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7fd442ea0a40]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7fd4448607d0]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(fn(2)))

#> [[1]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fd44aaa1bb8]
#> 
#> [[2]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fd445bc1f50]
#> 
#> [[3]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fd445bc1cb0]
#> 
#> [[4]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fd445bc15e8]
#> 
#> [[5]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fd445bc0f20]
#> 
#> [[6]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fd445bc4a40]
#> 
#> [[7]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fd445bc46c0]
#> 
#> [[8]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fd445bc3f18]
#> 
#> [[9]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fd445bc3c78]
#> 
#> [[10]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fd445bc3ab8]
#> 
#> [[11]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fd445bc39d8]
#> 
#> [[12]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fd445bdd700]
#> 
#> [[13]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fd442a5b880]
#> 
#> [[14]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fd442a8a240]
#> 
#> [[15]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fd4435f7d28]
#> 
#> [[16]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fd444ddac20]
#> 
#> [[17]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fd444f49288]
#> 
#> [[18]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fd4468eeba8]
#> 
#> [[19]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fd445275fa8]
#> 
#> [[20]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fd442ff0318]
#> 
#> [[21]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7fd442ea0a40]
#> 
#> [[22]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7fd4448607d0]
#> 
#> [[23]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g <- function(trim) identity(identity(fn(trim)))
g(2)

#> [[1]]
#> <frame 23> (22)
#> expr: g(2)
#> env:  [local 0x7fd445566160]
#> 
#> [[2]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fd44aaa1bb8]
#> 
#> [[3]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fd445565e18]
#> 
#> [[4]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fd445565bb0]
#> 
#> [[5]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fd445565440]
#> 
#> [[6]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fd445564e58]
#> 
#> [[7]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fd445564988]
#> 
#> [[8]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fd445564640]
#> 
#> [[9]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fd445567d70]
#> 
#> [[10]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fd445567ad0]
#> 
#> [[11]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fd445567948]
#> 
#> [[12]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fd445567830]
#> 
#> [[13]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fd4455cc4c0]
#> 
#> [[14]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fd442a5b880]
#> 
#> [[15]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fd442a8a240]
#> 
#> [[16]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fd4435f7d28]
#> 
#> [[17]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fd444ddac20]
#> 
#> [[18]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fd444f49288]
#> 
#> [[19]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fd4468eeba8]
#> 
#> [[20]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fd445275fa8]
#> 
#> [[21]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fd442ff0318]
#> 
#> [[22]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7fd442ea0a40]
#> 
#> [[23]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7fd4448607d0]
#> 
#> [[24]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g(3)

#> [[1]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fd44aaa1bb8]
#> 
#> [[2]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fd442df6fd0]
#> 
#> [[3]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fd442df6da0]
#> 
#> [[4]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fd442df65c0]
#> 
#> [[5]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fd442df98c8]
#> 
#> [[6]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fd442df9510]
#> 
#> [[7]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fd442df8fd0]
#> 
#> [[8]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fd442df8748]
#> 
#> [[9]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fd442df83c8]
#> 
#> [[10]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fd442dfb858]
#> 
#> [[11]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fd442dfa630]
#> 
#> [[12]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fd442e4a978]
#> 
#> [[13]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fd442a5b880]
#> 
#> [[14]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fd442a8a240]
#> 
#> [[15]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fd4435f7d28]
#> 
#> [[16]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fd444ddac20]
#> 
#> [[17]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fd444f49288]
#> 
#> [[18]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fd4468eeba8]
#> 
#> [[19]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fd445275fa8]
#> 
#> [[20]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fd442ff0318]
#> 
#> [[21]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7fd442ea0a40]
#> 
#> [[22]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7fd4448607d0]
#> 
#> [[23]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"