Call stack information

The eval_ and call_ families of functions provide a replacement for the base R functions prefixed with sys. (which are all about the context stack), as well as for parent.frame() (which is the only base R function for querying the call stack). The context stack includes all R-level evaluation contexts. It is linear in terms of execution history but due to lazy evaluation it is potentially nonlinear in terms of call history. The call stack history, on the other hand, is homogenous.

global_frame()

current_frame()

ctxt_frame(n = 1)

call_frame(n = 1, clean = TRUE)

ctxt_depth()

call_depth()

ctxt_stack(n = NULL, trim = 0)

call_stack(n = NULL, clean = TRUE)

Arguments

n
The number of frames to go back in the stack.

clean
Whether to post-process the call stack to clean non-standard frames. If TRUE, suboptimal call-stack entries by base::eval() will be cleaned up: the duplicate frame created by eval() is eliminated.

trim
The number of layers of intervening frames to trim off the stack. See stack_trim() and examples.

Details

ctxt_frame() and call_frame() return a frame object containing the following fields: expr and env (call expression and evaluation environment), pos and caller_pos (position of current frame in the context stack and position of the caller), and fun (function of the current frame). ctxt_stack() and call_stack() return a list of all context or call frames on the stack. Finally, ctxt_depth() and call_depth() report the current context position or the number of calling frames on the stack.

The base R functions take two sorts of arguments to indicate which frame to query: which and n. The n argument is straightforward: it's the number of frames to go down the stack, with n = 1 referring to the current context. The which argument is more complicated and changes meaning for values lower than 1. For the sake of consistency, the rlang functions all take the same kind of argument n. This argument has a single meaning (the number of frames to go down the stack) and cannot be lower than 1.

Note finally that parent.frame(1) corresponds to call_frame(2)$env, as n = 1 always refers to the current frame. This makes the _frame() and _stack() functions consistent: ctxt_frame(2) is the same as ctxt_stack()[[2]]. Also, ctxt_depth() returns one more frame than base::sys.nframe() because it counts the global frame. That is consistent with the _stack() functions which return the global frame as well. This way, call_stack(call_depth()) is the same as global_frame().

Life cycle

These functions are soft-deprecated and replaced by trace_back().

Examples

# Expressions within arguments count as contexts
identity(identity(ctxt_depth())) # returns 2

#> Warning: `ctxt_depth()` is soft-deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 38

# But they are not part of the call stack because arguments are
# evaluated within the calling function (or the global environment
# if called at top level)
identity(identity(call_depth())) # returns 0

#> Warning: `call_depth()` is soft-deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 14

# The context stacks includes all intervening execution frames. The
# call stack doesn't:
f <- function(x) identity(x)
f(f(ctxt_stack()))

#> [[1]]
#> <frame 39> (38)
#> expr: identity(x)
#> env:  [local 0x79db588]
#> 
#> [[2]]
#> <frame 38> (35)
#> expr: f(ctxt_stack())
#> env:  [local 0x79df2a0]
#> 
#> [[3]]
#> <frame 37> (36)
#> expr: identity(x)
#> env:  [local 0x79df188]
#> 
#> [[4]]
#> <frame 36> (35)
#> expr: f(f(ctxt_stack()))
#> env:  [local 0x79df038]
#> 
#> [[5]]
#> <frame 35> (34)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x812a480]
#> 
#> [[6]]
#> <frame 34> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x79dee78]
#> 
#> [[7]]
#> <frame 33> (24)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x79dec10]
#> 
#> [[8]]
#> <frame 32> (24)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x79de660]
#> 
#> [[9]]
#> <frame 31> (30)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x79de0e8]
#> 
#> [[10]]
#> <frame 30> (29)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x79ddda0]
#> 
#> [[11]]
#> <frame 29> (28)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x79dda58]
#> 
#> [[12]]
#> <frame 28> (27)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7a08040]
#> 
#> [[13]]
#> <frame 27> (26)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7a07e10]
#> 
#> [[14]]
#> <frame 26> (24)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7a07c88]
#> 
#> [[15]]
#> <frame 25> (24)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7a07ba8]
#> 
#> [[16]]
#> <frame 24> (23)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7a43b10]
#> 
#> [[17]]
#> <frame 23> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x7acafc8]
#> 
#> [[18]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x7acc688]
#> 
#> [[19]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x92459b0]
#> 
#> [[20]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x81a7c28]
#> 
#> [[21]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x81a85c8]
#> 
#> [[22]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x2621258]
#> 
#> [[23]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x3d49c48]
#> 
#> [[24]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x6c51558]
#> 
#> [[25]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x4fc3208]
#> 
#> [[26]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x184c128]
#> 
#> [[27]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x1cf0930]
#> 
#> [[28]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3558290]
#> 
#> [[29]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3559250]
#> 
#> [[30]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpLvKGyd/file3b3d6016e5ac"),  <...>
#> env:  [frame 0x35490b0]
#> 
#> [[31]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpLvKGyd/file3b3d6016e5ac"),  <...>
#> env:  [frame 0x3549e40]
#> 
#> [[32]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3544c18]
#> 
#> [[33]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3545dd0]
#> 
#> [[34]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x3546118]
#> 
#> [[35]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x35426a0]
#> 
#> [[36]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3542c50]
#> 
#> [[37]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3542f98]
#> 
#> [[38]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x35432e0]
#> 
#> [[39]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3543a50]
#> 
#> [[40]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(f(call_stack()))

#> Warning: `call_stack()` is soft-deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [[1]]
#> <frame 34> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x812a480]
#> 
#> [[2]]
#> <frame 24> (23)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x5b0d240]
#> 
#> [[3]]
#> <frame 23> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x7acafc8]
#> 
#> [[4]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x7acc688]
#> 
#> [[5]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x92459b0]
#> 
#> [[6]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x81a7c28]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x2621258]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x3d49c48]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x6c51558]
#> 
#> [[10]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x4fc3208]
#> 
#> [[11]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x184c128]
#> 
#> [[12]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x1cf0930]
#> 
#> [[13]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3558290]
#> 
#> [[14]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

g <- function(cmd) cmd()
f(g(ctxt_stack))

#> [[1]]
#> <frame 38> (35)
#> expr: g(ctxt_stack)
#> env:  [local 0x5563150]
#> 
#> [[2]]
#> <frame 37> (36)
#> expr: identity(x)
#> env:  [local 0x5563038]
#> 
#> [[3]]
#> <frame 36> (35)
#> expr: f(g(ctxt_stack))
#> env:  [local 0x5562f20]
#> 
#> [[4]]
#> <frame 35> (34)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x812a480]
#> 
#> [[5]]
#> <frame 34> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5562d60]
#> 
#> [[6]]
#> <frame 33> (24)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x5562b30]
#> 
#> [[7]]
#> <frame 32> (24)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x5562510]
#> 
#> [[8]]
#> <frame 31> (30)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x5561f98]
#> 
#> [[9]]
#> <frame 30> (29)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x5561c50]
#> 
#> [[10]]
#> <frame 29> (28)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x5561908]
#> 
#> [[11]]
#> <frame 28> (27)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x5564e40]
#> 
#> [[12]]
#> <frame 27> (26)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x5564c10]
#> 
#> [[13]]
#> <frame 26> (24)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x5564a88]
#> 
#> [[14]]
#> <frame 25> (24)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x55649a8]
#> 
#> [[15]]
#> <frame 24> (23)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x559a360]
#> 
#> [[16]]
#> <frame 23> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x7acafc8]
#> 
#> [[17]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x7acc688]
#> 
#> [[18]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x92459b0]
#> 
#> [[19]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x81a7c28]
#> 
#> [[20]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x81a85c8]
#> 
#> [[21]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x2621258]
#> 
#> [[22]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x3d49c48]
#> 
#> [[23]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x6c51558]
#> 
#> [[24]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x4fc3208]
#> 
#> [[25]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x184c128]
#> 
#> [[26]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x1cf0930]
#> 
#> [[27]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3558290]
#> 
#> [[28]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3559250]
#> 
#> [[29]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpLvKGyd/file3b3d6016e5ac"),  <...>
#> env:  [frame 0x35490b0]
#> 
#> [[30]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpLvKGyd/file3b3d6016e5ac"),  <...>
#> env:  [frame 0x3549e40]
#> 
#> [[31]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3544c18]
#> 
#> [[32]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3545dd0]
#> 
#> [[33]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x3546118]
#> 
#> [[34]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x35426a0]
#> 
#> [[35]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3542c50]
#> 
#> [[36]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3542f98]
#> 
#> [[37]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x35432e0]
#> 
#> [[38]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3543a50]
#> 
#> [[39]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(g(call_stack))

#> [[1]]
#> <frame 38> (34)
#> expr: g(call_stack)
#> env:  [local 0x1b98750]
#> 
#> [[2]]
#> <frame 34> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x812a480]
#> 
#> [[3]]
#> <frame 24> (23)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x26154e8]
#> 
#> [[4]]
#> <frame 23> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x7acafc8]
#> 
#> [[5]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x7acc688]
#> 
#> [[6]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x92459b0]
#> 
#> [[7]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x81a7c28]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x2621258]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x3d49c48]
#> 
#> [[10]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x6c51558]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x4fc3208]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x184c128]
#> 
#> [[13]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x1cf0930]
#> 
#> [[14]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3558290]
#> 
#> [[15]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

# The rlang _stack() functions return a list of frame
# objects. Use purrr::transpose() or index a field with
# purrr::map()'s to extract a particular field from a stack:

# stack <- f(f(call_stack()))
# purrr::map(stack, "env")
# purrr::transpose(stack)$expr

# current_frame() is an alias for ctxt_frame(1)
fn <- function() list(current = current_frame(), first = ctxt_frame(1))
fn()

#> Warning: `current_frame()` is soft-deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> $current
#> <frame 36> (35)
#> expr: fn()
#> env:  [local 0x7cb3150]
#> 
#> $first
#> <frame 36> (35)
#> expr: fn()
#> env:  [local 0x7cb3150]
#> 

# While current_frame() is the top of the stack, global_frame() is
# the bottom:
fn <- function() {
  n <- ctxt_depth()
  ctxt_frame(n)
}
identical(fn(), global_frame())

#> [1] TRUE

# ctxt_stack() returns a stack with all intervening frames. You can
# trim layers of intervening frames with the trim argument:
identity(identity(ctxt_stack()))

#> [[1]]
#> <frame 37> (35)
#> expr: identity(ctxt_stack())
#> env:  [local 0x90233f0]
#> 
#> [[2]]
#> <frame 36> (35)
#> expr: identity(identity(ctxt_stack()))
#> env:  [local 0x9023508]
#> 
#> [[3]]
#> <frame 35> (34)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x812a480]
#> 
#> [[4]]
#> <frame 34> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x90236c8]
#> 
#> [[5]]
#> <frame 33> (24)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x9023930]
#> 
#> [[6]]
#> <frame 32> (24)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x9023ee0]
#> 
#> [[7]]
#> <frame 31> (30)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x9024458]
#> 
#> [[8]]
#> <frame 30> (29)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x90247a0]
#> 
#> [[9]]
#> <frame 29> (28)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x9024ae8]
#> 
#> [[10]]
#> <frame 28> (27)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x9021380]
#> 
#> [[11]]
#> <frame 27> (26)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x90215b0]
#> 
#> [[12]]
#> <frame 26> (24)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x9021738]
#> 
#> [[13]]
#> <frame 25> (24)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x9021818]
#> 
#> [[14]]
#> <frame 24> (23)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x80081e8]
#> 
#> [[15]]
#> <frame 23> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x7acafc8]
#> 
#> [[16]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x7acc688]
#> 
#> [[17]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x92459b0]
#> 
#> [[18]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x81a7c28]
#> 
#> [[19]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x81a85c8]
#> 
#> [[20]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x2621258]
#> 
#> [[21]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x3d49c48]
#> 
#> [[22]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x6c51558]
#> 
#> [[23]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x4fc3208]
#> 
#> [[24]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x184c128]
#> 
#> [[25]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x1cf0930]
#> 
#> [[26]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3558290]
#> 
#> [[27]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3559250]
#> 
#> [[28]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpLvKGyd/file3b3d6016e5ac"),  <...>
#> env:  [frame 0x35490b0]
#> 
#> [[29]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpLvKGyd/file3b3d6016e5ac"),  <...>
#> env:  [frame 0x3549e40]
#> 
#> [[30]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3544c18]
#> 
#> [[31]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3545dd0]
#> 
#> [[32]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x3546118]
#> 
#> [[33]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x35426a0]
#> 
#> [[34]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3542c50]
#> 
#> [[35]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3542f98]
#> 
#> [[36]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x35432e0]
#> 
#> [[37]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3543a50]
#> 
#> [[38]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(ctxt_stack(trim = 1)))

#> [[1]]
#> <frame 35> (34)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x812a480]
#> 
#> [[2]]
#> <frame 34> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x80d8fb8]
#> 
#> [[3]]
#> <frame 33> (24)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x80d8d88]
#> 
#> [[4]]
#> <frame 32> (24)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x80d87d8]
#> 
#> [[5]]
#> <frame 31> (30)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x80d81b8]
#> 
#> [[6]]
#> <frame 30> (29)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x80d7dc8]
#> 
#> [[7]]
#> <frame 29> (28)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x80d7a48]
#> 
#> [[8]]
#> <frame 28> (27)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x83ef608]
#> 
#> [[9]]
#> <frame 27> (26)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x83ef330]
#> 
#> [[10]]
#> <frame 26> (24)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x83ef1a8]
#> 
#> [[11]]
#> <frame 25> (24)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x83ef090]
#> 
#> [[12]]
#> <frame 24> (23)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f0bbb0]
#> 
#> [[13]]
#> <frame 23> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x7acafc8]
#> 
#> [[14]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x7acc688]
#> 
#> [[15]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x92459b0]
#> 
#> [[16]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x81a7c28]
#> 
#> [[17]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x81a85c8]
#> 
#> [[18]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x2621258]
#> 
#> [[19]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x3d49c48]
#> 
#> [[20]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x6c51558]
#> 
#> [[21]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x4fc3208]
#> 
#> [[22]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x184c128]
#> 
#> [[23]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x1cf0930]
#> 
#> [[24]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3558290]
#> 
#> [[25]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3559250]
#> 
#> [[26]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpLvKGyd/file3b3d6016e5ac"),  <...>
#> env:  [frame 0x35490b0]
#> 
#> [[27]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpLvKGyd/file3b3d6016e5ac"),  <...>
#> env:  [frame 0x3549e40]
#> 
#> [[28]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3544c18]
#> 
#> [[29]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3545dd0]
#> 
#> [[30]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x3546118]
#> 
#> [[31]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x35426a0]
#> 
#> [[32]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3542c50]
#> 
#> [[33]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3542f98]
#> 
#> [[34]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x35432e0]
#> 
#> [[35]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3543a50]
#> 
#> [[36]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# ctxt_stack() is called within fn() with intervening frames:
fn <- function(trim) identity(identity(ctxt_stack(trim = trim)))
fn(0)

#> [[1]]
#> <frame 38> (36)
#> expr: identity(ctxt_stack(trim = trim))
#> env:  [local 0x622ab00]
#> 
#> [[2]]
#> <frame 37> (36)
#> expr: identity(identity(ctxt_stack(trim = trim)))
#> env:  [local 0x622a9e8]
#> 
#> [[3]]
#> <frame 36> (35)
#> expr: fn(0)
#> env:  [local 0x622a8d0]
#> 
#> [[4]]
#> <frame 35> (34)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x812a480]
#> 
#> [[5]]
#> <frame 34> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x622d7b0]
#> 
#> [[6]]
#> <frame 33> (24)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x622d580]
#> 
#> [[7]]
#> <frame 32> (24)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x622cf60]
#> 
#> [[8]]
#> <frame 31> (30)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x622c978]
#> 
#> [[9]]
#> <frame 30> (29)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x622c630]
#> 
#> [[10]]
#> <frame 29> (28)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x622c2e8]
#> 
#> [[11]]
#> <frame 28> (27)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x62314d8]
#> 
#> [[12]]
#> <frame 27> (26)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x62312a8]
#> 
#> [[13]]
#> <frame 26> (24)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x6231120]
#> 
#> [[14]]
#> <frame 25> (24)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x6230f98]
#> 
#> [[15]]
#> <frame 24> (23)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x6251e40]
#> 
#> [[16]]
#> <frame 23> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x7acafc8]
#> 
#> [[17]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x7acc688]
#> 
#> [[18]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x92459b0]
#> 
#> [[19]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x81a7c28]
#> 
#> [[20]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x81a85c8]
#> 
#> [[21]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x2621258]
#> 
#> [[22]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x3d49c48]
#> 
#> [[23]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x6c51558]
#> 
#> [[24]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x4fc3208]
#> 
#> [[25]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x184c128]
#> 
#> [[26]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x1cf0930]
#> 
#> [[27]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3558290]
#> 
#> [[28]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3559250]
#> 
#> [[29]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpLvKGyd/file3b3d6016e5ac"),  <...>
#> env:  [frame 0x35490b0]
#> 
#> [[30]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpLvKGyd/file3b3d6016e5ac"),  <...>
#> env:  [frame 0x3549e40]
#> 
#> [[31]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3544c18]
#> 
#> [[32]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3545dd0]
#> 
#> [[33]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x3546118]
#> 
#> [[34]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x35426a0]
#> 
#> [[35]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3542c50]
#> 
#> [[36]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3542f98]
#> 
#> [[37]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x35432e0]
#> 
#> [[38]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3543a50]
#> 
#> [[39]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# We can trim the first layer of those:
fn(1)

#> [[1]]
#> <frame 36> (35)
#> expr: fn(1)
#> env:  [local 0x573e7b0]
#> 
#> [[2]]
#> <frame 35> (34)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x812a480]
#> 
#> [[3]]
#> <frame 34> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x573e5f0]
#> 
#> [[4]]
#> <frame 33> (24)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x573e388]
#> 
#> [[5]]
#> <frame 32> (24)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x5741c08]
#> 
#> [[6]]
#> <frame 31> (30)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x5741658]
#> 
#> [[7]]
#> <frame 30> (29)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x5741310]
#> 
#> [[8]]
#> <frame 29> (28)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x5740f58]
#> 
#> [[9]]
#> <frame 28> (27)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x5740820]
#> 
#> [[10]]
#> <frame 27> (26)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x5740548]
#> 
#> [[11]]
#> <frame 26> (24)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x57403c0]
#> 
#> [[12]]
#> <frame 25> (24)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x57402e0]
#> 
#> [[13]]
#> <frame 24> (23)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x5779f60]
#> 
#> [[14]]
#> <frame 23> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x7acafc8]
#> 
#> [[15]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x7acc688]
#> 
#> [[16]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x92459b0]
#> 
#> [[17]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x81a7c28]
#> 
#> [[18]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x81a85c8]
#> 
#> [[19]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x2621258]
#> 
#> [[20]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x3d49c48]
#> 
#> [[21]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x6c51558]
#> 
#> [[22]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x4fc3208]
#> 
#> [[23]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x184c128]
#> 
#> [[24]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x1cf0930]
#> 
#> [[25]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3558290]
#> 
#> [[26]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3559250]
#> 
#> [[27]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpLvKGyd/file3b3d6016e5ac"),  <...>
#> env:  [frame 0x35490b0]
#> 
#> [[28]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpLvKGyd/file3b3d6016e5ac"),  <...>
#> env:  [frame 0x3549e40]
#> 
#> [[29]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3544c18]
#> 
#> [[30]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3545dd0]
#> 
#> [[31]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x3546118]
#> 
#> [[32]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x35426a0]
#> 
#> [[33]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3542c50]
#> 
#> [[34]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3542f98]
#> 
#> [[35]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x35432e0]
#> 
#> [[36]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3543a50]
#> 
#> [[37]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# The outside intervening frames (at the fn() call site) are still
# returned, but can be trimmed as well:
identity(identity(fn(1)))

#> [[1]]
#> <frame 38> (35)
#> expr: fn(1)
#> env:  [local 0x3d7a970]
#> 
#> [[2]]
#> <frame 37> (35)
#> expr: identity(fn(1))
#> env:  [local 0x3d7a858]
#> 
#> [[3]]
#> <frame 36> (35)
#> expr: identity(identity(fn(1)))
#> env:  [local 0x3d7a708]
#> 
#> [[4]]
#> <frame 35> (34)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x812a480]
#> 
#> [[5]]
#> <frame 34> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x3d7a548]
#> 
#> [[6]]
#> <frame 33> (24)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x3d7a270]
#> 
#> [[7]]
#> <frame 32> (24)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x3d7da80]
#> 
#> [[8]]
#> <frame 31> (30)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x3d7d000]
#> 
#> [[9]]
#> <frame 30> (29)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x3d7cc80]
#> 
#> [[10]]
#> <frame 29> (28)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x3d7c858]
#> 
#> [[11]]
#> <frame 28> (27)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x3d8b4c0]
#> 
#> [[12]]
#> <frame 27> (26)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x3d8b258]
#> 
#> [[13]]
#> <frame 26> (24)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x3d8b0d0]
#> 
#> [[14]]
#> <frame 25> (24)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x3d8aff0]
#> 
#> [[15]]
#> <frame 24> (23)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x3df7ad0]
#> 
#> [[16]]
#> <frame 23> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x7acafc8]
#> 
#> [[17]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x7acc688]
#> 
#> [[18]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x92459b0]
#> 
#> [[19]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x81a7c28]
#> 
#> [[20]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x81a85c8]
#> 
#> [[21]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x2621258]
#> 
#> [[22]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x3d49c48]
#> 
#> [[23]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x6c51558]
#> 
#> [[24]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x4fc3208]
#> 
#> [[25]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x184c128]
#> 
#> [[26]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x1cf0930]
#> 
#> [[27]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3558290]
#> 
#> [[28]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3559250]
#> 
#> [[29]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpLvKGyd/file3b3d6016e5ac"),  <...>
#> env:  [frame 0x35490b0]
#> 
#> [[30]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpLvKGyd/file3b3d6016e5ac"),  <...>
#> env:  [frame 0x3549e40]
#> 
#> [[31]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3544c18]
#> 
#> [[32]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3545dd0]
#> 
#> [[33]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x3546118]
#> 
#> [[34]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x35426a0]
#> 
#> [[35]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3542c50]
#> 
#> [[36]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3542f98]
#> 
#> [[37]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x35432e0]
#> 
#> [[38]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3543a50]
#> 
#> [[39]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(fn(2)))

#> [[1]]
#> <frame 35> (34)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x812a480]
#> 
#> [[2]]
#> <frame 34> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x65016d0]
#> 
#> [[3]]
#> <frame 33> (24)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x6501900]
#> 
#> [[4]]
#> <frame 32> (24)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x6501eb0]
#> 
#> [[5]]
#> <frame 31> (30)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x6502428]
#> 
#> [[6]]
#> <frame 30> (29)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x6502770]
#> 
#> [[7]]
#> <frame 29> (28)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x6502ab8]
#> 
#> [[8]]
#> <frame 28> (27)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x8f7e6a0]
#> 
#> [[9]]
#> <frame 27> (26)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x8f7e8d0]
#> 
#> [[10]]
#> <frame 26> (24)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x8f7ea58]
#> 
#> [[11]]
#> <frame 25> (24)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x8f7eb38]
#> 
#> [[12]]
#> <frame 24> (23)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x881bc50]
#> 
#> [[13]]
#> <frame 23> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x7acafc8]
#> 
#> [[14]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x7acc688]
#> 
#> [[15]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x92459b0]
#> 
#> [[16]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x81a7c28]
#> 
#> [[17]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x81a85c8]
#> 
#> [[18]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x2621258]
#> 
#> [[19]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x3d49c48]
#> 
#> [[20]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x6c51558]
#> 
#> [[21]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x4fc3208]
#> 
#> [[22]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x184c128]
#> 
#> [[23]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x1cf0930]
#> 
#> [[24]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3558290]
#> 
#> [[25]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3559250]
#> 
#> [[26]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpLvKGyd/file3b3d6016e5ac"),  <...>
#> env:  [frame 0x35490b0]
#> 
#> [[27]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpLvKGyd/file3b3d6016e5ac"),  <...>
#> env:  [frame 0x3549e40]
#> 
#> [[28]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3544c18]
#> 
#> [[29]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3545dd0]
#> 
#> [[30]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x3546118]
#> 
#> [[31]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x35426a0]
#> 
#> [[32]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3542c50]
#> 
#> [[33]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3542f98]
#> 
#> [[34]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x35432e0]
#> 
#> [[35]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3543a50]
#> 
#> [[36]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g <- function(trim) identity(identity(fn(trim)))
g(2)

#> [[1]]
#> <frame 36> (35)
#> expr: g(2)
#> env:  [local 0x850d8f8]
#> 
#> [[2]]
#> <frame 35> (34)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x812a480]
#> 
#> [[3]]
#> <frame 34> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x850dab8]
#> 
#> [[4]]
#> <frame 33> (24)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x850dce8]
#> 
#> [[5]]
#> <frame 32> (24)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x850e298]
#> 
#> [[6]]
#> <frame 31> (30)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x850e810]
#> 
#> [[7]]
#> <frame 30> (29)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x850eb58]
#> 
#> [[8]]
#> <frame 29> (28)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x850eea0]
#> 
#> [[9]]
#> <frame 28> (27)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x83bab08]
#> 
#> [[10]]
#> <frame 27> (26)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x83bad38]
#> 
#> [[11]]
#> <frame 26> (24)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x83baec0]
#> 
#> [[12]]
#> <frame 25> (24)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x83bafa0]
#> 
#> [[13]]
#> <frame 24> (23)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x8467ec8]
#> 
#> [[14]]
#> <frame 23> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x7acafc8]
#> 
#> [[15]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x7acc688]
#> 
#> [[16]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x92459b0]
#> 
#> [[17]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x81a7c28]
#> 
#> [[18]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x81a85c8]
#> 
#> [[19]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x2621258]
#> 
#> [[20]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x3d49c48]
#> 
#> [[21]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x6c51558]
#> 
#> [[22]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x4fc3208]
#> 
#> [[23]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x184c128]
#> 
#> [[24]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x1cf0930]
#> 
#> [[25]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3558290]
#> 
#> [[26]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3559250]
#> 
#> [[27]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpLvKGyd/file3b3d6016e5ac"),  <...>
#> env:  [frame 0x35490b0]
#> 
#> [[28]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpLvKGyd/file3b3d6016e5ac"),  <...>
#> env:  [frame 0x3549e40]
#> 
#> [[29]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3544c18]
#> 
#> [[30]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3545dd0]
#> 
#> [[31]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x3546118]
#> 
#> [[32]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x35426a0]
#> 
#> [[33]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3542c50]
#> 
#> [[34]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3542f98]
#> 
#> [[35]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x35432e0]
#> 
#> [[36]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3543a50]
#> 
#> [[37]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g(3)

#> [[1]]
#> <frame 35> (34)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x812a480]
#> 
#> [[2]]
#> <frame 34> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7e5dd08]
#> 
#> [[3]]
#> <frame 33> (24)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7e5daa0]
#> 
#> [[4]]
#> <frame 32> (24)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7e5d4b8]
#> 
#> [[5]]
#> <frame 31> (30)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x74785b0]
#> 
#> [[6]]
#> <frame 30> (29)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7478268]
#> 
#> [[7]]
#> <frame 29> (28)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7477f20]
#> 
#> [[8]]
#> <frame 28> (27)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7477858]
#> 
#> [[9]]
#> <frame 27> (26)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7477628]
#> 
#> [[10]]
#> <frame 26> (24)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x74774a0]
#> 
#> [[11]]
#> <frame 25> (24)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x74773c0]
#> 
#> [[12]]
#> <frame 24> (23)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7c82998]
#> 
#> [[13]]
#> <frame 23> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x7acafc8]
#> 
#> [[14]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x7acc688]
#> 
#> [[15]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x92459b0]
#> 
#> [[16]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x81a7c28]
#> 
#> [[17]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x81a85c8]
#> 
#> [[18]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x2621258]
#> 
#> [[19]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x3d49c48]
#> 
#> [[20]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x6c51558]
#> 
#> [[21]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x4fc3208]
#> 
#> [[22]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x184c128]
#> 
#> [[23]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x1cf0930]
#> 
#> [[24]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3558290]
#> 
#> [[25]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3559250]
#> 
#> [[26]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpLvKGyd/file3b3d6016e5ac"),  <...>
#> env:  [frame 0x35490b0]
#> 
#> [[27]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpLvKGyd/file3b3d6016e5ac"),  <...>
#> env:  [frame 0x3549e40]
#> 
#> [[28]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3544c18]
#> 
#> [[29]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3545dd0]
#> 
#> [[30]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x3546118]
#> 
#> [[31]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x35426a0]
#> 
#> [[32]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3542c50]
#> 
#> [[33]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3542f98]
#> 
#> [[34]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x35432e0]
#> 
#> [[35]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3543a50]
#> 
#> [[36]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"