Call stack information

The eval_ and call_ families of functions provide a replacement for the base R functions prefixed with sys. (which are all about the context stack), as well as for parent.frame() (which is the only base R function for querying the call stack). The context stack includes all R-level evaluation contexts. It is linear in terms of execution history but due to lazy evaluation it is potentially nonlinear in terms of call history. The call stack history, on the other hand, is homogenous.

global_frame()

current_frame()

ctxt_frame(n = 1)

call_frame(n = 1, clean = TRUE)

ctxt_depth()

call_depth()

ctxt_stack(n = NULL, trim = 0)

call_stack(n = NULL, clean = TRUE)

Arguments

n
The number of frames to go back in the stack.

clean
Whether to post-process the call stack to clean non-standard frames. If TRUE, suboptimal call-stack entries by base::eval() will be cleaned up: the duplicate frame created by eval() is eliminated.

trim
The number of layers of intervening frames to trim off the stack. See stack_trim() and examples.

Details

ctxt_frame() and call_frame() return a frame object containing the following fields: expr and env (call expression and evaluation environment), pos and caller_pos (position of current frame in the context stack and position of the caller), and fun (function of the current frame). ctxt_stack() and call_stack() return a list of all context or call frames on the stack. Finally, ctxt_depth() and call_depth() report the current context position or the number of calling frames on the stack.

The base R functions take two sorts of arguments to indicate which frame to query: which and n. The n argument is straightforward: it's the number of frames to go down the stack, with n = 1 referring to the current context. The which argument is more complicated and changes meaning for values lower than 1. For the sake of consistency, the rlang functions all take the same kind of argument n. This argument has a single meaning (the number of frames to go down the stack) and cannot be lower than 1.

Note finally that parent.frame(1) corresponds to call_frame(2)$env, as n = 1 always refers to the current frame. This makes the _frame() and _stack() functions consistent: ctxt_frame(2) is the same as ctxt_stack()[[2]]. Also, ctxt_depth() returns one more frame than base::sys.nframe() because it counts the global frame. That is consistent with the _stack() functions which return the global frame as well. This way, call_stack(call_depth()) is the same as global_frame().

Life cycle

These functions are soft-deprecated and replaced by trace_back().

Examples

# Expressions within arguments count as contexts
identity(identity(ctxt_depth())) # returns 2

#> Warning: `ctxt_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 36

# But they are not part of the call stack because arguments are
# evaluated within the calling function (or the global environment
# if called at top level)
identity(identity(call_depth())) # returns 0

#> Warning: `call_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 13

# The context stacks includes all intervening execution frames. The
# call stack doesn't:
f <- function(x) identity(x)
f(f(ctxt_stack()))

#> [[1]]
#> <frame 37> (36)
#> expr: identity(x)
#> env:  [local 0x97f4b60]
#> 
#> [[2]]
#> <frame 36> (33)
#> expr: f(ctxt_stack())
#> env:  [local 0x97f4c78]
#> 
#> [[3]]
#> <frame 35> (34)
#> expr: identity(x)
#> env:  [local 0x97f4d90]
#> 
#> [[4]]
#> <frame 34> (33)
#> expr: f(f(ctxt_stack()))
#> env:  [local 0x97f4ea8]
#> 
#> [[5]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x71a89a8]
#> 
#> [[6]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x97f5068]
#> 
#> [[7]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x97f5298]
#> 
#> [[8]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x97f5848]
#> 
#> [[9]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x97f5dc0]
#> 
#> [[10]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x97f22d8]
#> 
#> [[11]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x97f2620]
#> 
#> [[12]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x97f2ce8]
#> 
#> [[13]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x97f2f18]
#> 
#> [[14]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x97f30a0]
#> 
#> [[15]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x97f3180]
#> 
#> [[16]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x97e3458]
#> 
#> [[17]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x5d904e0]
#> 
#> [[18]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x6492e18]
#> 
#> [[19]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x69056f8]
#> 
#> [[20]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x2082920]
#> 
#> [[21]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x1a9c0d0]
#> 
#> [[22]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x5b02c38]
#> 
#> [[23]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x5ed9d58]
#> 
#> [[24]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x350d370]
#> 
#> [[25]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x34f4f10]
#> 
#> [[26]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3cbe088]
#> 
#> [[27]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3cbf198]
#> 
#> [[28]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/Rtmps09wZ5/file4f501cd4d1f3"),  <...>
#> env:  [frame 0x322a470]
#> 
#> [[29]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/Rtmps09wZ5/file4f501cd4d1f3"),  <...>
#> env:  [frame 0x322a7f0]
#> 
#> [[30]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3225538]
#> 
#> [[31]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3221d98]
#> 
#> [[32]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x32220e0]
#> 
#> [[33]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3222498]
#> 
#> [[34]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3222a48]
#> 
#> [[35]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3222d90]
#> 
#> [[36]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x32230d8]
#> 
#> [[37]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3223810]
#> 
#> [[38]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(f(call_stack()))

#> Warning: `call_stack()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [[1]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x71a89a8]
#> 
#> [[2]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x722a610]
#> 
#> [[3]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x5d904e0]
#> 
#> [[4]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x6492e18]
#> 
#> [[5]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x69056f8]
#> 
#> [[6]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x2082920]
#> 
#> [[7]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x1a9c0d0]
#> 
#> [[8]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x5b02c38]
#> 
#> [[9]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x5ed9d58]
#> 
#> [[10]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x350d370]
#> 
#> [[11]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x34f4f10]
#> 
#> [[12]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3cbe088]
#> 
#> [[13]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

g <- function(cmd) cmd()
f(g(ctxt_stack))

#> [[1]]
#> <frame 36> (33)
#> expr: g(ctxt_stack)
#> env:  [local 0x8f17368]
#> 
#> [[2]]
#> <frame 35> (34)
#> expr: identity(x)
#> env:  [local 0x8f17480]
#> 
#> [[3]]
#> <frame 34> (33)
#> expr: f(g(ctxt_stack))
#> env:  [local 0x8f17598]
#> 
#> [[4]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x71a89a8]
#> 
#> [[5]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x8f17758]
#> 
#> [[6]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x8f17988]
#> 
#> [[7]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x8f17f70]
#> 
#> [[8]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x8f184e8]
#> 
#> [[9]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x8f18830]
#> 
#> [[10]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x8f18b78]
#> 
#> [[11]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x8f15410]
#> 
#> [[12]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x8f15640]
#> 
#> [[13]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x8f157c8]
#> 
#> [[14]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x8f158a8]
#> 
#> [[15]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x8f05c60]
#> 
#> [[16]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x5d904e0]
#> 
#> [[17]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x6492e18]
#> 
#> [[18]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x69056f8]
#> 
#> [[19]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x2082920]
#> 
#> [[20]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x1a9c0d0]
#> 
#> [[21]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x5b02c38]
#> 
#> [[22]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x5ed9d58]
#> 
#> [[23]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x350d370]
#> 
#> [[24]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x34f4f10]
#> 
#> [[25]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3cbe088]
#> 
#> [[26]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3cbf198]
#> 
#> [[27]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/Rtmps09wZ5/file4f501cd4d1f3"),  <...>
#> env:  [frame 0x322a470]
#> 
#> [[28]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/Rtmps09wZ5/file4f501cd4d1f3"),  <...>
#> env:  [frame 0x322a7f0]
#> 
#> [[29]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3225538]
#> 
#> [[30]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3221d98]
#> 
#> [[31]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x32220e0]
#> 
#> [[32]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3222498]
#> 
#> [[33]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3222a48]
#> 
#> [[34]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3222d90]
#> 
#> [[35]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x32230d8]
#> 
#> [[36]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3223810]
#> 
#> [[37]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(g(call_stack))

#> [[1]]
#> <frame 36> (32)
#> expr: g(call_stack)
#> env:  [local 0x92e3678]
#> 
#> [[2]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x71a89a8]
#> 
#> [[3]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x92d2280]
#> 
#> [[4]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x5d904e0]
#> 
#> [[5]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x6492e18]
#> 
#> [[6]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x69056f8]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x2082920]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x1a9c0d0]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x5b02c38]
#> 
#> [[10]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x5ed9d58]
#> 
#> [[11]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x350d370]
#> 
#> [[12]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x34f4f10]
#> 
#> [[13]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3cbe088]
#> 
#> [[14]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

# The rlang _stack() functions return a list of frame
# objects. Use purrr::transpose() or index a field with
# purrr::map()'s to extract a particular field from a stack:

# stack <- f(f(call_stack()))
# purrr::map(stack, "env")
# purrr::transpose(stack)$expr

# current_frame() is an alias for ctxt_frame(1)
fn <- function() list(current = current_frame(), first = ctxt_frame(1))
fn()

#> Warning: `current_frame()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> $current
#> <frame 34> (33)
#> expr: fn()
#> env:  [local 0x6251668]
#> 
#> $first
#> <frame 34> (33)
#> expr: fn()
#> env:  [local 0x6251668]
#> 

# While current_frame() is the top of the stack, global_frame() is
# the bottom:
fn <- function() {
  n <- ctxt_depth()
  ctxt_frame(n)
}
identical(fn(), global_frame())

#> [1] TRUE

# ctxt_stack() returns a stack with all intervening frames. You can
# trim layers of intervening frames with the trim argument:
identity(identity(ctxt_stack()))

#> [[1]]
#> <frame 35> (33)
#> expr: identity(ctxt_stack())
#> env:  [local 0x647eda8]
#> 
#> [[2]]
#> <frame 34> (33)
#> expr: identity(identity(ctxt_stack()))
#> env:  [local 0x647ec90]
#> 
#> [[3]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x71a89a8]
#> 
#> [[4]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x647ead0]
#> 
#> [[5]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x647e8a0]
#> 
#> [[6]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x6482e50]
#> 
#> [[7]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x6485048]
#> 
#> [[8]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x6484d00]
#> 
#> [[9]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x64849b8]
#> 
#> [[10]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x64842f0]
#> 
#> [[11]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x64840c0]
#> 
#> [[12]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x6483f38]
#> 
#> [[13]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x6483e58]
#> 
#> [[14]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x64e2ab8]
#> 
#> [[15]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x5d904e0]
#> 
#> [[16]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x6492e18]
#> 
#> [[17]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x69056f8]
#> 
#> [[18]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x2082920]
#> 
#> [[19]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x1a9c0d0]
#> 
#> [[20]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x5b02c38]
#> 
#> [[21]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x5ed9d58]
#> 
#> [[22]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x350d370]
#> 
#> [[23]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x34f4f10]
#> 
#> [[24]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3cbe088]
#> 
#> [[25]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3cbf198]
#> 
#> [[26]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/Rtmps09wZ5/file4f501cd4d1f3"),  <...>
#> env:  [frame 0x322a470]
#> 
#> [[27]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/Rtmps09wZ5/file4f501cd4d1f3"),  <...>
#> env:  [frame 0x322a7f0]
#> 
#> [[28]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3225538]
#> 
#> [[29]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3221d98]
#> 
#> [[30]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x32220e0]
#> 
#> [[31]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3222498]
#> 
#> [[32]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3222a48]
#> 
#> [[33]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3222d90]
#> 
#> [[34]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x32230d8]
#> 
#> [[35]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3223810]
#> 
#> [[36]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(ctxt_stack(trim = 1)))

#> [[1]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x71a89a8]
#> 
#> [[2]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5add2e8]
#> 
#> [[3]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x5add0b8]
#> 
#> [[4]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x5adcb08]
#> 
#> [[5]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x5ae0298]
#> 
#> [[6]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x5adff50]
#> 
#> [[7]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x5adfbd0]
#> 
#> [[8]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x5ae2a40]
#> 
#> [[9]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x5ae2618]
#> 
#> [[10]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x5ae2420]
#> 
#> [[11]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x5ae2340]
#> 
#> [[12]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x5b1eb58]
#> 
#> [[13]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x5d904e0]
#> 
#> [[14]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x6492e18]
#> 
#> [[15]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x69056f8]
#> 
#> [[16]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x2082920]
#> 
#> [[17]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x1a9c0d0]
#> 
#> [[18]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x5b02c38]
#> 
#> [[19]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x5ed9d58]
#> 
#> [[20]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x350d370]
#> 
#> [[21]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x34f4f10]
#> 
#> [[22]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3cbe088]
#> 
#> [[23]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3cbf198]
#> 
#> [[24]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/Rtmps09wZ5/file4f501cd4d1f3"),  <...>
#> env:  [frame 0x322a470]
#> 
#> [[25]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/Rtmps09wZ5/file4f501cd4d1f3"),  <...>
#> env:  [frame 0x322a7f0]
#> 
#> [[26]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3225538]
#> 
#> [[27]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3221d98]
#> 
#> [[28]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x32220e0]
#> 
#> [[29]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3222498]
#> 
#> [[30]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3222a48]
#> 
#> [[31]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3222d90]
#> 
#> [[32]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x32230d8]
#> 
#> [[33]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3223810]
#> 
#> [[34]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# ctxt_stack() is called within fn() with intervening frames:
fn <- function(trim) identity(identity(ctxt_stack(trim = trim)))
fn(0)

#> [[1]]
#> <frame 36> (34)
#> expr: identity(ctxt_stack(trim = trim))
#> env:  [local 0x4e8df38]
#> 
#> [[2]]
#> <frame 35> (34)
#> expr: identity(identity(ctxt_stack(trim = trim)))
#> env:  [local 0x4e8de20]
#> 
#> [[3]]
#> <frame 34> (33)
#> expr: fn(0)
#> env:  [local 0x4e8dcd0]
#> 
#> [[4]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x71a89a8]
#> 
#> [[5]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x4e8db10]
#> 
#> [[6]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x4e95410]
#> 
#> [[7]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x4e98c58]
#> 
#> [[8]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x4e986a8]
#> 
#> [[9]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x4e9cc58]
#> 
#> [[10]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x4e9c910]
#> 
#> [[11]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x4e9bb80]
#> 
#> [[12]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x4e9b918]
#> 
#> [[13]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x4e9b790]
#> 
#> [[14]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x4ea0830]
#> 
#> [[15]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x4ebee48]
#> 
#> [[16]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x5d904e0]
#> 
#> [[17]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x6492e18]
#> 
#> [[18]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x69056f8]
#> 
#> [[19]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x2082920]
#> 
#> [[20]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x1a9c0d0]
#> 
#> [[21]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x5b02c38]
#> 
#> [[22]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x5ed9d58]
#> 
#> [[23]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x350d370]
#> 
#> [[24]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x34f4f10]
#> 
#> [[25]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3cbe088]
#> 
#> [[26]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3cbf198]
#> 
#> [[27]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/Rtmps09wZ5/file4f501cd4d1f3"),  <...>
#> env:  [frame 0x322a470]
#> 
#> [[28]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/Rtmps09wZ5/file4f501cd4d1f3"),  <...>
#> env:  [frame 0x322a7f0]
#> 
#> [[29]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3225538]
#> 
#> [[30]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3221d98]
#> 
#> [[31]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x32220e0]
#> 
#> [[32]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3222498]
#> 
#> [[33]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3222a48]
#> 
#> [[34]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3222d90]
#> 
#> [[35]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x32230d8]
#> 
#> [[36]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3223810]
#> 
#> [[37]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# We can trim the first layer of those:
fn(1)

#> [[1]]
#> <frame 34> (33)
#> expr: fn(1)
#> env:  [local 0x3eb1db8]
#> 
#> [[2]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x71a89a8]
#> 
#> [[3]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x3eb1bf8]
#> 
#> [[4]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x3eb19c8]
#> 
#> [[5]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x3eb1418]
#> 
#> [[6]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x3ea17a0]
#> 
#> [[7]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x3ea1458]
#> 
#> [[8]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x3ea1110]
#> 
#> [[9]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x3ea0a48]
#> 
#> [[10]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x3ea0818]
#> 
#> [[11]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x3ea0690]
#> 
#> [[12]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x3ea05b0]
#> 
#> [[13]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x3ecec38]
#> 
#> [[14]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x5d904e0]
#> 
#> [[15]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x6492e18]
#> 
#> [[16]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x69056f8]
#> 
#> [[17]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x2082920]
#> 
#> [[18]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x1a9c0d0]
#> 
#> [[19]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x5b02c38]
#> 
#> [[20]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x5ed9d58]
#> 
#> [[21]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x350d370]
#> 
#> [[22]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x34f4f10]
#> 
#> [[23]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3cbe088]
#> 
#> [[24]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3cbf198]
#> 
#> [[25]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/Rtmps09wZ5/file4f501cd4d1f3"),  <...>
#> env:  [frame 0x322a470]
#> 
#> [[26]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/Rtmps09wZ5/file4f501cd4d1f3"),  <...>
#> env:  [frame 0x322a7f0]
#> 
#> [[27]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3225538]
#> 
#> [[28]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3221d98]
#> 
#> [[29]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x32220e0]
#> 
#> [[30]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3222498]
#> 
#> [[31]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3222a48]
#> 
#> [[32]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3222d90]
#> 
#> [[33]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x32230d8]
#> 
#> [[34]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3223810]
#> 
#> [[35]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# The outside intervening frames (at the fn() call site) are still
# returned, but can be trimmed as well:
identity(identity(fn(1)))

#> [[1]]
#> <frame 36> (33)
#> expr: fn(1)
#> env:  [local 0x1b7a4e0]
#> 
#> [[2]]
#> <frame 35> (33)
#> expr: identity(fn(1))
#> env:  [local 0x1b7a3c8]
#> 
#> [[3]]
#> <frame 34> (33)
#> expr: identity(identity(fn(1)))
#> env:  [local 0x1b7a278]
#> 
#> [[4]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x71a89a8]
#> 
#> [[5]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x1b7a0b8]
#> 
#> [[6]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x1b79c58]
#> 
#> [[7]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x1b7d468]
#> 
#> [[8]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x1b7ce80]
#> 
#> [[9]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x1b7cb38]
#> 
#> [[10]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x1b7c780]
#> 
#> [[11]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x1b7bd70]
#> 
#> [[12]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x1b7bb40]
#> 
#> [[13]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x1b7b948]
#> 
#> [[14]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x1b7b868]
#> 
#> [[15]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x2f32958]
#> 
#> [[16]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x5d904e0]
#> 
#> [[17]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x6492e18]
#> 
#> [[18]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x69056f8]
#> 
#> [[19]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x2082920]
#> 
#> [[20]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x1a9c0d0]
#> 
#> [[21]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x5b02c38]
#> 
#> [[22]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x5ed9d58]
#> 
#> [[23]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x350d370]
#> 
#> [[24]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x34f4f10]
#> 
#> [[25]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3cbe088]
#> 
#> [[26]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3cbf198]
#> 
#> [[27]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/Rtmps09wZ5/file4f501cd4d1f3"),  <...>
#> env:  [frame 0x322a470]
#> 
#> [[28]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/Rtmps09wZ5/file4f501cd4d1f3"),  <...>
#> env:  [frame 0x322a7f0]
#> 
#> [[29]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3225538]
#> 
#> [[30]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3221d98]
#> 
#> [[31]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x32220e0]
#> 
#> [[32]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3222498]
#> 
#> [[33]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3222a48]
#> 
#> [[34]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3222d90]
#> 
#> [[35]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x32230d8]
#> 
#> [[36]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3223810]
#> 
#> [[37]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(fn(2)))

#> [[1]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x71a89a8]
#> 
#> [[2]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x62aa050]
#> 
#> [[3]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x62aa280]
#> 
#> [[4]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x4d64ca0]
#> 
#> [[5]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x4d65218]
#> 
#> [[6]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x4d65560]
#> 
#> [[7]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x4d658a8]
#> 
#> [[8]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x4d65f70]
#> 
#> [[9]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x4d661a0]
#> 
#> [[10]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x4d66328]
#> 
#> [[11]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x4d66408]
#> 
#> [[12]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x5ac2820]
#> 
#> [[13]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x5d904e0]
#> 
#> [[14]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x6492e18]
#> 
#> [[15]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x69056f8]
#> 
#> [[16]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x2082920]
#> 
#> [[17]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x1a9c0d0]
#> 
#> [[18]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x5b02c38]
#> 
#> [[19]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x5ed9d58]
#> 
#> [[20]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x350d370]
#> 
#> [[21]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x34f4f10]
#> 
#> [[22]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3cbe088]
#> 
#> [[23]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3cbf198]
#> 
#> [[24]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/Rtmps09wZ5/file4f501cd4d1f3"),  <...>
#> env:  [frame 0x322a470]
#> 
#> [[25]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/Rtmps09wZ5/file4f501cd4d1f3"),  <...>
#> env:  [frame 0x322a7f0]
#> 
#> [[26]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3225538]
#> 
#> [[27]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3221d98]
#> 
#> [[28]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x32220e0]
#> 
#> [[29]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3222498]
#> 
#> [[30]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3222a48]
#> 
#> [[31]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3222d90]
#> 
#> [[32]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x32230d8]
#> 
#> [[33]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3223810]
#> 
#> [[34]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g <- function(trim) identity(identity(fn(trim)))
g(2)

#> [[1]]
#> <frame 34> (33)
#> expr: g(2)
#> env:  [local 0x6cdd7c0]
#> 
#> [[2]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x71a89a8]
#> 
#> [[3]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x6cdd980]
#> 
#> [[4]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x6cddbb0]
#> 
#> [[5]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x6cde160]
#> 
#> [[6]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x667f078]
#> 
#> [[7]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x667f3c0]
#> 
#> [[8]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x667f708]
#> 
#> [[9]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x667fdd0]
#> 
#> [[10]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x6680000]
#> 
#> [[11]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x6680188]
#> 
#> [[12]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x6680268]
#> 
#> [[13]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x74b9dc8]
#> 
#> [[14]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x5d904e0]
#> 
#> [[15]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x6492e18]
#> 
#> [[16]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x69056f8]
#> 
#> [[17]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x2082920]
#> 
#> [[18]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x1a9c0d0]
#> 
#> [[19]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x5b02c38]
#> 
#> [[20]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x5ed9d58]
#> 
#> [[21]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x350d370]
#> 
#> [[22]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x34f4f10]
#> 
#> [[23]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3cbe088]
#> 
#> [[24]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3cbf198]
#> 
#> [[25]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/Rtmps09wZ5/file4f501cd4d1f3"),  <...>
#> env:  [frame 0x322a470]
#> 
#> [[26]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/Rtmps09wZ5/file4f501cd4d1f3"),  <...>
#> env:  [frame 0x322a7f0]
#> 
#> [[27]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3225538]
#> 
#> [[28]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3221d98]
#> 
#> [[29]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x32220e0]
#> 
#> [[30]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3222498]
#> 
#> [[31]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3222a48]
#> 
#> [[32]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3222d90]
#> 
#> [[33]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x32230d8]
#> 
#> [[34]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3223810]
#> 
#> [[35]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g(3)

#> [[1]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x71a89a8]
#> 
#> [[2]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x52b1170]
#> 
#> [[3]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x52b13a0]
#> 
#> [[4]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x52b1950]
#> 
#> [[5]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x52b1ec8]
#> 
#> [[6]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x52b2210]
#> 
#> [[7]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x52ae728]
#> 
#> [[8]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x52aedf0]
#> 
#> [[9]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x52af020]
#> 
#> [[10]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x52af1a8]
#> 
#> [[11]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x52af288]
#> 
#> [[12]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x4de17a0]
#> 
#> [[13]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x5d904e0]
#> 
#> [[14]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x6492e18]
#> 
#> [[15]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x69056f8]
#> 
#> [[16]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x2082920]
#> 
#> [[17]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x1a9c0d0]
#> 
#> [[18]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x5b02c38]
#> 
#> [[19]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x5ed9d58]
#> 
#> [[20]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x350d370]
#> 
#> [[21]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x34f4f10]
#> 
#> [[22]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3cbe088]
#> 
#> [[23]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3cbf198]
#> 
#> [[24]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/Rtmps09wZ5/file4f501cd4d1f3"),  <...>
#> env:  [frame 0x322a470]
#> 
#> [[25]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/Rtmps09wZ5/file4f501cd4d1f3"),  <...>
#> env:  [frame 0x322a7f0]
#> 
#> [[26]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3225538]
#> 
#> [[27]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3221d98]
#> 
#> [[28]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x32220e0]
#> 
#> [[29]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3222498]
#> 
#> [[30]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3222a48]
#> 
#> [[31]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3222d90]
#> 
#> [[32]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x32230d8]
#> 
#> [[33]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3223810]
#> 
#> [[34]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"