Call stack information

The eval_ and call_ families of functions provide a replacement for the base R functions prefixed with sys. (which are all about the context stack), as well as for parent.frame() (which is the only base R function for querying the call stack). The context stack includes all R-level evaluation contexts. It is linear in terms of execution history but due to lazy evaluation it is potentially nonlinear in terms of call history. The call stack history, on the other hand, is homogenous.

global_frame()

current_frame()

ctxt_frame(n = 1)

call_frame(n = 1, clean = TRUE)

ctxt_depth()

call_depth()

ctxt_stack(n = NULL, trim = 0)

call_stack(n = NULL, clean = TRUE)

Arguments

n
The number of frames to go back in the stack.

clean
Whether to post-process the call stack to clean non-standard frames. If TRUE, suboptimal call-stack entries by base::eval() will be cleaned up: the duplicate frame created by eval() is eliminated.

trim
The number of layers of intervening frames to trim off the stack. See stack_trim() and examples.

Details

ctxt_frame() and call_frame() return a frame object containing the following fields: expr and env (call expression and evaluation environment), pos and caller_pos (position of current frame in the context stack and position of the caller), and fun (function of the current frame). ctxt_stack() and call_stack() return a list of all context or call frames on the stack. Finally, ctxt_depth() and call_depth() report the current context position or the number of calling frames on the stack.

The base R functions take two sorts of arguments to indicate which frame to query: which and n. The n argument is straightforward: it's the number of frames to go down the stack, with n = 1 referring to the current context. The which argument is more complicated and changes meaning for values lower than 1. For the sake of consistency, the rlang functions all take the same kind of argument n. This argument has a single meaning (the number of frames to go down the stack) and cannot be lower than 1.

Note finally that parent.frame(1) corresponds to call_frame(2)$env, as n = 1 always refers to the current frame. This makes the _frame() and _stack() functions consistent: ctxt_frame(2) is the same as ctxt_stack()[[2]]. Also, ctxt_depth() returns one more frame than base::sys.nframe() because it counts the global frame. That is consistent with the _stack() functions which return the global frame as well. This way, call_stack(call_depth()) is the same as global_frame().

Life cycle

These functions are soft-deprecated and replaced by trace_back().

Examples

# Expressions within arguments count as contexts
identity(identity(ctxt_depth())) # returns 2

#> Warning: `ctxt_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 36

# But they are not part of the call stack because arguments are
# evaluated within the calling function (or the global environment
# if called at top level)
identity(identity(call_depth())) # returns 0

#> Warning: `call_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 13

# The context stacks includes all intervening execution frames. The
# call stack doesn't:
f <- function(x) identity(x)
f(f(ctxt_stack()))

#> [[1]]
#> <frame 37> (36)
#> expr: identity(x)
#> env:  [local 0x903dc80]
#> 
#> [[2]]
#> <frame 36> (33)
#> expr: f(ctxt_stack())
#> env:  [local 0x903dd98]
#> 
#> [[3]]
#> <frame 35> (34)
#> expr: identity(x)
#> env:  [local 0x903deb0]
#> 
#> [[4]]
#> <frame 34> (33)
#> expr: f(f(ctxt_stack()))
#> env:  [local 0x903dfc8]
#> 
#> [[5]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x915e708]
#> 
#> [[6]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x903e188]
#> 
#> [[7]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x903e3b8]
#> 
#> [[8]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x903e968]
#> 
#> [[9]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x903eee0]
#> 
#> [[10]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x903f228]
#> 
#> [[11]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x903f570]
#> 
#> [[12]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x903be08]
#> 
#> [[13]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x903c038]
#> 
#> [[14]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x903c1c0]
#> 
#> [[15]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x903c2a0]
#> 
#> [[16]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x8ff2028]
#> 
#> [[17]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x55c8508]
#> 
#> [[18]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x5b7edf8]
#> 
#> [[19]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x49f3f28]
#> 
#> [[20]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x32332e0]
#> 
#> [[21]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x33d19d8]
#> 
#> [[22]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x628fb98]
#> 
#> [[23]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x6237b90]
#> 
#> [[24]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x34d9e00]
#> 
#> [[25]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x34d6ed8]
#> 
#> [[26]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x31cfdd8]
#> 
#> [[27]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3c63428]
#> 
#> [[28]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpYRBQpT/file506ea8b0436"),  <...>
#> env:  [frame 0x31cb470]
#> 
#> [[29]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpYRBQpT/file506ea8b0436"),  <...>
#> env:  [frame 0x31cb7f0]
#> 
#> [[30]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x31c6538]
#> 
#> [[31]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x31c2d98]
#> 
#> [[32]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x31c30e0]
#> 
#> [[33]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x31c3498]
#> 
#> [[34]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x31c3a48]
#> 
#> [[35]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x31c3d90]
#> 
#> [[36]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x31c40d8]
#> 
#> [[37]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x31c4810]
#> 
#> [[38]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(f(call_stack()))

#> Warning: `call_stack()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [[1]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x915e708]
#> 
#> [[2]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x945ec58]
#> 
#> [[3]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x55c8508]
#> 
#> [[4]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x5b7edf8]
#> 
#> [[5]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x49f3f28]
#> 
#> [[6]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x32332e0]
#> 
#> [[7]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x33d19d8]
#> 
#> [[8]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x628fb98]
#> 
#> [[9]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x6237b90]
#> 
#> [[10]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x34d9e00]
#> 
#> [[11]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x34d6ed8]
#> 
#> [[12]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x31cfdd8]
#> 
#> [[13]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

g <- function(cmd) cmd()
f(g(ctxt_stack))

#> [[1]]
#> <frame 36> (33)
#> expr: g(ctxt_stack)
#> env:  [local 0x92cc820]
#> 
#> [[2]]
#> <frame 35> (34)
#> expr: identity(x)
#> env:  [local 0x92cc970]
#> 
#> [[3]]
#> <frame 34> (33)
#> expr: f(g(ctxt_stack))
#> env:  [local 0x92cca88]
#> 
#> [[4]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x915e708]
#> 
#> [[5]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x92ccc48]
#> 
#> [[6]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x92cce78]
#> 
#> [[7]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x92c95f8]
#> 
#> [[8]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x92c9b70]
#> 
#> [[9]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x92c9eb8]
#> 
#> [[10]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x92ca200]
#> 
#> [[11]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x92ca8c8]
#> 
#> [[12]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x92caaf8]
#> 
#> [[13]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x92cac80]
#> 
#> [[14]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x92cad60]
#> 
#> [[15]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x92bb038]
#> 
#> [[16]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x55c8508]
#> 
#> [[17]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x5b7edf8]
#> 
#> [[18]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x49f3f28]
#> 
#> [[19]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x32332e0]
#> 
#> [[20]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x33d19d8]
#> 
#> [[21]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x628fb98]
#> 
#> [[22]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x6237b90]
#> 
#> [[23]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x34d9e00]
#> 
#> [[24]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x34d6ed8]
#> 
#> [[25]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x31cfdd8]
#> 
#> [[26]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3c63428]
#> 
#> [[27]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpYRBQpT/file506ea8b0436"),  <...>
#> env:  [frame 0x31cb470]
#> 
#> [[28]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpYRBQpT/file506ea8b0436"),  <...>
#> env:  [frame 0x31cb7f0]
#> 
#> [[29]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x31c6538]
#> 
#> [[30]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x31c2d98]
#> 
#> [[31]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x31c30e0]
#> 
#> [[32]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x31c3498]
#> 
#> [[33]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x31c3a48]
#> 
#> [[34]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x31c3d90]
#> 
#> [[35]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x31c40d8]
#> 
#> [[36]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x31c4810]
#> 
#> [[37]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(g(call_stack))

#> [[1]]
#> <frame 36> (32)
#> expr: g(call_stack)
#> env:  [local 0x9321448]
#> 
#> [[2]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x915e708]
#> 
#> [[3]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x930c0d0]
#> 
#> [[4]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x55c8508]
#> 
#> [[5]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x5b7edf8]
#> 
#> [[6]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x49f3f28]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x32332e0]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x33d19d8]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x628fb98]
#> 
#> [[10]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x6237b90]
#> 
#> [[11]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x34d9e00]
#> 
#> [[12]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x34d6ed8]
#> 
#> [[13]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x31cfdd8]
#> 
#> [[14]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

# The rlang _stack() functions return a list of frame
# objects. Use purrr::transpose() or index a field with
# purrr::map()'s to extract a particular field from a stack:

# stack <- f(f(call_stack()))
# purrr::map(stack, "env")
# purrr::transpose(stack)$expr

# current_frame() is an alias for ctxt_frame(1)
fn <- function() list(current = current_frame(), first = ctxt_frame(1))
fn()

#> Warning: `current_frame()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> $current
#> <frame 34> (33)
#> expr: fn()
#> env:  [local 0x9033c70]
#> 
#> $first
#> <frame 34> (33)
#> expr: fn()
#> env:  [local 0x9033c70]
#> 

# While current_frame() is the top of the stack, global_frame() is
# the bottom:
fn <- function() {
  n <- ctxt_depth()
  ctxt_frame(n)
}
identical(fn(), global_frame())

#> [1] TRUE

# ctxt_stack() returns a stack with all intervening frames. You can
# trim layers of intervening frames with the trim argument:
identity(identity(ctxt_stack()))

#> [[1]]
#> <frame 35> (33)
#> expr: identity(ctxt_stack())
#> env:  [local 0x55b3308]
#> 
#> [[2]]
#> <frame 34> (33)
#> expr: identity(identity(ctxt_stack()))
#> env:  [local 0x55b31f0]
#> 
#> [[3]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x915e708]
#> 
#> [[4]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x55b3030]
#> 
#> [[5]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x55b2e00]
#> 
#> [[6]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x566b010]
#> 
#> [[7]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x566aa98]
#> 
#> [[8]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x566a478]
#> 
#> [[9]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x566a0c0]
#> 
#> [[10]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x56699f8]
#> 
#> [[11]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x56697c8]
#> 
#> [[12]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x5669640]
#> 
#> [[13]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x5d1af00]
#> 
#> [[14]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x4bb0348]
#> 
#> [[15]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x55c8508]
#> 
#> [[16]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x5b7edf8]
#> 
#> [[17]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x49f3f28]
#> 
#> [[18]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x32332e0]
#> 
#> [[19]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x33d19d8]
#> 
#> [[20]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x628fb98]
#> 
#> [[21]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x6237b90]
#> 
#> [[22]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x34d9e00]
#> 
#> [[23]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x34d6ed8]
#> 
#> [[24]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x31cfdd8]
#> 
#> [[25]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3c63428]
#> 
#> [[26]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpYRBQpT/file506ea8b0436"),  <...>
#> env:  [frame 0x31cb470]
#> 
#> [[27]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpYRBQpT/file506ea8b0436"),  <...>
#> env:  [frame 0x31cb7f0]
#> 
#> [[28]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x31c6538]
#> 
#> [[29]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x31c2d98]
#> 
#> [[30]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x31c30e0]
#> 
#> [[31]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x31c3498]
#> 
#> [[32]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x31c3a48]
#> 
#> [[33]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x31c3d90]
#> 
#> [[34]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x31c40d8]
#> 
#> [[35]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x31c4810]
#> 
#> [[36]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(ctxt_stack(trim = 1)))

#> [[1]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x915e708]
#> 
#> [[2]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x60f6020]
#> 
#> [[3]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x60f5d48]
#> 
#> [[4]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x60f5798]
#> 
#> [[5]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x60f9018]
#> 
#> [[6]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x60f8cd0]
#> 
#> [[7]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x60f8950]
#> 
#> [[8]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x6139d00]
#> 
#> [[9]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x6139ad0]
#> 
#> [[10]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x6139948]
#> 
#> [[11]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x6139868]
#> 
#> [[12]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x617e458]
#> 
#> [[13]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x55c8508]
#> 
#> [[14]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x5b7edf8]
#> 
#> [[15]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x49f3f28]
#> 
#> [[16]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x32332e0]
#> 
#> [[17]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x33d19d8]
#> 
#> [[18]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x628fb98]
#> 
#> [[19]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x6237b90]
#> 
#> [[20]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x34d9e00]
#> 
#> [[21]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x34d6ed8]
#> 
#> [[22]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x31cfdd8]
#> 
#> [[23]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3c63428]
#> 
#> [[24]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpYRBQpT/file506ea8b0436"),  <...>
#> env:  [frame 0x31cb470]
#> 
#> [[25]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpYRBQpT/file506ea8b0436"),  <...>
#> env:  [frame 0x31cb7f0]
#> 
#> [[26]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x31c6538]
#> 
#> [[27]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x31c2d98]
#> 
#> [[28]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x31c30e0]
#> 
#> [[29]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x31c3498]
#> 
#> [[30]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x31c3a48]
#> 
#> [[31]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x31c3d90]
#> 
#> [[32]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x31c40d8]
#> 
#> [[33]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x31c4810]
#> 
#> [[34]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# ctxt_stack() is called within fn() with intervening frames:
fn <- function(trim) identity(identity(ctxt_stack(trim = trim)))
fn(0)

#> [[1]]
#> <frame 36> (34)
#> expr: identity(ctxt_stack(trim = trim))
#> env:  [local 0x54a52b0]
#> 
#> [[2]]
#> <frame 35> (34)
#> expr: identity(identity(ctxt_stack(trim = trim)))
#> env:  [local 0x54a5198]
#> 
#> [[3]]
#> <frame 34> (33)
#> expr: fn(0)
#> env:  [local 0x54a8e78]
#> 
#> [[4]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x915e708]
#> 
#> [[5]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x54a8cb8]
#> 
#> [[6]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x54a8a88]
#> 
#> [[7]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x54a84d8]
#> 
#> [[8]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x54a7f28]
#> 
#> [[9]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x54a7be0]
#> 
#> [[10]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x54a7898]
#> 
#> [[11]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x54a71d0]
#> 
#> [[12]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x54abd30]
#> 
#> [[13]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x54abba8]
#> 
#> [[14]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x54abac8]
#> 
#> [[15]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x54d8b50]
#> 
#> [[16]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x55c8508]
#> 
#> [[17]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x5b7edf8]
#> 
#> [[18]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x49f3f28]
#> 
#> [[19]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x32332e0]
#> 
#> [[20]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x33d19d8]
#> 
#> [[21]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x628fb98]
#> 
#> [[22]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x6237b90]
#> 
#> [[23]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x34d9e00]
#> 
#> [[24]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x34d6ed8]
#> 
#> [[25]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x31cfdd8]
#> 
#> [[26]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3c63428]
#> 
#> [[27]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpYRBQpT/file506ea8b0436"),  <...>
#> env:  [frame 0x31cb470]
#> 
#> [[28]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpYRBQpT/file506ea8b0436"),  <...>
#> env:  [frame 0x31cb7f0]
#> 
#> [[29]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x31c6538]
#> 
#> [[30]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x31c2d98]
#> 
#> [[31]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x31c30e0]
#> 
#> [[32]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x31c3498]
#> 
#> [[33]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x31c3a48]
#> 
#> [[34]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x31c3d90]
#> 
#> [[35]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x31c40d8]
#> 
#> [[36]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x31c4810]
#> 
#> [[37]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# We can trim the first layer of those:
fn(1)

#> [[1]]
#> <frame 34> (33)
#> expr: fn(1)
#> env:  [local 0x4c06d28]
#> 
#> [[2]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x915e708]
#> 
#> [[3]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x4c06b68]
#> 
#> [[4]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x4c06938]
#> 
#> [[5]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x4c06318]
#> 
#> [[6]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x4c05da0]
#> 
#> [[7]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x4c05a58]
#> 
#> [[8]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x4c0b460]
#> 
#> [[9]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x4c0ad60]
#> 
#> [[10]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x4c0ab30]
#> 
#> [[11]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x4c0a9a8]
#> 
#> [[12]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x4c0a890]
#> 
#> [[13]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x4c24fb0]
#> 
#> [[14]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x55c8508]
#> 
#> [[15]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x5b7edf8]
#> 
#> [[16]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x49f3f28]
#> 
#> [[17]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x32332e0]
#> 
#> [[18]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x33d19d8]
#> 
#> [[19]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x628fb98]
#> 
#> [[20]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x6237b90]
#> 
#> [[21]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x34d9e00]
#> 
#> [[22]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x34d6ed8]
#> 
#> [[23]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x31cfdd8]
#> 
#> [[24]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3c63428]
#> 
#> [[25]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpYRBQpT/file506ea8b0436"),  <...>
#> env:  [frame 0x31cb470]
#> 
#> [[26]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpYRBQpT/file506ea8b0436"),  <...>
#> env:  [frame 0x31cb7f0]
#> 
#> [[27]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x31c6538]
#> 
#> [[28]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x31c2d98]
#> 
#> [[29]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x31c30e0]
#> 
#> [[30]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x31c3498]
#> 
#> [[31]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x31c3a48]
#> 
#> [[32]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x31c3d90]
#> 
#> [[33]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x31c40d8]
#> 
#> [[34]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x31c4810]
#> 
#> [[35]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# The outside intervening frames (at the fn() call site) are still
# returned, but can be trimmed as well:
identity(identity(fn(1)))

#> [[1]]
#> <frame 36> (33)
#> expr: fn(1)
#> env:  [local 0x35d6338]
#> 
#> [[2]]
#> <frame 35> (33)
#> expr: identity(fn(1))
#> env:  [local 0x35d6220]
#> 
#> [[3]]
#> <frame 34> (33)
#> expr: identity(identity(fn(1)))
#> env:  [local 0x35d6108]
#> 
#> [[4]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x915e708]
#> 
#> [[5]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x35d5ea0]
#> 
#> [[6]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x35d5c70]
#> 
#> [[7]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x35d56c0]
#> 
#> [[8]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x35d8f78]
#> 
#> [[9]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x35d8c30]
#> 
#> [[10]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x35d8680]
#> 
#> [[11]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x35d7f10]
#> 
#> [[12]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x35d7ce0]
#> 
#> [[13]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x35d7b58]
#> 
#> [[14]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x35d7a78]
#> 
#> [[15]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x35f16a8]
#> 
#> [[16]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x55c8508]
#> 
#> [[17]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x5b7edf8]
#> 
#> [[18]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x49f3f28]
#> 
#> [[19]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x32332e0]
#> 
#> [[20]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x33d19d8]
#> 
#> [[21]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x628fb98]
#> 
#> [[22]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x6237b90]
#> 
#> [[23]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x34d9e00]
#> 
#> [[24]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x34d6ed8]
#> 
#> [[25]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x31cfdd8]
#> 
#> [[26]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3c63428]
#> 
#> [[27]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpYRBQpT/file506ea8b0436"),  <...>
#> env:  [frame 0x31cb470]
#> 
#> [[28]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpYRBQpT/file506ea8b0436"),  <...>
#> env:  [frame 0x31cb7f0]
#> 
#> [[29]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x31c6538]
#> 
#> [[30]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x31c2d98]
#> 
#> [[31]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x31c30e0]
#> 
#> [[32]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x31c3498]
#> 
#> [[33]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x31c3a48]
#> 
#> [[34]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x31c3d90]
#> 
#> [[35]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x31c40d8]
#> 
#> [[36]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x31c4810]
#> 
#> [[37]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(fn(2)))

#> [[1]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x915e708]
#> 
#> [[2]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x2122588]
#> 
#> [[3]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x21222b0]
#> 
#> [[4]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x2121910]
#> 
#> [[5]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x21212f0]
#> 
#> [[6]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x2126b70]
#> 
#> [[7]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x2126780]
#> 
#> [[8]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x2126010]
#> 
#> [[9]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x2125de0]
#> 
#> [[10]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x2125c58]
#> 
#> [[11]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x2125b40]
#> 
#> [[12]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x22e6180]
#> 
#> [[13]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x55c8508]
#> 
#> [[14]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x5b7edf8]
#> 
#> [[15]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x49f3f28]
#> 
#> [[16]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x32332e0]
#> 
#> [[17]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x33d19d8]
#> 
#> [[18]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x628fb98]
#> 
#> [[19]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x6237b90]
#> 
#> [[20]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x34d9e00]
#> 
#> [[21]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x34d6ed8]
#> 
#> [[22]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x31cfdd8]
#> 
#> [[23]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3c63428]
#> 
#> [[24]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpYRBQpT/file506ea8b0436"),  <...>
#> env:  [frame 0x31cb470]
#> 
#> [[25]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpYRBQpT/file506ea8b0436"),  <...>
#> env:  [frame 0x31cb7f0]
#> 
#> [[26]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x31c6538]
#> 
#> [[27]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x31c2d98]
#> 
#> [[28]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x31c30e0]
#> 
#> [[29]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x31c3498]
#> 
#> [[30]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x31c3a48]
#> 
#> [[31]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x31c3d90]
#> 
#> [[32]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x31c40d8]
#> 
#> [[33]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x31c4810]
#> 
#> [[34]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g <- function(trim) identity(identity(fn(trim)))
g(2)

#> [[1]]
#> <frame 34> (33)
#> expr: g(2)
#> env:  [local 0x53fea70]
#> 
#> [[2]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x915e708]
#> 
#> [[3]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x53fec30]
#> 
#> [[4]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x53fee60]
#> 
#> [[5]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x53ff410]
#> 
#> [[6]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x53ff988]
#> 
#> [[7]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x53ffcd0]
#> 
#> [[8]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x5400018]
#> 
#> [[9]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x538be90]
#> 
#> [[10]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x538c0c0]
#> 
#> [[11]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x538c248]
#> 
#> [[12]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x538c328]
#> 
#> [[13]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x5250f60]
#> 
#> [[14]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x55c8508]
#> 
#> [[15]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x5b7edf8]
#> 
#> [[16]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x49f3f28]
#> 
#> [[17]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x32332e0]
#> 
#> [[18]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x33d19d8]
#> 
#> [[19]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x628fb98]
#> 
#> [[20]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x6237b90]
#> 
#> [[21]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x34d9e00]
#> 
#> [[22]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x34d6ed8]
#> 
#> [[23]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x31cfdd8]
#> 
#> [[24]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3c63428]
#> 
#> [[25]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpYRBQpT/file506ea8b0436"),  <...>
#> env:  [frame 0x31cb470]
#> 
#> [[26]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpYRBQpT/file506ea8b0436"),  <...>
#> env:  [frame 0x31cb7f0]
#> 
#> [[27]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x31c6538]
#> 
#> [[28]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x31c2d98]
#> 
#> [[29]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x31c30e0]
#> 
#> [[30]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x31c3498]
#> 
#> [[31]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x31c3a48]
#> 
#> [[32]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x31c3d90]
#> 
#> [[33]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x31c40d8]
#> 
#> [[34]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x31c4810]
#> 
#> [[35]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g(3)

#> [[1]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x915e708]
#> 
#> [[2]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x6449d80]
#> 
#> [[3]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x4a77530]
#> 
#> [[4]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x4a77ae0]
#> 
#> [[5]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x4a78058]
#> 
#> [[6]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x4a783a0]
#> 
#> [[7]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x4a786e8]
#> 
#> [[8]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x4a78db0]
#> 
#> [[9]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x4a78fe0]
#> 
#> [[10]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x4a79168]
#> 
#> [[11]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x4a79248]
#> 
#> [[12]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x6102bf0]
#> 
#> [[13]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x55c8508]
#> 
#> [[14]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x5b7edf8]
#> 
#> [[15]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x49f3f28]
#> 
#> [[16]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x32332e0]
#> 
#> [[17]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x33d19d8]
#> 
#> [[18]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x628fb98]
#> 
#> [[19]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x6237b90]
#> 
#> [[20]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x34d9e00]
#> 
#> [[21]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x34d6ed8]
#> 
#> [[22]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x31cfdd8]
#> 
#> [[23]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3c63428]
#> 
#> [[24]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpYRBQpT/file506ea8b0436"),  <...>
#> env:  [frame 0x31cb470]
#> 
#> [[25]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpYRBQpT/file506ea8b0436"),  <...>
#> env:  [frame 0x31cb7f0]
#> 
#> [[26]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x31c6538]
#> 
#> [[27]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x31c2d98]
#> 
#> [[28]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x31c30e0]
#> 
#> [[29]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x31c3498]
#> 
#> [[30]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x31c3a48]
#> 
#> [[31]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x31c3d90]
#> 
#> [[32]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x31c40d8]
#> 
#> [[33]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x31c4810]
#> 
#> [[34]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"