Call stack information

The eval_ and call_ families of functions provide a replacement for the base R functions prefixed with sys. (which are all about the context stack), as well as for parent.frame() (which is the only base R function for querying the call stack). The context stack includes all R-level evaluation contexts. It is linear in terms of execution history but due to lazy evaluation it is potentially nonlinear in terms of call history. The call stack history, on the other hand, is homogenous.

global_frame()

current_frame()

ctxt_frame(n = 1)

call_frame(n = 1, clean = TRUE)

ctxt_depth()

call_depth()

ctxt_stack(n = NULL, trim = 0)

call_stack(n = NULL, clean = TRUE)

Arguments

n
The number of frames to go back in the stack.

clean
Whether to post-process the call stack to clean non-standard frames. If TRUE, suboptimal call-stack entries by base::eval() will be cleaned up: the duplicate frame created by eval() is eliminated.

trim
The number of layers of intervening frames to trim off the stack. See stack_trim() and examples.

Details

ctxt_frame() and call_frame() return a frame object containing the following fields: expr and env (call expression and evaluation environment), pos and caller_pos (position of current frame in the context stack and position of the caller), and fun (function of the current frame). ctxt_stack() and call_stack() return a list of all context or call frames on the stack. Finally, ctxt_depth() and call_depth() report the current context position or the number of calling frames on the stack.

The base R functions take two sorts of arguments to indicate which frame to query: which and n. The n argument is straightforward: it's the number of frames to go down the stack, with n = 1 referring to the current context. The which argument is more complicated and changes meaning for values lower than 1. For the sake of consistency, the rlang functions all take the same kind of argument n. This argument has a single meaning (the number of frames to go down the stack) and cannot be lower than 1.

Note finally that parent.frame(1) corresponds to call_frame(2)$env, as n = 1 always refers to the current frame. This makes the _frame() and _stack() functions consistent: ctxt_frame(2) is the same as ctxt_stack()[[2]]. Also, ctxt_depth() returns one more frame than base::sys.nframe() because it counts the global frame. That is consistent with the _stack() functions which return the global frame as well. This way, call_stack(call_depth()) is the same as global_frame().

Life cycle

These functions are soft-deprecated and replaced by trace_back().

Examples

# Expressions within arguments count as contexts
identity(identity(ctxt_depth())) # returns 2

#> Warning: `ctxt_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 27

# But they are not part of the call stack because arguments are
# evaluated within the calling function (or the global environment
# if called at top level)
identity(identity(call_depth())) # returns 0

#> Warning: `call_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 14

# The context stacks includes all intervening execution frames. The
# call stack doesn't:
f <- function(x) identity(x)
f(f(ctxt_stack()))

#> [[1]]
#> <frame 28> (27)
#> expr: identity(x)
#> env:  [local 0x7f83d91cf990]
#> 
#> [[2]]
#> <frame 27> (24)
#> expr: f(ctxt_stack())
#> env:  [local 0x7f83d91cfaa8]
#> 
#> [[3]]
#> <frame 26> (25)
#> expr: identity(x)
#> env:  [local 0x7f83d91cbd58]
#> 
#> [[4]]
#> <frame 25> (24)
#> expr: f(f(ctxt_stack()))
#> env:  [local 0x7f83d91cbe70]
#> 
#> [[5]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f83d93d7920]
#> 
#> [[6]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f83d91cc068]
#> 
#> [[7]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f83d91cc298]
#> 
#> [[8]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f83d91ccbc8]
#> 
#> [[9]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f83d91cd178]
#> 
#> [[10]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f83d91cd530]
#> 
#> [[11]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f83d91cd878]
#> 
#> [[12]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f83d91ca0a0]
#> 
#> [[13]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f83d91ca2d0]
#> 
#> [[14]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f83d91ca4c8]
#> 
#> [[15]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f83d91ca5a8]
#> 
#> [[16]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f83d920d3a8]
#> 
#> [[17]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f83da58c720]
#> 
#> [[18]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7f83da587178]
#> 
#> [[19]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f83da060528]
#> 
#> [[20]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f83da42fe08]
#> 
#> [[21]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f83da1c6f50]
#> 
#> [[22]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7f83da1c7650]
#> 
#> [[23]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f83da62be30]
#> 
#> [[24]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f83d53a7fe0]
#> 
#> [[25]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f83d5585df8]
#> 
#> [[26]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f83d492d068]
#> 
#> [[27]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f83d492e038]
#> 
#> [[28]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f83d4886850]
#> 
#> [[29]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(f(call_stack()))

#> Warning: `call_stack()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [[1]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f83d93d7920]
#> 
#> [[2]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f83d9ac5238]
#> 
#> [[3]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f83da58c720]
#> 
#> [[4]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7f83da587178]
#> 
#> [[5]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f83da060528]
#> 
#> [[6]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f83da42fe08]
#> 
#> [[7]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f83da1c6f50]
#> 
#> [[8]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f83da62be30]
#> 
#> [[9]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f83d53a7fe0]
#> 
#> [[10]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f83d5585df8]
#> 
#> [[11]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f83d492d068]
#> 
#> [[12]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f83d492e038]
#> 
#> [[13]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f83d4886850]
#> 
#> [[14]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

g <- function(cmd) cmd()
f(g(ctxt_stack))

#> [[1]]
#> <frame 27> (24)
#> expr: g(ctxt_stack)
#> env:  [local 0x7f83d78ef5f8]
#> 
#> [[2]]
#> <frame 26> (25)
#> expr: identity(x)
#> env:  [local 0x7f83d78ef4e0]
#> 
#> [[3]]
#> <frame 25> (24)
#> expr: f(g(ctxt_stack))
#> env:  [local 0x7f83d78ef3c8]
#> 
#> [[4]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f83d93d7920]
#> 
#> [[5]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f83d78ef208]
#> 
#> [[6]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f83d78eefd8]
#> 
#> [[7]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f83d78eea28]
#> 
#> [[8]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f83d78ee4b0]
#> 
#> [[9]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f83d78f2078]
#> 
#> [[10]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f83d78f1cf8]
#> 
#> [[11]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f83d78f1630]
#> 
#> [[12]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f83d78f1400]
#> 
#> [[13]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f83d78f1278]
#> 
#> [[14]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f83d78f1198]
#> 
#> [[15]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f83d7908d70]
#> 
#> [[16]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f83da58c720]
#> 
#> [[17]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7f83da587178]
#> 
#> [[18]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f83da060528]
#> 
#> [[19]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f83da42fe08]
#> 
#> [[20]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f83da1c6f50]
#> 
#> [[21]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7f83da1c7650]
#> 
#> [[22]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f83da62be30]
#> 
#> [[23]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f83d53a7fe0]
#> 
#> [[24]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f83d5585df8]
#> 
#> [[25]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f83d492d068]
#> 
#> [[26]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f83d492e038]
#> 
#> [[27]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f83d4886850]
#> 
#> [[28]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(g(call_stack))

#> [[1]]
#> <frame 27> (23)
#> expr: g(call_stack)
#> env:  [local 0x7f83d6741ca0]
#> 
#> [[2]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f83d93d7920]
#> 
#> [[3]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f83d6757808]
#> 
#> [[4]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f83da58c720]
#> 
#> [[5]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7f83da587178]
#> 
#> [[6]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f83da060528]
#> 
#> [[7]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f83da42fe08]
#> 
#> [[8]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f83da1c6f50]
#> 
#> [[9]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f83da62be30]
#> 
#> [[10]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f83d53a7fe0]
#> 
#> [[11]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f83d5585df8]
#> 
#> [[12]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f83d492d068]
#> 
#> [[13]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f83d492e038]
#> 
#> [[14]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f83d4886850]
#> 
#> [[15]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

# The rlang _stack() functions return a list of frame
# objects. Use purrr::transpose() or index a field with
# purrr::map()'s to extract a particular field from a stack:

# stack <- f(f(call_stack()))
# purrr::map(stack, "env")
# purrr::transpose(stack)$expr

# current_frame() is an alias for ctxt_frame(1)
fn <- function() list(current = current_frame(), first = ctxt_frame(1))
fn()

#> Warning: `current_frame()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> $current
#> <frame 25> (24)
#> expr: fn()
#> env:  [local 0x7f83d55aa418]
#> 
#> $first
#> <frame 25> (24)
#> expr: fn()
#> env:  [local 0x7f83d55aa418]
#> 

# While current_frame() is the top of the stack, global_frame() is
# the bottom:
fn <- function() {
  n <- ctxt_depth()
  ctxt_frame(n)
}
identical(fn(), global_frame())

#> [1] TRUE

# ctxt_stack() returns a stack with all intervening frames. You can
# trim layers of intervening frames with the trim argument:
identity(identity(ctxt_stack()))

#> [[1]]
#> <frame 26> (24)
#> expr: identity(ctxt_stack())
#> env:  [local 0x7f83d66cbd48]
#> 
#> [[2]]
#> <frame 25> (24)
#> expr: identity(identity(ctxt_stack()))
#> env:  [local 0x7f83d66cbc30]
#> 
#> [[3]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f83d93d7920]
#> 
#> [[4]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f83d66cba70]
#> 
#> [[5]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f83d66cb840]
#> 
#> [[6]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f83d66cb290]
#> 
#> [[7]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f83d66cad18]
#> 
#> [[8]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f83d66d08e0]
#> 
#> [[9]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f83d66d0598]
#> 
#> [[10]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f83d66cfed0]
#> 
#> [[11]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f83d66cfca0]
#> 
#> [[12]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f83d66cfae0]
#> 
#> [[13]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f83d66cfa00]
#> 
#> [[14]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f83d66e52c8]
#> 
#> [[15]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f83da58c720]
#> 
#> [[16]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7f83da587178]
#> 
#> [[17]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f83da060528]
#> 
#> [[18]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f83da42fe08]
#> 
#> [[19]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f83da1c6f50]
#> 
#> [[20]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7f83da1c7650]
#> 
#> [[21]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f83da62be30]
#> 
#> [[22]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f83d53a7fe0]
#> 
#> [[23]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f83d5585df8]
#> 
#> [[24]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f83d492d068]
#> 
#> [[25]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f83d492e038]
#> 
#> [[26]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f83d4886850]
#> 
#> [[27]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(ctxt_stack(trim = 1)))

#> [[1]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f83d93d7920]
#> 
#> [[2]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f83d61042e0]
#> 
#> [[3]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f83d61040b0]
#> 
#> [[4]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f83d61039e8]
#> 
#> [[5]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f83d6103400]
#> 
#> [[6]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f83d6103080]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f83d6102d38]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f83d6102600]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f83d61062a8]
#> 
#> [[10]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f83d6106120]
#> 
#> [[11]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f83d6106040]
#> 
#> [[12]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f83d6118cc8]
#> 
#> [[13]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f83da58c720]
#> 
#> [[14]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7f83da587178]
#> 
#> [[15]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f83da060528]
#> 
#> [[16]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f83da42fe08]
#> 
#> [[17]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f83da1c6f50]
#> 
#> [[18]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7f83da1c7650]
#> 
#> [[19]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f83da62be30]
#> 
#> [[20]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f83d53a7fe0]
#> 
#> [[21]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f83d5585df8]
#> 
#> [[22]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f83d492d068]
#> 
#> [[23]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f83d492e038]
#> 
#> [[24]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f83d4886850]
#> 
#> [[25]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# ctxt_stack() is called within fn() with intervening frames:
fn <- function(trim) identity(identity(ctxt_stack(trim = trim)))
fn(0)

#> [[1]]
#> <frame 27> (25)
#> expr: identity(ctxt_stack(trim = trim))
#> env:  [local 0x7f83d6c3ba58]
#> 
#> [[2]]
#> <frame 26> (25)
#> expr: identity(identity(ctxt_stack(trim = trim)))
#> env:  [local 0x7f83d6c3b940]
#> 
#> [[3]]
#> <frame 25> (24)
#> expr: fn(0)
#> env:  [local 0x7f83d6c3b828]
#> 
#> [[4]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f83d93d7920]
#> 
#> [[5]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f83d6c3b668]
#> 
#> [[6]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f83d6c3b438]
#> 
#> [[7]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f83d6c3ae50]
#> 
#> [[8]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f83d6c3a8a0]
#> 
#> [[9]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f83d6c3a558]
#> 
#> [[10]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f83d6c3e120]
#> 
#> [[11]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f83d6c3da58]
#> 
#> [[12]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f83d6c3d7f0]
#> 
#> [[13]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f83d6c3d668]
#> 
#> [[14]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f83d6c3d588]
#> 
#> [[15]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f83d6c51320]
#> 
#> [[16]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f83da58c720]
#> 
#> [[17]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7f83da587178]
#> 
#> [[18]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f83da060528]
#> 
#> [[19]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f83da42fe08]
#> 
#> [[20]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f83da1c6f50]
#> 
#> [[21]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7f83da1c7650]
#> 
#> [[22]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f83da62be30]
#> 
#> [[23]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f83d53a7fe0]
#> 
#> [[24]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f83d5585df8]
#> 
#> [[25]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f83d492d068]
#> 
#> [[26]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f83d492e038]
#> 
#> [[27]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f83d4886850]
#> 
#> [[28]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# We can trim the first layer of those:
fn(1)

#> [[1]]
#> <frame 25> (24)
#> expr: fn(1)
#> env:  [local 0x7f83d68d2ef0]
#> 
#> [[2]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f83d93d7920]
#> 
#> [[3]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f83d68d2d30]
#> 
#> [[4]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f83d68d2b00]
#> 
#> [[5]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f83d68d2518]
#> 
#> [[6]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f83d68d5e78]
#> 
#> [[7]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f83d68d5af8]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f83d68d5778]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f83d68d4f98]
#> 
#> [[10]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f83d68d4d68]
#> 
#> [[11]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f83d68d4ba8]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f83d68d4ac8]
#> 
#> [[13]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f83d68ea550]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f83da58c720]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7f83da587178]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f83da060528]
#> 
#> [[17]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f83da42fe08]
#> 
#> [[18]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f83da1c6f50]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7f83da1c7650]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f83da62be30]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f83d53a7fe0]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f83d5585df8]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f83d492d068]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f83d492e038]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f83d4886850]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# The outside intervening frames (at the fn() call site) are still
# returned, but can be trimmed as well:
identity(identity(fn(1)))

#> [[1]]
#> <frame 27> (24)
#> expr: fn(1)
#> env:  [local 0x7f83d4d041d0]
#> 
#> [[2]]
#> <frame 26> (24)
#> expr: identity(fn(1))
#> env:  [local 0x7f83d4d040b8]
#> 
#> [[3]]
#> <frame 25> (24)
#> expr: identity(identity(fn(1)))
#> env:  [local 0x7f83d4d03fa0]
#> 
#> [[4]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f83d93d7920]
#> 
#> [[5]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f83d4d03de0]
#> 
#> [[6]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f83d4d03bb0]
#> 
#> [[7]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f83d4d03590]
#> 
#> [[8]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f83d4d06a20]
#> 
#> [[9]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f83d4d066a0]
#> 
#> [[10]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f83d4d06320]
#> 
#> [[11]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f83d4d05c58]
#> 
#> [[12]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f83d4d059f0]
#> 
#> [[13]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f83d4d057c0]
#> 
#> [[14]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f83d4d056e0]
#> 
#> [[15]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f83d4d35fd8]
#> 
#> [[16]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f83da58c720]
#> 
#> [[17]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7f83da587178]
#> 
#> [[18]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f83da060528]
#> 
#> [[19]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f83da42fe08]
#> 
#> [[20]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f83da1c6f50]
#> 
#> [[21]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7f83da1c7650]
#> 
#> [[22]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f83da62be30]
#> 
#> [[23]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f83d53a7fe0]
#> 
#> [[24]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f83d5585df8]
#> 
#> [[25]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f83d492d068]
#> 
#> [[26]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f83d492e038]
#> 
#> [[27]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f83d4886850]
#> 
#> [[28]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(fn(2)))

#> [[1]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f83d93d7920]
#> 
#> [[2]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f83d1721680]
#> 
#> [[3]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f83d1721418]
#> 
#> [[4]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f83d1727c68]
#> 
#> [[5]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f83d172a560]
#> 
#> [[6]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f83d1729220]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f83d1728ea0]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f83d172e560]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f83d172e288]
#> 
#> [[10]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f83d172e0c8]
#> 
#> [[11]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f83d172dd10]
#> 
#> [[12]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f83d174e208]
#> 
#> [[13]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f83da58c720]
#> 
#> [[14]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7f83da587178]
#> 
#> [[15]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f83da060528]
#> 
#> [[16]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f83da42fe08]
#> 
#> [[17]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f83da1c6f50]
#> 
#> [[18]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7f83da1c7650]
#> 
#> [[19]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f83da62be30]
#> 
#> [[20]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f83d53a7fe0]
#> 
#> [[21]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f83d5585df8]
#> 
#> [[22]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f83d492d068]
#> 
#> [[23]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f83d492e038]
#> 
#> [[24]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f83d4886850]
#> 
#> [[25]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g <- function(trim) identity(identity(fn(trim)))
g(2)

#> [[1]]
#> <frame 25> (24)
#> expr: g(2)
#> env:  [local 0x7f83d2a64d90]
#> 
#> [[2]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f83d93d7920]
#> 
#> [[3]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f83d2a64af0]
#> 
#> [[4]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f83d2a64268]
#> 
#> [[5]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f83d2a63430]
#> 
#> [[6]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f83d2a46360]
#> 
#> [[7]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f83d2a45ec8]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f83d2a45b48]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f83d29fdb58]
#> 
#> [[10]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f83d29fd8b8]
#> 
#> [[11]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f83d29fd6f8]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f83d29fd5a8]
#> 
#> [[13]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f83d263c070]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f83da58c720]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7f83da587178]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f83da060528]
#> 
#> [[17]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f83da42fe08]
#> 
#> [[18]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f83da1c6f50]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7f83da1c7650]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f83da62be30]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f83d53a7fe0]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f83d5585df8]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f83d492d068]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f83d492e038]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f83d4886850]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g(3)

#> [[1]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f83d93d7920]
#> 
#> [[2]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f83da61a338]
#> 
#> [[3]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f83da61a568]
#> 
#> [[4]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f83da61ab18]
#> 
#> [[5]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f83da61b090]
#> 
#> [[6]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f83da61b3d8]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f83da61b720]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f83da6086d8]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f83da608908]
#> 
#> [[10]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f83da608a90]
#> 
#> [[11]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f83da608b70]
#> 
#> [[12]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f83da5f8668]
#> 
#> [[13]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f83da58c720]
#> 
#> [[14]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7f83da587178]
#> 
#> [[15]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f83da060528]
#> 
#> [[16]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f83da42fe08]
#> 
#> [[17]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f83da1c6f50]
#> 
#> [[18]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7f83da1c7650]
#> 
#> [[19]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f83da62be30]
#> 
#> [[20]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f83d53a7fe0]
#> 
#> [[21]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f83d5585df8]
#> 
#> [[22]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f83d492d068]
#> 
#> [[23]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f83d492e038]
#> 
#> [[24]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f83d4886850]
#> 
#> [[25]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"