Call stack information

The eval_ and call_ families of functions provide a replacement for the base R functions prefixed with sys. (which are all about the context stack), as well as for parent.frame() (which is the only base R function for querying the call stack). The context stack includes all R-level evaluation contexts. It is linear in terms of execution history but due to lazy evaluation it is potentially nonlinear in terms of call history. The call stack history, on the other hand, is homogenous. See vignette("stack") for more information.

global_frame()

current_frame()

ctxt_frame(n = 1)

call_frame(n = 1, clean = TRUE)

ctxt_depth()

call_depth()

ctxt_stack(n = NULL, trim = 0)

call_stack(n = NULL, clean = TRUE)

Arguments

n
The number of frames to go back in the stack.

clean
Whether to post-process the call stack to clean non-standard frames. If TRUE, suboptimal call-stack entries by base::eval() will be cleaned up: the duplicate frame created by eval() is eliminated.

trim
The number of layers of intervening frames to trim off the stack. See stack_trim() and examples.

Details

ctxt_frame() and call_frame() return a frame object containing the following fields: expr and env (call expression and evaluation environment), pos and caller_pos (position of current frame in the context stack and position of the caller), and fun (function of the current frame). ctxt_stack() and call_stack() return a list of all context or call frames on the stack. Finally, ctxt_depth() and call_depth() report the current context position or the number of calling frames on the stack.

The base R functions take two sorts of arguments to indicate which frame to query: which and n. The n argument is straightforward: it's the number of frames to go down the stack, with n = 1 referring to the current context. The which argument is more complicated and changes meaning for values lower than

For the sake of consistency, the lazyeval functions all take the same kind of argument n. This argument has a single meaning (the number of frames to go down the stack) and cannot be lower than 1.

Note finally that parent.frame(1) corresponds to call_frame(2)$env, as n = 1 always refers to the current frame. This makes the _frame() and _stack() functions consistent: ctxt_frame(2) is the same as ctxt_stack()[[2]]. Also, ctxt_depth() returns one more frame than [base::sys.nframe()] because it counts the global frame. That is consistent with the _stack() functions which return the global frame as well. This way, call_stack(call_depth()) is the same as global_frame().

[[2]: R:[2 [base::sys.nframe()]: R:base::sys.nframe()

Examples

# Expressions within arguments count as contexts
identity(identity(ctxt_depth())) # returns 2

#> [1] 34

# But they are not part of the call stack because arguments are
# evaluated within the calling function (or the global environment
# if called at top level)
identity(identity(call_depth())) # returns 0

#> [1] 15

# The context stacks includes all intervening execution frames. The
# call stack doesn't:
f <- function(x) identity(x)
f(f(ctxt_stack()))

#> [[1]]
#> <frame 35> (34)
#> expr: identity(x)
#> env:  [local 0x7f8493d85fc0]
#> 
#> [[2]]
#> <frame 34> (31)
#> expr: f(ctxt_stack())
#> env:  [local 0x7f8493d85ea8]
#> 
#> [[3]]
#> <frame 33> (32)
#> expr: identity(x)
#> env:  [local 0x7f8493d85d58]
#> 
#> [[4]]
#> <frame 32> (31)
#> expr: f(f(ctxt_stack()))
#> env:  [local 0x7f8493d85c40]
#> 
#> [[5]]
#> <frame 31> (30)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f84904784a8]
#> 
#> [[6]]
#> <frame 30> (20)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f8493d86a28]
#> 
#> [[7]]
#> <frame 29> (20)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f8493d867c0]
#> 
#> [[8]]
#> <frame 28> (20)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f8493d870a0]
#> 
#> [[9]]
#> <frame 27> (26)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f8493998a28]
#> 
#> [[10]]
#> <frame 26> (25)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f84939986a8]
#> 
#> [[11]]
#> <frame 25> (24)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f849399a2d0]
#> 
#> [[12]]
#> <frame 24> (23)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f84918cb140]
#> 
#> [[13]]
#> <frame 23> (22)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f84918caed8]
#> 
#> [[14]]
#> <frame 22> (20)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f84918caca8]
#> 
#> [[15]]
#> <frame 21> (20)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f84918cab20]
#> 
#> [[16]]
#> <frame 20> (19)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f8491b78518]
#> 
#> [[17]]
#> <frame 19> (18)
#> expr: evaluate::evaluate(text, code_env, new_device = TRUE)
#> env:  [frame 0x7f848e79fa60]
#> 
#> [[18]]
#> <frame 18> (16)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f8494f0bb80]
#> 
#> [[19]]
#> <frame 17> (16)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f8494f0a988]
#> 
#> [[20]]
#> <frame 16> (12)
#> expr: data_reference_topic(topic, pkg, path = path, examples = examples,  <...>
#> env:  [frame 0x7f848d27df40]
#> 
#> [[21]]
#> <frame 15> (14)
#> expr: stopifnot(is.list(x), is.list(val))
#> env:  [frame 0x7f848d27d990]
#> 
#> [[22]]
#> <frame 14> (13)
#> expr: utils::modifyList(data, data_template(pkg, depth = depth))
#> env:  [frame 0x7f848d280740]
#> 
#> [[23]]
#> <frame 13> (12)
#> expr: render_page(pkg, "reference-topic", data = data_reference_topic(topic,  <...>
#> env:  [frame 0x7f848d2e9318]
#> 
#> [[24]]
#> <frame 12> (11)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f848d2ed2a8]
#> 
#> [[25]]
#> <frame 11> (10)
#> expr: purrr::map(., build_reference_topic, path, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f848aba6200]
#> 
#> [[26]]
#> <frame 10> (8)
#> expr: function_list[[k]](value)
#> env:  [frame 0x7f848abc34d8]
#> 
#> [[27]]
#> <frame 9> (8)
#> expr: withVisible(function_list[[k]](value))
#> env:  [frame 0x7f848abc33c0]
#> 
#> [[28]]
#> <frame 8> (7)
#> expr: freduce(value, `_function_list`)
#> env:  [frame 0x7f848abd1468]
#> 
#> [[29]]
#> <frame 7> (6)
#> expr: `_fseq`(`_lhs`)
#> env:  [frame 0x7f848abd12a8]
#> 
#> [[30]]
#> <frame 6> (5)
#> expr: eval(expr, envir, enclos)
#> env:  [frame 0x7f848adb4230]
#> 
#> [[31]]
#> <frame 5> (3)
#> expr: eval(quote(`_fseq`(`_lhs`)), env, env)
#> env:  [frame 0x7f848abd1078]
#> 
#> [[32]]
#> <frame 4> (3)
#> expr: withVisible(eval(quote(`_fseq`(`_lhs`)), env, env))
#> env:  [frame 0x7f848abd25f0]
#> 
#> [[33]]
#> <frame 3> (2)
#> expr: pkg$topics %>% purrr::transpose() %>% purrr::map(build_reference_topic,  <...>
#> env:  [frame 0x7f848adb4d78]
#> 
#> [[34]]
#> <frame 2> (1)
#> expr: build_reference(pkg, lazy = FALSE, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f8492ff3f40]
#> 
#> [[35]]
#> <frame 1> (0)
#> expr: pkgdown::build_site("/Users/yoplait/Dropbox/Projects/R/hadley/rlang")
#> env:  [frame 0x7f848fb31d48]
#> 
#> [[36]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(f(call_stack()))

#> [[1]]
#> <frame 30> (20)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f84904784a8]
#> 
#> [[2]]
#> <frame 20> (19)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f848bb9ab50]
#> 
#> [[3]]
#> <frame 19> (18)
#> expr: evaluate::evaluate(text, code_env, new_device = TRUE)
#> env:  [frame 0x7f848e79fa60]
#> 
#> [[4]]
#> <frame 18> (16)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f8494f0bb80]
#> 
#> [[5]]
#> <frame 16> (12)
#> expr: data_reference_topic(topic, pkg, path = path, examples = examples,  <...>
#> env:  [frame 0x7f848d27df40]
#> 
#> [[6]]
#> <frame 12> (11)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f848d2ed2a8]
#> 
#> [[7]]
#> <frame 11> (10)
#> expr: purrr::map(., build_reference_topic, path, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f848aba6200]
#> 
#> [[8]]
#> <frame 10> (8)
#> expr: function_list[[k]](value)
#> env:  [frame 0x7f848abc34d8]
#> 
#> [[9]]
#> <frame 8> (7)
#> expr: freduce(value, `_function_list`)
#> env:  [frame 0x7f848abd1468]
#> 
#> [[10]]
#> <frame 7> (5)
#> expr: `_fseq`(`_lhs`)
#> env:  [frame 0x7f848abd12a8]
#> 
#> [[11]]
#> <frame 5> (3)
#> expr: eval(quote(`_fseq`(`_lhs`)), env, env)
#> env:  [frame 0x7f848adb4230]
#> 
#> [[12]]
#> <frame 3> (2)
#> expr: pkg$topics %>% purrr::transpose() %>% purrr::map(build_reference_topic,  <...>
#> env:  [frame 0x7f848adb4d78]
#> 
#> [[13]]
#> <frame 2> (1)
#> expr: build_reference(pkg, lazy = FALSE, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f8492ff3f40]
#> 
#> [[14]]
#> <frame 1> (0)
#> expr: pkgdown::build_site("/Users/yoplait/Dropbox/Projects/R/hadley/rlang")
#> env:  [frame 0x7f848fb31d48]
#> 
#> [[15]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

g <- function(cmd) cmd()
f(g(ctxt_stack))

#> [[1]]
#> <frame 34> (31)
#> expr: g(ctxt_stack)
#> env:  [local 0x7f84920ef430]
#> 
#> [[2]]
#> <frame 33> (32)
#> expr: identity(x)
#> env:  [local 0x7f84920ef2e0]
#> 
#> [[3]]
#> <frame 32> (31)
#> expr: f(g(ctxt_stack))
#> env:  [local 0x7f84920ef1c8]
#> 
#> [[4]]
#> <frame 31> (30)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f84904784a8]
#> 
#> [[5]]
#> <frame 30> (20)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f8492e1a978]
#> 
#> [[6]]
#> <frame 29> (20)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f8492e1a6a0]
#> 
#> [[7]]
#> <frame 28> (20)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f8492e1af48]
#> 
#> [[8]]
#> <frame 27> (26)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f84929bc2d0]
#> 
#> [[9]]
#> <frame 26> (25)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f84929bbf18]
#> 
#> [[10]]
#> <frame 25> (24)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f8492ab5340]
#> 
#> [[11]]
#> <frame 24> (23)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f8493764b08]
#> 
#> [[12]]
#> <frame 23> (22)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f84937648d8]
#> 
#> [[13]]
#> <frame 22> (20)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f84937646e0]
#> 
#> [[14]]
#> <frame 21> (20)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f8493764590]
#> 
#> [[15]]
#> <frame 20> (19)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f848e0063c0]
#> 
#> [[16]]
#> <frame 19> (18)
#> expr: evaluate::evaluate(text, code_env, new_device = TRUE)
#> env:  [frame 0x7f848e79fa60]
#> 
#> [[17]]
#> <frame 18> (16)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f8494f0bb80]
#> 
#> [[18]]
#> <frame 17> (16)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f8494f0a988]
#> 
#> [[19]]
#> <frame 16> (12)
#> expr: data_reference_topic(topic, pkg, path = path, examples = examples,  <...>
#> env:  [frame 0x7f848d27df40]
#> 
#> [[20]]
#> <frame 15> (14)
#> expr: stopifnot(is.list(x), is.list(val))
#> env:  [frame 0x7f848d27d990]
#> 
#> [[21]]
#> <frame 14> (13)
#> expr: utils::modifyList(data, data_template(pkg, depth = depth))
#> env:  [frame 0x7f848d280740]
#> 
#> [[22]]
#> <frame 13> (12)
#> expr: render_page(pkg, "reference-topic", data = data_reference_topic(topic,  <...>
#> env:  [frame 0x7f848d2e9318]
#> 
#> [[23]]
#> <frame 12> (11)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f848d2ed2a8]
#> 
#> [[24]]
#> <frame 11> (10)
#> expr: purrr::map(., build_reference_topic, path, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f848aba6200]
#> 
#> [[25]]
#> <frame 10> (8)
#> expr: function_list[[k]](value)
#> env:  [frame 0x7f848abc34d8]
#> 
#> [[26]]
#> <frame 9> (8)
#> expr: withVisible(function_list[[k]](value))
#> env:  [frame 0x7f848abc33c0]
#> 
#> [[27]]
#> <frame 8> (7)
#> expr: freduce(value, `_function_list`)
#> env:  [frame 0x7f848abd1468]
#> 
#> [[28]]
#> <frame 7> (6)
#> expr: `_fseq`(`_lhs`)
#> env:  [frame 0x7f848abd12a8]
#> 
#> [[29]]
#> <frame 6> (5)
#> expr: eval(expr, envir, enclos)
#> env:  [frame 0x7f848adb4230]
#> 
#> [[30]]
#> <frame 5> (3)
#> expr: eval(quote(`_fseq`(`_lhs`)), env, env)
#> env:  [frame 0x7f848abd1078]
#> 
#> [[31]]
#> <frame 4> (3)
#> expr: withVisible(eval(quote(`_fseq`(`_lhs`)), env, env))
#> env:  [frame 0x7f848abd25f0]
#> 
#> [[32]]
#> <frame 3> (2)
#> expr: pkg$topics %>% purrr::transpose() %>% purrr::map(build_reference_topic,  <...>
#> env:  [frame 0x7f848adb4d78]
#> 
#> [[33]]
#> <frame 2> (1)
#> expr: build_reference(pkg, lazy = FALSE, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f8492ff3f40]
#> 
#> [[34]]
#> <frame 1> (0)
#> expr: pkgdown::build_site("/Users/yoplait/Dropbox/Projects/R/hadley/rlang")
#> env:  [frame 0x7f848fb31d48]
#> 
#> [[35]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(g(call_stack))

#> [[1]]
#> <frame 34> (30)
#> expr: g(call_stack)
#> env:  [local 0x7f8492aaec40]
#> 
#> [[2]]
#> <frame 30> (20)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f84904784a8]
#> 
#> [[3]]
#> <frame 20> (19)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f8492dfdee8]
#> 
#> [[4]]
#> <frame 19> (18)
#> expr: evaluate::evaluate(text, code_env, new_device = TRUE)
#> env:  [frame 0x7f848e79fa60]
#> 
#> [[5]]
#> <frame 18> (16)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f8494f0bb80]
#> 
#> [[6]]
#> <frame 16> (12)
#> expr: data_reference_topic(topic, pkg, path = path, examples = examples,  <...>
#> env:  [frame 0x7f848d27df40]
#> 
#> [[7]]
#> <frame 12> (11)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f848d2ed2a8]
#> 
#> [[8]]
#> <frame 11> (10)
#> expr: purrr::map(., build_reference_topic, path, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f848aba6200]
#> 
#> [[9]]
#> <frame 10> (8)
#> expr: function_list[[k]](value)
#> env:  [frame 0x7f848abc34d8]
#> 
#> [[10]]
#> <frame 8> (7)
#> expr: freduce(value, `_function_list`)
#> env:  [frame 0x7f848abd1468]
#> 
#> [[11]]
#> <frame 7> (5)
#> expr: `_fseq`(`_lhs`)
#> env:  [frame 0x7f848abd12a8]
#> 
#> [[12]]
#> <frame 5> (3)
#> expr: eval(quote(`_fseq`(`_lhs`)), env, env)
#> env:  [frame 0x7f848adb4230]
#> 
#> [[13]]
#> <frame 3> (2)
#> expr: pkg$topics %>% purrr::transpose() %>% purrr::map(build_reference_topic,  <...>
#> env:  [frame 0x7f848adb4d78]
#> 
#> [[14]]
#> <frame 2> (1)
#> expr: build_reference(pkg, lazy = FALSE, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f8492ff3f40]
#> 
#> [[15]]
#> <frame 1> (0)
#> expr: pkgdown::build_site("/Users/yoplait/Dropbox/Projects/R/hadley/rlang")
#> env:  [frame 0x7f848fb31d48]
#> 
#> [[16]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

# The lazyeval _stack() functions return a list of frame
# objects. Use purrr::transpose() or index a field with
# purrr::map()'s to extract a particular field from a stack:

# stack <- f(f(call_stack()))
# purrr::map(stack, "env")
# purrr::transpose(stack)$expr

# current_frame() is an alias for ctxt_frame(1)
fn <- function() list(current = current_frame(), first = ctxt_frame(1))
fn()

#> $current
#> <frame 32> (31)
#> expr: fn()
#> env:  [local 0x7f848befe190]
#> 
#> $first
#> <frame 32> (31)
#> expr: fn()
#> env:  [local 0x7f848befe190]
#> 

# While current_frame() is the top of the stack, global_frame() is
# the bottom:
fn <- function() {
  n <- ctxt_depth()
  ctxt_frame(n)
}
identical(fn(), global_frame())

#> [1] TRUE

# ctxt_stack() returns a stack with all intervening frames. You can
# trim layers of intervening frames with the trim argument:
identity(identity(ctxt_stack()))

#> [[1]]
#> <frame 33> (31)
#> expr: identity(ctxt_stack())
#> env:  [local 0x7f848be965b8]
#> 
#> [[2]]
#> <frame 32> (31)
#> expr: identity(identity(ctxt_stack()))
#> env:  [local 0x7f848be96468]
#> 
#> [[3]]
#> <frame 31> (30)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f84904784a8]
#> 
#> [[4]]
#> <frame 30> (20)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f848be960e8]
#> 
#> [[5]]
#> <frame 29> (20)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f848be98db8]
#> 
#> [[6]]
#> <frame 28> (20)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f848beaced0]
#> 
#> [[7]]
#> <frame 27> (26)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f848beac8b0]
#> 
#> [[8]]
#> <frame 26> (25)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f848beadbf8]
#> 
#> [[9]]
#> <frame 25> (24)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f848beb0628]
#> 
#> [[10]]
#> <frame 24> (23)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f848bebc708]
#> 
#> [[11]]
#> <frame 23> (22)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f848bebc3f8]
#> 
#> [[12]]
#> <frame 22> (20)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f848bebc1c8]
#> 
#> [[13]]
#> <frame 21> (20)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f848bebc078]
#> 
#> [[14]]
#> <frame 20> (19)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f848abf3c68]
#> 
#> [[15]]
#> <frame 19> (18)
#> expr: evaluate::evaluate(text, code_env, new_device = TRUE)
#> env:  [frame 0x7f848e79fa60]
#> 
#> [[16]]
#> <frame 18> (16)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f8494f0bb80]
#> 
#> [[17]]
#> <frame 17> (16)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f8494f0a988]
#> 
#> [[18]]
#> <frame 16> (12)
#> expr: data_reference_topic(topic, pkg, path = path, examples = examples,  <...>
#> env:  [frame 0x7f848d27df40]
#> 
#> [[19]]
#> <frame 15> (14)
#> expr: stopifnot(is.list(x), is.list(val))
#> env:  [frame 0x7f848d27d990]
#> 
#> [[20]]
#> <frame 14> (13)
#> expr: utils::modifyList(data, data_template(pkg, depth = depth))
#> env:  [frame 0x7f848d280740]
#> 
#> [[21]]
#> <frame 13> (12)
#> expr: render_page(pkg, "reference-topic", data = data_reference_topic(topic,  <...>
#> env:  [frame 0x7f848d2e9318]
#> 
#> [[22]]
#> <frame 12> (11)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f848d2ed2a8]
#> 
#> [[23]]
#> <frame 11> (10)
#> expr: purrr::map(., build_reference_topic, path, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f848aba6200]
#> 
#> [[24]]
#> <frame 10> (8)
#> expr: function_list[[k]](value)
#> env:  [frame 0x7f848abc34d8]
#> 
#> [[25]]
#> <frame 9> (8)
#> expr: withVisible(function_list[[k]](value))
#> env:  [frame 0x7f848abc33c0]
#> 
#> [[26]]
#> <frame 8> (7)
#> expr: freduce(value, `_function_list`)
#> env:  [frame 0x7f848abd1468]
#> 
#> [[27]]
#> <frame 7> (6)
#> expr: `_fseq`(`_lhs`)
#> env:  [frame 0x7f848abd12a8]
#> 
#> [[28]]
#> <frame 6> (5)
#> expr: eval(expr, envir, enclos)
#> env:  [frame 0x7f848adb4230]
#> 
#> [[29]]
#> <frame 5> (3)
#> expr: eval(quote(`_fseq`(`_lhs`)), env, env)
#> env:  [frame 0x7f848abd1078]
#> 
#> [[30]]
#> <frame 4> (3)
#> expr: withVisible(eval(quote(`_fseq`(`_lhs`)), env, env))
#> env:  [frame 0x7f848abd25f0]
#> 
#> [[31]]
#> <frame 3> (2)
#> expr: pkg$topics %>% purrr::transpose() %>% purrr::map(build_reference_topic,  <...>
#> env:  [frame 0x7f848adb4d78]
#> 
#> [[32]]
#> <frame 2> (1)
#> expr: build_reference(pkg, lazy = FALSE, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f8492ff3f40]
#> 
#> [[33]]
#> <frame 1> (0)
#> expr: pkgdown::build_site("/Users/yoplait/Dropbox/Projects/R/hadley/rlang")
#> env:  [frame 0x7f848fb31d48]
#> 
#> [[34]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(ctxt_stack(trim = 1)))

#> [[1]]
#> <frame 31> (30)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f84904784a8]
#> 
#> [[2]]
#> <frame 30> (20)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f8493a96428]
#> 
#> [[3]]
#> <frame 29> (20)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f848c0ac2e0]
#> 
#> [[4]]
#> <frame 28> (20)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f8491d256b0]
#> 
#> [[5]]
#> <frame 27> (26)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f8491d25c60]
#> 
#> [[6]]
#> <frame 26> (25)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f84918bff50]
#> 
#> [[7]]
#> <frame 25> (24)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f84918c0298]
#> 
#> [[8]]
#> <frame 24> (23)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f8491a508d0]
#> 
#> [[9]]
#> <frame 23> (22)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f8491a63f58]
#> 
#> [[10]]
#> <frame 22> (20)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f8491a64118]
#> 
#> [[11]]
#> <frame 21> (20)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f8491a64230]
#> 
#> [[12]]
#> <frame 20> (19)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f848a094150]
#> 
#> [[13]]
#> <frame 19> (18)
#> expr: evaluate::evaluate(text, code_env, new_device = TRUE)
#> env:  [frame 0x7f848e79fa60]
#> 
#> [[14]]
#> <frame 18> (16)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f8494f0bb80]
#> 
#> [[15]]
#> <frame 17> (16)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f8494f0a988]
#> 
#> [[16]]
#> <frame 16> (12)
#> expr: data_reference_topic(topic, pkg, path = path, examples = examples,  <...>
#> env:  [frame 0x7f848d27df40]
#> 
#> [[17]]
#> <frame 15> (14)
#> expr: stopifnot(is.list(x), is.list(val))
#> env:  [frame 0x7f848d27d990]
#> 
#> [[18]]
#> <frame 14> (13)
#> expr: utils::modifyList(data, data_template(pkg, depth = depth))
#> env:  [frame 0x7f848d280740]
#> 
#> [[19]]
#> <frame 13> (12)
#> expr: render_page(pkg, "reference-topic", data = data_reference_topic(topic,  <...>
#> env:  [frame 0x7f848d2e9318]
#> 
#> [[20]]
#> <frame 12> (11)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f848d2ed2a8]
#> 
#> [[21]]
#> <frame 11> (10)
#> expr: purrr::map(., build_reference_topic, path, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f848aba6200]
#> 
#> [[22]]
#> <frame 10> (8)
#> expr: function_list[[k]](value)
#> env:  [frame 0x7f848abc34d8]
#> 
#> [[23]]
#> <frame 9> (8)
#> expr: withVisible(function_list[[k]](value))
#> env:  [frame 0x7f848abc33c0]
#> 
#> [[24]]
#> <frame 8> (7)
#> expr: freduce(value, `_function_list`)
#> env:  [frame 0x7f848abd1468]
#> 
#> [[25]]
#> <frame 7> (6)
#> expr: `_fseq`(`_lhs`)
#> env:  [frame 0x7f848abd12a8]
#> 
#> [[26]]
#> <frame 6> (5)
#> expr: eval(expr, envir, enclos)
#> env:  [frame 0x7f848adb4230]
#> 
#> [[27]]
#> <frame 5> (3)
#> expr: eval(quote(`_fseq`(`_lhs`)), env, env)
#> env:  [frame 0x7f848abd1078]
#> 
#> [[28]]
#> <frame 4> (3)
#> expr: withVisible(eval(quote(`_fseq`(`_lhs`)), env, env))
#> env:  [frame 0x7f848abd25f0]
#> 
#> [[29]]
#> <frame 3> (2)
#> expr: pkg$topics %>% purrr::transpose() %>% purrr::map(build_reference_topic,  <...>
#> env:  [frame 0x7f848adb4d78]
#> 
#> [[30]]
#> <frame 2> (1)
#> expr: build_reference(pkg, lazy = FALSE, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f8492ff3f40]
#> 
#> [[31]]
#> <frame 1> (0)
#> expr: pkgdown::build_site("/Users/yoplait/Dropbox/Projects/R/hadley/rlang")
#> env:  [frame 0x7f848fb31d48]
#> 
#> [[32]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# ctxt_stack() is called within fn() with intervening frames:
fn <- function(trim) identity(identity(ctxt_stack(trim = trim)))
fn(0)

#> [[1]]
#> <frame 34> (32)
#> expr: identity(ctxt_stack(trim = trim))
#> env:  [local 0x7f848ded8fc0]
#> 
#> [[2]]
#> <frame 33> (32)
#> expr: identity(identity(ctxt_stack(trim = trim)))
#> env:  [local 0x7f848ded9110]
#> 
#> [[3]]
#> <frame 32> (31)
#> expr: fn(0)
#> env:  [local 0x7f848ded9228]
#> 
#> [[4]]
#> <frame 31> (30)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f84904784a8]
#> 
#> [[5]]
#> <frame 30> (20)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f848df15240]
#> 
#> [[6]]
#> <frame 29> (20)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f848df154a8]
#> 
#> [[7]]
#> <frame 28> (20)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f848df3f6e8]
#> 
#> [[8]]
#> <frame 27> (26)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f848df3fc98]
#> 
#> [[9]]
#> <frame 26> (25)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f848df43c38]
#> 
#> [[10]]
#> <frame 25> (24)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f848df43f80]
#> 
#> [[11]]
#> <frame 24> (23)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f848df97ad8]
#> 
#> [[12]]
#> <frame 23> (22)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f848df97cd0]
#> 
#> [[13]]
#> <frame 22> (20)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f848dfd20e8]
#> 
#> [[14]]
#> <frame 21> (20)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f848dfd2200]
#> 
#> [[15]]
#> <frame 20> (19)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f848c67c868]
#> 
#> [[16]]
#> <frame 19> (18)
#> expr: evaluate::evaluate(text, code_env, new_device = TRUE)
#> env:  [frame 0x7f848e79fa60]
#> 
#> [[17]]
#> <frame 18> (16)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f8494f0bb80]
#> 
#> [[18]]
#> <frame 17> (16)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f8494f0a988]
#> 
#> [[19]]
#> <frame 16> (12)
#> expr: data_reference_topic(topic, pkg, path = path, examples = examples,  <...>
#> env:  [frame 0x7f848d27df40]
#> 
#> [[20]]
#> <frame 15> (14)
#> expr: stopifnot(is.list(x), is.list(val))
#> env:  [frame 0x7f848d27d990]
#> 
#> [[21]]
#> <frame 14> (13)
#> expr: utils::modifyList(data, data_template(pkg, depth = depth))
#> env:  [frame 0x7f848d280740]
#> 
#> [[22]]
#> <frame 13> (12)
#> expr: render_page(pkg, "reference-topic", data = data_reference_topic(topic,  <...>
#> env:  [frame 0x7f848d2e9318]
#> 
#> [[23]]
#> <frame 12> (11)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f848d2ed2a8]
#> 
#> [[24]]
#> <frame 11> (10)
#> expr: purrr::map(., build_reference_topic, path, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f848aba6200]
#> 
#> [[25]]
#> <frame 10> (8)
#> expr: function_list[[k]](value)
#> env:  [frame 0x7f848abc34d8]
#> 
#> [[26]]
#> <frame 9> (8)
#> expr: withVisible(function_list[[k]](value))
#> env:  [frame 0x7f848abc33c0]
#> 
#> [[27]]
#> <frame 8> (7)
#> expr: freduce(value, `_function_list`)
#> env:  [frame 0x7f848abd1468]
#> 
#> [[28]]
#> <frame 7> (6)
#> expr: `_fseq`(`_lhs`)
#> env:  [frame 0x7f848abd12a8]
#> 
#> [[29]]
#> <frame 6> (5)
#> expr: eval(expr, envir, enclos)
#> env:  [frame 0x7f848adb4230]
#> 
#> [[30]]
#> <frame 5> (3)
#> expr: eval(quote(`_fseq`(`_lhs`)), env, env)
#> env:  [frame 0x7f848abd1078]
#> 
#> [[31]]
#> <frame 4> (3)
#> expr: withVisible(eval(quote(`_fseq`(`_lhs`)), env, env))
#> env:  [frame 0x7f848abd25f0]
#> 
#> [[32]]
#> <frame 3> (2)
#> expr: pkg$topics %>% purrr::transpose() %>% purrr::map(build_reference_topic,  <...>
#> env:  [frame 0x7f848adb4d78]
#> 
#> [[33]]
#> <frame 2> (1)
#> expr: build_reference(pkg, lazy = FALSE, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f8492ff3f40]
#> 
#> [[34]]
#> <frame 1> (0)
#> expr: pkgdown::build_site("/Users/yoplait/Dropbox/Projects/R/hadley/rlang")
#> env:  [frame 0x7f848fb31d48]
#> 
#> [[35]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# We can trim the first layer of those:
fn(1)

#> [[1]]
#> <frame 32> (31)
#> expr: fn(1)
#> env:  [local 0x7f84904ee7c0]
#> 
#> [[2]]
#> <frame 31> (30)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f84904784a8]
#> 
#> [[3]]
#> <frame 30> (20)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f84904ee600]
#> 
#> [[4]]
#> <frame 29> (20)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f84904d7a28]
#> 
#> [[5]]
#> <frame 28> (20)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f84904d66a8]
#> 
#> [[6]]
#> <frame 27> (26)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f84904cd030]
#> 
#> [[7]]
#> <frame 26> (25)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f84904ccce8]
#> 
#> [[8]]
#> <frame 25> (24)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f84904a8148]
#> 
#> [[9]]
#> <frame 24> (23)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f84904a21b8]
#> 
#> [[10]]
#> <frame 23> (22)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f84904a1fc0]
#> 
#> [[11]]
#> <frame 22> (20)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f84904a1d90]
#> 
#> [[12]]
#> <frame 21> (20)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f84904a1c78]
#> 
#> [[13]]
#> <frame 20> (19)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f848e646a08]
#> 
#> [[14]]
#> <frame 19> (18)
#> expr: evaluate::evaluate(text, code_env, new_device = TRUE)
#> env:  [frame 0x7f848e79fa60]
#> 
#> [[15]]
#> <frame 18> (16)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f8494f0bb80]
#> 
#> [[16]]
#> <frame 17> (16)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f8494f0a988]
#> 
#> [[17]]
#> <frame 16> (12)
#> expr: data_reference_topic(topic, pkg, path = path, examples = examples,  <...>
#> env:  [frame 0x7f848d27df40]
#> 
#> [[18]]
#> <frame 15> (14)
#> expr: stopifnot(is.list(x), is.list(val))
#> env:  [frame 0x7f848d27d990]
#> 
#> [[19]]
#> <frame 14> (13)
#> expr: utils::modifyList(data, data_template(pkg, depth = depth))
#> env:  [frame 0x7f848d280740]
#> 
#> [[20]]
#> <frame 13> (12)
#> expr: render_page(pkg, "reference-topic", data = data_reference_topic(topic,  <...>
#> env:  [frame 0x7f848d2e9318]
#> 
#> [[21]]
#> <frame 12> (11)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f848d2ed2a8]
#> 
#> [[22]]
#> <frame 11> (10)
#> expr: purrr::map(., build_reference_topic, path, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f848aba6200]
#> 
#> [[23]]
#> <frame 10> (8)
#> expr: function_list[[k]](value)
#> env:  [frame 0x7f848abc34d8]
#> 
#> [[24]]
#> <frame 9> (8)
#> expr: withVisible(function_list[[k]](value))
#> env:  [frame 0x7f848abc33c0]
#> 
#> [[25]]
#> <frame 8> (7)
#> expr: freduce(value, `_function_list`)
#> env:  [frame 0x7f848abd1468]
#> 
#> [[26]]
#> <frame 7> (6)
#> expr: `_fseq`(`_lhs`)
#> env:  [frame 0x7f848abd12a8]
#> 
#> [[27]]
#> <frame 6> (5)
#> expr: eval(expr, envir, enclos)
#> env:  [frame 0x7f848adb4230]
#> 
#> [[28]]
#> <frame 5> (3)
#> expr: eval(quote(`_fseq`(`_lhs`)), env, env)
#> env:  [frame 0x7f848abd1078]
#> 
#> [[29]]
#> <frame 4> (3)
#> expr: withVisible(eval(quote(`_fseq`(`_lhs`)), env, env))
#> env:  [frame 0x7f848abd25f0]
#> 
#> [[30]]
#> <frame 3> (2)
#> expr: pkg$topics %>% purrr::transpose() %>% purrr::map(build_reference_topic,  <...>
#> env:  [frame 0x7f848adb4d78]
#> 
#> [[31]]
#> <frame 2> (1)
#> expr: build_reference(pkg, lazy = FALSE, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f8492ff3f40]
#> 
#> [[32]]
#> <frame 1> (0)
#> expr: pkgdown::build_site("/Users/yoplait/Dropbox/Projects/R/hadley/rlang")
#> env:  [frame 0x7f848fb31d48]
#> 
#> [[33]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# The outside intervening frames (at the fn() call site) are still
# returned, but can be trimmed as well:
identity(identity(fn(1)))

#> [[1]]
#> <frame 34> (31)
#> expr: fn(1)
#> env:  [local 0x7f848e0050b0]
#> 
#> [[2]]
#> <frame 33> (31)
#> expr: identity(fn(1))
#> env:  [local 0x7f848e535940]
#> 
#> [[3]]
#> <frame 32> (31)
#> expr: identity(identity(fn(1)))
#> env:  [local 0x7f848e535828]
#> 
#> [[4]]
#> <frame 31> (30)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f84904784a8]
#> 
#> [[5]]
#> <frame 30> (20)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f848e535668]
#> 
#> [[6]]
#> <frame 29> (20)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f848e535390]
#> 
#> [[7]]
#> <frame 28> (20)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f848ad33f18]
#> 
#> [[8]]
#> <frame 27> (26)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f848ad348a0]
#> 
#> [[9]]
#> <frame 26> (25)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f848ad34558]
#> 
#> [[10]]
#> <frame 25> (24)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f849483ad80]
#> 
#> [[11]]
#> <frame 24> (23)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f8494c6b628]
#> 
#> [[12]]
#> <frame 23> (22)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f8494c6b350]
#> 
#> [[13]]
#> <frame 22> (20)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f8494c6b190]
#> 
#> [[14]]
#> <frame 21> (20)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f8494c6b078]
#> 
#> [[15]]
#> <frame 20> (19)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f8494dc4860]
#> 
#> [[16]]
#> <frame 19> (18)
#> expr: evaluate::evaluate(text, code_env, new_device = TRUE)
#> env:  [frame 0x7f848e79fa60]
#> 
#> [[17]]
#> <frame 18> (16)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f8494f0bb80]
#> 
#> [[18]]
#> <frame 17> (16)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f8494f0a988]
#> 
#> [[19]]
#> <frame 16> (12)
#> expr: data_reference_topic(topic, pkg, path = path, examples = examples,  <...>
#> env:  [frame 0x7f848d27df40]
#> 
#> [[20]]
#> <frame 15> (14)
#> expr: stopifnot(is.list(x), is.list(val))
#> env:  [frame 0x7f848d27d990]
#> 
#> [[21]]
#> <frame 14> (13)
#> expr: utils::modifyList(data, data_template(pkg, depth = depth))
#> env:  [frame 0x7f848d280740]
#> 
#> [[22]]
#> <frame 13> (12)
#> expr: render_page(pkg, "reference-topic", data = data_reference_topic(topic,  <...>
#> env:  [frame 0x7f848d2e9318]
#> 
#> [[23]]
#> <frame 12> (11)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f848d2ed2a8]
#> 
#> [[24]]
#> <frame 11> (10)
#> expr: purrr::map(., build_reference_topic, path, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f848aba6200]
#> 
#> [[25]]
#> <frame 10> (8)
#> expr: function_list[[k]](value)
#> env:  [frame 0x7f848abc34d8]
#> 
#> [[26]]
#> <frame 9> (8)
#> expr: withVisible(function_list[[k]](value))
#> env:  [frame 0x7f848abc33c0]
#> 
#> [[27]]
#> <frame 8> (7)
#> expr: freduce(value, `_function_list`)
#> env:  [frame 0x7f848abd1468]
#> 
#> [[28]]
#> <frame 7> (6)
#> expr: `_fseq`(`_lhs`)
#> env:  [frame 0x7f848abd12a8]
#> 
#> [[29]]
#> <frame 6> (5)
#> expr: eval(expr, envir, enclos)
#> env:  [frame 0x7f848adb4230]
#> 
#> [[30]]
#> <frame 5> (3)
#> expr: eval(quote(`_fseq`(`_lhs`)), env, env)
#> env:  [frame 0x7f848abd1078]
#> 
#> [[31]]
#> <frame 4> (3)
#> expr: withVisible(eval(quote(`_fseq`(`_lhs`)), env, env))
#> env:  [frame 0x7f848abd25f0]
#> 
#> [[32]]
#> <frame 3> (2)
#> expr: pkg$topics %>% purrr::transpose() %>% purrr::map(build_reference_topic,  <...>
#> env:  [frame 0x7f848adb4d78]
#> 
#> [[33]]
#> <frame 2> (1)
#> expr: build_reference(pkg, lazy = FALSE, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f8492ff3f40]
#> 
#> [[34]]
#> <frame 1> (0)
#> expr: pkgdown::build_site("/Users/yoplait/Dropbox/Projects/R/hadley/rlang")
#> env:  [frame 0x7f848fb31d48]
#> 
#> [[35]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(fn(2)))

#> [[1]]
#> <frame 31> (30)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f84904784a8]
#> 
#> [[2]]
#> <frame 30> (20)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f8493881698]
#> 
#> [[3]]
#> <frame 29> (20)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f8493881430]
#> 
#> [[4]]
#> <frame 28> (20)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f8493881d10]
#> 
#> [[5]]
#> <frame 27> (26)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f8493ec5098]
#> 
#> [[6]]
#> <frame 26> (25)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f8493ec4ce0]
#> 
#> [[7]]
#> <frame 25> (24)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f848a4cfed0]
#> 
#> [[8]]
#> <frame 24> (23)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f848a2c4540]
#> 
#> [[9]]
#> <frame 23> (22)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f848a2c4310]
#> 
#> [[10]]
#> <frame 22> (20)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f848a2c4150]
#> 
#> [[11]]
#> <frame 21> (20)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f848a2c4038]
#> 
#> [[12]]
#> <frame 20> (19)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f8493fae038]
#> 
#> [[13]]
#> <frame 19> (18)
#> expr: evaluate::evaluate(text, code_env, new_device = TRUE)
#> env:  [frame 0x7f848e79fa60]
#> 
#> [[14]]
#> <frame 18> (16)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f8494f0bb80]
#> 
#> [[15]]
#> <frame 17> (16)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f8494f0a988]
#> 
#> [[16]]
#> <frame 16> (12)
#> expr: data_reference_topic(topic, pkg, path = path, examples = examples,  <...>
#> env:  [frame 0x7f848d27df40]
#> 
#> [[17]]
#> <frame 15> (14)
#> expr: stopifnot(is.list(x), is.list(val))
#> env:  [frame 0x7f848d27d990]
#> 
#> [[18]]
#> <frame 14> (13)
#> expr: utils::modifyList(data, data_template(pkg, depth = depth))
#> env:  [frame 0x7f848d280740]
#> 
#> [[19]]
#> <frame 13> (12)
#> expr: render_page(pkg, "reference-topic", data = data_reference_topic(topic,  <...>
#> env:  [frame 0x7f848d2e9318]
#> 
#> [[20]]
#> <frame 12> (11)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f848d2ed2a8]
#> 
#> [[21]]
#> <frame 11> (10)
#> expr: purrr::map(., build_reference_topic, path, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f848aba6200]
#> 
#> [[22]]
#> <frame 10> (8)
#> expr: function_list[[k]](value)
#> env:  [frame 0x7f848abc34d8]
#> 
#> [[23]]
#> <frame 9> (8)
#> expr: withVisible(function_list[[k]](value))
#> env:  [frame 0x7f848abc33c0]
#> 
#> [[24]]
#> <frame 8> (7)
#> expr: freduce(value, `_function_list`)
#> env:  [frame 0x7f848abd1468]
#> 
#> [[25]]
#> <frame 7> (6)
#> expr: `_fseq`(`_lhs`)
#> env:  [frame 0x7f848abd12a8]
#> 
#> [[26]]
#> <frame 6> (5)
#> expr: eval(expr, envir, enclos)
#> env:  [frame 0x7f848adb4230]
#> 
#> [[27]]
#> <frame 5> (3)
#> expr: eval(quote(`_fseq`(`_lhs`)), env, env)
#> env:  [frame 0x7f848abd1078]
#> 
#> [[28]]
#> <frame 4> (3)
#> expr: withVisible(eval(quote(`_fseq`(`_lhs`)), env, env))
#> env:  [frame 0x7f848abd25f0]
#> 
#> [[29]]
#> <frame 3> (2)
#> expr: pkg$topics %>% purrr::transpose() %>% purrr::map(build_reference_topic,  <...>
#> env:  [frame 0x7f848adb4d78]
#> 
#> [[30]]
#> <frame 2> (1)
#> expr: build_reference(pkg, lazy = FALSE, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f8492ff3f40]
#> 
#> [[31]]
#> <frame 1> (0)
#> expr: pkgdown::build_site("/Users/yoplait/Dropbox/Projects/R/hadley/rlang")
#> env:  [frame 0x7f848fb31d48]
#> 
#> [[32]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g <- function(trim) identity(identity(fn(trim)))
g(2)

#> [[1]]
#> <frame 32> (31)
#> expr: g(2)
#> env:  [local 0x7f848d6847b8]
#> 
#> [[2]]
#> <frame 31> (30)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f84904784a8]
#> 
#> [[3]]
#> <frame 30> (20)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f848d6845f8]
#> 
#> [[4]]
#> <frame 29> (20)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f848d6842e8]
#> 
#> [[5]]
#> <frame 28> (20)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f848a4a4660]
#> 
#> [[6]]
#> <frame 27> (26)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f848a4a4078]
#> 
#> [[7]]
#> <frame 26> (25)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f848a4a4bf8]
#> 
#> [[8]]
#> <frame 25> (24)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f848a4a4878]
#> 
#> [[9]]
#> <frame 24> (23)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f848b4b5628]
#> 
#> [[10]]
#> <frame 23> (22)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f848b4b53f8]
#> 
#> [[11]]
#> <frame 22> (20)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f848b4b5200]
#> 
#> [[12]]
#> <frame 21> (20)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f848b4b50e8]
#> 
#> [[13]]
#> <frame 20> (19)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f848b204188]
#> 
#> [[14]]
#> <frame 19> (18)
#> expr: evaluate::evaluate(text, code_env, new_device = TRUE)
#> env:  [frame 0x7f848e79fa60]
#> 
#> [[15]]
#> <frame 18> (16)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f8494f0bb80]
#> 
#> [[16]]
#> <frame 17> (16)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f8494f0a988]
#> 
#> [[17]]
#> <frame 16> (12)
#> expr: data_reference_topic(topic, pkg, path = path, examples = examples,  <...>
#> env:  [frame 0x7f848d27df40]
#> 
#> [[18]]
#> <frame 15> (14)
#> expr: stopifnot(is.list(x), is.list(val))
#> env:  [frame 0x7f848d27d990]
#> 
#> [[19]]
#> <frame 14> (13)
#> expr: utils::modifyList(data, data_template(pkg, depth = depth))
#> env:  [frame 0x7f848d280740]
#> 
#> [[20]]
#> <frame 13> (12)
#> expr: render_page(pkg, "reference-topic", data = data_reference_topic(topic,  <...>
#> env:  [frame 0x7f848d2e9318]
#> 
#> [[21]]
#> <frame 12> (11)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f848d2ed2a8]
#> 
#> [[22]]
#> <frame 11> (10)
#> expr: purrr::map(., build_reference_topic, path, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f848aba6200]
#> 
#> [[23]]
#> <frame 10> (8)
#> expr: function_list[[k]](value)
#> env:  [frame 0x7f848abc34d8]
#> 
#> [[24]]
#> <frame 9> (8)
#> expr: withVisible(function_list[[k]](value))
#> env:  [frame 0x7f848abc33c0]
#> 
#> [[25]]
#> <frame 8> (7)
#> expr: freduce(value, `_function_list`)
#> env:  [frame 0x7f848abd1468]
#> 
#> [[26]]
#> <frame 7> (6)
#> expr: `_fseq`(`_lhs`)
#> env:  [frame 0x7f848abd12a8]
#> 
#> [[27]]
#> <frame 6> (5)
#> expr: eval(expr, envir, enclos)
#> env:  [frame 0x7f848adb4230]
#> 
#> [[28]]
#> <frame 5> (3)
#> expr: eval(quote(`_fseq`(`_lhs`)), env, env)
#> env:  [frame 0x7f848abd1078]
#> 
#> [[29]]
#> <frame 4> (3)
#> expr: withVisible(eval(quote(`_fseq`(`_lhs`)), env, env))
#> env:  [frame 0x7f848abd25f0]
#> 
#> [[30]]
#> <frame 3> (2)
#> expr: pkg$topics %>% purrr::transpose() %>% purrr::map(build_reference_topic,  <...>
#> env:  [frame 0x7f848adb4d78]
#> 
#> [[31]]
#> <frame 2> (1)
#> expr: build_reference(pkg, lazy = FALSE, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f8492ff3f40]
#> 
#> [[32]]
#> <frame 1> (0)
#> expr: pkgdown::build_site("/Users/yoplait/Dropbox/Projects/R/hadley/rlang")
#> env:  [frame 0x7f848fb31d48]
#> 
#> [[33]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g(3)

#> [[1]]
#> <frame 31> (30)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f84904784a8]
#> 
#> [[2]]
#> <frame 30> (20)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f848c221040]
#> 
#> [[3]]
#> <frame 29> (20)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f848c2bcd48]
#> 
#> [[4]]
#> <frame 28> (20)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f848c2f3bf8]
#> 
#> [[5]]
#> <frame 27> (26)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f848c3520d8]
#> 
#> [[6]]
#> <frame 26> (25)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f848c351d90]
#> 
#> [[7]]
#> <frame 25> (24)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f848c425348]
#> 
#> [[8]]
#> <frame 24> (23)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f848c0bcbc0]
#> 
#> [[9]]
#> <frame 23> (22)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f848c0bc958]
#> 
#> [[10]]
#> <frame 22> (20)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f848c0d5708]
#> 
#> [[11]]
#> <frame 21> (20)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f848c0d5510]
#> 
#> [[12]]
#> <frame 20> (19)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f848d597390]
#> 
#> [[13]]
#> <frame 19> (18)
#> expr: evaluate::evaluate(text, code_env, new_device = TRUE)
#> env:  [frame 0x7f848e79fa60]
#> 
#> [[14]]
#> <frame 18> (16)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f8494f0bb80]
#> 
#> [[15]]
#> <frame 17> (16)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f8494f0a988]
#> 
#> [[16]]
#> <frame 16> (12)
#> expr: data_reference_topic(topic, pkg, path = path, examples = examples,  <...>
#> env:  [frame 0x7f848d27df40]
#> 
#> [[17]]
#> <frame 15> (14)
#> expr: stopifnot(is.list(x), is.list(val))
#> env:  [frame 0x7f848d27d990]
#> 
#> [[18]]
#> <frame 14> (13)
#> expr: utils::modifyList(data, data_template(pkg, depth = depth))
#> env:  [frame 0x7f848d280740]
#> 
#> [[19]]
#> <frame 13> (12)
#> expr: render_page(pkg, "reference-topic", data = data_reference_topic(topic,  <...>
#> env:  [frame 0x7f848d2e9318]
#> 
#> [[20]]
#> <frame 12> (11)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f848d2ed2a8]
#> 
#> [[21]]
#> <frame 11> (10)
#> expr: purrr::map(., build_reference_topic, path, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f848aba6200]
#> 
#> [[22]]
#> <frame 10> (8)
#> expr: function_list[[k]](value)
#> env:  [frame 0x7f848abc34d8]
#> 
#> [[23]]
#> <frame 9> (8)
#> expr: withVisible(function_list[[k]](value))
#> env:  [frame 0x7f848abc33c0]
#> 
#> [[24]]
#> <frame 8> (7)
#> expr: freduce(value, `_function_list`)
#> env:  [frame 0x7f848abd1468]
#> 
#> [[25]]
#> <frame 7> (6)
#> expr: `_fseq`(`_lhs`)
#> env:  [frame 0x7f848abd12a8]
#> 
#> [[26]]
#> <frame 6> (5)
#> expr: eval(expr, envir, enclos)
#> env:  [frame 0x7f848adb4230]
#> 
#> [[27]]
#> <frame 5> (3)
#> expr: eval(quote(`_fseq`(`_lhs`)), env, env)
#> env:  [frame 0x7f848abd1078]
#> 
#> [[28]]
#> <frame 4> (3)
#> expr: withVisible(eval(quote(`_fseq`(`_lhs`)), env, env))
#> env:  [frame 0x7f848abd25f0]
#> 
#> [[29]]
#> <frame 3> (2)
#> expr: pkg$topics %>% purrr::transpose() %>% purrr::map(build_reference_topic,  <...>
#> env:  [frame 0x7f848adb4d78]
#> 
#> [[30]]
#> <frame 2> (1)
#> expr: build_reference(pkg, lazy = FALSE, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f8492ff3f40]
#> 
#> [[31]]
#> <frame 1> (0)
#> expr: pkgdown::build_site("/Users/yoplait/Dropbox/Projects/R/hadley/rlang")
#> env:  [frame 0x7f848fb31d48]
#> 
#> [[32]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"