Call stack information

The eval_ and call_ families of functions provide a replacement for the base R functions prefixed with sys. (which are all about the context stack), as well as for parent.frame() (which is the only base R function for querying the call stack). The context stack includes all R-level evaluation contexts. It is linear in terms of execution history but due to lazy evaluation it is potentially nonlinear in terms of call history. The call stack history, on the other hand, is homogenous.

global_frame()

current_frame()

ctxt_frame(n = 1)

call_frame(n = 1, clean = TRUE)

ctxt_depth()

call_depth()

ctxt_stack(n = NULL, trim = 0)

call_stack(n = NULL, clean = TRUE)

Arguments

n
The number of frames to go back in the stack.

clean
Whether to post-process the call stack to clean non-standard frames. If TRUE, suboptimal call-stack entries by base::eval() will be cleaned up: the duplicate frame created by eval() is eliminated.

trim
The number of layers of intervening frames to trim off the stack. See stack_trim() and examples.

Details

ctxt_frame() and call_frame() return a frame object containing the following fields: expr and env (call expression and evaluation environment), pos and caller_pos (position of current frame in the context stack and position of the caller), and fun (function of the current frame). ctxt_stack() and call_stack() return a list of all context or call frames on the stack. Finally, ctxt_depth() and call_depth() report the current context position or the number of calling frames on the stack.

The base R functions take two sorts of arguments to indicate which frame to query: which and n. The n argument is straightforward: it's the number of frames to go down the stack, with n = 1 referring to the current context. The which argument is more complicated and changes meaning for values lower than 1. For the sake of consistency, the rlang functions all take the same kind of argument n. This argument has a single meaning (the number of frames to go down the stack) and cannot be lower than 1.

Note finally that parent.frame(1) corresponds to call_frame(2)$env, as n = 1 always refers to the current frame. This makes the _frame() and _stack() functions consistent: ctxt_frame(2) is the same as ctxt_stack()[[2]]. Also, ctxt_depth() returns one more frame than base::sys.nframe() because it counts the global frame. That is consistent with the _stack() functions which return the global frame as well. This way, call_stack(call_depth()) is the same as global_frame().

Life cycle

These functions are soft-deprecated and replaced by trace_back().

Examples

# Expressions within arguments count as contexts
identity(identity(ctxt_depth())) # returns 2

#> Warning: `ctxt_depth()` is soft-deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 40

# But they are not part of the call stack because arguments are
# evaluated within the calling function (or the global environment
# if called at top level)
identity(identity(call_depth())) # returns 0

#> Warning: `call_depth()` is soft-deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 14

# The context stacks includes all intervening execution frames. The
# call stack doesn't:
f <- function(x) identity(x)
f(f(ctxt_stack()))

#> [[1]]
#> <frame 41> (40)
#> expr: identity(x)
#> env:  [local 0x8c562a0]
#> 
#> [[2]]
#> <frame 40> (37)
#> expr: f(ctxt_stack())
#> env:  [local 0x8c52be0]
#> 
#> [[3]]
#> <frame 39> (38)
#> expr: identity(x)
#> env:  [local 0x8c52d30]
#> 
#> [[4]]
#> <frame 38> (37)
#> expr: f(f(ctxt_stack()))
#> env:  [local 0x8c52ef0]
#> 
#> [[5]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x8ad6148]
#> 
#> [[6]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x8c533f8]
#> 
#> [[7]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x8c53f90]
#> 
#> [[8]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x8c51318]
#> 
#> [[9]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x8c51900]
#> 
#> [[10]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x8c51c48]
#> 
#> [[11]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x8c51f90]
#> 
#> [[12]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x8c52658]
#> 
#> [[13]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x8c52888]
#> 
#> [[14]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x91dd750]
#> 
#> [[15]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x91dd830]
#> 
#> [[16]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x9b8ed30]
#> 
#> [[17]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x2dce3a8]
#> 
#> [[18]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x2ddd2b8]
#> 
#> [[19]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x2de6e18]
#> 
#> [[20]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x2ded128]
#> 
#> [[21]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x2dfd2d0]
#> 
#> [[22]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x3cd38f8]
#> 
#> [[23]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x3cd6b20]
#> 
#> [[24]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x6642be8]
#> 
#> [[25]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x67c68c0]
#> 
#> [[26]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x78af898]
#> 
#> [[27]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x5676658]
#> 
#> [[28]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x30f7608]
#> 
#> [[29]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x3101378]
#> 
#> [[30]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x40ee2d0]
#> 
#> [[31]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x40ef290]
#> 
#> [[32]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/Rtmp31Sth2/file3d5d3bfe99e9"),  <...>
#> env:  [frame 0x40df0f0]
#> 
#> [[33]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/Rtmp31Sth2/file3d5d3bfe99e9"),  <...>
#> env:  [frame 0x40dfe80]
#> 
#> [[34]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x40dac58]
#> 
#> [[35]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x40dbe10]
#> 
#> [[36]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x40dc158]
#> 
#> [[37]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x40d86e0]
#> 
#> [[38]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x40d8c90]
#> 
#> [[39]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x40d8fd8]
#> 
#> [[40]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x40d9320]
#> 
#> [[41]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x40d9a90]
#> 
#> [[42]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(f(call_stack()))

#> Warning: `call_stack()` is soft-deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [[1]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x8ad6148]
#> 
#> [[2]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x8dd6c58]
#> 
#> [[3]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x2dce3a8]
#> 
#> [[4]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x2ded128]
#> 
#> [[5]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x2dfd2d0]
#> 
#> [[6]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x3cd38f8]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x6642be8]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x67c68c0]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x78af898]
#> 
#> [[10]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x5676658]
#> 
#> [[11]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x30f7608]
#> 
#> [[12]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x3101378]
#> 
#> [[13]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x40ee2d0]
#> 
#> [[14]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

g <- function(cmd) cmd()
f(g(ctxt_stack))

#> [[1]]
#> <frame 40> (37)
#> expr: g(ctxt_stack)
#> env:  [local 0x7e704a0]
#> 
#> [[2]]
#> <frame 39> (38)
#> expr: identity(x)
#> env:  [local 0x7e70388]
#> 
#> [[3]]
#> <frame 38> (37)
#> expr: f(g(ctxt_stack))
#> env:  [local 0x7e70270]
#> 
#> [[4]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x8ad6148]
#> 
#> [[5]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7e700b0]
#> 
#> [[6]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7e6fe80]
#> 
#> [[7]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7e6f8d0]
#> 
#> [[8]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7e6f358]
#> 
#> [[9]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f7b398]
#> 
#> [[10]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f7b050]
#> 
#> [[11]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f7a3d8]
#> 
#> [[12]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f7a1a8]
#> 
#> [[13]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f7a020]
#> 
#> [[14]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f79f40]
#> 
#> [[15]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x768df00]
#> 
#> [[16]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x2dce3a8]
#> 
#> [[17]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x2ddd2b8]
#> 
#> [[18]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x2de6e18]
#> 
#> [[19]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x2ded128]
#> 
#> [[20]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x2dfd2d0]
#> 
#> [[21]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x3cd38f8]
#> 
#> [[22]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x3cd6b20]
#> 
#> [[23]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x6642be8]
#> 
#> [[24]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x67c68c0]
#> 
#> [[25]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x78af898]
#> 
#> [[26]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x5676658]
#> 
#> [[27]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x30f7608]
#> 
#> [[28]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x3101378]
#> 
#> [[29]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x40ee2d0]
#> 
#> [[30]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x40ef290]
#> 
#> [[31]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/Rtmp31Sth2/file3d5d3bfe99e9"),  <...>
#> env:  [frame 0x40df0f0]
#> 
#> [[32]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/Rtmp31Sth2/file3d5d3bfe99e9"),  <...>
#> env:  [frame 0x40dfe80]
#> 
#> [[33]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x40dac58]
#> 
#> [[34]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x40dbe10]
#> 
#> [[35]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x40dc158]
#> 
#> [[36]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x40d86e0]
#> 
#> [[37]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x40d8c90]
#> 
#> [[38]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x40d8fd8]
#> 
#> [[39]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x40d9320]
#> 
#> [[40]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x40d9a90]
#> 
#> [[41]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(g(call_stack))

#> [[1]]
#> <frame 40> (36)
#> expr: g(call_stack)
#> env:  [local 0x63eaea0]
#> 
#> [[2]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x8ad6148]
#> 
#> [[3]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x6497790]
#> 
#> [[4]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x2dce3a8]
#> 
#> [[5]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x2ded128]
#> 
#> [[6]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x2dfd2d0]
#> 
#> [[7]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x3cd38f8]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x6642be8]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x67c68c0]
#> 
#> [[10]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x78af898]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x5676658]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x30f7608]
#> 
#> [[13]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x3101378]
#> 
#> [[14]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x40ee2d0]
#> 
#> [[15]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

# The rlang _stack() functions return a list of frame
# objects. Use purrr::transpose() or index a field with
# purrr::map()'s to extract a particular field from a stack:

# stack <- f(f(call_stack()))
# purrr::map(stack, "env")
# purrr::transpose(stack)$expr

# current_frame() is an alias for ctxt_frame(1)
fn <- function() list(current = current_frame(), first = ctxt_frame(1))
fn()

#> Warning: `current_frame()` is soft-deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> $current
#> <frame 38> (37)
#> expr: fn()
#> env:  [local 0x3295c58]
#> 
#> $first
#> <frame 38> (37)
#> expr: fn()
#> env:  [local 0x3295c58]
#> 

# While current_frame() is the top of the stack, global_frame() is
# the bottom:
fn <- function() {
  n <- ctxt_depth()
  ctxt_frame(n)
}
identical(fn(), global_frame())

#> Warning: `global_frame()` is soft-deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] TRUE

# ctxt_stack() returns a stack with all intervening frames. You can
# trim layers of intervening frames with the trim argument:
identity(identity(ctxt_stack()))

#> [[1]]
#> <frame 39> (37)
#> expr: identity(ctxt_stack())
#> env:  [local 0x8fd6f10]
#> 
#> [[2]]
#> <frame 38> (37)
#> expr: identity(identity(ctxt_stack()))
#> env:  [local 0x8fd7028]
#> 
#> [[3]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x8ad6148]
#> 
#> [[4]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x8fd71e8]
#> 
#> [[5]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x8fd7418]
#> 
#> [[6]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x8fd79c8]
#> 
#> [[7]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x8fd7f40]
#> 
#> [[8]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x8fd8288]
#> 
#> [[9]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x8fcb5f0]
#> 
#> [[10]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x8fcbcb8]
#> 
#> [[11]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x8fcbf20]
#> 
#> [[12]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x8fcc0a8]
#> 
#> [[13]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x8fcc188]
#> 
#> [[14]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x8e15450]
#> 
#> [[15]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x2dce3a8]
#> 
#> [[16]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x2ddd2b8]
#> 
#> [[17]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x2de6e18]
#> 
#> [[18]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x2ded128]
#> 
#> [[19]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x2dfd2d0]
#> 
#> [[20]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x3cd38f8]
#> 
#> [[21]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x3cd6b20]
#> 
#> [[22]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x6642be8]
#> 
#> [[23]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x67c68c0]
#> 
#> [[24]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x78af898]
#> 
#> [[25]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x5676658]
#> 
#> [[26]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x30f7608]
#> 
#> [[27]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x3101378]
#> 
#> [[28]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x40ee2d0]
#> 
#> [[29]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x40ef290]
#> 
#> [[30]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/Rtmp31Sth2/file3d5d3bfe99e9"),  <...>
#> env:  [frame 0x40df0f0]
#> 
#> [[31]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/Rtmp31Sth2/file3d5d3bfe99e9"),  <...>
#> env:  [frame 0x40dfe80]
#> 
#> [[32]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x40dac58]
#> 
#> [[33]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x40dbe10]
#> 
#> [[34]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x40dc158]
#> 
#> [[35]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x40d86e0]
#> 
#> [[36]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x40d8c90]
#> 
#> [[37]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x40d8fd8]
#> 
#> [[38]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x40d9320]
#> 
#> [[39]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x40d9a90]
#> 
#> [[40]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(ctxt_stack(trim = 1)))

#> [[1]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x8ad6148]
#> 
#> [[2]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x858ad38]
#> 
#> [[3]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x858af68]
#> 
#> [[4]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x858b518]
#> 
#> [[5]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x8587c60]
#> 
#> [[6]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x8587fe0]
#> 
#> [[7]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x8588328]
#> 
#> [[8]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x85889f0]
#> 
#> [[9]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x8588c20]
#> 
#> [[10]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x8588da8]
#> 
#> [[11]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x8588e88]
#> 
#> [[12]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x8ec7a28]
#> 
#> [[13]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x2dce3a8]
#> 
#> [[14]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x2ddd2b8]
#> 
#> [[15]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x2de6e18]
#> 
#> [[16]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x2ded128]
#> 
#> [[17]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x2dfd2d0]
#> 
#> [[18]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x3cd38f8]
#> 
#> [[19]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x3cd6b20]
#> 
#> [[20]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x6642be8]
#> 
#> [[21]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x67c68c0]
#> 
#> [[22]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x78af898]
#> 
#> [[23]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x5676658]
#> 
#> [[24]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x30f7608]
#> 
#> [[25]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x3101378]
#> 
#> [[26]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x40ee2d0]
#> 
#> [[27]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x40ef290]
#> 
#> [[28]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/Rtmp31Sth2/file3d5d3bfe99e9"),  <...>
#> env:  [frame 0x40df0f0]
#> 
#> [[29]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/Rtmp31Sth2/file3d5d3bfe99e9"),  <...>
#> env:  [frame 0x40dfe80]
#> 
#> [[30]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x40dac58]
#> 
#> [[31]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x40dbe10]
#> 
#> [[32]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x40dc158]
#> 
#> [[33]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x40d86e0]
#> 
#> [[34]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x40d8c90]
#> 
#> [[35]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x40d8fd8]
#> 
#> [[36]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x40d9320]
#> 
#> [[37]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x40d9a90]
#> 
#> [[38]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# ctxt_stack() is called within fn() with intervening frames:
fn <- function(trim) identity(identity(ctxt_stack(trim = trim)))
fn(0)

#> [[1]]
#> <frame 40> (38)
#> expr: identity(ctxt_stack(trim = trim))
#> env:  [local 0x9461ae0]
#> 
#> [[2]]
#> <frame 39> (38)
#> expr: identity(identity(ctxt_stack(trim = trim)))
#> env:  [local 0x94619c8]
#> 
#> [[3]]
#> <frame 38> (37)
#> expr: fn(0)
#> env:  [local 0x94618b0]
#> 
#> [[4]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x8ad6148]
#> 
#> [[5]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x94616f0]
#> 
#> [[6]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x94614c0]
#> 
#> [[7]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x9460ed8]
#> 
#> [[8]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x9460960]
#> 
#> [[9]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x9460618]
#> 
#> [[10]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x946e720]
#> 
#> [[11]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x946e058]
#> 
#> [[12]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x946de28]
#> 
#> [[13]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x946dca0]
#> 
#> [[14]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x946dbc0]
#> 
#> [[15]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x87e4458]
#> 
#> [[16]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x2dce3a8]
#> 
#> [[17]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x2ddd2b8]
#> 
#> [[18]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x2de6e18]
#> 
#> [[19]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x2ded128]
#> 
#> [[20]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x2dfd2d0]
#> 
#> [[21]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x3cd38f8]
#> 
#> [[22]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x3cd6b20]
#> 
#> [[23]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x6642be8]
#> 
#> [[24]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x67c68c0]
#> 
#> [[25]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x78af898]
#> 
#> [[26]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x5676658]
#> 
#> [[27]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x30f7608]
#> 
#> [[28]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x3101378]
#> 
#> [[29]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x40ee2d0]
#> 
#> [[30]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x40ef290]
#> 
#> [[31]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/Rtmp31Sth2/file3d5d3bfe99e9"),  <...>
#> env:  [frame 0x40df0f0]
#> 
#> [[32]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/Rtmp31Sth2/file3d5d3bfe99e9"),  <...>
#> env:  [frame 0x40dfe80]
#> 
#> [[33]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x40dac58]
#> 
#> [[34]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x40dbe10]
#> 
#> [[35]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x40dc158]
#> 
#> [[36]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x40d86e0]
#> 
#> [[37]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x40d8c90]
#> 
#> [[38]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x40d8fd8]
#> 
#> [[39]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x40d9320]
#> 
#> [[40]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x40d9a90]
#> 
#> [[41]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# We can trim the first layer of those:
fn(1)

#> [[1]]
#> <frame 38> (37)
#> expr: fn(1)
#> env:  [local 0x6de8118]
#> 
#> [[2]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x8ad6148]
#> 
#> [[3]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x6de7f58]
#> 
#> [[4]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x6de7cf0]
#> 
#> [[5]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x6de76d0]
#> 
#> [[6]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x6de7158]
#> 
#> [[7]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x6de6dd8]
#> 
#> [[8]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x6de6a90]
#> 
#> [[9]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x6de90b0]
#> 
#> [[10]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x6de8198]
#> 
#> [[11]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x6dedc10]
#> 
#> [[12]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x6dedb30]
#> 
#> [[13]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x6e3b4b0]
#> 
#> [[14]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x2dce3a8]
#> 
#> [[15]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x2ddd2b8]
#> 
#> [[16]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x2de6e18]
#> 
#> [[17]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x2ded128]
#> 
#> [[18]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x2dfd2d0]
#> 
#> [[19]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x3cd38f8]
#> 
#> [[20]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x3cd6b20]
#> 
#> [[21]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x6642be8]
#> 
#> [[22]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x67c68c0]
#> 
#> [[23]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x78af898]
#> 
#> [[24]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x5676658]
#> 
#> [[25]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x30f7608]
#> 
#> [[26]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x3101378]
#> 
#> [[27]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x40ee2d0]
#> 
#> [[28]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x40ef290]
#> 
#> [[29]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/Rtmp31Sth2/file3d5d3bfe99e9"),  <...>
#> env:  [frame 0x40df0f0]
#> 
#> [[30]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/Rtmp31Sth2/file3d5d3bfe99e9"),  <...>
#> env:  [frame 0x40dfe80]
#> 
#> [[31]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x40dac58]
#> 
#> [[32]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x40dbe10]
#> 
#> [[33]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x40dc158]
#> 
#> [[34]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x40d86e0]
#> 
#> [[35]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x40d8c90]
#> 
#> [[36]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x40d8fd8]
#> 
#> [[37]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x40d9320]
#> 
#> [[38]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x40d9a90]
#> 
#> [[39]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# The outside intervening frames (at the fn() call site) are still
# returned, but can be trimmed as well:
identity(identity(fn(1)))

#> [[1]]
#> <frame 40> (37)
#> expr: fn(1)
#> env:  [local 0x5cbbc60]
#> 
#> [[2]]
#> <frame 39> (37)
#> expr: identity(fn(1))
#> env:  [local 0x5cbbb48]
#> 
#> [[3]]
#> <frame 38> (37)
#> expr: identity(identity(fn(1)))
#> env:  [local 0x5cbb950]
#> 
#> [[4]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x8ad6148]
#> 
#> [[5]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5cbb790]
#> 
#> [[6]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x5cbb560]
#> 
#> [[7]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x5cc0be8]
#> 
#> [[8]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x5cc05c8]
#> 
#> [[9]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x5cc0248]
#> 
#> [[10]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x5cbff00]
#> 
#> [[11]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x5cbf5d0]
#> 
#> [[12]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x5cbf3a0]
#> 
#> [[13]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x5cbf218]
#> 
#> [[14]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x5cbf100]
#> 
#> [[15]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x5ce8ed0]
#> 
#> [[16]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x2dce3a8]
#> 
#> [[17]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x2ddd2b8]
#> 
#> [[18]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x2de6e18]
#> 
#> [[19]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x2ded128]
#> 
#> [[20]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x2dfd2d0]
#> 
#> [[21]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x3cd38f8]
#> 
#> [[22]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x3cd6b20]
#> 
#> [[23]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x6642be8]
#> 
#> [[24]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x67c68c0]
#> 
#> [[25]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x78af898]
#> 
#> [[26]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x5676658]
#> 
#> [[27]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x30f7608]
#> 
#> [[28]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x3101378]
#> 
#> [[29]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x40ee2d0]
#> 
#> [[30]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x40ef290]
#> 
#> [[31]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/Rtmp31Sth2/file3d5d3bfe99e9"),  <...>
#> env:  [frame 0x40df0f0]
#> 
#> [[32]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/Rtmp31Sth2/file3d5d3bfe99e9"),  <...>
#> env:  [frame 0x40dfe80]
#> 
#> [[33]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x40dac58]
#> 
#> [[34]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x40dbe10]
#> 
#> [[35]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x40dc158]
#> 
#> [[36]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x40d86e0]
#> 
#> [[37]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x40d8c90]
#> 
#> [[38]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x40d8fd8]
#> 
#> [[39]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x40d9320]
#> 
#> [[40]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x40d9a90]
#> 
#> [[41]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(fn(2)))

#> [[1]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x8ad6148]
#> 
#> [[2]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x32b8298]
#> 
#> [[3]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x32b7ee0]
#> 
#> [[4]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x32b7620]
#> 
#> [[5]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x32bad50]
#> 
#> [[6]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x32ba810]
#> 
#> [[7]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x32ba490]
#> 
#> [[8]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x32b9770]
#> 
#> [[9]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x32b9498]
#> 
#> [[10]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x32b92a0]
#> 
#> [[11]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x32b9188]
#> 
#> [[12]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x2c6f448]
#> 
#> [[13]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x2dce3a8]
#> 
#> [[14]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x2ddd2b8]
#> 
#> [[15]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x2de6e18]
#> 
#> [[16]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x2ded128]
#> 
#> [[17]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x2dfd2d0]
#> 
#> [[18]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x3cd38f8]
#> 
#> [[19]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x3cd6b20]
#> 
#> [[20]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x6642be8]
#> 
#> [[21]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x67c68c0]
#> 
#> [[22]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x78af898]
#> 
#> [[23]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x5676658]
#> 
#> [[24]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x30f7608]
#> 
#> [[25]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x3101378]
#> 
#> [[26]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x40ee2d0]
#> 
#> [[27]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x40ef290]
#> 
#> [[28]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/Rtmp31Sth2/file3d5d3bfe99e9"),  <...>
#> env:  [frame 0x40df0f0]
#> 
#> [[29]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/Rtmp31Sth2/file3d5d3bfe99e9"),  <...>
#> env:  [frame 0x40dfe80]
#> 
#> [[30]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x40dac58]
#> 
#> [[31]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x40dbe10]
#> 
#> [[32]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x40dc158]
#> 
#> [[33]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x40d86e0]
#> 
#> [[34]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x40d8c90]
#> 
#> [[35]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x40d8fd8]
#> 
#> [[36]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x40d9320]
#> 
#> [[37]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x40d9a90]
#> 
#> [[38]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g <- function(trim) identity(identity(fn(trim)))
g(2)

#> [[1]]
#> <frame 38> (37)
#> expr: g(2)
#> env:  [local 0x8c383b8]
#> 
#> [[2]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x8ad6148]
#> 
#> [[3]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x8c38578]
#> 
#> [[4]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x8c387a8]
#> 
#> [[5]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x8c38d58]
#> 
#> [[6]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x8c392d0]
#> 
#> [[7]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x9646008]
#> 
#> [[8]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x9646350]
#> 
#> [[9]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x9646a18]
#> 
#> [[10]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x9646c48]
#> 
#> [[11]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x9646dd0]
#> 
#> [[12]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x9646eb0]
#> 
#> [[13]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fba288]
#> 
#> [[14]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x2dce3a8]
#> 
#> [[15]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x2ddd2b8]
#> 
#> [[16]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x2de6e18]
#> 
#> [[17]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x2ded128]
#> 
#> [[18]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x2dfd2d0]
#> 
#> [[19]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x3cd38f8]
#> 
#> [[20]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x3cd6b20]
#> 
#> [[21]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x6642be8]
#> 
#> [[22]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x67c68c0]
#> 
#> [[23]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x78af898]
#> 
#> [[24]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x5676658]
#> 
#> [[25]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x30f7608]
#> 
#> [[26]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x3101378]
#> 
#> [[27]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x40ee2d0]
#> 
#> [[28]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x40ef290]
#> 
#> [[29]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/Rtmp31Sth2/file3d5d3bfe99e9"),  <...>
#> env:  [frame 0x40df0f0]
#> 
#> [[30]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/Rtmp31Sth2/file3d5d3bfe99e9"),  <...>
#> env:  [frame 0x40dfe80]
#> 
#> [[31]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x40dac58]
#> 
#> [[32]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x40dbe10]
#> 
#> [[33]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x40dc158]
#> 
#> [[34]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x40d86e0]
#> 
#> [[35]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x40d8c90]
#> 
#> [[36]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x40d8fd8]
#> 
#> [[37]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x40d9320]
#> 
#> [[38]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x40d9a90]
#> 
#> [[39]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g(3)

#> [[1]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x8ad6148]
#> 
#> [[2]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x81fcf30]
#> 
#> [[3]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x81fcd00]
#> 
#> [[4]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x81fc750]
#> 
#> [[5]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x9b5c078]
#> 
#> [[6]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x9b5bcf8]
#> 
#> [[7]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x9b5b9b0]
#> 
#> [[8]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x9b5b278]
#> 
#> [[9]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x9b5b010]
#> 
#> [[10]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x9b5ae88]
#> 
#> [[11]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x9b5ada8]
#> 
#> [[12]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f5a080]
#> 
#> [[13]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x2dce3a8]
#> 
#> [[14]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x2ddd2b8]
#> 
#> [[15]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x2de6e18]
#> 
#> [[16]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x2ded128]
#> 
#> [[17]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x2dfd2d0]
#> 
#> [[18]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x3cd38f8]
#> 
#> [[19]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x3cd6b20]
#> 
#> [[20]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x6642be8]
#> 
#> [[21]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x67c68c0]
#> 
#> [[22]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x78af898]
#> 
#> [[23]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x5676658]
#> 
#> [[24]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x30f7608]
#> 
#> [[25]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x3101378]
#> 
#> [[26]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x40ee2d0]
#> 
#> [[27]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x40ef290]
#> 
#> [[28]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/Rtmp31Sth2/file3d5d3bfe99e9"),  <...>
#> env:  [frame 0x40df0f0]
#> 
#> [[29]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/Rtmp31Sth2/file3d5d3bfe99e9"),  <...>
#> env:  [frame 0x40dfe80]
#> 
#> [[30]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x40dac58]
#> 
#> [[31]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x40dbe10]
#> 
#> [[32]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x40dc158]
#> 
#> [[33]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x40d86e0]
#> 
#> [[34]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x40d8c90]
#> 
#> [[35]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x40d8fd8]
#> 
#> [[36]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x40d9320]
#> 
#> [[37]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x40d9a90]
#> 
#> [[38]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"