Call stack information

The eval_ and call_ families of functions provide a replacement for the base R functions prefixed with sys. (which are all about the context stack), as well as for parent.frame() (which is the only base R function for querying the call stack). The context stack includes all R-level evaluation contexts. It is linear in terms of execution history but due to lazy evaluation it is potentially nonlinear in terms of call history. The call stack history, on the other hand, is homogenous.

global_frame()

current_frame()

ctxt_frame(n = 1)

call_frame(n = 1, clean = TRUE)

ctxt_depth()

call_depth()

ctxt_stack(n = NULL, trim = 0)

call_stack(n = NULL, clean = TRUE)

Arguments

n
The number of frames to go back in the stack.

clean
Whether to post-process the call stack to clean non-standard frames. If TRUE, suboptimal call-stack entries by base::eval() will be cleaned up: the duplicate frame created by eval() is eliminated.

trim
The number of layers of intervening frames to trim off the stack. See stack_trim() and examples.

Details

ctxt_frame() and call_frame() return a frame object containing the following fields: expr and env (call expression and evaluation environment), pos and caller_pos (position of current frame in the context stack and position of the caller), and fun (function of the current frame). ctxt_stack() and call_stack() return a list of all context or call frames on the stack. Finally, ctxt_depth() and call_depth() report the current context position or the number of calling frames on the stack.

The base R functions take two sorts of arguments to indicate which frame to query: which and n. The n argument is straightforward: it's the number of frames to go down the stack, with n = 1 referring to the current context. The which argument is more complicated and changes meaning for values lower than 1. For the sake of consistency, the rlang functions all take the same kind of argument n. This argument has a single meaning (the number of frames to go down the stack) and cannot be lower than 1.

Note finally that parent.frame(1) corresponds to call_frame(2)$env, as n = 1 always refers to the current frame. This makes the _frame() and _stack() functions consistent: ctxt_frame(2) is the same as ctxt_stack()[[2]]. Also, ctxt_depth() returns one more frame than base::sys.nframe() because it counts the global frame. That is consistent with the _stack() functions which return the global frame as well. This way, call_stack(call_depth()) is the same as global_frame().

Life cycle

These functions are soft-deprecated and replaced by trace_back().

Examples

# Expressions within arguments count as contexts
identity(identity(ctxt_depth())) # returns 2

#> Warning: `ctxt_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 26

# But they are not part of the call stack because arguments are
# evaluated within the calling function (or the global environment
# if called at top level)
identity(identity(call_depth())) # returns 0

#> Warning: `call_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 13

# The context stacks includes all intervening execution frames. The
# call stack doesn't:
f <- function(x) identity(x)
f(f(ctxt_stack()))

#> [[1]]
#> <frame 27> (26)
#> expr: identity(x)
#> env:  [local 0x7fdcf6112268]
#> 
#> [[2]]
#> <frame 26> (23)
#> expr: f(ctxt_stack())
#> env:  [local 0x7fdcf6112150]
#> 
#> [[3]]
#> <frame 25> (24)
#> expr: identity(x)
#> env:  [local 0x7fdcf6112000]
#> 
#> [[4]]
#> <frame 24> (23)
#> expr: f(f(ctxt_stack()))
#> env:  [local 0x7fdcf6111e78]
#> 
#> [[5]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdcf5a5bd98]
#> 
#> [[6]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdcf611f000]
#> 
#> [[7]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fdcf611eaf8]
#> 
#> [[8]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fdcf611e200]
#> 
#> [[9]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fdcf6121968]
#> 
#> [[10]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fdcf6121428]
#> 
#> [[11]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fdcf6120fc8]
#> 
#> [[12]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fdcf612e870]
#> 
#> [[13]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fdcf612e598]
#> 
#> [[14]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fdcf612e368]
#> 
#> [[15]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fdcf612e218]
#> 
#> [[16]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fdcf62bb060]
#> 
#> [[17]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fdcf56485d8]
#> 
#> [[18]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fdcf5655058]
#> 
#> [[19]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fdcf918cf40]
#> 
#> [[20]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fdcf93d9d78]
#> 
#> [[21]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fdcf93d9608]
#> 
#> [[22]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fdcf9407170]
#> 
#> [[23]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fdcfa0d3f58]
#> 
#> [[24]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdcf95fcb18]
#> 
#> [[25]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdcf8267e08]
#> 
#> [[26]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fdcf8262e80]
#> 
#> [[27]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fdcf80ed488]
#> 
#> [[28]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(f(call_stack()))

#> Warning: `call_stack()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [[1]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdcf5a5bd98]
#> 
#> [[2]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fdcfa6c52b0]
#> 
#> [[3]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fdcf56485d8]
#> 
#> [[4]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fdcf5655058]
#> 
#> [[5]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fdcf918cf40]
#> 
#> [[6]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fdcf93d9d78]
#> 
#> [[7]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fdcf9407170]
#> 
#> [[8]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fdcfa0d3f58]
#> 
#> [[9]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdcf95fcb18]
#> 
#> [[10]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdcf8267e08]
#> 
#> [[11]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fdcf8262e80]
#> 
#> [[12]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fdcf80ed488]
#> 
#> [[13]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

g <- function(cmd) cmd()
f(g(ctxt_stack))

#> [[1]]
#> <frame 26> (23)
#> expr: g(ctxt_stack)
#> env:  [local 0x7fdcfb0b9898]
#> 
#> [[2]]
#> <frame 25> (24)
#> expr: identity(x)
#> env:  [local 0x7fdcfb0b99b0]
#> 
#> [[3]]
#> <frame 24> (23)
#> expr: f(g(ctxt_stack))
#> env:  [local 0x7fdcfb0b9ac8]
#> 
#> [[4]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdcf5a5bd98]
#> 
#> [[5]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdcfb0b9c88]
#> 
#> [[6]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fdcfb0b9eb8]
#> 
#> [[7]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fdcfb0b6558]
#> 
#> [[8]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fdcfb0b6ad0]
#> 
#> [[9]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fdcfb0b6e18]
#> 
#> [[10]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fdcfb0b7160]
#> 
#> [[11]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fdcfb0b7828]
#> 
#> [[12]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fdcfb0b7a58]
#> 
#> [[13]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fdcfb0b7be0]
#> 
#> [[14]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fdcfb0b7cc0]
#> 
#> [[15]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fdcfb5dead8]
#> 
#> [[16]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fdcf56485d8]
#> 
#> [[17]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fdcf5655058]
#> 
#> [[18]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fdcf918cf40]
#> 
#> [[19]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fdcf93d9d78]
#> 
#> [[20]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fdcf93d9608]
#> 
#> [[21]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fdcf9407170]
#> 
#> [[22]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fdcfa0d3f58]
#> 
#> [[23]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdcf95fcb18]
#> 
#> [[24]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdcf8267e08]
#> 
#> [[25]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fdcf8262e80]
#> 
#> [[26]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fdcf80ed488]
#> 
#> [[27]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(g(call_stack))

#> [[1]]
#> <frame 26> (22)
#> expr: g(call_stack)
#> env:  [local 0x7fdcfa474f00]
#> 
#> [[2]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdcf5a5bd98]
#> 
#> [[3]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fdcfa45fad0]
#> 
#> [[4]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fdcf56485d8]
#> 
#> [[5]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fdcf5655058]
#> 
#> [[6]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fdcf918cf40]
#> 
#> [[7]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fdcf93d9d78]
#> 
#> [[8]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fdcf9407170]
#> 
#> [[9]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fdcfa0d3f58]
#> 
#> [[10]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdcf95fcb18]
#> 
#> [[11]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdcf8267e08]
#> 
#> [[12]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fdcf8262e80]
#> 
#> [[13]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fdcf80ed488]
#> 
#> [[14]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

# The rlang _stack() functions return a list of frame
# objects. Use purrr::transpose() or index a field with
# purrr::map()'s to extract a particular field from a stack:

# stack <- f(f(call_stack()))
# purrr::map(stack, "env")
# purrr::transpose(stack)$expr

# current_frame() is an alias for ctxt_frame(1)
fn <- function() list(current = current_frame(), first = ctxt_frame(1))
fn()

#> Warning: `current_frame()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> $current
#> <frame 24> (23)
#> expr: fn()
#> env:  [local 0x7fdcfce86b88]
#> 
#> $first
#> <frame 24> (23)
#> expr: fn()
#> env:  [local 0x7fdcfce86b88]
#> 

# While current_frame() is the top of the stack, global_frame() is
# the bottom:
fn <- function() {
  n <- ctxt_depth()
  ctxt_frame(n)
}
identical(fn(), global_frame())

#> [1] TRUE

# ctxt_stack() returns a stack with all intervening frames. You can
# trim layers of intervening frames with the trim argument:
identity(identity(ctxt_stack()))

#> [[1]]
#> <frame 25> (23)
#> expr: identity(ctxt_stack())
#> env:  [local 0x7fdcfc77f9d0]
#> 
#> [[2]]
#> <frame 24> (23)
#> expr: identity(identity(ctxt_stack()))
#> env:  [local 0x7fdcfc77f8b8]
#> 
#> [[3]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdcf5a5bd98]
#> 
#> [[4]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdcfc77f6f8]
#> 
#> [[5]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fdcfc77f490]
#> 
#> [[6]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fdcfc77eee0]
#> 
#> [[7]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fdcfc786878]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fdcfc786530]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fdcfc7861e8]
#> 
#> [[10]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fdcfc785b20]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fdcfc7858f0]
#> 
#> [[12]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fdcfc785730]
#> 
#> [[13]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fdcfc785650]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fdcfc79e990]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fdcf56485d8]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fdcf5655058]
#> 
#> [[17]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fdcf918cf40]
#> 
#> [[18]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fdcf93d9d78]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fdcf93d9608]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fdcf9407170]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fdcfa0d3f58]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdcf95fcb18]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdcf8267e08]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fdcf8262e80]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fdcf80ed488]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(ctxt_stack(trim = 1)))

#> [[1]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdcf5a5bd98]
#> 
#> [[2]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdcfc35db98]
#> 
#> [[3]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fdcfc35d968]
#> 
#> [[4]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fdcfc35d3b8]
#> 
#> [[5]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fdcfc35ce40]
#> 
#> [[6]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fdcfc35cac0]
#> 
#> [[7]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fdcfc35c548]
#> 
#> [[8]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fdcfcbb8990]
#> 
#> [[9]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fdcfcbb8760]
#> 
#> [[10]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fdcfcbb85d8]
#> 
#> [[11]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fdcfcbb84f8]
#> 
#> [[12]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fdcfb19fe48]
#> 
#> [[13]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fdcf56485d8]
#> 
#> [[14]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fdcf5655058]
#> 
#> [[15]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fdcf918cf40]
#> 
#> [[16]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fdcf93d9d78]
#> 
#> [[17]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fdcf93d9608]
#> 
#> [[18]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fdcf9407170]
#> 
#> [[19]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fdcfa0d3f58]
#> 
#> [[20]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdcf95fcb18]
#> 
#> [[21]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdcf8267e08]
#> 
#> [[22]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fdcf8262e80]
#> 
#> [[23]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fdcf80ed488]
#> 
#> [[24]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# ctxt_stack() is called within fn() with intervening frames:
fn <- function(trim) identity(identity(ctxt_stack(trim = trim)))
fn(0)

#> [[1]]
#> <frame 26> (24)
#> expr: identity(ctxt_stack(trim = trim))
#> env:  [local 0x7fdcf5e1ce80]
#> 
#> [[2]]
#> <frame 25> (24)
#> expr: identity(identity(ctxt_stack(trim = trim)))
#> env:  [local 0x7fdcf5e1cd68]
#> 
#> [[3]]
#> <frame 24> (23)
#> expr: fn(0)
#> env:  [local 0x7fdcf5e1cc50]
#> 
#> [[4]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdcf5a5bd98]
#> 
#> [[5]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdcf5e1ca90]
#> 
#> [[6]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fdcf5e1c7f0]
#> 
#> [[7]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fdcf5e1c240]
#> 
#> [[8]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fdcf5e1fba0]
#> 
#> [[9]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fdcf5e1f858]
#> 
#> [[10]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fdcf5e1f510]
#> 
#> [[11]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fdcf5e1edd8]
#> 
#> [[12]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fdcf5e1eba8]
#> 
#> [[13]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fdcf5e1ea20]
#> 
#> [[14]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fdcf5e1e940]
#> 
#> [[15]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fdcf5e30b70]
#> 
#> [[16]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fdcf56485d8]
#> 
#> [[17]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fdcf5655058]
#> 
#> [[18]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fdcf918cf40]
#> 
#> [[19]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fdcf93d9d78]
#> 
#> [[20]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fdcf93d9608]
#> 
#> [[21]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fdcf9407170]
#> 
#> [[22]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fdcfa0d3f58]
#> 
#> [[23]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdcf95fcb18]
#> 
#> [[24]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdcf8267e08]
#> 
#> [[25]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fdcf8262e80]
#> 
#> [[26]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fdcf80ed488]
#> 
#> [[27]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# We can trim the first layer of those:
fn(1)

#> [[1]]
#> <frame 24> (23)
#> expr: fn(1)
#> env:  [local 0x7fdcf976b608]
#> 
#> [[2]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdcf5a5bd98]
#> 
#> [[3]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdcf976b3d8]
#> 
#> [[4]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fdcf976b1a8]
#> 
#> [[5]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fdcf976e3d0]
#> 
#> [[6]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fdcf976de58]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fdcf976d790]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fdcf976d410]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fdcf976fec8]
#> 
#> [[10]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fdcf9770bf8]
#> 
#> [[11]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fdcf9770a00]
#> 
#> [[12]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fdcf9770920]
#> 
#> [[13]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fdcf97a5de8]
#> 
#> [[14]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fdcf56485d8]
#> 
#> [[15]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fdcf5655058]
#> 
#> [[16]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fdcf918cf40]
#> 
#> [[17]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fdcf93d9d78]
#> 
#> [[18]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fdcf93d9608]
#> 
#> [[19]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fdcf9407170]
#> 
#> [[20]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fdcfa0d3f58]
#> 
#> [[21]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdcf95fcb18]
#> 
#> [[22]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdcf8267e08]
#> 
#> [[23]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fdcf8262e80]
#> 
#> [[24]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fdcf80ed488]
#> 
#> [[25]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# The outside intervening frames (at the fn() call site) are still
# returned, but can be trimmed as well:
identity(identity(fn(1)))

#> [[1]]
#> <frame 26> (23)
#> expr: fn(1)
#> env:  [local 0x7fdcf963c600]
#> 
#> [[2]]
#> <frame 25> (23)
#> expr: identity(fn(1))
#> env:  [local 0x7fdcf963c4e8]
#> 
#> [[3]]
#> <frame 24> (23)
#> expr: identity(identity(fn(1)))
#> env:  [local 0x7fdcf963c328]
#> 
#> [[4]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdcf5a5bd98]
#> 
#> [[5]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdcf963c168]
#> 
#> [[6]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fdcf963bf38]
#> 
#> [[7]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fdcf963b988]
#> 
#> [[8]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fdcf963b3d8]
#> 
#> [[9]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fdcf963ae28]
#> 
#> [[10]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fdcf963aaa8]
#> 
#> [[11]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fdcf963e2f0]
#> 
#> [[12]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fdcf963e0c0]
#> 
#> [[13]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fdcf963df38]
#> 
#> [[14]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fdcf963de58]
#> 
#> [[15]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fdcf9660fe8]
#> 
#> [[16]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fdcf56485d8]
#> 
#> [[17]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fdcf5655058]
#> 
#> [[18]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fdcf918cf40]
#> 
#> [[19]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fdcf93d9d78]
#> 
#> [[20]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fdcf93d9608]
#> 
#> [[21]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fdcf9407170]
#> 
#> [[22]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fdcfa0d3f58]
#> 
#> [[23]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdcf95fcb18]
#> 
#> [[24]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdcf8267e08]
#> 
#> [[25]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fdcf8262e80]
#> 
#> [[26]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fdcf80ed488]
#> 
#> [[27]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(fn(2)))

#> [[1]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdcf5a5bd98]
#> 
#> [[2]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdcf92ef170]
#> 
#> [[3]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fdcf92eef08]
#> 
#> [[4]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fdcf92ee958]
#> 
#> [[5]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fdcf92f22f0]
#> 
#> [[6]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fdcf92f1fa8]
#> 
#> [[7]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fdcf92f1c60]
#> 
#> [[8]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fdcf92f1560]
#> 
#> [[9]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fdcf92f1330]
#> 
#> [[10]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fdcf92f1170]
#> 
#> [[11]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fdcf92f1090]
#> 
#> [[12]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fdcf9305138]
#> 
#> [[13]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fdcf56485d8]
#> 
#> [[14]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fdcf5655058]
#> 
#> [[15]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fdcf918cf40]
#> 
#> [[16]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fdcf93d9d78]
#> 
#> [[17]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fdcf93d9608]
#> 
#> [[18]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fdcf9407170]
#> 
#> [[19]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fdcfa0d3f58]
#> 
#> [[20]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdcf95fcb18]
#> 
#> [[21]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdcf8267e08]
#> 
#> [[22]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fdcf8262e80]
#> 
#> [[23]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fdcf80ed488]
#> 
#> [[24]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g <- function(trim) identity(identity(fn(trim)))
g(2)

#> [[1]]
#> <frame 24> (23)
#> expr: g(2)
#> env:  [local 0x7fdcf9d05d80]
#> 
#> [[2]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdcf5a5bd98]
#> 
#> [[3]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdcf9d05b50]
#> 
#> [[4]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fdcf9d05920]
#> 
#> [[5]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fdcf9d05338]
#> 
#> [[6]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fdcf9d04d88]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fdcf9d04a08]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fdcf9d085d0]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fdcf9d07f08]
#> 
#> [[10]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fdcf9d07ca0]
#> 
#> [[11]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fdcf9d07b18]
#> 
#> [[12]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fdcf9d07a38]
#> 
#> [[13]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fdcf9d19418]
#> 
#> [[14]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fdcf56485d8]
#> 
#> [[15]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fdcf5655058]
#> 
#> [[16]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fdcf918cf40]
#> 
#> [[17]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fdcf93d9d78]
#> 
#> [[18]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fdcf93d9608]
#> 
#> [[19]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fdcf9407170]
#> 
#> [[20]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fdcfa0d3f58]
#> 
#> [[21]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdcf95fcb18]
#> 
#> [[22]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdcf8267e08]
#> 
#> [[23]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fdcf8262e80]
#> 
#> [[24]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fdcf80ed488]
#> 
#> [[25]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g(3)

#> [[1]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdcf5a5bd98]
#> 
#> [[2]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdcf993e270]
#> 
#> [[3]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fdcf993e040]
#> 
#> [[4]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fdcf993d358]
#> 
#> [[5]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fdcf993c8a0]
#> 
#> [[6]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fdcf9951358]
#> 
#> [[7]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fdcf9951010]
#> 
#> [[8]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fdcf9950788]
#> 
#> [[9]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fdcf99533c8]
#> 
#> [[10]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fdcf9953208]
#> 
#> [[11]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fdcf9953128]
#> 
#> [[12]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fdcf99753c8]
#> 
#> [[13]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fdcf56485d8]
#> 
#> [[14]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fdcf5655058]
#> 
#> [[15]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fdcf918cf40]
#> 
#> [[16]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fdcf93d9d78]
#> 
#> [[17]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fdcf93d9608]
#> 
#> [[18]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fdcf9407170]
#> 
#> [[19]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fdcfa0d3f58]
#> 
#> [[20]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdcf95fcb18]
#> 
#> [[21]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdcf8267e08]
#> 
#> [[22]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fdcf8262e80]
#> 
#> [[23]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fdcf80ed488]
#> 
#> [[24]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"