Call stack information

The eval_ and call_ families of functions provide a replacement for the base R functions prefixed with sys. (which are all about the context stack), as well as for parent.frame() (which is the only base R function for querying the call stack). The context stack includes all R-level evaluation contexts. It is linear in terms of execution history but due to lazy evaluation it is potentially nonlinear in terms of call history. The call stack history, on the other hand, is homogenous.

global_frame()

current_frame()

ctxt_frame(n = 1)

call_frame(n = 1, clean = TRUE)

ctxt_depth()

call_depth()

ctxt_stack(n = NULL, trim = 0)

call_stack(n = NULL, clean = TRUE)

Arguments

n
The number of frames to go back in the stack.

clean
Whether to post-process the call stack to clean non-standard frames. If TRUE, suboptimal call-stack entries by base::eval() will be cleaned up: the duplicate frame created by eval() is eliminated.

trim
The number of layers of intervening frames to trim off the stack. See stack_trim() and examples.

Details

ctxt_frame() and call_frame() return a frame object containing the following fields: expr and env (call expression and evaluation environment), pos and caller_pos (position of current frame in the context stack and position of the caller), and fun (function of the current frame). ctxt_stack() and call_stack() return a list of all context or call frames on the stack. Finally, ctxt_depth() and call_depth() report the current context position or the number of calling frames on the stack.

The base R functions take two sorts of arguments to indicate which frame to query: which and n. The n argument is straightforward: it's the number of frames to go down the stack, with n = 1 referring to the current context. The which argument is more complicated and changes meaning for values lower than 1. For the sake of consistency, the rlang functions all take the same kind of argument n. This argument has a single meaning (the number of frames to go down the stack) and cannot be lower than 1.

Note finally that parent.frame(1) corresponds to call_frame(2)$env, as n = 1 always refers to the current frame. This makes the _frame() and _stack() functions consistent: ctxt_frame(2) is the same as ctxt_stack()[[2]]. Also, ctxt_depth() returns one more frame than base::sys.nframe() because it counts the global frame. That is consistent with the _stack() functions which return the global frame as well. This way, call_stack(call_depth()) is the same as global_frame().

Life cycle

These functions are soft-deprecated and replaced by trace_back().

Examples

# Expressions within arguments count as contexts
identity(identity(ctxt_depth())) # returns 2

#> Warning: `ctxt_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 27

# But they are not part of the call stack because arguments are
# evaluated within the calling function (or the global environment
# if called at top level)
identity(identity(call_depth())) # returns 0

#> Warning: `call_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 14

# The context stacks includes all intervening execution frames. The
# call stack doesn't:
f <- function(x) identity(x)
f(f(ctxt_stack()))

#> [[1]]
#> <frame 28> (27)
#> expr: identity(x)
#> env:  [local 0x7fbc0edebf00]
#> 
#> [[2]]
#> <frame 27> (24)
#> expr: f(ctxt_stack())
#> env:  [local 0x7fbc0edec018]
#> 
#> [[3]]
#> <frame 26> (25)
#> expr: identity(x)
#> env:  [local 0x7fbc0edec130]
#> 
#> [[4]]
#> <frame 25> (24)
#> expr: f(f(ctxt_stack()))
#> env:  [local 0x7fbc0edec248]
#> 
#> [[5]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fbc0ec25400]
#> 
#> [[6]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fbc0edec408]
#> 
#> [[7]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fbc0edec638]
#> 
#> [[8]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fbc0edecbe8]
#> 
#> [[9]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fbc0eded198]
#> 
#> [[10]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fbc0eded4e0]
#> 
#> [[11]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fbc0ede9918]
#> 
#> [[12]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fbc0ede9fe0]
#> 
#> [[13]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fbc0edea210]
#> 
#> [[14]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fbc0edea398]
#> 
#> [[15]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fbc0edea478]
#> 
#> [[16]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fbc0edd7fa8]
#> 
#> [[17]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fbc0dd3ca60]
#> 
#> [[18]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7fbc0dd3b400]
#> 
#> [[19]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fbc0dd2f0f0]
#> 
#> [[20]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fbc0dcd92b0]
#> 
#> [[21]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fbc05f4bd80]
#> 
#> [[22]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fbc05f50040]
#> 
#> [[23]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fbc0602f6e0]
#> 
#> [[24]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fbc08e1b9b8]
#> 
#> [[25]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fbc0a1908d0]
#> 
#> [[26]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fbc079cc8f0]
#> 
#> [[27]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fbc079d32a0]
#> 
#> [[28]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fbc08137f08]
#> 
#> [[29]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(f(call_stack()))

#> Warning: `call_stack()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [[1]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fbc0ec25400]
#> 
#> [[2]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fbc0af2e710]
#> 
#> [[3]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fbc0dd3ca60]
#> 
#> [[4]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7fbc0dd3b400]
#> 
#> [[5]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fbc0dd2f0f0]
#> 
#> [[6]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fbc0dcd92b0]
#> 
#> [[7]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fbc05f4bd80]
#> 
#> [[8]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fbc0602f6e0]
#> 
#> [[9]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fbc08e1b9b8]
#> 
#> [[10]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fbc0a1908d0]
#> 
#> [[11]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fbc079cc8f0]
#> 
#> [[12]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fbc079d32a0]
#> 
#> [[13]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fbc08137f08]
#> 
#> [[14]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

g <- function(cmd) cmd()
f(g(ctxt_stack))

#> [[1]]
#> <frame 27> (24)
#> expr: g(ctxt_stack)
#> env:  [local 0x7fbc0adec4e0]
#> 
#> [[2]]
#> <frame 26> (25)
#> expr: identity(x)
#> env:  [local 0x7fbc0add4ef0]
#> 
#> [[3]]
#> <frame 25> (24)
#> expr: f(g(ctxt_stack))
#> env:  [local 0x7fbc0edd34e0]
#> 
#> [[4]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fbc0ec25400]
#> 
#> [[5]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fbc04c4c520]
#> 
#> [[6]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fbc09a8cea8]
#> 
#> [[7]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fbc0dd06c90]
#> 
#> [[8]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fbc0ae1f0e8]
#> 
#> [[9]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fbc0ae3add8]
#> 
#> [[10]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fbc0ef11048]
#> 
#> [[11]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fbc05afece8]
#> 
#> [[12]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fbc0ae1f5b8]
#> 
#> [[13]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fbc0ae1f2e0]
#> 
#> [[14]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fbc0ae1f4d8]
#> 
#> [[15]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fbc0afd2a90]
#> 
#> [[16]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fbc0dd3ca60]
#> 
#> [[17]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7fbc0dd3b400]
#> 
#> [[18]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fbc0dd2f0f0]
#> 
#> [[19]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fbc0dcd92b0]
#> 
#> [[20]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fbc05f4bd80]
#> 
#> [[21]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fbc05f50040]
#> 
#> [[22]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fbc0602f6e0]
#> 
#> [[23]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fbc08e1b9b8]
#> 
#> [[24]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fbc0a1908d0]
#> 
#> [[25]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fbc079cc8f0]
#> 
#> [[26]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fbc079d32a0]
#> 
#> [[27]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fbc08137f08]
#> 
#> [[28]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(g(call_stack))

#> [[1]]
#> <frame 27> (23)
#> expr: g(call_stack)
#> env:  [local 0x7fbc0aa0c338]
#> 
#> [[2]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fbc0ec25400]
#> 
#> [[3]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fbc087774d8]
#> 
#> [[4]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fbc0dd3ca60]
#> 
#> [[5]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7fbc0dd3b400]
#> 
#> [[6]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fbc0dd2f0f0]
#> 
#> [[7]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fbc0dcd92b0]
#> 
#> [[8]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fbc05f4bd80]
#> 
#> [[9]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fbc0602f6e0]
#> 
#> [[10]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fbc08e1b9b8]
#> 
#> [[11]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fbc0a1908d0]
#> 
#> [[12]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fbc079cc8f0]
#> 
#> [[13]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fbc079d32a0]
#> 
#> [[14]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fbc08137f08]
#> 
#> [[15]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

# The rlang _stack() functions return a list of frame
# objects. Use purrr::transpose() or index a field with
# purrr::map()'s to extract a particular field from a stack:

# stack <- f(f(call_stack()))
# purrr::map(stack, "env")
# purrr::transpose(stack)$expr

# current_frame() is an alias for ctxt_frame(1)
fn <- function() list(current = current_frame(), first = ctxt_frame(1))
fn()

#> Warning: `current_frame()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> $current
#> <frame 25> (24)
#> expr: fn()
#> env:  [local 0x7fbc0a9638d0]
#> 
#> $first
#> <frame 25> (24)
#> expr: fn()
#> env:  [local 0x7fbc0a9638d0]
#> 

# While current_frame() is the top of the stack, global_frame() is
# the bottom:
fn <- function() {
  n <- ctxt_depth()
  ctxt_frame(n)
}
identical(fn(), global_frame())

#> [1] TRUE

# ctxt_stack() returns a stack with all intervening frames. You can
# trim layers of intervening frames with the trim argument:
identity(identity(ctxt_stack()))

#> [[1]]
#> <frame 26> (24)
#> expr: identity(ctxt_stack())
#> env:  [local 0x7fbc0b158c20]
#> 
#> [[2]]
#> <frame 25> (24)
#> expr: identity(identity(ctxt_stack()))
#> env:  [local 0x7fbc0b158ad0]
#> 
#> [[3]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fbc0ec25400]
#> 
#> [[4]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fbc0b158910]
#> 
#> [[5]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fbc0b1586e0]
#> 
#> [[6]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fbc0b15c040]
#> 
#> [[7]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fbc0b15ba90]
#> 
#> [[8]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fbc0b15b748]
#> 
#> [[9]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fbc0b15b3c8]
#> 
#> [[10]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fbc0b15acc8]
#> 
#> [[11]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fbc0b15aa98]
#> 
#> [[12]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fbc0b15a8d8]
#> 
#> [[13]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fbc0b15a7f8]
#> 
#> [[14]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fbc0b175fd0]
#> 
#> [[15]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fbc0dd3ca60]
#> 
#> [[16]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7fbc0dd3b400]
#> 
#> [[17]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fbc0dd2f0f0]
#> 
#> [[18]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fbc0dcd92b0]
#> 
#> [[19]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fbc05f4bd80]
#> 
#> [[20]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fbc05f50040]
#> 
#> [[21]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fbc0602f6e0]
#> 
#> [[22]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fbc08e1b9b8]
#> 
#> [[23]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fbc0a1908d0]
#> 
#> [[24]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fbc079cc8f0]
#> 
#> [[25]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fbc079d32a0]
#> 
#> [[26]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fbc08137f08]
#> 
#> [[27]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(ctxt_stack(trim = 1)))

#> [[1]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fbc0ec25400]
#> 
#> [[2]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fbc07e9ae70]
#> 
#> [[3]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fbc07e9ac40]
#> 
#> [[4]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fbc07e9a690]
#> 
#> [[5]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fbc07e9a0a8]
#> 
#> [[6]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fbc07e99d28]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fbc07e999a8]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fbc07e992e0]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fbc07e990b0]
#> 
#> [[10]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fbc07e9ce00]
#> 
#> [[11]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fbc07e9cd20]
#> 
#> [[12]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fbc07eaf318]
#> 
#> [[13]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fbc0dd3ca60]
#> 
#> [[14]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7fbc0dd3b400]
#> 
#> [[15]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fbc0dd2f0f0]
#> 
#> [[16]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fbc0dcd92b0]
#> 
#> [[17]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fbc05f4bd80]
#> 
#> [[18]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fbc05f50040]
#> 
#> [[19]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fbc0602f6e0]
#> 
#> [[20]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fbc08e1b9b8]
#> 
#> [[21]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fbc0a1908d0]
#> 
#> [[22]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fbc079cc8f0]
#> 
#> [[23]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fbc079d32a0]
#> 
#> [[24]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fbc08137f08]
#> 
#> [[25]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# ctxt_stack() is called within fn() with intervening frames:
fn <- function(trim) identity(identity(ctxt_stack(trim = trim)))
fn(0)

#> [[1]]
#> <frame 27> (25)
#> expr: identity(ctxt_stack(trim = trim))
#> env:  [local 0x7fbc0a875bd8]
#> 
#> [[2]]
#> <frame 26> (25)
#> expr: identity(identity(ctxt_stack(trim = trim)))
#> env:  [local 0x7fbc0a875ac0]
#> 
#> [[3]]
#> <frame 25> (24)
#> expr: fn(0)
#> env:  [local 0x7fbc0a8759a8]
#> 
#> [[4]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fbc0ec25400]
#> 
#> [[5]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fbc0a874fd0]
#> 
#> [[6]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fbc0a874d68]
#> 
#> [[7]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fbc0a88deb0]
#> 
#> [[8]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fbc0a88d938]
#> 
#> [[9]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fbc0a88d5f0]
#> 
#> [[10]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fbc0a88d2a8]
#> 
#> [[11]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fbc0a88cb70]
#> 
#> [[12]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fbc0a88c940]
#> 
#> [[13]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fbc0a88c7b8]
#> 
#> [[14]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fbc0a88c6a0]
#> 
#> [[15]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fbc0a8a4eb8]
#> 
#> [[16]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fbc0dd3ca60]
#> 
#> [[17]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7fbc0dd3b400]
#> 
#> [[18]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fbc0dd2f0f0]
#> 
#> [[19]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fbc0dcd92b0]
#> 
#> [[20]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fbc05f4bd80]
#> 
#> [[21]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fbc05f50040]
#> 
#> [[22]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fbc0602f6e0]
#> 
#> [[23]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fbc08e1b9b8]
#> 
#> [[24]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fbc0a1908d0]
#> 
#> [[25]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fbc079cc8f0]
#> 
#> [[26]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fbc079d32a0]
#> 
#> [[27]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fbc08137f08]
#> 
#> [[28]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# We can trim the first layer of those:
fn(1)

#> [[1]]
#> <frame 25> (24)
#> expr: fn(1)
#> env:  [local 0x7fbc09abcb60]
#> 
#> [[2]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fbc0ec25400]
#> 
#> [[3]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fbc09abc968]
#> 
#> [[4]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fbc09abc738]
#> 
#> [[5]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fbc09abc188]
#> 
#> [[6]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fbc09abbbd8]
#> 
#> [[7]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fbc09abb890]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fbc09abb548]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fbc09abed58]
#> 
#> [[10]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fbc09abeb28]
#> 
#> [[11]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fbc09abe9a0]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fbc09abe8c0]
#> 
#> [[13]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fbc09ad3fc8]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fbc0dd3ca60]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7fbc0dd3b400]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fbc0dd2f0f0]
#> 
#> [[17]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fbc0dcd92b0]
#> 
#> [[18]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fbc05f4bd80]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fbc05f50040]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fbc0602f6e0]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fbc08e1b9b8]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fbc0a1908d0]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fbc079cc8f0]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fbc079d32a0]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fbc08137f08]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# The outside intervening frames (at the fn() call site) are still
# returned, but can be trimmed as well:
identity(identity(fn(1)))

#> [[1]]
#> <frame 27> (24)
#> expr: fn(1)
#> env:  [local 0x7fbc0a606710]
#> 
#> [[2]]
#> <frame 26> (24)
#> expr: identity(fn(1))
#> env:  [local 0x7fbc0a6065f8]
#> 
#> [[3]]
#> <frame 25> (24)
#> expr: identity(identity(fn(1)))
#> env:  [local 0x7fbc0a6064e0]
#> 
#> [[4]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fbc0ec25400]
#> 
#> [[5]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fbc0a606320]
#> 
#> [[6]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fbc0a60c000]
#> 
#> [[7]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fbc0a60b858]
#> 
#> [[8]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fbc0a60b1c8]
#> 
#> [[9]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fbc0a60ae80]
#> 
#> [[10]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fbc0a60ab38]
#> 
#> [[11]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fbc0a60a3c8]
#> 
#> [[12]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fbc0a60e038]
#> 
#> [[13]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fbc0a60deb0]
#> 
#> [[14]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fbc0a60ddd0]
#> 
#> [[15]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fbc0a6237b0]
#> 
#> [[16]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fbc0dd3ca60]
#> 
#> [[17]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7fbc0dd3b400]
#> 
#> [[18]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fbc0dd2f0f0]
#> 
#> [[19]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fbc0dcd92b0]
#> 
#> [[20]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fbc05f4bd80]
#> 
#> [[21]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fbc05f50040]
#> 
#> [[22]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fbc0602f6e0]
#> 
#> [[23]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fbc08e1b9b8]
#> 
#> [[24]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fbc0a1908d0]
#> 
#> [[25]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fbc079cc8f0]
#> 
#> [[26]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fbc079d32a0]
#> 
#> [[27]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fbc08137f08]
#> 
#> [[28]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(fn(2)))

#> [[1]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fbc0ec25400]
#> 
#> [[2]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fbc0a26e070]
#> 
#> [[3]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fbc0a26de08]
#> 
#> [[4]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fbc0a26d820]
#> 
#> [[5]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fbc0a26c588]
#> 
#> [[6]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fbc0a26c240]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fbc0a26fe08]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fbc0a26f200]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fbc0a26ef60]
#> 
#> [[10]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fbc0a26ed68]
#> 
#> [[11]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fbc0a26ec88]
#> 
#> [[12]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fbc0a2888d0]
#> 
#> [[13]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fbc0dd3ca60]
#> 
#> [[14]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7fbc0dd3b400]
#> 
#> [[15]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fbc0dd2f0f0]
#> 
#> [[16]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fbc0dcd92b0]
#> 
#> [[17]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fbc05f4bd80]
#> 
#> [[18]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fbc05f50040]
#> 
#> [[19]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fbc0602f6e0]
#> 
#> [[20]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fbc08e1b9b8]
#> 
#> [[21]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fbc0a1908d0]
#> 
#> [[22]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fbc079cc8f0]
#> 
#> [[23]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fbc079d32a0]
#> 
#> [[24]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fbc08137f08]
#> 
#> [[25]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g <- function(trim) identity(identity(fn(trim)))
g(2)

#> [[1]]
#> <frame 25> (24)
#> expr: g(2)
#> env:  [local 0x7fbc089f78a8]
#> 
#> [[2]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fbc0ec25400]
#> 
#> [[3]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fbc089f76e8]
#> 
#> [[4]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fbc089fb3c8]
#> 
#> [[5]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fbc089fad38]
#> 
#> [[6]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fbc089fa750]
#> 
#> [[7]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fbc089fa408]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fbc089fa088]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fbc089f98e0]
#> 
#> [[10]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fbc089f9678]
#> 
#> [[11]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fbc089fd400]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fbc089fd2e8]
#> 
#> [[13]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fbc08a163d0]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fbc0dd3ca60]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7fbc0dd3b400]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fbc0dd2f0f0]
#> 
#> [[17]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fbc0dcd92b0]
#> 
#> [[18]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fbc05f4bd80]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fbc05f50040]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fbc0602f6e0]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fbc08e1b9b8]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fbc0a1908d0]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fbc079cc8f0]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fbc079d32a0]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fbc08137f08]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g(3)

#> [[1]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fbc0ec25400]
#> 
#> [[2]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fbc076b7f50]
#> 
#> [[3]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fbc076b7c08]
#> 
#> [[4]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fbc076c1448]
#> 
#> [[5]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fbc076c0e60]
#> 
#> [[6]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fbc076c0a70]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fbc076c65c8]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fbc076c5e20]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fbc076c5b80]
#> 
#> [[10]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fbc076c58e0]
#> 
#> [[11]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fbc076c5790]
#> 
#> [[12]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fbc07729218]
#> 
#> [[13]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fbc0dd3ca60]
#> 
#> [[14]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7fbc0dd3b400]
#> 
#> [[15]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fbc0dd2f0f0]
#> 
#> [[16]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fbc0dcd92b0]
#> 
#> [[17]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fbc05f4bd80]
#> 
#> [[18]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fbc05f50040]
#> 
#> [[19]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fbc0602f6e0]
#> 
#> [[20]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fbc08e1b9b8]
#> 
#> [[21]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fbc0a1908d0]
#> 
#> [[22]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fbc079cc8f0]
#> 
#> [[23]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fbc079d32a0]
#> 
#> [[24]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fbc08137f08]
#> 
#> [[25]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"