Call stack information

The eval_ and call_ families of functions provide a replacement for the base R functions prefixed with sys. (which are all about the context stack), as well as for parent.frame() (which is the only base R function for querying the call stack). The context stack includes all R-level evaluation contexts. It is linear in terms of execution history but due to lazy evaluation it is potentially nonlinear in terms of call history. The call stack history, on the other hand, is homogenous.

global_frame()

current_frame()

ctxt_frame(n = 1)

call_frame(n = 1, clean = TRUE)

ctxt_depth()

call_depth()

ctxt_stack(n = NULL, trim = 0)

call_stack(n = NULL, clean = TRUE)

Arguments

n
The number of frames to go back in the stack.

clean
Whether to post-process the call stack to clean non-standard frames. If TRUE, suboptimal call-stack entries by base::eval() will be cleaned up: the duplicate frame created by eval() is eliminated.

trim
The number of layers of intervening frames to trim off the stack. See stack_trim() and examples.

Details

ctxt_frame() and call_frame() return a frame object containing the following fields: expr and env (call expression and evaluation environment), pos and caller_pos (position of current frame in the context stack and position of the caller), and fun (function of the current frame). ctxt_stack() and call_stack() return a list of all context or call frames on the stack. Finally, ctxt_depth() and call_depth() report the current context position or the number of calling frames on the stack.

The base R functions take two sorts of arguments to indicate which frame to query: which and n. The n argument is straightforward: it's the number of frames to go down the stack, with n = 1 referring to the current context. The which argument is more complicated and changes meaning for values lower than 1. For the sake of consistency, the rlang functions all take the same kind of argument n. This argument has a single meaning (the number of frames to go down the stack) and cannot be lower than 1.

Note finally that parent.frame(1) corresponds to call_frame(2)$env, as n = 1 always refers to the current frame. This makes the _frame() and _stack() functions consistent: ctxt_frame(2) is the same as ctxt_stack()[[2]]. Also, ctxt_depth() returns one more frame than base::sys.nframe() because it counts the global frame. That is consistent with the _stack() functions which return the global frame as well. This way, call_stack(call_depth()) is the same as global_frame().

Life cycle

These functions are soft-deprecated and replaced by trace_back().

Examples

# Expressions within arguments count as contexts
identity(identity(ctxt_depth())) # returns 2

#> Warning: `ctxt_depth()` is soft-deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 40

# But they are not part of the call stack because arguments are
# evaluated within the calling function (or the global environment
# if called at top level)
identity(identity(call_depth())) # returns 0

#> Warning: `call_depth()` is soft-deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 14

# The context stacks includes all intervening execution frames. The
# call stack doesn't:
f <- function(x) identity(x)
f(f(ctxt_stack()))

#> [[1]]
#> <frame 41> (40)
#> expr: identity(x)
#> env:  [local 0x7023638]
#> 
#> [[2]]
#> <frame 40> (37)
#> expr: f(ctxt_stack())
#> env:  [local 0x70234e8]
#> 
#> [[3]]
#> <frame 39> (38)
#> expr: identity(x)
#> env:  [local 0x70233d0]
#> 
#> [[4]]
#> <frame 38> (37)
#> expr: f(f(ctxt_stack()))
#> env:  [local 0x70232b8]
#> 
#> [[5]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7337078]
#> 
#> [[6]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x70230f8]
#> 
#> [[7]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7022e90]
#> 
#> [[8]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x70228a8]
#> 
#> [[9]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x70220c8]
#> 
#> [[10]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x71dfc68]
#> 
#> [[11]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x71df7d0]
#> 
#> [[12]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x71df0d0]
#> 
#> [[13]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x71deea0]
#> 
#> [[14]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x71ded18]
#> 
#> [[15]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x71dec00]
#> 
#> [[16]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x80b0890]
#> 
#> [[17]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x179bab8]
#> 
#> [[18]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x83c2d20]
#> 
#> [[19]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x4ea9300]
#> 
#> [[20]]
#> <frame 22> (21)
#> expr: .f(code = .l[[1L]][[1L]], run = .l[[2L]][[1L]], show = .l[[3L]][[1L]],  <...>
#> env:  [frame 0x8db9df8]
#> 
#> [[21]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x79cf5d8]
#> 
#> [[22]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x899ddf8]
#> 
#> [[23]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x899eb50]
#> 
#> [[24]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x2451a10]
#> 
#> [[25]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x21797f0]
#> 
#> [[26]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x4f052a0]
#> 
#> [[27]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x4983528]
#> 
#> [[28]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x16ff988]
#> 
#> [[29]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x16fee50]
#> 
#> [[30]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3009790]
#> 
#> [[31]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3006920]
#> 
#> [[32]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpcgPkcb/file3eb7c87bb2d"),  <...>
#> env:  [frame 0x3004760]
#> 
#> [[33]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpcgPkcb/file3eb7c87bb2d"),  <...>
#> env:  [frame 0x30054f0]
#> 
#> [[34]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x30040f8]
#> 
#> [[35]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3001480]
#> 
#> [[36]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x30017c8]
#> 
#> [[37]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3001b80]
#> 
#> [[38]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3002130]
#> 
#> [[39]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3002478]
#> 
#> [[40]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x2ffe990]
#> 
#> [[41]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x2fff100]
#> 
#> [[42]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(f(call_stack()))

#> Warning: `call_stack()` is soft-deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [[1]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7337078]
#> 
#> [[2]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x56033f0]
#> 
#> [[3]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x179bab8]
#> 
#> [[4]]
#> <frame 22> (21)
#> expr: .f(code = .l[[1L]][[1L]], run = .l[[2L]][[1L]], show = .l[[3L]][[1L]],  <...>
#> env:  [frame 0x8db9df8]
#> 
#> [[5]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x79cf5d8]
#> 
#> [[6]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x899ddf8]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x2451a10]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x21797f0]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x4f052a0]
#> 
#> [[10]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x4983528]
#> 
#> [[11]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x16ff988]
#> 
#> [[12]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x16fee50]
#> 
#> [[13]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3009790]
#> 
#> [[14]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

g <- function(cmd) cmd()
f(g(ctxt_stack))

#> [[1]]
#> <frame 40> (37)
#> expr: g(ctxt_stack)
#> env:  [local 0x4d403e0]
#> 
#> [[2]]
#> <frame 39> (38)
#> expr: identity(x)
#> env:  [local 0x4d40258]
#> 
#> [[3]]
#> <frame 38> (37)
#> expr: f(g(ctxt_stack))
#> env:  [local 0x4d40140]
#> 
#> [[4]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7337078]
#> 
#> [[5]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x4d3fe68]
#> 
#> [[6]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x4d3fc00]
#> 
#> [[7]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x4d43448]
#> 
#> [[8]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x4d42cd8]
#> 
#> [[9]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x4d42990]
#> 
#> [[10]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x4d42610]
#> 
#> [[11]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x4d41ed8]
#> 
#> [[12]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x4d41c70]
#> 
#> [[13]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x4d41a78]
#> 
#> [[14]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x4d457c8]
#> 
#> [[15]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x4d7d9d0]
#> 
#> [[16]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x179bab8]
#> 
#> [[17]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x83c2d20]
#> 
#> [[18]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x4ea9300]
#> 
#> [[19]]
#> <frame 22> (21)
#> expr: .f(code = .l[[1L]][[1L]], run = .l[[2L]][[1L]], show = .l[[3L]][[1L]],  <...>
#> env:  [frame 0x8db9df8]
#> 
#> [[20]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x79cf5d8]
#> 
#> [[21]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x899ddf8]
#> 
#> [[22]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x899eb50]
#> 
#> [[23]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x2451a10]
#> 
#> [[24]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x21797f0]
#> 
#> [[25]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x4f052a0]
#> 
#> [[26]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x4983528]
#> 
#> [[27]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x16ff988]
#> 
#> [[28]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x16fee50]
#> 
#> [[29]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3009790]
#> 
#> [[30]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3006920]
#> 
#> [[31]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpcgPkcb/file3eb7c87bb2d"),  <...>
#> env:  [frame 0x3004760]
#> 
#> [[32]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpcgPkcb/file3eb7c87bb2d"),  <...>
#> env:  [frame 0x30054f0]
#> 
#> [[33]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x30040f8]
#> 
#> [[34]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3001480]
#> 
#> [[35]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x30017c8]
#> 
#> [[36]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3001b80]
#> 
#> [[37]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3002130]
#> 
#> [[38]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3002478]
#> 
#> [[39]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x2ffe990]
#> 
#> [[40]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x2fff100]
#> 
#> [[41]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(g(call_stack))

#> [[1]]
#> <frame 40> (36)
#> expr: g(call_stack)
#> env:  [local 0x1b66450]
#> 
#> [[2]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7337078]
#> 
#> [[3]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x2248f30]
#> 
#> [[4]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x179bab8]
#> 
#> [[5]]
#> <frame 22> (21)
#> expr: .f(code = .l[[1L]][[1L]], run = .l[[2L]][[1L]], show = .l[[3L]][[1L]],  <...>
#> env:  [frame 0x8db9df8]
#> 
#> [[6]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x79cf5d8]
#> 
#> [[7]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x899ddf8]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x2451a10]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x21797f0]
#> 
#> [[10]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x4f052a0]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x4983528]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x16ff988]
#> 
#> [[13]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x16fee50]
#> 
#> [[14]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3009790]
#> 
#> [[15]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

# The rlang _stack() functions return a list of frame
# objects. Use purrr::transpose() or index a field with
# purrr::map()'s to extract a particular field from a stack:

# stack <- f(f(call_stack()))
# purrr::map(stack, "env")
# purrr::transpose(stack)$expr

# current_frame() is an alias for ctxt_frame(1)
fn <- function() list(current = current_frame(), first = ctxt_frame(1))
fn()

#> Warning: `current_frame()` is soft-deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> $current
#> <frame 38> (37)
#> expr: fn()
#> env:  [local 0x7cc4148]
#> 
#> $first
#> <frame 38> (37)
#> expr: fn()
#> env:  [local 0x7cc4148]
#> 

# While current_frame() is the top of the stack, global_frame() is
# the bottom:
fn <- function() {
  n <- ctxt_depth()
  ctxt_frame(n)
}
identical(fn(), global_frame())

#> Warning: `global_frame()` is soft-deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] TRUE

# ctxt_stack() returns a stack with all intervening frames. You can
# trim layers of intervening frames with the trim argument:
identity(identity(ctxt_stack()))

#> [[1]]
#> <frame 39> (37)
#> expr: identity(ctxt_stack())
#> env:  [local 0x1c64a70]
#> 
#> [[2]]
#> <frame 38> (37)
#> expr: identity(identity(ctxt_stack()))
#> env:  [local 0x47f3ba0]
#> 
#> [[3]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7337078]
#> 
#> [[4]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5ab9798]
#> 
#> [[5]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x817baf8]
#> 
#> [[6]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7a1ce90]
#> 
#> [[7]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x8104f98]
#> 
#> [[8]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x13794f0]
#> 
#> [[9]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x67bcd30]
#> 
#> [[10]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7ee5980]
#> 
#> [[11]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x87bae20]
#> 
#> [[12]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x8104b70]
#> 
#> [[13]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x8d56ca0]
#> 
#> [[14]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x92c0618]
#> 
#> [[15]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x179bab8]
#> 
#> [[16]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x83c2d20]
#> 
#> [[17]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x4ea9300]
#> 
#> [[18]]
#> <frame 22> (21)
#> expr: .f(code = .l[[1L]][[1L]], run = .l[[2L]][[1L]], show = .l[[3L]][[1L]],  <...>
#> env:  [frame 0x8db9df8]
#> 
#> [[19]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x79cf5d8]
#> 
#> [[20]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x899ddf8]
#> 
#> [[21]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x899eb50]
#> 
#> [[22]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x2451a10]
#> 
#> [[23]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x21797f0]
#> 
#> [[24]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x4f052a0]
#> 
#> [[25]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x4983528]
#> 
#> [[26]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x16ff988]
#> 
#> [[27]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x16fee50]
#> 
#> [[28]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3009790]
#> 
#> [[29]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3006920]
#> 
#> [[30]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpcgPkcb/file3eb7c87bb2d"),  <...>
#> env:  [frame 0x3004760]
#> 
#> [[31]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpcgPkcb/file3eb7c87bb2d"),  <...>
#> env:  [frame 0x30054f0]
#> 
#> [[32]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x30040f8]
#> 
#> [[33]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3001480]
#> 
#> [[34]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x30017c8]
#> 
#> [[35]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3001b80]
#> 
#> [[36]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3002130]
#> 
#> [[37]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3002478]
#> 
#> [[38]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x2ffe990]
#> 
#> [[39]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x2fff100]
#> 
#> [[40]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(ctxt_stack(trim = 1)))

#> [[1]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7337078]
#> 
#> [[2]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x883ab98]
#> 
#> [[3]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x883a930]
#> 
#> [[4]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x883fef0]
#> 
#> [[5]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x883f898]
#> 
#> [[6]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x883f550]
#> 
#> [[7]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x883f208]
#> 
#> [[8]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x883eb40]
#> 
#> [[9]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x883e910]
#> 
#> [[10]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x883e788]
#> 
#> [[11]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x883e6a8]
#> 
#> [[12]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x88ace10]
#> 
#> [[13]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x179bab8]
#> 
#> [[14]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x83c2d20]
#> 
#> [[15]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x4ea9300]
#> 
#> [[16]]
#> <frame 22> (21)
#> expr: .f(code = .l[[1L]][[1L]], run = .l[[2L]][[1L]], show = .l[[3L]][[1L]],  <...>
#> env:  [frame 0x8db9df8]
#> 
#> [[17]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x79cf5d8]
#> 
#> [[18]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x899ddf8]
#> 
#> [[19]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x899eb50]
#> 
#> [[20]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x2451a10]
#> 
#> [[21]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x21797f0]
#> 
#> [[22]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x4f052a0]
#> 
#> [[23]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x4983528]
#> 
#> [[24]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x16ff988]
#> 
#> [[25]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x16fee50]
#> 
#> [[26]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3009790]
#> 
#> [[27]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3006920]
#> 
#> [[28]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpcgPkcb/file3eb7c87bb2d"),  <...>
#> env:  [frame 0x3004760]
#> 
#> [[29]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpcgPkcb/file3eb7c87bb2d"),  <...>
#> env:  [frame 0x30054f0]
#> 
#> [[30]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x30040f8]
#> 
#> [[31]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3001480]
#> 
#> [[32]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x30017c8]
#> 
#> [[33]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3001b80]
#> 
#> [[34]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3002130]
#> 
#> [[35]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3002478]
#> 
#> [[36]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x2ffe990]
#> 
#> [[37]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x2fff100]
#> 
#> [[38]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# ctxt_stack() is called within fn() with intervening frames:
fn <- function(trim) identity(identity(ctxt_stack(trim = trim)))
fn(0)

#> [[1]]
#> <frame 40> (38)
#> expr: identity(ctxt_stack(trim = trim))
#> env:  [local 0x5d69a30]
#> 
#> [[2]]
#> <frame 39> (38)
#> expr: identity(identity(ctxt_stack(trim = trim)))
#> env:  [local 0x5d69918]
#> 
#> [[3]]
#> <frame 38> (37)
#> expr: fn(0)
#> env:  [local 0x5d69800]
#> 
#> [[4]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7337078]
#> 
#> [[5]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5d69640]
#> 
#> [[6]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x5d6d208]
#> 
#> [[7]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x5d6cc58]
#> 
#> [[8]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x5d6c638]
#> 
#> [[9]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x5d6c2b8]
#> 
#> [[10]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x5d6bf70]
#> 
#> [[11]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x5d6b838]
#> 
#> [[12]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x5d6b608]
#> 
#> [[13]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x5d6b480]
#> 
#> [[14]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x5d6f198]
#> 
#> [[15]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x5d93320]
#> 
#> [[16]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x179bab8]
#> 
#> [[17]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x83c2d20]
#> 
#> [[18]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x4ea9300]
#> 
#> [[19]]
#> <frame 22> (21)
#> expr: .f(code = .l[[1L]][[1L]], run = .l[[2L]][[1L]], show = .l[[3L]][[1L]],  <...>
#> env:  [frame 0x8db9df8]
#> 
#> [[20]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x79cf5d8]
#> 
#> [[21]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x899ddf8]
#> 
#> [[22]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x899eb50]
#> 
#> [[23]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x2451a10]
#> 
#> [[24]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x21797f0]
#> 
#> [[25]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x4f052a0]
#> 
#> [[26]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x4983528]
#> 
#> [[27]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x16ff988]
#> 
#> [[28]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x16fee50]
#> 
#> [[29]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3009790]
#> 
#> [[30]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3006920]
#> 
#> [[31]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpcgPkcb/file3eb7c87bb2d"),  <...>
#> env:  [frame 0x3004760]
#> 
#> [[32]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpcgPkcb/file3eb7c87bb2d"),  <...>
#> env:  [frame 0x30054f0]
#> 
#> [[33]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x30040f8]
#> 
#> [[34]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3001480]
#> 
#> [[35]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x30017c8]
#> 
#> [[36]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3001b80]
#> 
#> [[37]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3002130]
#> 
#> [[38]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3002478]
#> 
#> [[39]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x2ffe990]
#> 
#> [[40]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x2fff100]
#> 
#> [[41]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# We can trim the first layer of those:
fn(1)

#> [[1]]
#> <frame 38> (37)
#> expr: fn(1)
#> env:  [local 0x52a8ec8]
#> 
#> [[2]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7337078]
#> 
#> [[3]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x52a8d08]
#> 
#> [[4]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x52a8aa0]
#> 
#> [[5]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x52a8410]
#> 
#> [[6]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x52a7d80]
#> 
#> [[7]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x52a79c8]
#> 
#> [[8]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x52acfa8]
#> 
#> [[9]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x52b38b0]
#> 
#> [[10]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x52b3680]
#> 
#> [[11]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x52b31b0]
#> 
#> [[12]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x52b30d0]
#> 
#> [[13]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x5300090]
#> 
#> [[14]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x179bab8]
#> 
#> [[15]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x83c2d20]
#> 
#> [[16]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x4ea9300]
#> 
#> [[17]]
#> <frame 22> (21)
#> expr: .f(code = .l[[1L]][[1L]], run = .l[[2L]][[1L]], show = .l[[3L]][[1L]],  <...>
#> env:  [frame 0x8db9df8]
#> 
#> [[18]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x79cf5d8]
#> 
#> [[19]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x899ddf8]
#> 
#> [[20]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x899eb50]
#> 
#> [[21]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x2451a10]
#> 
#> [[22]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x21797f0]
#> 
#> [[23]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x4f052a0]
#> 
#> [[24]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x4983528]
#> 
#> [[25]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x16ff988]
#> 
#> [[26]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x16fee50]
#> 
#> [[27]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3009790]
#> 
#> [[28]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3006920]
#> 
#> [[29]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpcgPkcb/file3eb7c87bb2d"),  <...>
#> env:  [frame 0x3004760]
#> 
#> [[30]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpcgPkcb/file3eb7c87bb2d"),  <...>
#> env:  [frame 0x30054f0]
#> 
#> [[31]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x30040f8]
#> 
#> [[32]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3001480]
#> 
#> [[33]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x30017c8]
#> 
#> [[34]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3001b80]
#> 
#> [[35]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3002130]
#> 
#> [[36]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3002478]
#> 
#> [[37]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x2ffe990]
#> 
#> [[38]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x2fff100]
#> 
#> [[39]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# The outside intervening frames (at the fn() call site) are still
# returned, but can be trimmed as well:
identity(identity(fn(1)))

#> [[1]]
#> <frame 40> (37)
#> expr: fn(1)
#> env:  [local 0x2ef8d00]
#> 
#> [[2]]
#> <frame 39> (37)
#> expr: identity(fn(1))
#> env:  [local 0x2ef8be8]
#> 
#> [[3]]
#> <frame 38> (37)
#> expr: identity(identity(fn(1)))
#> env:  [local 0x2ef8ad0]
#> 
#> [[4]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7337078]
#> 
#> [[5]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x2ef88d8]
#> 
#> [[6]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x2ef8520]
#> 
#> [[7]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x2ef7410]
#> 
#> [[8]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x2efa088]
#> 
#> [[9]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x2ef99f8]
#> 
#> [[10]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x2efcb40]
#> 
#> [[11]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x2efb5d0]
#> 
#> [[12]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x2efb138]
#> 
#> [[13]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x2efad80]
#> 
#> [[14]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x2efe9f0]
#> 
#> [[15]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x3011020]
#> 
#> [[16]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x179bab8]
#> 
#> [[17]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x83c2d20]
#> 
#> [[18]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x4ea9300]
#> 
#> [[19]]
#> <frame 22> (21)
#> expr: .f(code = .l[[1L]][[1L]], run = .l[[2L]][[1L]], show = .l[[3L]][[1L]],  <...>
#> env:  [frame 0x8db9df8]
#> 
#> [[20]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x79cf5d8]
#> 
#> [[21]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x899ddf8]
#> 
#> [[22]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x899eb50]
#> 
#> [[23]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x2451a10]
#> 
#> [[24]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x21797f0]
#> 
#> [[25]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x4f052a0]
#> 
#> [[26]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x4983528]
#> 
#> [[27]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x16ff988]
#> 
#> [[28]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x16fee50]
#> 
#> [[29]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3009790]
#> 
#> [[30]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3006920]
#> 
#> [[31]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpcgPkcb/file3eb7c87bb2d"),  <...>
#> env:  [frame 0x3004760]
#> 
#> [[32]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpcgPkcb/file3eb7c87bb2d"),  <...>
#> env:  [frame 0x30054f0]
#> 
#> [[33]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x30040f8]
#> 
#> [[34]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3001480]
#> 
#> [[35]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x30017c8]
#> 
#> [[36]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3001b80]
#> 
#> [[37]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3002130]
#> 
#> [[38]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3002478]
#> 
#> [[39]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x2ffe990]
#> 
#> [[40]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x2fff100]
#> 
#> [[41]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(fn(2)))

#> [[1]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7337078]
#> 
#> [[2]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x8163538]
#> 
#> [[3]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x8163768]
#> 
#> [[4]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7145b68]
#> 
#> [[5]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x71460e0]
#> 
#> [[6]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7146460]
#> 
#> [[7]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x71467a8]
#> 
#> [[8]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7146e70]
#> 
#> [[9]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x71470a0]
#> 
#> [[10]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7147260]
#> 
#> [[11]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7147378]
#> 
#> [[12]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x83f25e0]
#> 
#> [[13]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x179bab8]
#> 
#> [[14]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x83c2d20]
#> 
#> [[15]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x4ea9300]
#> 
#> [[16]]
#> <frame 22> (21)
#> expr: .f(code = .l[[1L]][[1L]], run = .l[[2L]][[1L]], show = .l[[3L]][[1L]],  <...>
#> env:  [frame 0x8db9df8]
#> 
#> [[17]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x79cf5d8]
#> 
#> [[18]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x899ddf8]
#> 
#> [[19]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x899eb50]
#> 
#> [[20]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x2451a10]
#> 
#> [[21]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x21797f0]
#> 
#> [[22]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x4f052a0]
#> 
#> [[23]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x4983528]
#> 
#> [[24]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x16ff988]
#> 
#> [[25]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x16fee50]
#> 
#> [[26]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3009790]
#> 
#> [[27]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3006920]
#> 
#> [[28]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpcgPkcb/file3eb7c87bb2d"),  <...>
#> env:  [frame 0x3004760]
#> 
#> [[29]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpcgPkcb/file3eb7c87bb2d"),  <...>
#> env:  [frame 0x30054f0]
#> 
#> [[30]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x30040f8]
#> 
#> [[31]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3001480]
#> 
#> [[32]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x30017c8]
#> 
#> [[33]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3001b80]
#> 
#> [[34]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3002130]
#> 
#> [[35]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3002478]
#> 
#> [[36]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x2ffe990]
#> 
#> [[37]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x2fff100]
#> 
#> [[38]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g <- function(trim) identity(identity(fn(trim)))
g(2)

#> [[1]]
#> <frame 38> (37)
#> expr: g(2)
#> env:  [local 0x7f7b648]
#> 
#> [[2]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7337078]
#> 
#> [[3]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f7b808]
#> 
#> [[4]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f7ba38]
#> 
#> [[5]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f7bfe8]
#> 
#> [[6]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f7c598]
#> 
#> [[7]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f7c8e0]
#> 
#> [[8]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f78f10]
#> 
#> [[9]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f79648]
#> 
#> [[10]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f79878]
#> 
#> [[11]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f79a00]
#> 
#> [[12]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f79ae0]
#> 
#> [[13]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x81ba778]
#> 
#> [[14]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x179bab8]
#> 
#> [[15]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x83c2d20]
#> 
#> [[16]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x4ea9300]
#> 
#> [[17]]
#> <frame 22> (21)
#> expr: .f(code = .l[[1L]][[1L]], run = .l[[2L]][[1L]], show = .l[[3L]][[1L]],  <...>
#> env:  [frame 0x8db9df8]
#> 
#> [[18]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x79cf5d8]
#> 
#> [[19]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x899ddf8]
#> 
#> [[20]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x899eb50]
#> 
#> [[21]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x2451a10]
#> 
#> [[22]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x21797f0]
#> 
#> [[23]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x4f052a0]
#> 
#> [[24]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x4983528]
#> 
#> [[25]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x16ff988]
#> 
#> [[26]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x16fee50]
#> 
#> [[27]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3009790]
#> 
#> [[28]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3006920]
#> 
#> [[29]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpcgPkcb/file3eb7c87bb2d"),  <...>
#> env:  [frame 0x3004760]
#> 
#> [[30]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpcgPkcb/file3eb7c87bb2d"),  <...>
#> env:  [frame 0x30054f0]
#> 
#> [[31]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x30040f8]
#> 
#> [[32]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3001480]
#> 
#> [[33]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x30017c8]
#> 
#> [[34]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3001b80]
#> 
#> [[35]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3002130]
#> 
#> [[36]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3002478]
#> 
#> [[37]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x2ffe990]
#> 
#> [[38]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x2fff100]
#> 
#> [[39]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g(3)

#> [[1]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7337078]
#> 
#> [[2]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7d52080]
#> 
#> [[3]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7d51e50]
#> 
#> [[4]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7d518a0]
#> 
#> [[5]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7d51328]
#> 
#> [[6]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7d50fe0]
#> 
#> [[7]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7d50c98]
#> 
#> [[8]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7d505d0]
#> 
#> [[9]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7090580]
#> 
#> [[10]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x70903f8]
#> 
#> [[11]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7090318]
#> 
#> [[12]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x86df348]
#> 
#> [[13]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x179bab8]
#> 
#> [[14]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x83c2d20]
#> 
#> [[15]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x4ea9300]
#> 
#> [[16]]
#> <frame 22> (21)
#> expr: .f(code = .l[[1L]][[1L]], run = .l[[2L]][[1L]], show = .l[[3L]][[1L]],  <...>
#> env:  [frame 0x8db9df8]
#> 
#> [[17]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x79cf5d8]
#> 
#> [[18]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x899ddf8]
#> 
#> [[19]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x899eb50]
#> 
#> [[20]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x2451a10]
#> 
#> [[21]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x21797f0]
#> 
#> [[22]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x4f052a0]
#> 
#> [[23]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x4983528]
#> 
#> [[24]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x16ff988]
#> 
#> [[25]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x16fee50]
#> 
#> [[26]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3009790]
#> 
#> [[27]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3006920]
#> 
#> [[28]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpcgPkcb/file3eb7c87bb2d"),  <...>
#> env:  [frame 0x3004760]
#> 
#> [[29]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpcgPkcb/file3eb7c87bb2d"),  <...>
#> env:  [frame 0x30054f0]
#> 
#> [[30]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x30040f8]
#> 
#> [[31]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3001480]
#> 
#> [[32]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x30017c8]
#> 
#> [[33]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3001b80]
#> 
#> [[34]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3002130]
#> 
#> [[35]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3002478]
#> 
#> [[36]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x2ffe990]
#> 
#> [[37]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x2fff100]
#> 
#> [[38]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"