Call stack information

The eval_ and call_ families of functions provide a replacement for the base R functions prefixed with sys. (which are all about the context stack), as well as for parent.frame() (which is the only base R function for querying the call stack). The context stack includes all R-level evaluation contexts. It is linear in terms of execution history but due to lazy evaluation it is potentially nonlinear in terms of call history. The call stack history, on the other hand, is homogenous.

global_frame()

current_frame()

ctxt_frame(n = 1)

call_frame(n = 1, clean = TRUE)

ctxt_depth()

call_depth()

ctxt_stack(n = NULL, trim = 0)

call_stack(n = NULL, clean = TRUE)

Arguments

n
The number of frames to go back in the stack.

clean
Whether to post-process the call stack to clean non-standard frames. If TRUE, suboptimal call-stack entries by base::eval() will be cleaned up: the duplicate frame created by eval() is eliminated.

trim
The number of layers of intervening frames to trim off the stack. See stack_trim() and examples.

Details

ctxt_frame() and call_frame() return a frame object containing the following fields: expr and env (call expression and evaluation environment), pos and caller_pos (position of current frame in the context stack and position of the caller), and fun (function of the current frame). ctxt_stack() and call_stack() return a list of all context or call frames on the stack. Finally, ctxt_depth() and call_depth() report the current context position or the number of calling frames on the stack.

The base R functions take two sorts of arguments to indicate which frame to query: which and n. The n argument is straightforward: it's the number of frames to go down the stack, with n = 1 referring to the current context. The which argument is more complicated and changes meaning for values lower than 1. For the sake of consistency, the rlang functions all take the same kind of argument n. This argument has a single meaning (the number of frames to go down the stack) and cannot be lower than 1.

Note finally that parent.frame(1) corresponds to call_frame(2)$env, as n = 1 always refers to the current frame. This makes the _frame() and _stack() functions consistent: ctxt_frame(2) is the same as ctxt_stack()[[2]]. Also, ctxt_depth() returns one more frame than base::sys.nframe() because it counts the global frame. That is consistent with the _stack() functions which return the global frame as well. This way, call_stack(call_depth()) is the same as global_frame().

Life cycle

These functions are soft-deprecated and replaced by trace_back().

Examples

# Expressions within arguments count as contexts
identity(identity(ctxt_depth())) # returns 2

#> Warning: `ctxt_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 25

# But they are not part of the call stack because arguments are
# evaluated within the calling function (or the global environment
# if called at top level)
identity(identity(call_depth())) # returns 0

#> Warning: `call_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 13

# The context stacks includes all intervening execution frames. The
# call stack doesn't:
f <- function(x) identity(x)
f(f(ctxt_stack()))

#> [[1]]
#> <frame 26> (25)
#> expr: identity(x)
#> env:  [local 0x7fdae0450528]
#> 
#> [[2]]
#> <frame 25> (22)
#> expr: f(ctxt_stack())
#> env:  [local 0x7fdae0450640]
#> 
#> [[3]]
#> <frame 24> (23)
#> expr: identity(x)
#> env:  [local 0x7fdae044cc00]
#> 
#> [[4]]
#> <frame 23> (22)
#> expr: f(f(ctxt_stack()))
#> env:  [local 0x7fdae044cd18]
#> 
#> [[5]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdae0374d18]
#> 
#> [[6]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdae044ced8]
#> 
#> [[7]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fdae044d220]
#> 
#> [[8]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fdae044d7d0]
#> 
#> [[9]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fdae044ded0]
#> 
#> [[10]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fdae044e218]
#> 
#> [[11]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fdae044e560]
#> 
#> [[12]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fdae044ad50]
#> 
#> [[13]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fdae044af80]
#> 
#> [[14]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fdae044b108]
#> 
#> [[15]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fdae044b1e8]
#> 
#> [[16]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fdae043b220]
#> 
#> [[17]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fdadb393cf8]
#> 
#> [[18]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fdadb39d6d8]
#> 
#> [[19]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fdada2aa548]
#> 
#> [[20]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fdadb4acaa0]
#> 
#> [[21]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fdadb54e0b8]
#> 
#> [[22]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fdadcd98cf8]
#> 
#> [[23]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdadc1c6bd8]
#> 
#> [[24]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdadd8fb708]
#> 
#> [[25]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7fdadbd3ba08]
#> 
#> [[26]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7fdadbb75670]
#> 
#> [[27]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(f(call_stack()))

#> Warning: `call_stack()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [[1]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdae0374d18]
#> 
#> [[2]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fdae0161338]
#> 
#> [[3]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fdadb393cf8]
#> 
#> [[4]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fdadb39d6d8]
#> 
#> [[5]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fdada2aa548]
#> 
#> [[6]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fdadb4acaa0]
#> 
#> [[7]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fdadb54e0b8]
#> 
#> [[8]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fdadcd98cf8]
#> 
#> [[9]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdadc1c6bd8]
#> 
#> [[10]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdadd8fb708]
#> 
#> [[11]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7fdadbd3ba08]
#> 
#> [[12]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7fdadbb75670]
#> 
#> [[13]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

g <- function(cmd) cmd()
f(g(ctxt_stack))

#> [[1]]
#> <frame 25> (22)
#> expr: g(ctxt_stack)
#> env:  [local 0x7fdae0705b48]
#> 
#> [[2]]
#> <frame 24> (23)
#> expr: identity(x)
#> env:  [local 0x7fdae0705c60]
#> 
#> [[3]]
#> <frame 23> (22)
#> expr: f(g(ctxt_stack))
#> env:  [local 0x7fdae0705d78]
#> 
#> [[4]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdae0374d18]
#> 
#> [[5]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdae0705f38]
#> 
#> [[6]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fdae0706168]
#> 
#> [[7]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fdae0702808]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fdae0702d80]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fdae07030c8]
#> 
#> [[10]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fdae0703410]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fdae0703ad8]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fdae0703d08]
#> 
#> [[13]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fdae0703fa8]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fdae0704130]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fdae06ee990]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fdadb393cf8]
#> 
#> [[17]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fdadb39d6d8]
#> 
#> [[18]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fdada2aa548]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fdadb4acaa0]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fdadb54e0b8]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fdadcd98cf8]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdadc1c6bd8]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdadd8fb708]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7fdadbd3ba08]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7fdadbb75670]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(g(call_stack))

#> [[1]]
#> <frame 25> (21)
#> expr: g(call_stack)
#> env:  [local 0x7fdae1318ee8]
#> 
#> [[2]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdae0374d18]
#> 
#> [[3]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fdae13072a0]
#> 
#> [[4]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fdadb393cf8]
#> 
#> [[5]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fdadb39d6d8]
#> 
#> [[6]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fdada2aa548]
#> 
#> [[7]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fdadb4acaa0]
#> 
#> [[8]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fdadb54e0b8]
#> 
#> [[9]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fdadcd98cf8]
#> 
#> [[10]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdadc1c6bd8]
#> 
#> [[11]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdadd8fb708]
#> 
#> [[12]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7fdadbd3ba08]
#> 
#> [[13]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7fdadbb75670]
#> 
#> [[14]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

# The rlang _stack() functions return a list of frame
# objects. Use purrr::transpose() or index a field with
# purrr::map()'s to extract a particular field from a stack:

# stack <- f(f(call_stack()))
# purrr::map(stack, "env")
# purrr::transpose(stack)$expr

# current_frame() is an alias for ctxt_frame(1)
fn <- function() list(current = current_frame(), first = ctxt_frame(1))
fn()

#> Warning: `current_frame()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> $current
#> <frame 23> (22)
#> expr: fn()
#> env:  [local 0x7fdae14b2148]
#> 
#> $first
#> <frame 23> (22)
#> expr: fn()
#> env:  [local 0x7fdae14b2148]
#> 

# While current_frame() is the top of the stack, global_frame() is
# the bottom:
fn <- function() {
  n <- ctxt_depth()
  ctxt_frame(n)
}
identical(fn(), global_frame())

#> [1] TRUE

# ctxt_stack() returns a stack with all intervening frames. You can
# trim layers of intervening frames with the trim argument:
identity(identity(ctxt_stack()))

#> [[1]]
#> <frame 24> (22)
#> expr: identity(ctxt_stack())
#> env:  [local 0x7fdadfe34038]
#> 
#> [[2]]
#> <frame 23> (22)
#> expr: identity(identity(ctxt_stack()))
#> env:  [local 0x7fdadfe34150]
#> 
#> [[3]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdae0374d18]
#> 
#> [[4]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdadfe34310]
#> 
#> [[5]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fdadfe34540]
#> 
#> [[6]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fdadfe34af0]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fdadfe31158]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fdadfe314a0]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fdadfe31820]
#> 
#> [[10]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fdadfe31f58]
#> 
#> [[11]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fdadfe32188]
#> 
#> [[12]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fdadfe32310]
#> 
#> [[13]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fdadfe323f0]
#> 
#> [[14]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fdadfe22af0]
#> 
#> [[15]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fdadb393cf8]
#> 
#> [[16]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fdadb39d6d8]
#> 
#> [[17]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fdada2aa548]
#> 
#> [[18]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fdadb4acaa0]
#> 
#> [[19]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fdadb54e0b8]
#> 
#> [[20]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fdadcd98cf8]
#> 
#> [[21]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdadc1c6bd8]
#> 
#> [[22]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdadd8fb708]
#> 
#> [[23]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7fdadbd3ba08]
#> 
#> [[24]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7fdadbb75670]
#> 
#> [[25]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(ctxt_stack(trim = 1)))

#> [[1]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdae0374d18]
#> 
#> [[2]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdadfc170e0]
#> 
#> [[3]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fdadfc17310]
#> 
#> [[4]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fdadfc178f8]
#> 
#> [[5]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fdadfc17e70]
#> 
#> [[6]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fdadfc142a8]
#> 
#> [[7]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fdadfc14660]
#> 
#> [[8]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fdadfc14d28]
#> 
#> [[9]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fdadfc14f58]
#> 
#> [[10]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fdadfc150e0]
#> 
#> [[11]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fdadfc151c0]
#> 
#> [[12]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fdadfd60580]
#> 
#> [[13]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fdadb393cf8]
#> 
#> [[14]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fdadb39d6d8]
#> 
#> [[15]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fdada2aa548]
#> 
#> [[16]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fdadb4acaa0]
#> 
#> [[17]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fdadb54e0b8]
#> 
#> [[18]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fdadcd98cf8]
#> 
#> [[19]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdadc1c6bd8]
#> 
#> [[20]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdadd8fb708]
#> 
#> [[21]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7fdadbd3ba08]
#> 
#> [[22]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7fdadbb75670]
#> 
#> [[23]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# ctxt_stack() is called within fn() with intervening frames:
fn <- function(trim) identity(identity(ctxt_stack(trim = trim)))
fn(0)

#> [[1]]
#> <frame 25> (23)
#> expr: identity(ctxt_stack(trim = trim))
#> env:  [local 0x7fdade7f1e08]
#> 
#> [[2]]
#> <frame 24> (23)
#> expr: identity(identity(ctxt_stack(trim = trim)))
#> env:  [local 0x7fdaded3ec00]
#> 
#> [[3]]
#> <frame 23> (22)
#> expr: fn(0)
#> env:  [local 0x7fdaded3eae8]
#> 
#> [[4]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdae0374d18]
#> 
#> [[5]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdaded3e928]
#> 
#> [[6]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fdaded3e6f8]
#> 
#> [[7]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fdaded3de00]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fdaded3d818]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fdaded3d4d0]
#> 
#> [[10]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fdaded3d000]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fdadeac5448]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fdadeac5218]
#> 
#> [[13]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fdadeac5090]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fdadeac4fb0]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fdadf247830]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fdadb393cf8]
#> 
#> [[17]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fdadb39d6d8]
#> 
#> [[18]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fdada2aa548]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fdadb4acaa0]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fdadb54e0b8]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fdadcd98cf8]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdadc1c6bd8]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdadd8fb708]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7fdadbd3ba08]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7fdadbb75670]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# We can trim the first layer of those:
fn(1)

#> [[1]]
#> <frame 23> (22)
#> expr: fn(1)
#> env:  [local 0x7fdaddef04b8]
#> 
#> [[2]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdae0374d18]
#> 
#> [[3]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdaddef0288]
#> 
#> [[4]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fdaddef0058]
#> 
#> [[5]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fdaddeefaa8]
#> 
#> [[6]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fdaddeef530]
#> 
#> [[7]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fdaddeef1e8]
#> 
#> [[8]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fdaddeeeea0]
#> 
#> [[9]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fdaddef46b0]
#> 
#> [[10]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fdaddef4448]
#> 
#> [[11]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fdaddef42c0]
#> 
#> [[12]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fdaddef41e0]
#> 
#> [[13]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fdaddf05fe8]
#> 
#> [[14]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fdadb393cf8]
#> 
#> [[15]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fdadb39d6d8]
#> 
#> [[16]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fdada2aa548]
#> 
#> [[17]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fdadb4acaa0]
#> 
#> [[18]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fdadb54e0b8]
#> 
#> [[19]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fdadcd98cf8]
#> 
#> [[20]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdadc1c6bd8]
#> 
#> [[21]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdadd8fb708]
#> 
#> [[22]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7fdadbd3ba08]
#> 
#> [[23]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7fdadbb75670]
#> 
#> [[24]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# The outside intervening frames (at the fn() call site) are still
# returned, but can be trimmed as well:
identity(identity(fn(1)))

#> [[1]]
#> <frame 25> (22)
#> expr: fn(1)
#> env:  [local 0x7fdadb5c2e28]
#> 
#> [[2]]
#> <frame 24> (22)
#> expr: identity(fn(1))
#> env:  [local 0x7fdadb5c2d10]
#> 
#> [[3]]
#> <frame 23> (22)
#> expr: identity(identity(fn(1)))
#> env:  [local 0x7fdadb5c2bf8]
#> 
#> [[4]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdae0374d18]
#> 
#> [[5]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdadb5c2a38]
#> 
#> [[6]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fdadb5c2808]
#> 
#> [[7]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fdadb5c6130]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fdadb5c5b80]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fdadb5c57c8]
#> 
#> [[10]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fdadb5c5448]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fdadb5c4d10]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fdadb5c4ae0]
#> 
#> [[13]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fdadb5c4958]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fdadb5c4840]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fdadb5d9cd0]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fdadb393cf8]
#> 
#> [[17]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fdadb39d6d8]
#> 
#> [[18]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fdada2aa548]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fdadb4acaa0]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fdadb54e0b8]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fdadcd98cf8]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdadc1c6bd8]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdadd8fb708]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7fdadbd3ba08]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7fdadbb75670]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(fn(2)))

#> [[1]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdae0374d18]
#> 
#> [[2]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdaddd80e98]
#> 
#> [[3]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fdaddd80c30]
#> 
#> [[4]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fdaddd84558]
#> 
#> [[5]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fdaddd83f00]
#> 
#> [[6]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fdaddd83b80]
#> 
#> [[7]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fdaddd83838]
#> 
#> [[8]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fdaddd83100]
#> 
#> [[9]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fdaddd82e98]
#> 
#> [[10]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fdaddd82cd8]
#> 
#> [[11]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fdaddd82bf8]
#> 
#> [[12]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fdaddd9d678]
#> 
#> [[13]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fdadb393cf8]
#> 
#> [[14]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fdadb39d6d8]
#> 
#> [[15]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fdada2aa548]
#> 
#> [[16]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fdadb4acaa0]
#> 
#> [[17]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fdadb54e0b8]
#> 
#> [[18]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fdadcd98cf8]
#> 
#> [[19]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdadc1c6bd8]
#> 
#> [[20]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdadd8fb708]
#> 
#> [[21]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7fdadbd3ba08]
#> 
#> [[22]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7fdadbb75670]
#> 
#> [[23]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g <- function(trim) identity(identity(fn(trim)))
g(2)

#> [[1]]
#> <frame 23> (22)
#> expr: g(2)
#> env:  [local 0x7fdaddb80b80]
#> 
#> [[2]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdae0374d18]
#> 
#> [[3]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdaddb80988]
#> 
#> [[4]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fdaddb80758]
#> 
#> [[5]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fdaddb84080]
#> 
#> [[6]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fdaddb83ad0]
#> 
#> [[7]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fdaddb83718]
#> 
#> [[8]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fdaddb833d0]
#> 
#> [[9]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fdaddb82c98]
#> 
#> [[10]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fdaddb82a68]
#> 
#> [[11]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fdaddb828a8]
#> 
#> [[12]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fdaddb827c8]
#> 
#> [[13]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fdaddb9c0f0]
#> 
#> [[14]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fdadb393cf8]
#> 
#> [[15]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fdadb39d6d8]
#> 
#> [[16]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fdada2aa548]
#> 
#> [[17]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fdadb4acaa0]
#> 
#> [[18]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fdadb54e0b8]
#> 
#> [[19]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fdadcd98cf8]
#> 
#> [[20]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdadc1c6bd8]
#> 
#> [[21]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdadd8fb708]
#> 
#> [[22]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7fdadbd3ba08]
#> 
#> [[23]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7fdadbb75670]
#> 
#> [[24]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g(3)

#> [[1]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdae0374d18]
#> 
#> [[2]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fdada7cb2f8]
#> 
#> [[3]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fdada7cb0c8]
#> 
#> [[4]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fdada7caa38]
#> 
#> [[5]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fdada7d0360]
#> 
#> [[6]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fdada7d0018]
#> 
#> [[7]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fdada7cfcd0]
#> 
#> [[8]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fdada7cf608]
#> 
#> [[9]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fdada7cf3d8]
#> 
#> [[10]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fdada7cf218]
#> 
#> [[11]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fdada7cf138]
#> 
#> [[12]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fdadd838008]
#> 
#> [[13]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fdadb393cf8]
#> 
#> [[14]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fdadb39d6d8]
#> 
#> [[15]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fdada2aa548]
#> 
#> [[16]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fdadb4acaa0]
#> 
#> [[17]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fdadb54e0b8]
#> 
#> [[18]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fdadcd98cf8]
#> 
#> [[19]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdadc1c6bd8]
#> 
#> [[20]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fdadd8fb708]
#> 
#> [[21]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7fdadbd3ba08]
#> 
#> [[22]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7fdadbb75670]
#> 
#> [[23]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"