Call stack information

The eval_ and call_ families of functions provide a replacement for the base R functions prefixed with sys. (which are all about the context stack), as well as for parent.frame() (which is the only base R function for querying the call stack). The context stack includes all R-level evaluation contexts. It is linear in terms of execution history but due to lazy evaluation it is potentially nonlinear in terms of call history. The call stack history, on the other hand, is homogenous.

global_frame()

current_frame()

ctxt_frame(n = 1)

call_frame(n = 1, clean = TRUE)

ctxt_depth()

call_depth()

ctxt_stack(n = NULL, trim = 0)

call_stack(n = NULL, clean = TRUE)

Arguments

n
The number of frames to go back in the stack.

clean
Whether to post-process the call stack to clean non-standard frames. If TRUE, suboptimal call-stack entries by base::eval() will be cleaned up: the duplicate frame created by eval() is eliminated.

trim
The number of layers of intervening frames to trim off the stack. See stack_trim() and examples.

Details

ctxt_frame() and call_frame() return a frame object containing the following fields: expr and env (call expression and evaluation environment), pos and caller_pos (position of current frame in the context stack and position of the caller), and fun (function of the current frame). ctxt_stack() and call_stack() return a list of all context or call frames on the stack. Finally, ctxt_depth() and call_depth() report the current context position or the number of calling frames on the stack.

The base R functions take two sorts of arguments to indicate which frame to query: which and n. The n argument is straightforward: it's the number of frames to go down the stack, with n = 1 referring to the current context. The which argument is more complicated and changes meaning for values lower than 1. For the sake of consistency, the rlang functions all take the same kind of argument n. This argument has a single meaning (the number of frames to go down the stack) and cannot be lower than 1.

Note finally that parent.frame(1) corresponds to call_frame(2)$env, as n = 1 always refers to the current frame. This makes the _frame() and _stack() functions consistent: ctxt_frame(2) is the same as ctxt_stack()[[2]]. Also, ctxt_depth() returns one more frame than base::sys.nframe() because it counts the global frame. That is consistent with the _stack() functions which return the global frame as well. This way, call_stack(call_depth()) is the same as global_frame().

Life cycle

These functions are soft-deprecated and replaced by trace_back().

Examples

# Expressions within arguments count as contexts
identity(identity(ctxt_depth())) # returns 2

#> Warning: `ctxt_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 36

# But they are not part of the call stack because arguments are
# evaluated within the calling function (or the global environment
# if called at top level)
identity(identity(call_depth())) # returns 0

#> Warning: `call_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 13

# The context stacks includes all intervening execution frames. The
# call stack doesn't:
f <- function(x) identity(x)
f(f(ctxt_stack()))

#> [[1]]
#> <frame 37> (36)
#> expr: identity(x)
#> env:  [local 0x6047bf8]
#> 
#> [[2]]
#> <frame 36> (33)
#> expr: f(ctxt_stack())
#> env:  [local 0x6047d10]
#> 
#> [[3]]
#> <frame 35> (34)
#> expr: identity(x)
#> env:  [local 0x6047e28]
#> 
#> [[4]]
#> <frame 34> (33)
#> expr: f(f(ctxt_stack()))
#> env:  [local 0x6047f40]
#> 
#> [[5]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x6f1f3a8]
#> 
#> [[6]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x6048100]
#> 
#> [[7]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x6048330]
#> 
#> [[8]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x60488e0]
#> 
#> [[9]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x6045028]
#> 
#> [[10]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x6045370]
#> 
#> [[11]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x60456b8]
#> 
#> [[12]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x6045d80]
#> 
#> [[13]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x6045fb0]
#> 
#> [[14]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x6046138]
#> 
#> [[15]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x6046218]
#> 
#> [[16]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x6626680]
#> 
#> [[17]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x4425808]
#> 
#> [[18]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x442da10]
#> 
#> [[19]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x44de2b0]
#> 
#> [[20]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x6817e90]
#> 
#> [[21]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x6a131b8]
#> 
#> [[22]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x5e004c8]
#> 
#> [[23]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x478d450]
#> 
#> [[24]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x46a4fb8]
#> 
#> [[25]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x469c650]
#> 
#> [[26]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x437b270]
#> 
#> [[27]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x4e15730]
#> 
#> [[28]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/Rtmp9QljHr/file4fee31e5eb35"),  <...>
#> env:  [frame 0x4376470]
#> 
#> [[29]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/Rtmp9QljHr/file4fee31e5eb35"),  <...>
#> env:  [frame 0x43767f0]
#> 
#> [[30]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x4371538]
#> 
#> [[31]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x436dd98]
#> 
#> [[32]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x436e0e0]
#> 
#> [[33]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x436e498]
#> 
#> [[34]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x436ea48]
#> 
#> [[35]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x436ed90]
#> 
#> [[36]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x436f0d8]
#> 
#> [[37]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x436f810]
#> 
#> [[38]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(f(call_stack()))

#> Warning: `call_stack()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [[1]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x6f1f3a8]
#> 
#> [[2]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x6ab19d8]
#> 
#> [[3]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x4425808]
#> 
#> [[4]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x442da10]
#> 
#> [[5]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x44de2b0]
#> 
#> [[6]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x6817e90]
#> 
#> [[7]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x6a131b8]
#> 
#> [[8]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x5e004c8]
#> 
#> [[9]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x478d450]
#> 
#> [[10]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x46a4fb8]
#> 
#> [[11]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x469c650]
#> 
#> [[12]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x437b270]
#> 
#> [[13]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

g <- function(cmd) cmd()
f(g(ctxt_stack))

#> [[1]]
#> <frame 36> (33)
#> expr: g(ctxt_stack)
#> env:  [local 0x68f52c8]
#> 
#> [[2]]
#> <frame 35> (34)
#> expr: identity(x)
#> env:  [local 0x68f53e0]
#> 
#> [[3]]
#> <frame 34> (33)
#> expr: f(g(ctxt_stack))
#> env:  [local 0x68f54f8]
#> 
#> [[4]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x6f1f3a8]
#> 
#> [[5]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x68f56b8]
#> 
#> [[6]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x68f58e8]
#> 
#> [[7]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x68f5e98]
#> 
#> [[8]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x68f6410]
#> 
#> [[9]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x68f6758]
#> 
#> [[10]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x68f2c70]
#> 
#> [[11]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x68f3338]
#> 
#> [[12]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x68f3568]
#> 
#> [[13]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x68f3840]
#> 
#> [[14]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x68f3a70]
#> 
#> [[15]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x6a072d0]
#> 
#> [[16]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x4425808]
#> 
#> [[17]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x442da10]
#> 
#> [[18]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x44de2b0]
#> 
#> [[19]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x6817e90]
#> 
#> [[20]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x6a131b8]
#> 
#> [[21]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x5e004c8]
#> 
#> [[22]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x478d450]
#> 
#> [[23]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x46a4fb8]
#> 
#> [[24]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x469c650]
#> 
#> [[25]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x437b270]
#> 
#> [[26]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x4e15730]
#> 
#> [[27]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/Rtmp9QljHr/file4fee31e5eb35"),  <...>
#> env:  [frame 0x4376470]
#> 
#> [[28]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/Rtmp9QljHr/file4fee31e5eb35"),  <...>
#> env:  [frame 0x43767f0]
#> 
#> [[29]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x4371538]
#> 
#> [[30]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x436dd98]
#> 
#> [[31]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x436e0e0]
#> 
#> [[32]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x436e498]
#> 
#> [[33]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x436ea48]
#> 
#> [[34]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x436ed90]
#> 
#> [[35]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x436f0d8]
#> 
#> [[36]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x436f810]
#> 
#> [[37]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(g(call_stack))

#> [[1]]
#> <frame 36> (32)
#> expr: g(call_stack)
#> env:  [local 0xa22be68]
#> 
#> [[2]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x6f1f3a8]
#> 
#> [[3]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0xa8db318]
#> 
#> [[4]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x4425808]
#> 
#> [[5]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x442da10]
#> 
#> [[6]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x44de2b0]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x6817e90]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x6a131b8]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x5e004c8]
#> 
#> [[10]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x478d450]
#> 
#> [[11]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x46a4fb8]
#> 
#> [[12]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x469c650]
#> 
#> [[13]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x437b270]
#> 
#> [[14]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

# The rlang _stack() functions return a list of frame
# objects. Use purrr::transpose() or index a field with
# purrr::map()'s to extract a particular field from a stack:

# stack <- f(f(call_stack()))
# purrr::map(stack, "env")
# purrr::transpose(stack)$expr

# current_frame() is an alias for ctxt_frame(1)
fn <- function() list(current = current_frame(), first = ctxt_frame(1))
fn()

#> Warning: `current_frame()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> $current
#> <frame 34> (33)
#> expr: fn()
#> env:  [local 0xa84f138]
#> 
#> $first
#> <frame 34> (33)
#> expr: fn()
#> env:  [local 0xa84f138]
#> 

# While current_frame() is the top of the stack, global_frame() is
# the bottom:
fn <- function() {
  n <- ctxt_depth()
  ctxt_frame(n)
}
identical(fn(), global_frame())

#> [1] TRUE

# ctxt_stack() returns a stack with all intervening frames. You can
# trim layers of intervening frames with the trim argument:
identity(identity(ctxt_stack()))

#> [[1]]
#> <frame 35> (33)
#> expr: identity(ctxt_stack())
#> env:  [local 0xa44a0c8]
#> 
#> [[2]]
#> <frame 34> (33)
#> expr: identity(identity(ctxt_stack()))
#> env:  [local 0xa44a1e0]
#> 
#> [[3]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x6f1f3a8]
#> 
#> [[4]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0xa44a3a0]
#> 
#> [[5]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0xa44a5d0]
#> 
#> [[6]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0xa446d50]
#> 
#> [[7]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0xa4472c8]
#> 
#> [[8]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0xa447610]
#> 
#> [[9]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0xa447958]
#> 
#> [[10]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0xa448020]
#> 
#> [[11]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0xa448288]
#> 
#> [[12]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0xa448448]
#> 
#> [[13]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0xa448528]
#> 
#> [[14]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0xa438e90]
#> 
#> [[15]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x4425808]
#> 
#> [[16]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x442da10]
#> 
#> [[17]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x44de2b0]
#> 
#> [[18]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x6817e90]
#> 
#> [[19]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x6a131b8]
#> 
#> [[20]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x5e004c8]
#> 
#> [[21]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x478d450]
#> 
#> [[22]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x46a4fb8]
#> 
#> [[23]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x469c650]
#> 
#> [[24]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x437b270]
#> 
#> [[25]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x4e15730]
#> 
#> [[26]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/Rtmp9QljHr/file4fee31e5eb35"),  <...>
#> env:  [frame 0x4376470]
#> 
#> [[27]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/Rtmp9QljHr/file4fee31e5eb35"),  <...>
#> env:  [frame 0x43767f0]
#> 
#> [[28]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x4371538]
#> 
#> [[29]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x436dd98]
#> 
#> [[30]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x436e0e0]
#> 
#> [[31]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x436e498]
#> 
#> [[32]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x436ea48]
#> 
#> [[33]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x436ed90]
#> 
#> [[34]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x436f0d8]
#> 
#> [[35]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x436f810]
#> 
#> [[36]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(ctxt_stack(trim = 1)))

#> [[1]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x6f1f3a8]
#> 
#> [[2]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0xaf3e7d0]
#> 
#> [[3]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0xaf3ea00]
#> 
#> [[4]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0xaf3efb0]
#> 
#> [[5]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0xaf3f528]
#> 
#> [[6]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0xaf3f870]
#> 
#> [[7]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0xaf3fbb8]
#> 
#> [[8]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0xaf3a440]
#> 
#> [[9]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0xaf3a670]
#> 
#> [[10]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0xaf3a7f8]
#> 
#> [[11]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0xaf3a8d8]
#> 
#> [[12]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0xaf29258]
#> 
#> [[13]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x4425808]
#> 
#> [[14]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x442da10]
#> 
#> [[15]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x44de2b0]
#> 
#> [[16]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x6817e90]
#> 
#> [[17]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x6a131b8]
#> 
#> [[18]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x5e004c8]
#> 
#> [[19]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x478d450]
#> 
#> [[20]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x46a4fb8]
#> 
#> [[21]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x469c650]
#> 
#> [[22]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x437b270]
#> 
#> [[23]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x4e15730]
#> 
#> [[24]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/Rtmp9QljHr/file4fee31e5eb35"),  <...>
#> env:  [frame 0x4376470]
#> 
#> [[25]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/Rtmp9QljHr/file4fee31e5eb35"),  <...>
#> env:  [frame 0x43767f0]
#> 
#> [[26]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x4371538]
#> 
#> [[27]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x436dd98]
#> 
#> [[28]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x436e0e0]
#> 
#> [[29]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x436e498]
#> 
#> [[30]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x436ea48]
#> 
#> [[31]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x436ed90]
#> 
#> [[32]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x436f0d8]
#> 
#> [[33]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x436f810]
#> 
#> [[34]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# ctxt_stack() is called within fn() with intervening frames:
fn <- function(trim) identity(identity(ctxt_stack(trim = trim)))
fn(0)

#> [[1]]
#> <frame 36> (34)
#> expr: identity(ctxt_stack(trim = trim))
#> env:  [local 0xab34518]
#> 
#> [[2]]
#> <frame 35> (34)
#> expr: identity(identity(ctxt_stack(trim = trim)))
#> env:  [local 0xab34400]
#> 
#> [[3]]
#> <frame 34> (33)
#> expr: fn(0)
#> env:  [local 0xab342e8]
#> 
#> [[4]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x6f1f3a8]
#> 
#> [[5]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0xab34128]
#> 
#> [[6]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0xab3f9a8]
#> 
#> [[7]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0xab3f3f8]
#> 
#> [[8]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0xab3ee10]
#> 
#> [[9]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0xab3e9e8]
#> 
#> [[10]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0xab3e3c8]
#> 
#> [[11]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0xab3dc58]
#> 
#> [[12]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0xab41858]
#> 
#> [[13]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0xab416d0]
#> 
#> [[14]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0xab415b8]
#> 
#> [[15]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0xa7f23f8]
#> 
#> [[16]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x4425808]
#> 
#> [[17]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x442da10]
#> 
#> [[18]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x44de2b0]
#> 
#> [[19]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x6817e90]
#> 
#> [[20]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x6a131b8]
#> 
#> [[21]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x5e004c8]
#> 
#> [[22]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x478d450]
#> 
#> [[23]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x46a4fb8]
#> 
#> [[24]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x469c650]
#> 
#> [[25]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x437b270]
#> 
#> [[26]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x4e15730]
#> 
#> [[27]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/Rtmp9QljHr/file4fee31e5eb35"),  <...>
#> env:  [frame 0x4376470]
#> 
#> [[28]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/Rtmp9QljHr/file4fee31e5eb35"),  <...>
#> env:  [frame 0x43767f0]
#> 
#> [[29]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x4371538]
#> 
#> [[30]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x436dd98]
#> 
#> [[31]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x436e0e0]
#> 
#> [[32]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x436e498]
#> 
#> [[33]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x436ea48]
#> 
#> [[34]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x436ed90]
#> 
#> [[35]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x436f0d8]
#> 
#> [[36]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x436f810]
#> 
#> [[37]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# We can trim the first layer of those:
fn(1)

#> [[1]]
#> <frame 34> (33)
#> expr: fn(1)
#> env:  [local 0x77e37f8]
#> 
#> [[2]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x6f1f3a8]
#> 
#> [[3]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x77e3600]
#> 
#> [[4]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x77e3360]
#> 
#> [[5]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x77e2db0]
#> 
#> [[6]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x77e6668]
#> 
#> [[7]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x77e6320]
#> 
#> [[8]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x77e5fd8]
#> 
#> [[9]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x77e5280]
#> 
#> [[10]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x77e5018]
#> 
#> [[11]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x77e4e90]
#> 
#> [[12]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x77e4db0]
#> 
#> [[13]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x780e9c0]
#> 
#> [[14]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x4425808]
#> 
#> [[15]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x442da10]
#> 
#> [[16]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x44de2b0]
#> 
#> [[17]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x6817e90]
#> 
#> [[18]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x6a131b8]
#> 
#> [[19]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x5e004c8]
#> 
#> [[20]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x478d450]
#> 
#> [[21]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x46a4fb8]
#> 
#> [[22]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x469c650]
#> 
#> [[23]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x437b270]
#> 
#> [[24]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x4e15730]
#> 
#> [[25]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/Rtmp9QljHr/file4fee31e5eb35"),  <...>
#> env:  [frame 0x4376470]
#> 
#> [[26]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/Rtmp9QljHr/file4fee31e5eb35"),  <...>
#> env:  [frame 0x43767f0]
#> 
#> [[27]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x4371538]
#> 
#> [[28]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x436dd98]
#> 
#> [[29]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x436e0e0]
#> 
#> [[30]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x436e498]
#> 
#> [[31]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x436ea48]
#> 
#> [[32]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x436ed90]
#> 
#> [[33]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x436f0d8]
#> 
#> [[34]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x436f810]
#> 
#> [[35]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# The outside intervening frames (at the fn() call site) are still
# returned, but can be trimmed as well:
identity(identity(fn(1)))

#> [[1]]
#> <frame 36> (33)
#> expr: fn(1)
#> env:  [local 0x6c12860]
#> 
#> [[2]]
#> <frame 35> (33)
#> expr: identity(fn(1))
#> env:  [local 0x6c12748]
#> 
#> [[3]]
#> <frame 34> (33)
#> expr: identity(identity(fn(1)))
#> env:  [local 0x6c12630]
#> 
#> [[4]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x6f1f3a8]
#> 
#> [[5]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x6c12470]
#> 
#> [[6]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x6c12208]
#> 
#> [[7]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x6c118a0]
#> 
#> [[8]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x6c110c0]
#> 
#> [[9]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x6c14ba8]
#> 
#> [[10]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x6c14860]
#> 
#> [[11]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x6c14198]
#> 
#> [[12]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x6c13f68]
#> 
#> [[13]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x6c13de0]
#> 
#> [[14]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x6c13d00]
#> 
#> [[15]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x6c33e88]
#> 
#> [[16]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x4425808]
#> 
#> [[17]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x442da10]
#> 
#> [[18]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x44de2b0]
#> 
#> [[19]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x6817e90]
#> 
#> [[20]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x6a131b8]
#> 
#> [[21]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x5e004c8]
#> 
#> [[22]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x478d450]
#> 
#> [[23]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x46a4fb8]
#> 
#> [[24]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x469c650]
#> 
#> [[25]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x437b270]
#> 
#> [[26]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x4e15730]
#> 
#> [[27]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/Rtmp9QljHr/file4fee31e5eb35"),  <...>
#> env:  [frame 0x4376470]
#> 
#> [[28]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/Rtmp9QljHr/file4fee31e5eb35"),  <...>
#> env:  [frame 0x43767f0]
#> 
#> [[29]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x4371538]
#> 
#> [[30]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x436dd98]
#> 
#> [[31]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x436e0e0]
#> 
#> [[32]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x436e498]
#> 
#> [[33]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x436ea48]
#> 
#> [[34]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x436ed90]
#> 
#> [[35]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x436f0d8]
#> 
#> [[36]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x436f810]
#> 
#> [[37]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(fn(2)))

#> [[1]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x6f1f3a8]
#> 
#> [[2]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x626c440]
#> 
#> [[3]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x626c210]
#> 
#> [[4]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x626bc60]
#> 
#> [[5]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x626b6e8]
#> 
#> [[6]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x626b3a0]
#> 
#> [[7]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x626b058]
#> 
#> [[8]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x626a958]
#> 
#> [[9]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x626e558]
#> 
#> [[10]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x626e3d0]
#> 
#> [[11]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x626e2f0]
#> 
#> [[12]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x6290370]
#> 
#> [[13]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x4425808]
#> 
#> [[14]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x442da10]
#> 
#> [[15]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x44de2b0]
#> 
#> [[16]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x6817e90]
#> 
#> [[17]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x6a131b8]
#> 
#> [[18]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x5e004c8]
#> 
#> [[19]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x478d450]
#> 
#> [[20]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x46a4fb8]
#> 
#> [[21]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x469c650]
#> 
#> [[22]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x437b270]
#> 
#> [[23]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x4e15730]
#> 
#> [[24]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/Rtmp9QljHr/file4fee31e5eb35"),  <...>
#> env:  [frame 0x4376470]
#> 
#> [[25]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/Rtmp9QljHr/file4fee31e5eb35"),  <...>
#> env:  [frame 0x43767f0]
#> 
#> [[26]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x4371538]
#> 
#> [[27]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x436dd98]
#> 
#> [[28]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x436e0e0]
#> 
#> [[29]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x436e498]
#> 
#> [[30]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x436ea48]
#> 
#> [[31]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x436ed90]
#> 
#> [[32]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x436f0d8]
#> 
#> [[33]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x436f810]
#> 
#> [[34]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g <- function(trim) identity(identity(fn(trim)))
g(2)

#> [[1]]
#> <frame 34> (33)
#> expr: g(2)
#> env:  [local 0x4fd5fd0]
#> 
#> [[2]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x6f1f3a8]
#> 
#> [[3]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x4fd5e10]
#> 
#> [[4]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x4fd5be0]
#> 
#> [[5]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x4fd9460]
#> 
#> [[6]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x4fd8ee8]
#> 
#> [[7]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x4fd8b68]
#> 
#> [[8]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x4fd7a58]
#> 
#> [[9]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x4fdd758]
#> 
#> [[10]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x4fdd528]
#> 
#> [[11]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x4fdd3a0]
#> 
#> [[12]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x4fdd2c0]
#> 
#> [[13]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x4ff4f08]
#> 
#> [[14]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x4425808]
#> 
#> [[15]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x442da10]
#> 
#> [[16]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x44de2b0]
#> 
#> [[17]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x6817e90]
#> 
#> [[18]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x6a131b8]
#> 
#> [[19]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x5e004c8]
#> 
#> [[20]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x478d450]
#> 
#> [[21]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x46a4fb8]
#> 
#> [[22]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x469c650]
#> 
#> [[23]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x437b270]
#> 
#> [[24]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x4e15730]
#> 
#> [[25]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/Rtmp9QljHr/file4fee31e5eb35"),  <...>
#> env:  [frame 0x4376470]
#> 
#> [[26]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/Rtmp9QljHr/file4fee31e5eb35"),  <...>
#> env:  [frame 0x43767f0]
#> 
#> [[27]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x4371538]
#> 
#> [[28]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x436dd98]
#> 
#> [[29]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x436e0e0]
#> 
#> [[30]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x436e498]
#> 
#> [[31]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x436ea48]
#> 
#> [[32]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x436ed90]
#> 
#> [[33]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x436f0d8]
#> 
#> [[34]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x436f810]
#> 
#> [[35]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g(3)

#> [[1]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x6f1f3a8]
#> 
#> [[2]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x42aa418]
#> 
#> [[3]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x42b20c8]
#> 
#> [[4]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x42b1b18]
#> 
#> [[5]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x42b15a0]
#> 
#> [[6]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x42b1258]
#> 
#> [[7]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x42b0ed8]
#> 
#> [[8]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x42b07d8]
#> 
#> [[9]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x42b05a8]
#> 
#> [[10]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x42b0420]
#> 
#> [[11]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x42b0340]
#> 
#> [[12]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x42c61c0]
#> 
#> [[13]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x4425808]
#> 
#> [[14]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x442da10]
#> 
#> [[15]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x44de2b0]
#> 
#> [[16]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x6817e90]
#> 
#> [[17]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x6a131b8]
#> 
#> [[18]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x5e004c8]
#> 
#> [[19]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x478d450]
#> 
#> [[20]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x46a4fb8]
#> 
#> [[21]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x469c650]
#> 
#> [[22]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x437b270]
#> 
#> [[23]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x4e15730]
#> 
#> [[24]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/Rtmp9QljHr/file4fee31e5eb35"),  <...>
#> env:  [frame 0x4376470]
#> 
#> [[25]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/Rtmp9QljHr/file4fee31e5eb35"),  <...>
#> env:  [frame 0x43767f0]
#> 
#> [[26]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x4371538]
#> 
#> [[27]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x436dd98]
#> 
#> [[28]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x436e0e0]
#> 
#> [[29]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x436e498]
#> 
#> [[30]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x436ea48]
#> 
#> [[31]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x436ed90]
#> 
#> [[32]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x436f0d8]
#> 
#> [[33]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x436f810]
#> 
#> [[34]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"