Call stack information

The eval_ and call_ families of functions provide a replacement for the base R functions prefixed with sys. (which are all about the context stack), as well as for parent.frame() (which is the only base R function for querying the call stack). The context stack includes all R-level evaluation contexts. It is linear in terms of execution history but due to lazy evaluation it is potentially nonlinear in terms of call history. The call stack history, on the other hand, is homogenous.

global_frame()

current_frame()

ctxt_frame(n = 1)

call_frame(n = 1, clean = TRUE)

ctxt_depth()

call_depth()

ctxt_stack(n = NULL, trim = 0)

call_stack(n = NULL, clean = TRUE)

Arguments

n
The number of frames to go back in the stack.

clean
Whether to post-process the call stack to clean non-standard frames. If TRUE, suboptimal call-stack entries by base::eval() will be cleaned up: the duplicate frame created by eval() is eliminated.

trim
The number of layers of intervening frames to trim off the stack. See stack_trim() and examples.

Details

ctxt_frame() and call_frame() return a frame object containing the following fields: expr and env (call expression and evaluation environment), pos and caller_pos (position of current frame in the context stack and position of the caller), and fun (function of the current frame). ctxt_stack() and call_stack() return a list of all context or call frames on the stack. Finally, ctxt_depth() and call_depth() report the current context position or the number of calling frames on the stack.

The base R functions take two sorts of arguments to indicate which frame to query: which and n. The n argument is straightforward: it's the number of frames to go down the stack, with n = 1 referring to the current context. The which argument is more complicated and changes meaning for values lower than 1. For the sake of consistency, the rlang functions all take the same kind of argument n. This argument has a single meaning (the number of frames to go down the stack) and cannot be lower than 1.

Note finally that parent.frame(1) corresponds to call_frame(2)$env, as n = 1 always refers to the current frame. This makes the _frame() and _stack() functions consistent: ctxt_frame(2) is the same as ctxt_stack()[[2]]. Also, ctxt_depth() returns one more frame than base::sys.nframe() because it counts the global frame. That is consistent with the _stack() functions which return the global frame as well. This way, call_stack(call_depth()) is the same as global_frame().

Life cycle

These functions are soft-deprecated and replaced by trace_back().

Examples

# Expressions within arguments count as contexts
identity(identity(ctxt_depth())) # returns 2

#> Warning: `ctxt_depth()` is soft-deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 40

# But they are not part of the call stack because arguments are
# evaluated within the calling function (or the global environment
# if called at top level)
identity(identity(call_depth())) # returns 0

#> Warning: `call_depth()` is soft-deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 14

# The context stacks includes all intervening execution frames. The
# call stack doesn't:
f <- function(x) identity(x)
f(f(ctxt_stack()))

#> [[1]]
#> <frame 41> (40)
#> expr: identity(x)
#> env:  [local 0x8277a58]
#> 
#> [[2]]
#> <frame 40> (37)
#> expr: f(ctxt_stack())
#> env:  [local 0x8277940]
#> 
#> [[3]]
#> <frame 39> (38)
#> expr: identity(x)
#> env:  [local 0x8277828]
#> 
#> [[4]]
#> <frame 38> (37)
#> expr: f(f(ctxt_stack()))
#> env:  [local 0x8277710]
#> 
#> [[5]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x8c34bc8]
#> 
#> [[6]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x8277550]
#> 
#> [[7]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x8277320]
#> 
#> [[8]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x8276d70]
#> 
#> [[9]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x82767f8]
#> 
#> [[10]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x96fddc0]
#> 
#> [[11]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x96fda78]
#> 
#> [[12]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x96fd3b0]
#> 
#> [[13]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x96fd180]
#> 
#> [[14]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x96fcff8]
#> 
#> [[15]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x96fcf18]
#> 
#> [[16]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x8539800]
#> 
#> [[17]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x991d768]
#> 
#> [[18]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x991c070]
#> 
#> [[19]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x991e620]
#> 
#> [[20]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x8e421a0]
#> 
#> [[21]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x8311900]
#> 
#> [[22]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x27c4f98]
#> 
#> [[23]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x28039f8]
#> 
#> [[24]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x5ee6710]
#> 
#> [[25]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x5fba0b8]
#> 
#> [[26]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x73b9ed0]
#> 
#> [[27]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x5071f50]
#> 
#> [[28]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x2e66e50]
#> 
#> [[29]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x2e6f6d8]
#> 
#> [[30]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3b152d0]
#> 
#> [[31]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3b16290]
#> 
#> [[32]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpMyYvaN/file3d1f24aeb05f"),  <...>
#> env:  [frame 0x3b060f0]
#> 
#> [[33]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpMyYvaN/file3d1f24aeb05f"),  <...>
#> env:  [frame 0x3b06e80]
#> 
#> [[34]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3b01c58]
#> 
#> [[35]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3b02e10]
#> 
#> [[36]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x3b03158]
#> 
#> [[37]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3aff6e0]
#> 
#> [[38]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3affc90]
#> 
#> [[39]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3afffd8]
#> 
#> [[40]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x3b00320]
#> 
#> [[41]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3b00a90]
#> 
#> [[42]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(f(call_stack()))

#> Warning: `call_stack()` is soft-deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [[1]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x8c34bc8]
#> 
#> [[2]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x93f96b8]
#> 
#> [[3]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x991d768]
#> 
#> [[4]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x8e421a0]
#> 
#> [[5]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x8311900]
#> 
#> [[6]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x27c4f98]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x5ee6710]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x5fba0b8]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x73b9ed0]
#> 
#> [[10]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x5071f50]
#> 
#> [[11]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x2e66e50]
#> 
#> [[12]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x2e6f6d8]
#> 
#> [[13]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3b152d0]
#> 
#> [[14]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

g <- function(cmd) cmd()
f(g(ctxt_stack))

#> [[1]]
#> <frame 40> (37)
#> expr: g(ctxt_stack)
#> env:  [local 0x84fa788]
#> 
#> [[2]]
#> <frame 39> (38)
#> expr: identity(x)
#> env:  [local 0x84fa670]
#> 
#> [[3]]
#> <frame 38> (37)
#> expr: f(g(ctxt_stack))
#> env:  [local 0x84fa558]
#> 
#> [[4]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x8c34bc8]
#> 
#> [[5]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x84fa398]
#> 
#> [[6]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x84fa168]
#> 
#> [[7]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x84f9bb8]
#> 
#> [[8]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x84fe8e0]
#> 
#> [[9]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x84fe598]
#> 
#> [[10]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x84fe250]
#> 
#> [[11]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x84fdb88]
#> 
#> [[12]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x84fd958]
#> 
#> [[13]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x84fd7d0]
#> 
#> [[14]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x84fd6f0]
#> 
#> [[15]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x8608ec8]
#> 
#> [[16]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x991d768]
#> 
#> [[17]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x991c070]
#> 
#> [[18]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x991e620]
#> 
#> [[19]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x8e421a0]
#> 
#> [[20]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x8311900]
#> 
#> [[21]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x27c4f98]
#> 
#> [[22]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x28039f8]
#> 
#> [[23]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x5ee6710]
#> 
#> [[24]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x5fba0b8]
#> 
#> [[25]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x73b9ed0]
#> 
#> [[26]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x5071f50]
#> 
#> [[27]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x2e66e50]
#> 
#> [[28]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x2e6f6d8]
#> 
#> [[29]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3b152d0]
#> 
#> [[30]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3b16290]
#> 
#> [[31]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpMyYvaN/file3d1f24aeb05f"),  <...>
#> env:  [frame 0x3b060f0]
#> 
#> [[32]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpMyYvaN/file3d1f24aeb05f"),  <...>
#> env:  [frame 0x3b06e80]
#> 
#> [[33]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3b01c58]
#> 
#> [[34]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3b02e10]
#> 
#> [[35]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x3b03158]
#> 
#> [[36]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3aff6e0]
#> 
#> [[37]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3affc90]
#> 
#> [[38]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3afffd8]
#> 
#> [[39]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x3b00320]
#> 
#> [[40]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3b00a90]
#> 
#> [[41]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(g(call_stack))

#> [[1]]
#> <frame 40> (36)
#> expr: g(call_stack)
#> env:  [local 0x61b0558]
#> 
#> [[2]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x8c34bc8]
#> 
#> [[3]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x61d6488]
#> 
#> [[4]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x991d768]
#> 
#> [[5]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x8e421a0]
#> 
#> [[6]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x8311900]
#> 
#> [[7]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x27c4f98]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x5ee6710]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x5fba0b8]
#> 
#> [[10]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x73b9ed0]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x5071f50]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x2e66e50]
#> 
#> [[13]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x2e6f6d8]
#> 
#> [[14]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3b152d0]
#> 
#> [[15]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

# The rlang _stack() functions return a list of frame
# objects. Use purrr::transpose() or index a field with
# purrr::map()'s to extract a particular field from a stack:

# stack <- f(f(call_stack()))
# purrr::map(stack, "env")
# purrr::transpose(stack)$expr

# current_frame() is an alias for ctxt_frame(1)
fn <- function() list(current = current_frame(), first = ctxt_frame(1))
fn()

#> Warning: `current_frame()` is soft-deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> $current
#> <frame 38> (37)
#> expr: fn()
#> env:  [local 0x3c28b60]
#> 
#> $first
#> <frame 38> (37)
#> expr: fn()
#> env:  [local 0x3c28b60]
#> 

# While current_frame() is the top of the stack, global_frame() is
# the bottom:
fn <- function() {
  n <- ctxt_depth()
  ctxt_frame(n)
}
identical(fn(), global_frame())

#> Warning: `global_frame()` is soft-deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] TRUE

# ctxt_stack() returns a stack with all intervening frames. You can
# trim layers of intervening frames with the trim argument:
identity(identity(ctxt_stack()))

#> [[1]]
#> <frame 39> (37)
#> expr: identity(ctxt_stack())
#> env:  [local 0x2a31ff8]
#> 
#> [[2]]
#> <frame 38> (37)
#> expr: identity(identity(ctxt_stack()))
#> env:  [local 0x2a31ea8]
#> 
#> [[3]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x8c34bc8]
#> 
#> [[4]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x2a31ce8]
#> 
#> [[5]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x2a31a80]
#> 
#> [[6]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x2a313b8]
#> 
#> [[7]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x2a30d98]
#> 
#> [[8]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x2a30a50]
#> 
#> [[9]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x2a305f0]
#> 
#> [[10]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x2a33c78]
#> 
#> [[11]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x2a33a10]
#> 
#> [[12]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x2a33850]
#> 
#> [[13]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x2a33770]
#> 
#> [[14]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x2aceff0]
#> 
#> [[15]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x991d768]
#> 
#> [[16]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x991c070]
#> 
#> [[17]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x991e620]
#> 
#> [[18]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x8e421a0]
#> 
#> [[19]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x8311900]
#> 
#> [[20]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x27c4f98]
#> 
#> [[21]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x28039f8]
#> 
#> [[22]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x5ee6710]
#> 
#> [[23]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x5fba0b8]
#> 
#> [[24]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x73b9ed0]
#> 
#> [[25]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x5071f50]
#> 
#> [[26]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x2e66e50]
#> 
#> [[27]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x2e6f6d8]
#> 
#> [[28]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3b152d0]
#> 
#> [[29]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3b16290]
#> 
#> [[30]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpMyYvaN/file3d1f24aeb05f"),  <...>
#> env:  [frame 0x3b060f0]
#> 
#> [[31]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpMyYvaN/file3d1f24aeb05f"),  <...>
#> env:  [frame 0x3b06e80]
#> 
#> [[32]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3b01c58]
#> 
#> [[33]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3b02e10]
#> 
#> [[34]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x3b03158]
#> 
#> [[35]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3aff6e0]
#> 
#> [[36]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3affc90]
#> 
#> [[37]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3afffd8]
#> 
#> [[38]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x3b00320]
#> 
#> [[39]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3b00a90]
#> 
#> [[40]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(ctxt_stack(trim = 1)))

#> [[1]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x8c34bc8]
#> 
#> [[2]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x97938f0]
#> 
#> [[3]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x97936c0]
#> 
#> [[4]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x9793110]
#> 
#> [[5]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x9792b98]
#> 
#> [[6]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x9792850]
#> 
#> [[7]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x9796338]
#> 
#> [[8]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x9795c70]
#> 
#> [[9]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x9795a40]
#> 
#> [[10]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x97958b8]
#> 
#> [[11]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x97957d8]
#> 
#> [[12]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x8b61a28]
#> 
#> [[13]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x991d768]
#> 
#> [[14]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x991c070]
#> 
#> [[15]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x991e620]
#> 
#> [[16]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x8e421a0]
#> 
#> [[17]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x8311900]
#> 
#> [[18]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x27c4f98]
#> 
#> [[19]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x28039f8]
#> 
#> [[20]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x5ee6710]
#> 
#> [[21]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x5fba0b8]
#> 
#> [[22]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x73b9ed0]
#> 
#> [[23]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x5071f50]
#> 
#> [[24]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x2e66e50]
#> 
#> [[25]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x2e6f6d8]
#> 
#> [[26]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3b152d0]
#> 
#> [[27]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3b16290]
#> 
#> [[28]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpMyYvaN/file3d1f24aeb05f"),  <...>
#> env:  [frame 0x3b060f0]
#> 
#> [[29]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpMyYvaN/file3d1f24aeb05f"),  <...>
#> env:  [frame 0x3b06e80]
#> 
#> [[30]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3b01c58]
#> 
#> [[31]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3b02e10]
#> 
#> [[32]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x3b03158]
#> 
#> [[33]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3aff6e0]
#> 
#> [[34]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3affc90]
#> 
#> [[35]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3afffd8]
#> 
#> [[36]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x3b00320]
#> 
#> [[37]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3b00a90]
#> 
#> [[38]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# ctxt_stack() is called within fn() with intervening frames:
fn <- function(trim) identity(identity(ctxt_stack(trim = trim)))
fn(0)

#> [[1]]
#> <frame 40> (38)
#> expr: identity(ctxt_stack(trim = trim))
#> env:  [local 0x97c48f8]
#> 
#> [[2]]
#> <frame 39> (38)
#> expr: identity(identity(ctxt_stack(trim = trim)))
#> env:  [local 0x97c47e0]
#> 
#> [[3]]
#> <frame 38> (37)
#> expr: fn(0)
#> env:  [local 0x97c46c8]
#> 
#> [[4]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x8c34bc8]
#> 
#> [[5]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x97c4380]
#> 
#> [[6]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x97c4150]
#> 
#> [[7]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x97c3ba0]
#> 
#> [[8]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x97cb298]
#> 
#> [[9]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x97caf50]
#> 
#> [[10]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x97cab60]
#> 
#> [[11]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x97ca428]
#> 
#> [[12]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x97ca1f8]
#> 
#> [[13]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x97ca038]
#> 
#> [[14]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x97c9e40]
#> 
#> [[15]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x89fe500]
#> 
#> [[16]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x991d768]
#> 
#> [[17]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x991c070]
#> 
#> [[18]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x991e620]
#> 
#> [[19]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x8e421a0]
#> 
#> [[20]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x8311900]
#> 
#> [[21]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x27c4f98]
#> 
#> [[22]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x28039f8]
#> 
#> [[23]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x5ee6710]
#> 
#> [[24]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x5fba0b8]
#> 
#> [[25]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x73b9ed0]
#> 
#> [[26]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x5071f50]
#> 
#> [[27]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x2e66e50]
#> 
#> [[28]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x2e6f6d8]
#> 
#> [[29]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3b152d0]
#> 
#> [[30]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3b16290]
#> 
#> [[31]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpMyYvaN/file3d1f24aeb05f"),  <...>
#> env:  [frame 0x3b060f0]
#> 
#> [[32]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpMyYvaN/file3d1f24aeb05f"),  <...>
#> env:  [frame 0x3b06e80]
#> 
#> [[33]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3b01c58]
#> 
#> [[34]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3b02e10]
#> 
#> [[35]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x3b03158]
#> 
#> [[36]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3aff6e0]
#> 
#> [[37]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3affc90]
#> 
#> [[38]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3afffd8]
#> 
#> [[39]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x3b00320]
#> 
#> [[40]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3b00a90]
#> 
#> [[41]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# We can trim the first layer of those:
fn(1)

#> [[1]]
#> <frame 38> (37)
#> expr: fn(1)
#> env:  [local 0x8a92320]
#> 
#> [[2]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x8c34bc8]
#> 
#> [[3]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x8a92128]
#> 
#> [[4]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x8a95cf0]
#> 
#> [[5]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x8a95740]
#> 
#> [[6]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x8a951c8]
#> 
#> [[7]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x8a94e80]
#> 
#> [[8]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x8a94b00]
#> 
#> [[9]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x8a94438]
#> 
#> [[10]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x8a94208]
#> 
#> [[11]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x8a94080]
#> 
#> [[12]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x8a93fa0]
#> 
#> [[13]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x8b73398]
#> 
#> [[14]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x991d768]
#> 
#> [[15]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x991c070]
#> 
#> [[16]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x991e620]
#> 
#> [[17]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x8e421a0]
#> 
#> [[18]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x8311900]
#> 
#> [[19]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x27c4f98]
#> 
#> [[20]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x28039f8]
#> 
#> [[21]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x5ee6710]
#> 
#> [[22]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x5fba0b8]
#> 
#> [[23]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x73b9ed0]
#> 
#> [[24]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x5071f50]
#> 
#> [[25]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x2e66e50]
#> 
#> [[26]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x2e6f6d8]
#> 
#> [[27]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3b152d0]
#> 
#> [[28]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3b16290]
#> 
#> [[29]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpMyYvaN/file3d1f24aeb05f"),  <...>
#> env:  [frame 0x3b060f0]
#> 
#> [[30]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpMyYvaN/file3d1f24aeb05f"),  <...>
#> env:  [frame 0x3b06e80]
#> 
#> [[31]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3b01c58]
#> 
#> [[32]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3b02e10]
#> 
#> [[33]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x3b03158]
#> 
#> [[34]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3aff6e0]
#> 
#> [[35]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3affc90]
#> 
#> [[36]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3afffd8]
#> 
#> [[37]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x3b00320]
#> 
#> [[38]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3b00a90]
#> 
#> [[39]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# The outside intervening frames (at the fn() call site) are still
# returned, but can be trimmed as well:
identity(identity(fn(1)))

#> [[1]]
#> <frame 40> (37)
#> expr: fn(1)
#> env:  [local 0x75c9218]
#> 
#> [[2]]
#> <frame 39> (37)
#> expr: identity(fn(1))
#> env:  [local 0x75c9100]
#> 
#> [[3]]
#> <frame 38> (37)
#> expr: identity(identity(fn(1)))
#> env:  [local 0x75c8fe8]
#> 
#> [[4]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x8c34bc8]
#> 
#> [[5]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x75c8e28]
#> 
#> [[6]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x75c8bf8]
#> 
#> [[7]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x75c8648]
#> 
#> [[8]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x5b97cd8]
#> 
#> [[9]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x5b97990]
#> 
#> [[10]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x5b97648]
#> 
#> [[11]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x5b96f80]
#> 
#> [[12]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x5b96d50]
#> 
#> [[13]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x5b96bc8]
#> 
#> [[14]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x5b96ae8]
#> 
#> [[15]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x70a9890]
#> 
#> [[16]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x991d768]
#> 
#> [[17]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x991c070]
#> 
#> [[18]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x991e620]
#> 
#> [[19]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x8e421a0]
#> 
#> [[20]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x8311900]
#> 
#> [[21]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x27c4f98]
#> 
#> [[22]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x28039f8]
#> 
#> [[23]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x5ee6710]
#> 
#> [[24]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x5fba0b8]
#> 
#> [[25]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x73b9ed0]
#> 
#> [[26]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x5071f50]
#> 
#> [[27]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x2e66e50]
#> 
#> [[28]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x2e6f6d8]
#> 
#> [[29]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3b152d0]
#> 
#> [[30]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3b16290]
#> 
#> [[31]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpMyYvaN/file3d1f24aeb05f"),  <...>
#> env:  [frame 0x3b060f0]
#> 
#> [[32]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpMyYvaN/file3d1f24aeb05f"),  <...>
#> env:  [frame 0x3b06e80]
#> 
#> [[33]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3b01c58]
#> 
#> [[34]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3b02e10]
#> 
#> [[35]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x3b03158]
#> 
#> [[36]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3aff6e0]
#> 
#> [[37]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3affc90]
#> 
#> [[38]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3afffd8]
#> 
#> [[39]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x3b00320]
#> 
#> [[40]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3b00a90]
#> 
#> [[41]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(fn(2)))

#> [[1]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x8c34bc8]
#> 
#> [[2]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5e9d708]
#> 
#> [[3]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x5e9d4d8]
#> 
#> [[4]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x5e9ceb8]
#> 
#> [[5]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x5e9c8d0]
#> 
#> [[6]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x5e9c4e0]
#> 
#> [[7]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x5e9fee8]
#> 
#> [[8]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x5e9f778]
#> 
#> [[9]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x5e9f548]
#> 
#> [[10]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x5e9f3c0]
#> 
#> [[11]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x5e9f2a8]
#> 
#> [[12]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x5ebae10]
#> 
#> [[13]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x991d768]
#> 
#> [[14]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x991c070]
#> 
#> [[15]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x991e620]
#> 
#> [[16]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x8e421a0]
#> 
#> [[17]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x8311900]
#> 
#> [[18]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x27c4f98]
#> 
#> [[19]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x28039f8]
#> 
#> [[20]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x5ee6710]
#> 
#> [[21]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x5fba0b8]
#> 
#> [[22]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x73b9ed0]
#> 
#> [[23]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x5071f50]
#> 
#> [[24]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x2e66e50]
#> 
#> [[25]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x2e6f6d8]
#> 
#> [[26]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3b152d0]
#> 
#> [[27]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3b16290]
#> 
#> [[28]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpMyYvaN/file3d1f24aeb05f"),  <...>
#> env:  [frame 0x3b060f0]
#> 
#> [[29]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpMyYvaN/file3d1f24aeb05f"),  <...>
#> env:  [frame 0x3b06e80]
#> 
#> [[30]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3b01c58]
#> 
#> [[31]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3b02e10]
#> 
#> [[32]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x3b03158]
#> 
#> [[33]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3aff6e0]
#> 
#> [[34]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3affc90]
#> 
#> [[35]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3afffd8]
#> 
#> [[36]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x3b00320]
#> 
#> [[37]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3b00a90]
#> 
#> [[38]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g <- function(trim) identity(identity(fn(trim)))
g(2)

#> [[1]]
#> <frame 38> (37)
#> expr: g(2)
#> env:  [local 0x4016d48]
#> 
#> [[2]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x8c34bc8]
#> 
#> [[3]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x4068d60]
#> 
#> [[4]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x40685b8]
#> 
#> [[5]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x4067d30]
#> 
#> [[6]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x406d188]
#> 
#> [[7]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x406c628]
#> 
#> [[8]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x4070f58]
#> 
#> [[9]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x40aa678]
#> 
#> [[10]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x40b0ea0]
#> 
#> [[11]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x40afcb0]
#> 
#> [[12]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x40b2710]
#> 
#> [[13]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x415c8c8]
#> 
#> [[14]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x991d768]
#> 
#> [[15]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x991c070]
#> 
#> [[16]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x991e620]
#> 
#> [[17]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x8e421a0]
#> 
#> [[18]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x8311900]
#> 
#> [[19]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x27c4f98]
#> 
#> [[20]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x28039f8]
#> 
#> [[21]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x5ee6710]
#> 
#> [[22]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x5fba0b8]
#> 
#> [[23]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x73b9ed0]
#> 
#> [[24]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x5071f50]
#> 
#> [[25]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x2e66e50]
#> 
#> [[26]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x2e6f6d8]
#> 
#> [[27]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3b152d0]
#> 
#> [[28]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3b16290]
#> 
#> [[29]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpMyYvaN/file3d1f24aeb05f"),  <...>
#> env:  [frame 0x3b060f0]
#> 
#> [[30]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpMyYvaN/file3d1f24aeb05f"),  <...>
#> env:  [frame 0x3b06e80]
#> 
#> [[31]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3b01c58]
#> 
#> [[32]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3b02e10]
#> 
#> [[33]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x3b03158]
#> 
#> [[34]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3aff6e0]
#> 
#> [[35]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3affc90]
#> 
#> [[36]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3afffd8]
#> 
#> [[37]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x3b00320]
#> 
#> [[38]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3b00a90]
#> 
#> [[39]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g(3)

#> [[1]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x8c34bc8]
#> 
#> [[2]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x98fbe80]
#> 
#> [[3]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x98fbbe0]
#> 
#> [[4]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x98fb518]
#> 
#> [[5]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x98feb30]
#> 
#> [[6]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x98fe7e8]
#> 
#> [[7]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x98fe468]
#> 
#> [[8]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x98fdda0]
#> 
#> [[9]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x98fdb70]
#> 
#> [[10]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x98fd9e8]
#> 
#> [[11]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x98fd908]
#> 
#> [[12]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x8a3df80]
#> 
#> [[13]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x991d768]
#> 
#> [[14]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x991c070]
#> 
#> [[15]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x991e620]
#> 
#> [[16]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x8e421a0]
#> 
#> [[17]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x8311900]
#> 
#> [[18]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x27c4f98]
#> 
#> [[19]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x28039f8]
#> 
#> [[20]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x5ee6710]
#> 
#> [[21]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x5fba0b8]
#> 
#> [[22]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x73b9ed0]
#> 
#> [[23]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x5071f50]
#> 
#> [[24]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x2e66e50]
#> 
#> [[25]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x2e6f6d8]
#> 
#> [[26]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x3b152d0]
#> 
#> [[27]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3b16290]
#> 
#> [[28]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpMyYvaN/file3d1f24aeb05f"),  <...>
#> env:  [frame 0x3b060f0]
#> 
#> [[29]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpMyYvaN/file3d1f24aeb05f"),  <...>
#> env:  [frame 0x3b06e80]
#> 
#> [[30]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3b01c58]
#> 
#> [[31]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3b02e10]
#> 
#> [[32]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x3b03158]
#> 
#> [[33]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3aff6e0]
#> 
#> [[34]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3affc90]
#> 
#> [[35]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3afffd8]
#> 
#> [[36]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x3b00320]
#> 
#> [[37]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3b00a90]
#> 
#> [[38]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"