Call stack information

The eval_ and call_ families of functions provide a replacement for the base R functions prefixed with sys. (which are all about the context stack), as well as for parent.frame() (which is the only base R function for querying the call stack). The context stack includes all R-level evaluation contexts. It is linear in terms of execution history but due to lazy evaluation it is potentially nonlinear in terms of call history. The call stack history, on the other hand, is homogenous.

global_frame()

current_frame()

ctxt_frame(n = 1)

call_frame(n = 1, clean = TRUE)

ctxt_depth()

call_depth()

ctxt_stack(n = NULL, trim = 0)

call_stack(n = NULL, clean = TRUE)

Arguments

n
The number of frames to go back in the stack.

clean
Whether to post-process the call stack to clean non-standard frames. If TRUE, suboptimal call-stack entries by base::eval() will be cleaned up: the duplicate frame created by eval() is eliminated.

trim
The number of layers of intervening frames to trim off the stack. See stack_trim() and examples.

Details

ctxt_frame() and call_frame() return a frame object containing the following fields: expr and env (call expression and evaluation environment), pos and caller_pos (position of current frame in the context stack and position of the caller), and fun (function of the current frame). ctxt_stack() and call_stack() return a list of all context or call frames on the stack. Finally, ctxt_depth() and call_depth() report the current context position or the number of calling frames on the stack.

The base R functions take two sorts of arguments to indicate which frame to query: which and n. The n argument is straightforward: it's the number of frames to go down the stack, with n = 1 referring to the current context. The which argument is more complicated and changes meaning for values lower than 1. For the sake of consistency, the rlang functions all take the same kind of argument n. This argument has a single meaning (the number of frames to go down the stack) and cannot be lower than 1.

Note finally that parent.frame(1) corresponds to call_frame(2)$env, as n = 1 always refers to the current frame. This makes the _frame() and _stack() functions consistent: ctxt_frame(2) is the same as ctxt_stack()[[2]]. Also, ctxt_depth() returns one more frame than base::sys.nframe() because it counts the global frame. That is consistent with the _stack() functions which return the global frame as well. This way, call_stack(call_depth()) is the same as global_frame().

Life cycle

These functions are soft-deprecated and replaced by trace_back().

Examples

# Expressions within arguments count as contexts
identity(identity(ctxt_depth())) # returns 2

#> Warning: `ctxt_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 26

# But they are not part of the call stack because arguments are
# evaluated within the calling function (or the global environment
# if called at top level)
identity(identity(call_depth())) # returns 0

#> Warning: `call_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 13

# The context stacks includes all intervening execution frames. The
# call stack doesn't:
f <- function(x) identity(x)
f(f(ctxt_stack()))

#> [[1]]
#> <frame 27> (26)
#> expr: identity(x)
#> env:  [local 0x7f93b9ef4738]
#> 
#> [[2]]
#> <frame 26> (23)
#> expr: f(ctxt_stack())
#> env:  [local 0x7f93b9ef4850]
#> 
#> [[3]]
#> <frame 25> (24)
#> expr: identity(x)
#> env:  [local 0x7f93b9ef4968]
#> 
#> [[4]]
#> <frame 24> (23)
#> expr: f(f(ctxt_stack()))
#> env:  [local 0x7f93b9ef4a80]
#> 
#> [[5]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f93bc3590c0]
#> 
#> [[6]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f93b9ef4c78]
#> 
#> [[7]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f93b9ef4ea8]
#> 
#> [[8]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f93b9ef1548]
#> 
#> [[9]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f93b9ef1ac0]
#> 
#> [[10]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f93b9ef1e78]
#> 
#> [[11]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f93b9ef21c0]
#> 
#> [[12]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f93b9ef2888]
#> 
#> [[13]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f93b9ef2ab8]
#> 
#> [[14]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f93b9ef2c40]
#> 
#> [[15]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f93b9ef2d20]
#> 
#> [[16]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f93b9e447a8]
#> 
#> [[17]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f93b927a190]
#> 
#> [[18]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f93b9289da8]
#> 
#> [[19]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f93b549afe0]
#> 
#> [[20]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f93b66a9790]
#> 
#> [[21]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7f93b66a9020]
#> 
#> [[22]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f93b66cc8e8]
#> 
#> [[23]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f93b8b0c478]
#> 
#> [[24]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f93b826a178]
#> 
#> [[25]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f93b7271cd8]
#> 
#> [[26]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f93b72764b8]
#> 
#> [[27]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f93b5504038]
#> 
#> [[28]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(f(call_stack()))

#> Warning: `call_stack()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [[1]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f93bc3590c0]
#> 
#> [[2]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f93be1e3558]
#> 
#> [[3]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f93b927a190]
#> 
#> [[4]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f93b9289da8]
#> 
#> [[5]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f93b549afe0]
#> 
#> [[6]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f93b66a9790]
#> 
#> [[7]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f93b66cc8e8]
#> 
#> [[8]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f93b8b0c478]
#> 
#> [[9]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f93b826a178]
#> 
#> [[10]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f93b7271cd8]
#> 
#> [[11]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f93b72764b8]
#> 
#> [[12]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f93b5504038]
#> 
#> [[13]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

g <- function(cmd) cmd()
f(g(ctxt_stack))

#> [[1]]
#> <frame 26> (23)
#> expr: g(ctxt_stack)
#> env:  [local 0x7f93be8ba038]
#> 
#> [[2]]
#> <frame 25> (24)
#> expr: identity(x)
#> env:  [local 0x7f93be8b9f20]
#> 
#> [[3]]
#> <frame 24> (23)
#> expr: f(g(ctxt_stack))
#> env:  [local 0x7f93be8b9e08]
#> 
#> [[4]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f93bc3590c0]
#> 
#> [[5]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f93be8bfb58]
#> 
#> [[6]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f93be8bf928]
#> 
#> [[7]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f93be8bf378]
#> 
#> [[8]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f93be8bedc8]
#> 
#> [[9]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f93be8bea80]
#> 
#> [[10]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f93be8be738]
#> 
#> [[11]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f93be8be070]
#> 
#> [[12]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f93be8bde40]
#> 
#> [[13]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f93be8c3bc8]
#> 
#> [[14]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f93be8c3ae8]
#> 
#> [[15]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f93be95ee38]
#> 
#> [[16]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f93b927a190]
#> 
#> [[17]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f93b9289da8]
#> 
#> [[18]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f93b549afe0]
#> 
#> [[19]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f93b66a9790]
#> 
#> [[20]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7f93b66a9020]
#> 
#> [[21]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f93b66cc8e8]
#> 
#> [[22]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f93b8b0c478]
#> 
#> [[23]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f93b826a178]
#> 
#> [[24]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f93b7271cd8]
#> 
#> [[25]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f93b72764b8]
#> 
#> [[26]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f93b5504038]
#> 
#> [[27]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(g(call_stack))

#> [[1]]
#> <frame 26> (22)
#> expr: g(call_stack)
#> env:  [local 0x7f93bd0276d0]
#> 
#> [[2]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f93bc3590c0]
#> 
#> [[3]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f93bd0c7628]
#> 
#> [[4]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f93b927a190]
#> 
#> [[5]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f93b9289da8]
#> 
#> [[6]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f93b549afe0]
#> 
#> [[7]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f93b66a9790]
#> 
#> [[8]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f93b66cc8e8]
#> 
#> [[9]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f93b8b0c478]
#> 
#> [[10]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f93b826a178]
#> 
#> [[11]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f93b7271cd8]
#> 
#> [[12]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f93b72764b8]
#> 
#> [[13]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f93b5504038]
#> 
#> [[14]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

# The rlang _stack() functions return a list of frame
# objects. Use purrr::transpose() or index a field with
# purrr::map()'s to extract a particular field from a stack:

# stack <- f(f(call_stack()))
# purrr::map(stack, "env")
# purrr::transpose(stack)$expr

# current_frame() is an alias for ctxt_frame(1)
fn <- function() list(current = current_frame(), first = ctxt_frame(1))
fn()

#> Warning: `current_frame()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> $current
#> <frame 24> (23)
#> expr: fn()
#> env:  [local 0x7f93bbaad950]
#> 
#> $first
#> <frame 24> (23)
#> expr: fn()
#> env:  [local 0x7f93bbaad950]
#> 

# While current_frame() is the top of the stack, global_frame() is
# the bottom:
fn <- function() {
  n <- ctxt_depth()
  ctxt_frame(n)
}
identical(fn(), global_frame())

#> [1] TRUE

# ctxt_stack() returns a stack with all intervening frames. You can
# trim layers of intervening frames with the trim argument:
identity(identity(ctxt_stack()))

#> [[1]]
#> <frame 25> (23)
#> expr: identity(ctxt_stack())
#> env:  [local 0x7f93b954a2a8]
#> 
#> [[2]]
#> <frame 24> (23)
#> expr: identity(identity(ctxt_stack()))
#> env:  [local 0x7f93b954a158]
#> 
#> [[3]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f93bc3590c0]
#> 
#> [[4]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f93b954dea8]
#> 
#> [[5]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f93b954dc78]
#> 
#> [[6]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f93b954d6c8]
#> 
#> [[7]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f93b954d150]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f93b954ce08]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f93b954cac0]
#> 
#> [[10]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f93b954c3c0]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f93b954c190]
#> 
#> [[12]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f93b954c008]
#> 
#> [[13]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f93b9551e38]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f93b9583430]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f93b927a190]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f93b9289da8]
#> 
#> [[17]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f93b549afe0]
#> 
#> [[18]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f93b66a9790]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7f93b66a9020]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f93b66cc8e8]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f93b8b0c478]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f93b826a178]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f93b7271cd8]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f93b72764b8]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f93b5504038]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(ctxt_stack(trim = 1)))

#> [[1]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f93bc3590c0]
#> 
#> [[2]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f93bb6afdc0]
#> 
#> [[3]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f93bb6afa40]
#> 
#> [[4]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f93bb6af458]
#> 
#> [[5]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f93bb6aedc8]
#> 
#> [[6]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f93bb6b4990]
#> 
#> [[7]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f93bb6b4648]
#> 
#> [[8]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f93bb6b3f80]
#> 
#> [[9]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f93bb6b3d50]
#> 
#> [[10]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f93bb6b3bc8]
#> 
#> [[11]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f93bb6b3ae8]
#> 
#> [[12]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f93bb6c5d18]
#> 
#> [[13]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f93b927a190]
#> 
#> [[14]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f93b9289da8]
#> 
#> [[15]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f93b549afe0]
#> 
#> [[16]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f93b66a9790]
#> 
#> [[17]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7f93b66a9020]
#> 
#> [[18]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f93b66cc8e8]
#> 
#> [[19]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f93b8b0c478]
#> 
#> [[20]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f93b826a178]
#> 
#> [[21]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f93b7271cd8]
#> 
#> [[22]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f93b72764b8]
#> 
#> [[23]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f93b5504038]
#> 
#> [[24]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# ctxt_stack() is called within fn() with intervening frames:
fn <- function(trim) identity(identity(ctxt_stack(trim = trim)))
fn(0)

#> [[1]]
#> <frame 26> (24)
#> expr: identity(ctxt_stack(trim = trim))
#> env:  [local 0x7f93bb3139b0]
#> 
#> [[2]]
#> <frame 25> (24)
#> expr: identity(identity(ctxt_stack(trim = trim)))
#> env:  [local 0x7f93bb313898]
#> 
#> [[3]]
#> <frame 24> (23)
#> expr: fn(0)
#> env:  [local 0x7f93bb313780]
#> 
#> [[4]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f93bc3590c0]
#> 
#> [[5]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f93bb3135c0]
#> 
#> [[6]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f93bb313358]
#> 
#> [[7]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f93bb312d70]
#> 
#> [[8]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f93bb3127f8]
#> 
#> [[9]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f93bb3124b0]
#> 
#> [[10]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f93bb316040]
#> 
#> [[11]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f93bb315978]
#> 
#> [[12]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f93bb315748]
#> 
#> [[13]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f93bb315588]
#> 
#> [[14]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f93bb3154a8]
#> 
#> [[15]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f93bb32f400]
#> 
#> [[16]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f93b927a190]
#> 
#> [[17]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f93b9289da8]
#> 
#> [[18]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f93b549afe0]
#> 
#> [[19]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f93b66a9790]
#> 
#> [[20]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7f93b66a9020]
#> 
#> [[21]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f93b66cc8e8]
#> 
#> [[22]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f93b8b0c478]
#> 
#> [[23]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f93b826a178]
#> 
#> [[24]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f93b7271cd8]
#> 
#> [[25]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f93b72764b8]
#> 
#> [[26]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f93b5504038]
#> 
#> [[27]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# We can trim the first layer of those:
fn(1)

#> [[1]]
#> <frame 24> (23)
#> expr: fn(1)
#> env:  [local 0x7f93baa6d4e0]
#> 
#> [[2]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f93bc3590c0]
#> 
#> [[3]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f93baa6d320]
#> 
#> [[4]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f93baa6d0f0]
#> 
#> [[5]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f93baa6cb40]
#> 
#> [[6]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f93baa6c5c8]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f93baa72120]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f93baa71dd8]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f93baa716d8]
#> 
#> [[10]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f93baa714a8]
#> 
#> [[11]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f93baa71320]
#> 
#> [[12]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f93baa71240]
#> 
#> [[13]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f93baa85208]
#> 
#> [[14]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f93b927a190]
#> 
#> [[15]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f93b9289da8]
#> 
#> [[16]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f93b549afe0]
#> 
#> [[17]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f93b66a9790]
#> 
#> [[18]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7f93b66a9020]
#> 
#> [[19]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f93b66cc8e8]
#> 
#> [[20]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f93b8b0c478]
#> 
#> [[21]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f93b826a178]
#> 
#> [[22]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f93b7271cd8]
#> 
#> [[23]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f93b72764b8]
#> 
#> [[24]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f93b5504038]
#> 
#> [[25]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# The outside intervening frames (at the fn() call site) are still
# returned, but can be trimmed as well:
identity(identity(fn(1)))

#> [[1]]
#> <frame 26> (23)
#> expr: fn(1)
#> env:  [local 0x7f93ba8c7e10]
#> 
#> [[2]]
#> <frame 25> (23)
#> expr: identity(fn(1))
#> env:  [local 0x7f93ba8c7cf8]
#> 
#> [[3]]
#> <frame 24> (23)
#> expr: identity(identity(fn(1)))
#> env:  [local 0x7f93ba8c7be0]
#> 
#> [[4]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f93bc3590c0]
#> 
#> [[5]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f93ba8c7a20]
#> 
#> [[6]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f93ba8c77f0]
#> 
#> [[7]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f93ba8c7208]
#> 
#> [[8]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f93ba8c6c90]
#> 
#> [[9]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f93ba8c6948]
#> 
#> [[10]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f93ba8c6600]
#> 
#> [[11]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f93ba8cbe48]
#> 
#> [[12]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f93ba8cbbe0]
#> 
#> [[13]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f93ba8cba58]
#> 
#> [[14]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f93ba8cb978]
#> 
#> [[15]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f93ba8ecde0]
#> 
#> [[16]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f93b927a190]
#> 
#> [[17]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f93b9289da8]
#> 
#> [[18]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f93b549afe0]
#> 
#> [[19]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f93b66a9790]
#> 
#> [[20]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7f93b66a9020]
#> 
#> [[21]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f93b66cc8e8]
#> 
#> [[22]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f93b8b0c478]
#> 
#> [[23]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f93b826a178]
#> 
#> [[24]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f93b7271cd8]
#> 
#> [[25]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f93b72764b8]
#> 
#> [[26]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f93b5504038]
#> 
#> [[27]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(fn(2)))

#> [[1]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f93bc3590c0]
#> 
#> [[2]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f93b6583b80]
#> 
#> [[3]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f93b6583918]
#> 
#> [[4]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f93b6583330]
#> 
#> [[5]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f93b6582db8]
#> 
#> [[6]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f93b6582a70]
#> 
#> [[7]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f93b6586638]
#> 
#> [[8]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f93b6585f70]
#> 
#> [[9]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f93b6585d08]
#> 
#> [[10]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f93b6585b80]
#> 
#> [[11]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f93b6585aa0]
#> 
#> [[12]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f93b659aa70]
#> 
#> [[13]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f93b927a190]
#> 
#> [[14]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f93b9289da8]
#> 
#> [[15]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f93b549afe0]
#> 
#> [[16]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f93b66a9790]
#> 
#> [[17]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7f93b66a9020]
#> 
#> [[18]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f93b66cc8e8]
#> 
#> [[19]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f93b8b0c478]
#> 
#> [[20]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f93b826a178]
#> 
#> [[21]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f93b7271cd8]
#> 
#> [[22]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f93b72764b8]
#> 
#> [[23]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f93b5504038]
#> 
#> [[24]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g <- function(trim) identity(identity(fn(trim)))
g(2)

#> [[1]]
#> <frame 24> (23)
#> expr: g(2)
#> env:  [local 0x7f93b6cc1108]
#> 
#> [[2]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f93bc3590c0]
#> 
#> [[3]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f93b6cc0f48]
#> 
#> [[4]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f93b6cc0d18]
#> 
#> [[5]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f93b6cc0378]
#> 
#> [[6]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f93b6cbfe00]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f93b6cc39c8]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f93b6cc3610]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f93b6cc5990]
#> 
#> [[10]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f93b6cc50d0]
#> 
#> [[11]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f93b6cc4f48]
#> 
#> [[12]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f93b6cc4e30]
#> 
#> [[13]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f93b6ce7028]
#> 
#> [[14]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f93b927a190]
#> 
#> [[15]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f93b9289da8]
#> 
#> [[16]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f93b549afe0]
#> 
#> [[17]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f93b66a9790]
#> 
#> [[18]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7f93b66a9020]
#> 
#> [[19]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f93b66cc8e8]
#> 
#> [[20]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f93b8b0c478]
#> 
#> [[21]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f93b826a178]
#> 
#> [[22]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f93b7271cd8]
#> 
#> [[23]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f93b72764b8]
#> 
#> [[24]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f93b5504038]
#> 
#> [[25]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g(3)

#> [[1]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f93bc3590c0]
#> 
#> [[2]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f93b993a780]
#> 
#> [[3]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f93b993d8c8]
#> 
#> [[4]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f93b993d2e0]
#> 
#> [[5]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f93b993fcf0]
#> 
#> [[6]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f93b993f970]
#> 
#> [[7]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f93b993f628]
#> 
#> [[8]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f93b993ee48]
#> 
#> [[9]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f93b993e9e8]
#> 
#> [[10]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f93b993e828]
#> 
#> [[11]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f93b993e748]
#> 
#> [[12]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f93b99c69e8]
#> 
#> [[13]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f93b927a190]
#> 
#> [[14]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f93b9289da8]
#> 
#> [[15]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f93b549afe0]
#> 
#> [[16]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f93b66a9790]
#> 
#> [[17]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7f93b66a9020]
#> 
#> [[18]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f93b66cc8e8]
#> 
#> [[19]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f93b8b0c478]
#> 
#> [[20]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f93b826a178]
#> 
#> [[21]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f93b7271cd8]
#> 
#> [[22]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f93b72764b8]
#> 
#> [[23]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f93b5504038]
#> 
#> [[24]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"