Call stack information

The eval_ and call_ families of functions provide a replacement for the base R functions prefixed with sys. (which are all about the context stack), as well as for parent.frame() (which is the only base R function for querying the call stack). The context stack includes all R-level evaluation contexts. It is linear in terms of execution history but due to lazy evaluation it is potentially nonlinear in terms of call history. The call stack history, on the other hand, is homogenous.

global_frame()

current_frame()

ctxt_frame(n = 1)

call_frame(n = 1, clean = TRUE)

ctxt_depth()

call_depth()

ctxt_stack(n = NULL, trim = 0)

call_stack(n = NULL, clean = TRUE)

Arguments

n
The number of frames to go back in the stack.

clean
Whether to post-process the call stack to clean non-standard frames. If TRUE, suboptimal call-stack entries by base::eval() will be cleaned up: the duplicate frame created by eval() is eliminated.

trim
The number of layers of intervening frames to trim off the stack. See stack_trim() and examples.

Details

ctxt_frame() and call_frame() return a frame object containing the following fields: expr and env (call expression and evaluation environment), pos and caller_pos (position of current frame in the context stack and position of the caller), and fun (function of the current frame). ctxt_stack() and call_stack() return a list of all context or call frames on the stack. Finally, ctxt_depth() and call_depth() report the current context position or the number of calling frames on the stack.

The base R functions take two sorts of arguments to indicate which frame to query: which and n. The n argument is straightforward: it's the number of frames to go down the stack, with n = 1 referring to the current context. The which argument is more complicated and changes meaning for values lower than 1. For the sake of consistency, the rlang functions all take the same kind of argument n. This argument has a single meaning (the number of frames to go down the stack) and cannot be lower than 1.

Note finally that parent.frame(1) corresponds to call_frame(2)$env, as n = 1 always refers to the current frame. This makes the _frame() and _stack() functions consistent: ctxt_frame(2) is the same as ctxt_stack()[[2]]. Also, ctxt_depth() returns one more frame than base::sys.nframe() because it counts the global frame. That is consistent with the _stack() functions which return the global frame as well. This way, call_stack(call_depth()) is the same as global_frame().

Life cycle

These functions are soft-deprecated and replaced by trace_back().

Examples

# Expressions within arguments count as contexts
identity(identity(ctxt_depth())) # returns 2

#> Warning: `ctxt_depth()` is soft-deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 40

# But they are not part of the call stack because arguments are
# evaluated within the calling function (or the global environment
# if called at top level)
identity(identity(call_depth())) # returns 0

#> Warning: `call_depth()` is soft-deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 14

# The context stacks includes all intervening execution frames. The
# call stack doesn't:
f <- function(x) identity(x)
f(f(ctxt_stack()))

#> [[1]]
#> <frame 41> (40)
#> expr: identity(x)
#> env:  [local 0x4b18450]
#> 
#> [[2]]
#> <frame 40> (37)
#> expr: f(ctxt_stack())
#> env:  [local 0x4b18338]
#> 
#> [[3]]
#> <frame 39> (38)
#> expr: identity(x)
#> env:  [local 0x4b18220]
#> 
#> [[4]]
#> <frame 38> (37)
#> expr: f(f(ctxt_stack()))
#> env:  [local 0x4b18108]
#> 
#> [[5]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5064ef8]
#> 
#> [[6]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x4b17f48]
#> 
#> [[7]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x4b17ca8]
#> 
#> [[8]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x4b17618]
#> 
#> [[9]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x4b1cdf0]
#> 
#> [[10]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x4b1caa8]
#> 
#> [[11]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x4b1c6b8]
#> 
#> [[12]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x4b1b768]
#> 
#> [[13]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x4b20f78]
#> 
#> [[14]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x4b20db8]
#> 
#> [[15]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x4b20ca0]
#> 
#> [[16]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x4b58cb0]
#> 
#> [[17]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x565efe0]
#> 
#> [[18]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x6a1c808]
#> 
#> [[19]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x6261910]
#> 
#> [[20]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x753c7d0]
#> 
#> [[21]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x6d87af0]
#> 
#> [[22]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6ffcef0]
#> 
#> [[23]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6ffc4e0]
#> 
#> [[24]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7056778]
#> 
#> [[25]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7835f50]
#> 
#> [[26]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x64585d8]
#> 
#> [[27]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x41f8ee0]
#> 
#> [[28]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0xc17748]
#> 
#> [[29]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x740438]
#> 
#> [[30]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x244c290]
#> 
#> [[31]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x244d250]
#> 
#> [[32]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpCL1qTH/file3b624f4c2b38"),  <...>
#> env:  [frame 0x243d0b0]
#> 
#> [[33]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpCL1qTH/file3b624f4c2b38"),  <...>
#> env:  [frame 0x243de40]
#> 
#> [[34]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x2438c18]
#> 
#> [[35]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x2439dd0]
#> 
#> [[36]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x243a118]
#> 
#> [[37]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x24366a0]
#> 
#> [[38]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x2436c50]
#> 
#> [[39]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x2436f98]
#> 
#> [[40]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x24372e0]
#> 
#> [[41]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x2437a50]
#> 
#> [[42]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(f(call_stack()))

#> Warning: `call_stack()` is soft-deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [[1]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5064ef8]
#> 
#> [[2]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x3b78d28]
#> 
#> [[3]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x565efe0]
#> 
#> [[4]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x753c7d0]
#> 
#> [[5]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x6d87af0]
#> 
#> [[6]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6ffcef0]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7056778]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7835f50]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x64585d8]
#> 
#> [[10]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x41f8ee0]
#> 
#> [[11]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0xc17748]
#> 
#> [[12]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x740438]
#> 
#> [[13]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x244c290]
#> 
#> [[14]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

g <- function(cmd) cmd()
f(g(ctxt_stack))

#> [[1]]
#> <frame 40> (37)
#> expr: g(ctxt_stack)
#> env:  [local 0x2396920]
#> 
#> [[2]]
#> <frame 39> (38)
#> expr: identity(x)
#> env:  [local 0x2396450]
#> 
#> [[3]]
#> <frame 38> (37)
#> expr: f(g(ctxt_stack))
#> env:  [local 0x2399bb8]
#> 
#> [[4]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5064ef8]
#> 
#> [[5]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x2399480]
#> 
#> [[6]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x2398c30]
#> 
#> [[7]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x239aa70]
#> 
#> [[8]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x239dbf0]
#> 
#> [[9]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x239cf78]
#> 
#> [[10]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x239be68]
#> 
#> [[11]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x239eaa8]
#> 
#> [[12]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x239e418]
#> 
#> [[13]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x23a15d0]
#> 
#> [[14]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x23a1410]
#> 
#> [[15]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x24992d8]
#> 
#> [[16]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x565efe0]
#> 
#> [[17]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x6a1c808]
#> 
#> [[18]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x6261910]
#> 
#> [[19]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x753c7d0]
#> 
#> [[20]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x6d87af0]
#> 
#> [[21]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6ffcef0]
#> 
#> [[22]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6ffc4e0]
#> 
#> [[23]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7056778]
#> 
#> [[24]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7835f50]
#> 
#> [[25]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x64585d8]
#> 
#> [[26]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x41f8ee0]
#> 
#> [[27]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0xc17748]
#> 
#> [[28]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x740438]
#> 
#> [[29]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x244c290]
#> 
#> [[30]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x244d250]
#> 
#> [[31]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpCL1qTH/file3b624f4c2b38"),  <...>
#> env:  [frame 0x243d0b0]
#> 
#> [[32]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpCL1qTH/file3b624f4c2b38"),  <...>
#> env:  [frame 0x243de40]
#> 
#> [[33]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x2438c18]
#> 
#> [[34]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x2439dd0]
#> 
#> [[35]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x243a118]
#> 
#> [[36]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x24366a0]
#> 
#> [[37]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x2436c50]
#> 
#> [[38]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x2436f98]
#> 
#> [[39]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x24372e0]
#> 
#> [[40]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x2437a50]
#> 
#> [[41]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(g(call_stack))

#> [[1]]
#> <frame 40> (36)
#> expr: g(call_stack)
#> env:  [local 0x6839f70]
#> 
#> [[2]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5064ef8]
#> 
#> [[3]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x69a6b18]
#> 
#> [[4]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x565efe0]
#> 
#> [[5]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x753c7d0]
#> 
#> [[6]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x6d87af0]
#> 
#> [[7]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6ffcef0]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7056778]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7835f50]
#> 
#> [[10]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x64585d8]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x41f8ee0]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0xc17748]
#> 
#> [[13]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x740438]
#> 
#> [[14]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x244c290]
#> 
#> [[15]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

# The rlang _stack() functions return a list of frame
# objects. Use purrr::transpose() or index a field with
# purrr::map()'s to extract a particular field from a stack:

# stack <- f(f(call_stack()))
# purrr::map(stack, "env")
# purrr::transpose(stack)$expr

# current_frame() is an alias for ctxt_frame(1)
fn <- function() list(current = current_frame(), first = ctxt_frame(1))
fn()

#> Warning: `current_frame()` is soft-deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> $current
#> <frame 38> (37)
#> expr: fn()
#> env:  [local 0x71e49e0]
#> 
#> $first
#> <frame 38> (37)
#> expr: fn()
#> env:  [local 0x71e49e0]
#> 

# While current_frame() is the top of the stack, global_frame() is
# the bottom:
fn <- function() {
  n <- ctxt_depth()
  ctxt_frame(n)
}
identical(fn(), global_frame())

#> [1] TRUE

# ctxt_stack() returns a stack with all intervening frames. You can
# trim layers of intervening frames with the trim argument:
identity(identity(ctxt_stack()))

#> [[1]]
#> <frame 39> (37)
#> expr: identity(ctxt_stack())
#> env:  [local 0x7175b00]
#> 
#> [[2]]
#> <frame 38> (37)
#> expr: identity(identity(ctxt_stack()))
#> env:  [local 0x71759e8]
#> 
#> [[3]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5064ef8]
#> 
#> [[4]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7175828]
#> 
#> [[5]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x71755c0]
#> 
#> [[6]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7174fa0]
#> 
#> [[7]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7174a28]
#> 
#> [[8]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7432640]
#> 
#> [[9]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x74322f8]
#> 
#> [[10]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7431c30]
#> 
#> [[11]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7431a00]
#> 
#> [[12]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7431878]
#> 
#> [[13]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7431798]
#> 
#> [[14]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x6a93bd8]
#> 
#> [[15]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x565efe0]
#> 
#> [[16]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x6a1c808]
#> 
#> [[17]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x6261910]
#> 
#> [[18]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x753c7d0]
#> 
#> [[19]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x6d87af0]
#> 
#> [[20]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6ffcef0]
#> 
#> [[21]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6ffc4e0]
#> 
#> [[22]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7056778]
#> 
#> [[23]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7835f50]
#> 
#> [[24]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x64585d8]
#> 
#> [[25]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x41f8ee0]
#> 
#> [[26]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0xc17748]
#> 
#> [[27]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x740438]
#> 
#> [[28]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x244c290]
#> 
#> [[29]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x244d250]
#> 
#> [[30]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpCL1qTH/file3b624f4c2b38"),  <...>
#> env:  [frame 0x243d0b0]
#> 
#> [[31]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpCL1qTH/file3b624f4c2b38"),  <...>
#> env:  [frame 0x243de40]
#> 
#> [[32]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x2438c18]
#> 
#> [[33]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x2439dd0]
#> 
#> [[34]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x243a118]
#> 
#> [[35]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x24366a0]
#> 
#> [[36]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x2436c50]
#> 
#> [[37]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x2436f98]
#> 
#> [[38]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x24372e0]
#> 
#> [[39]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x2437a50]
#> 
#> [[40]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(ctxt_stack(trim = 1)))

#> [[1]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5064ef8]
#> 
#> [[2]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x547a578]
#> 
#> [[3]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x547a268]
#> 
#> [[4]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x5436d98]
#> 
#> [[5]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x54367b0]
#> 
#> [[6]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x54363c0]
#> 
#> [[7]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x5435d68]
#> 
#> [[8]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x5435630]
#> 
#> [[9]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x54353c8]
#> 
#> [[10]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x54350b8]
#> 
#> [[11]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x5422840]
#> 
#> [[12]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x5428678]
#> 
#> [[13]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x565efe0]
#> 
#> [[14]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x6a1c808]
#> 
#> [[15]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x6261910]
#> 
#> [[16]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x753c7d0]
#> 
#> [[17]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x6d87af0]
#> 
#> [[18]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6ffcef0]
#> 
#> [[19]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6ffc4e0]
#> 
#> [[20]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7056778]
#> 
#> [[21]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7835f50]
#> 
#> [[22]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x64585d8]
#> 
#> [[23]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x41f8ee0]
#> 
#> [[24]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0xc17748]
#> 
#> [[25]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x740438]
#> 
#> [[26]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x244c290]
#> 
#> [[27]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x244d250]
#> 
#> [[28]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpCL1qTH/file3b624f4c2b38"),  <...>
#> env:  [frame 0x243d0b0]
#> 
#> [[29]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpCL1qTH/file3b624f4c2b38"),  <...>
#> env:  [frame 0x243de40]
#> 
#> [[30]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x2438c18]
#> 
#> [[31]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x2439dd0]
#> 
#> [[32]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x243a118]
#> 
#> [[33]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x24366a0]
#> 
#> [[34]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x2436c50]
#> 
#> [[35]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x2436f98]
#> 
#> [[36]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x24372e0]
#> 
#> [[37]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x2437a50]
#> 
#> [[38]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# ctxt_stack() is called within fn() with intervening frames:
fn <- function(trim) identity(identity(ctxt_stack(trim = trim)))
fn(0)

#> [[1]]
#> <frame 40> (38)
#> expr: identity(ctxt_stack(trim = trim))
#> env:  [local 0x4597410]
#> 
#> [[2]]
#> <frame 39> (38)
#> expr: identity(identity(ctxt_stack(trim = trim)))
#> env:  [local 0x45972f8]
#> 
#> [[3]]
#> <frame 38> (37)
#> expr: fn(0)
#> env:  [local 0x4597170]
#> 
#> [[4]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5064ef8]
#> 
#> [[5]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x4596d10]
#> 
#> [[6]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x459a910]
#> 
#> [[7]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x4599138]
#> 
#> [[8]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x4598b50]
#> 
#> [[9]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x459c638]
#> 
#> [[10]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x459c2f0]
#> 
#> [[11]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x459bc28]
#> 
#> [[12]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x459b9f8]
#> 
#> [[13]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x459b800]
#> 
#> [[14]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x459b720]
#> 
#> [[15]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x45bde30]
#> 
#> [[16]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x565efe0]
#> 
#> [[17]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x6a1c808]
#> 
#> [[18]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x6261910]
#> 
#> [[19]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x753c7d0]
#> 
#> [[20]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x6d87af0]
#> 
#> [[21]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6ffcef0]
#> 
#> [[22]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6ffc4e0]
#> 
#> [[23]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7056778]
#> 
#> [[24]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7835f50]
#> 
#> [[25]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x64585d8]
#> 
#> [[26]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x41f8ee0]
#> 
#> [[27]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0xc17748]
#> 
#> [[28]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x740438]
#> 
#> [[29]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x244c290]
#> 
#> [[30]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x244d250]
#> 
#> [[31]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpCL1qTH/file3b624f4c2b38"),  <...>
#> env:  [frame 0x243d0b0]
#> 
#> [[32]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpCL1qTH/file3b624f4c2b38"),  <...>
#> env:  [frame 0x243de40]
#> 
#> [[33]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x2438c18]
#> 
#> [[34]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x2439dd0]
#> 
#> [[35]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x243a118]
#> 
#> [[36]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x24366a0]
#> 
#> [[37]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x2436c50]
#> 
#> [[38]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x2436f98]
#> 
#> [[39]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x24372e0]
#> 
#> [[40]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x2437a50]
#> 
#> [[41]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# We can trim the first layer of those:
fn(1)

#> [[1]]
#> <frame 38> (37)
#> expr: fn(1)
#> env:  [local 0x247daa0]
#> 
#> [[2]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5064ef8]
#> 
#> [[3]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x247d410]
#> 
#> [[4]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x247cdf0]
#> 
#> [[5]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x24870d0]
#> 
#> [[6]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x247fc98]
#> 
#> [[7]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x247f170]
#> 
#> [[8]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x2481720]
#> 
#> [[9]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x24805d8]
#> 
#> [[10]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x248a740]
#> 
#> [[11]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x248a468]
#> 
#> [[12]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x248a0b0]
#> 
#> [[13]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x24e6498]
#> 
#> [[14]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x565efe0]
#> 
#> [[15]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x6a1c808]
#> 
#> [[16]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x6261910]
#> 
#> [[17]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x753c7d0]
#> 
#> [[18]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x6d87af0]
#> 
#> [[19]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6ffcef0]
#> 
#> [[20]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6ffc4e0]
#> 
#> [[21]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7056778]
#> 
#> [[22]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7835f50]
#> 
#> [[23]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x64585d8]
#> 
#> [[24]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x41f8ee0]
#> 
#> [[25]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0xc17748]
#> 
#> [[26]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x740438]
#> 
#> [[27]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x244c290]
#> 
#> [[28]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x244d250]
#> 
#> [[29]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpCL1qTH/file3b624f4c2b38"),  <...>
#> env:  [frame 0x243d0b0]
#> 
#> [[30]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpCL1qTH/file3b624f4c2b38"),  <...>
#> env:  [frame 0x243de40]
#> 
#> [[31]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x2438c18]
#> 
#> [[32]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x2439dd0]
#> 
#> [[33]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x243a118]
#> 
#> [[34]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x24366a0]
#> 
#> [[35]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x2436c50]
#> 
#> [[36]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x2436f98]
#> 
#> [[37]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x24372e0]
#> 
#> [[38]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x2437a50]
#> 
#> [[39]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# The outside intervening frames (at the fn() call site) are still
# returned, but can be trimmed as well:
identity(identity(fn(1)))

#> [[1]]
#> <frame 40> (37)
#> expr: fn(1)
#> env:  [local 0x741d088]
#> 
#> [[2]]
#> <frame 39> (37)
#> expr: identity(fn(1))
#> env:  [local 0x741d1a0]
#> 
#> [[3]]
#> <frame 38> (37)
#> expr: identity(identity(fn(1)))
#> env:  [local 0x71d5718]
#> 
#> [[4]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5064ef8]
#> 
#> [[5]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x71d58d8]
#> 
#> [[6]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x71d5b08]
#> 
#> [[7]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x71d60b8]
#> 
#> [[8]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x71d6630]
#> 
#> [[9]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x71d6978]
#> 
#> [[10]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x71d6cc0]
#> 
#> [[11]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x71d7388]
#> 
#> [[12]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x71d75b8]
#> 
#> [[13]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7889580]
#> 
#> [[14]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7889660]
#> 
#> [[15]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7114fc8]
#> 
#> [[16]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x565efe0]
#> 
#> [[17]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x6a1c808]
#> 
#> [[18]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x6261910]
#> 
#> [[19]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x753c7d0]
#> 
#> [[20]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x6d87af0]
#> 
#> [[21]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6ffcef0]
#> 
#> [[22]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6ffc4e0]
#> 
#> [[23]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7056778]
#> 
#> [[24]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7835f50]
#> 
#> [[25]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x64585d8]
#> 
#> [[26]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x41f8ee0]
#> 
#> [[27]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0xc17748]
#> 
#> [[28]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x740438]
#> 
#> [[29]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x244c290]
#> 
#> [[30]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x244d250]
#> 
#> [[31]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpCL1qTH/file3b624f4c2b38"),  <...>
#> env:  [frame 0x243d0b0]
#> 
#> [[32]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpCL1qTH/file3b624f4c2b38"),  <...>
#> env:  [frame 0x243de40]
#> 
#> [[33]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x2438c18]
#> 
#> [[34]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x2439dd0]
#> 
#> [[35]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x243a118]
#> 
#> [[36]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x24366a0]
#> 
#> [[37]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x2436c50]
#> 
#> [[38]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x2436f98]
#> 
#> [[39]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x24372e0]
#> 
#> [[40]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x2437a50]
#> 
#> [[41]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(fn(2)))

#> [[1]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5064ef8]
#> 
#> [[2]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x68cc0d0]
#> 
#> [[3]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x68cc300]
#> 
#> [[4]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x8280e80]
#> 
#> [[5]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x82813f8]
#> 
#> [[6]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x8281740]
#> 
#> [[7]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x8281a88]
#> 
#> [[8]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x8282150]
#> 
#> [[9]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x8282380]
#> 
#> [[10]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x8282508]
#> 
#> [[11]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x82825e8]
#> 
#> [[12]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x8076cc0]
#> 
#> [[13]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x565efe0]
#> 
#> [[14]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x6a1c808]
#> 
#> [[15]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x6261910]
#> 
#> [[16]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x753c7d0]
#> 
#> [[17]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x6d87af0]
#> 
#> [[18]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6ffcef0]
#> 
#> [[19]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6ffc4e0]
#> 
#> [[20]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7056778]
#> 
#> [[21]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7835f50]
#> 
#> [[22]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x64585d8]
#> 
#> [[23]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x41f8ee0]
#> 
#> [[24]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0xc17748]
#> 
#> [[25]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x740438]
#> 
#> [[26]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x244c290]
#> 
#> [[27]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x244d250]
#> 
#> [[28]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpCL1qTH/file3b624f4c2b38"),  <...>
#> env:  [frame 0x243d0b0]
#> 
#> [[29]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpCL1qTH/file3b624f4c2b38"),  <...>
#> env:  [frame 0x243de40]
#> 
#> [[30]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x2438c18]
#> 
#> [[31]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x2439dd0]
#> 
#> [[32]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x243a118]
#> 
#> [[33]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x24366a0]
#> 
#> [[34]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x2436c50]
#> 
#> [[35]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x2436f98]
#> 
#> [[36]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x24372e0]
#> 
#> [[37]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x2437a50]
#> 
#> [[38]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g <- function(trim) identity(identity(fn(trim)))
g(2)

#> [[1]]
#> <frame 38> (37)
#> expr: g(2)
#> env:  [local 0x75b20a8]
#> 
#> [[2]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5064ef8]
#> 
#> [[3]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x75b1ee8]
#> 
#> [[4]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x75c7308]
#> 
#> [[5]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x75c6d20]
#> 
#> [[6]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x75c67a8]
#> 
#> [[7]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x75c6428]
#> 
#> [[8]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x75c60a8]
#> 
#> [[9]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x75c59a8]
#> 
#> [[10]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x75c5778]
#> 
#> [[11]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x75c55f0]
#> 
#> [[12]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x75c5510]
#> 
#> [[13]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7687d60]
#> 
#> [[14]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x565efe0]
#> 
#> [[15]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x6a1c808]
#> 
#> [[16]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x6261910]
#> 
#> [[17]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x753c7d0]
#> 
#> [[18]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x6d87af0]
#> 
#> [[19]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6ffcef0]
#> 
#> [[20]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6ffc4e0]
#> 
#> [[21]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7056778]
#> 
#> [[22]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7835f50]
#> 
#> [[23]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x64585d8]
#> 
#> [[24]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x41f8ee0]
#> 
#> [[25]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0xc17748]
#> 
#> [[26]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x740438]
#> 
#> [[27]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x244c290]
#> 
#> [[28]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x244d250]
#> 
#> [[29]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpCL1qTH/file3b624f4c2b38"),  <...>
#> env:  [frame 0x243d0b0]
#> 
#> [[30]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpCL1qTH/file3b624f4c2b38"),  <...>
#> env:  [frame 0x243de40]
#> 
#> [[31]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x2438c18]
#> 
#> [[32]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x2439dd0]
#> 
#> [[33]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x243a118]
#> 
#> [[34]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x24366a0]
#> 
#> [[35]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x2436c50]
#> 
#> [[36]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x2436f98]
#> 
#> [[37]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x24372e0]
#> 
#> [[38]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x2437a50]
#> 
#> [[39]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g(3)

#> [[1]]
#> <frame 37> (36)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5064ef8]
#> 
#> [[2]]
#> <frame 36> (26)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x4fbe830]
#> 
#> [[3]]
#> <frame 35> (26)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x4fbe600]
#> 
#> [[4]]
#> <frame 34> (26)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x4fbe050]
#> 
#> [[5]]
#> <frame 33> (32)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x4fc1898]
#> 
#> [[6]]
#> <frame 32> (31)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x4fc1550]
#> 
#> [[7]]
#> <frame 31> (30)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x4fc1208]
#> 
#> [[8]]
#> <frame 30> (29)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x4fc0a98]
#> 
#> [[9]]
#> <frame 29> (28)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x4fc0868]
#> 
#> [[10]]
#> <frame 28> (26)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x4fc06e0]
#> 
#> [[11]]
#> <frame 27> (26)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x4fc0600]
#> 
#> [[12]]
#> <frame 26> (25)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x4fe7a20]
#> 
#> [[13]]
#> <frame 25> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x565efe0]
#> 
#> [[14]]
#> <frame 24> (23)
#> expr: force(code)
#> env:  [frame 0x6a1c808]
#> 
#> [[15]]
#> <frame 23> (22)
#> expr: withr::with_options(list(crayon.enabled = getOption("crayon.enabled",  <...>
#> env:  [frame 0x6261910]
#> 
#> [[16]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x753c7d0]
#> 
#> [[17]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x6d87af0]
#> 
#> [[18]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6ffcef0]
#> 
#> [[19]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6ffc4e0]
#> 
#> [[20]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7056778]
#> 
#> [[21]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7835f50]
#> 
#> [[22]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x64585d8]
#> 
#> [[23]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x41f8ee0]
#> 
#> [[24]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0xc17748]
#> 
#> [[25]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x740438]
#> 
#> [[26]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x244c290]
#> 
#> [[27]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x244d250]
#> 
#> [[28]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpCL1qTH/file3b624f4c2b38"),  <...>
#> env:  [frame 0x243d0b0]
#> 
#> [[29]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpCL1qTH/file3b624f4c2b38"),  <...>
#> env:  [frame 0x243de40]
#> 
#> [[30]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x2438c18]
#> 
#> [[31]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x2439dd0]
#> 
#> [[32]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x243a118]
#> 
#> [[33]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x24366a0]
#> 
#> [[34]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x2436c50]
#> 
#> [[35]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x2436f98]
#> 
#> [[36]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x24372e0]
#> 
#> [[37]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x2437a50]
#> 
#> [[38]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"