Call stack information

The eval_ and call_ families of functions provide a replacement for the base R functions prefixed with sys. (which are all about the context stack), as well as for parent.frame() (which is the only base R function for querying the call stack). The context stack includes all R-level evaluation contexts. It is linear in terms of execution history but due to lazy evaluation it is potentially nonlinear in terms of call history. The call stack history, on the other hand, is homogenous.

global_frame()

current_frame()

ctxt_frame(n = 1)

call_frame(n = 1, clean = TRUE)

ctxt_depth()

call_depth()

ctxt_stack(n = NULL, trim = 0)

call_stack(n = NULL, clean = TRUE)

Arguments

n
The number of frames to go back in the stack.

clean
Whether to post-process the call stack to clean non-standard frames. If TRUE, suboptimal call-stack entries by base::eval() will be cleaned up: the duplicate frame created by eval() is eliminated.

trim
The number of layers of intervening frames to trim off the stack. See stack_trim() and examples.

Details

ctxt_frame() and call_frame() return a frame object containing the following fields: expr and env (call expression and evaluation environment), pos and caller_pos (position of current frame in the context stack and position of the caller), and fun (function of the current frame). ctxt_stack() and call_stack() return a list of all context or call frames on the stack. Finally, ctxt_depth() and call_depth() report the current context position or the number of calling frames on the stack.

The base R functions take two sorts of arguments to indicate which frame to query: which and n. The n argument is straightforward: it's the number of frames to go down the stack, with n = 1 referring to the current context. The which argument is more complicated and changes meaning for values lower than 1. For the sake of consistency, the rlang functions all take the same kind of argument n. This argument has a single meaning (the number of frames to go down the stack) and cannot be lower than 1.

Note finally that parent.frame(1) corresponds to call_frame(2)$env, as n = 1 always refers to the current frame. This makes the _frame() and _stack() functions consistent: ctxt_frame(2) is the same as ctxt_stack()[[2]]. Also, ctxt_depth() returns one more frame than base::sys.nframe() because it counts the global frame. That is consistent with the _stack() functions which return the global frame as well. This way, call_stack(call_depth()) is the same as global_frame().

Life cycle

These functions are soft-deprecated and replaced by trace_back().

Examples

# Expressions within arguments count as contexts
identity(identity(ctxt_depth())) # returns 2

#> Warning: `ctxt_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 36

# But they are not part of the call stack because arguments are
# evaluated within the calling function (or the global environment
# if called at top level)
identity(identity(call_depth())) # returns 0

#> Warning: `call_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 13

# The context stacks includes all intervening execution frames. The
# call stack doesn't:
f <- function(x) identity(x)
f(f(ctxt_stack()))

#> [[1]]
#> <frame 37> (36)
#> expr: identity(x)
#> env:  [local 0x9df08d8]
#> 
#> [[2]]
#> <frame 36> (33)
#> expr: f(ctxt_stack())
#> env:  [local 0x9df09f0]
#> 
#> [[3]]
#> <frame 35> (34)
#> expr: identity(x)
#> env:  [local 0x9deccd8]
#> 
#> [[4]]
#> <frame 34> (33)
#> expr: f(f(ctxt_stack()))
#> env:  [local 0x9decdf0]
#> 
#> [[5]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x68b4778]
#> 
#> [[6]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x9decfb0]
#> 
#> [[7]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x9ded1e0]
#> 
#> [[8]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x9ded790]
#> 
#> [[9]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x9dedd08]
#> 
#> [[10]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x9dee050]
#> 
#> [[11]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x9dee398]
#> 
#> [[12]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x9deea60]
#> 
#> [[13]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x9deae60]
#> 
#> [[14]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x9deafe8]
#> 
#> [[15]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x9deb0c8]
#> 
#> [[16]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x73454e8]
#> 
#> [[17]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x6ce9c70]
#> 
#> [[18]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7b21098]
#> 
#> [[19]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6328878]
#> 
#> [[20]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x3f036e0]
#> 
#> [[21]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x49d7b50]
#> 
#> [[22]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x60639f8]
#> 
#> [[23]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x60a2350]
#> 
#> [[24]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x3645df8]
#> 
#> [[25]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x363f030]
#> 
#> [[26]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x336f358]
#> 
#> [[27]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3df7338]
#> 
#> [[28]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpO1GfL0/file4f8948c60f87"),  <...>
#> env:  [frame 0x3369470]
#> 
#> [[29]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpO1GfL0/file4f8948c60f87"),  <...>
#> env:  [frame 0x33697f0]
#> 
#> [[30]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3364538]
#> 
#> [[31]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3360d98]
#> 
#> [[32]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x33610e0]
#> 
#> [[33]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3361498]
#> 
#> [[34]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3361a48]
#> 
#> [[35]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3361d90]
#> 
#> [[36]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x33620d8]
#> 
#> [[37]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3362810]
#> 
#> [[38]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(f(call_stack()))

#> Warning: `call_stack()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [[1]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x68b4778]
#> 
#> [[2]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x799fa58]
#> 
#> [[3]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x6ce9c70]
#> 
#> [[4]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7b21098]
#> 
#> [[5]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6328878]
#> 
#> [[6]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x3f036e0]
#> 
#> [[7]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x49d7b50]
#> 
#> [[8]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x60639f8]
#> 
#> [[9]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x60a2350]
#> 
#> [[10]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x3645df8]
#> 
#> [[11]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x363f030]
#> 
#> [[12]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x336f358]
#> 
#> [[13]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

g <- function(cmd) cmd()
f(g(ctxt_stack))

#> [[1]]
#> <frame 36> (33)
#> expr: g(ctxt_stack)
#> env:  [local 0x9b0f940]
#> 
#> [[2]]
#> <frame 35> (34)
#> expr: identity(x)
#> env:  [local 0x9b0fa90]
#> 
#> [[3]]
#> <frame 34> (33)
#> expr: f(g(ctxt_stack))
#> env:  [local 0x9b0fba8]
#> 
#> [[4]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x68b4778]
#> 
#> [[5]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x9b0fd68]
#> 
#> [[6]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x9b0ffd0]
#> 
#> [[7]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x9e2ee38]
#> 
#> [[8]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x9e2f3b0]
#> 
#> [[9]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x9e2f6f8]
#> 
#> [[10]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x9e2fa40]
#> 
#> [[11]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x9e30108]
#> 
#> [[12]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x9e30338]
#> 
#> [[13]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x9e304c0]
#> 
#> [[14]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x9e305a0]
#> 
#> [[15]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x9e20cd8]
#> 
#> [[16]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x6ce9c70]
#> 
#> [[17]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7b21098]
#> 
#> [[18]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6328878]
#> 
#> [[19]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x3f036e0]
#> 
#> [[20]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x49d7b50]
#> 
#> [[21]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x60639f8]
#> 
#> [[22]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x60a2350]
#> 
#> [[23]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x3645df8]
#> 
#> [[24]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x363f030]
#> 
#> [[25]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x336f358]
#> 
#> [[26]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3df7338]
#> 
#> [[27]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpO1GfL0/file4f8948c60f87"),  <...>
#> env:  [frame 0x3369470]
#> 
#> [[28]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpO1GfL0/file4f8948c60f87"),  <...>
#> env:  [frame 0x33697f0]
#> 
#> [[29]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3364538]
#> 
#> [[30]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3360d98]
#> 
#> [[31]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x33610e0]
#> 
#> [[32]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3361498]
#> 
#> [[33]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3361a48]
#> 
#> [[34]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3361d90]
#> 
#> [[35]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x33620d8]
#> 
#> [[36]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3362810]
#> 
#> [[37]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(g(call_stack))

#> [[1]]
#> <frame 36> (32)
#> expr: g(call_stack)
#> env:  [local 0x7a7e160]
#> 
#> [[2]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x68b4778]
#> 
#> [[3]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x742a088]
#> 
#> [[4]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x6ce9c70]
#> 
#> [[5]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7b21098]
#> 
#> [[6]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6328878]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x3f036e0]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x49d7b50]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x60639f8]
#> 
#> [[10]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x60a2350]
#> 
#> [[11]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x3645df8]
#> 
#> [[12]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x363f030]
#> 
#> [[13]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x336f358]
#> 
#> [[14]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

# The rlang _stack() functions return a list of frame
# objects. Use purrr::transpose() or index a field with
# purrr::map()'s to extract a particular field from a stack:

# stack <- f(f(call_stack()))
# purrr::map(stack, "env")
# purrr::transpose(stack)$expr

# current_frame() is an alias for ctxt_frame(1)
fn <- function() list(current = current_frame(), first = ctxt_frame(1))
fn()

#> Warning: `current_frame()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> $current
#> <frame 34> (33)
#> expr: fn()
#> env:  [local 0x9caa098]
#> 
#> $first
#> <frame 34> (33)
#> expr: fn()
#> env:  [local 0x9caa098]
#> 

# While current_frame() is the top of the stack, global_frame() is
# the bottom:
fn <- function() {
  n <- ctxt_depth()
  ctxt_frame(n)
}
identical(fn(), global_frame())

#> [1] TRUE

# ctxt_stack() returns a stack with all intervening frames. You can
# trim layers of intervening frames with the trim argument:
identity(identity(ctxt_stack()))

#> [[1]]
#> <frame 35> (33)
#> expr: identity(ctxt_stack())
#> env:  [local 0xa085448]
#> 
#> [[2]]
#> <frame 34> (33)
#> expr: identity(identity(ctxt_stack()))
#> env:  [local 0xa085560]
#> 
#> [[3]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x68b4778]
#> 
#> [[4]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0xa085720]
#> 
#> [[5]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0xa085950]
#> 
#> [[6]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0xa085f00]
#> 
#> [[7]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0xa082648]
#> 
#> [[8]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0xa082990]
#> 
#> [[9]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0xa082cd8]
#> 
#> [[10]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0xa0833a0]
#> 
#> [[11]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0xa0835d0]
#> 
#> [[12]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0xa083758]
#> 
#> [[13]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0xa083838]
#> 
#> [[14]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0xa0709c8]
#> 
#> [[15]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x6ce9c70]
#> 
#> [[16]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7b21098]
#> 
#> [[17]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6328878]
#> 
#> [[18]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x3f036e0]
#> 
#> [[19]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x49d7b50]
#> 
#> [[20]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x60639f8]
#> 
#> [[21]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x60a2350]
#> 
#> [[22]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x3645df8]
#> 
#> [[23]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x363f030]
#> 
#> [[24]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x336f358]
#> 
#> [[25]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3df7338]
#> 
#> [[26]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpO1GfL0/file4f8948c60f87"),  <...>
#> env:  [frame 0x3369470]
#> 
#> [[27]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpO1GfL0/file4f8948c60f87"),  <...>
#> env:  [frame 0x33697f0]
#> 
#> [[28]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3364538]
#> 
#> [[29]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3360d98]
#> 
#> [[30]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x33610e0]
#> 
#> [[31]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3361498]
#> 
#> [[32]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3361a48]
#> 
#> [[33]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3361d90]
#> 
#> [[34]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x33620d8]
#> 
#> [[35]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3362810]
#> 
#> [[36]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(ctxt_stack(trim = 1)))

#> [[1]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x68b4778]
#> 
#> [[2]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x75baa90]
#> 
#> [[3]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x75ba390]
#> 
#> [[4]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x75b9910]
#> 
#> [[5]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7596c90]
#> 
#> [[6]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x75966e0]
#> 
#> [[7]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7596328]
#> 
#> [[8]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7595560]
#> 
#> [[9]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7595330]
#> 
#> [[10]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x76522f8]
#> 
#> [[11]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7652218]
#> 
#> [[12]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x48ec668]
#> 
#> [[13]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x6ce9c70]
#> 
#> [[14]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7b21098]
#> 
#> [[15]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6328878]
#> 
#> [[16]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x3f036e0]
#> 
#> [[17]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x49d7b50]
#> 
#> [[18]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x60639f8]
#> 
#> [[19]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x60a2350]
#> 
#> [[20]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x3645df8]
#> 
#> [[21]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x363f030]
#> 
#> [[22]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x336f358]
#> 
#> [[23]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3df7338]
#> 
#> [[24]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpO1GfL0/file4f8948c60f87"),  <...>
#> env:  [frame 0x3369470]
#> 
#> [[25]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpO1GfL0/file4f8948c60f87"),  <...>
#> env:  [frame 0x33697f0]
#> 
#> [[26]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3364538]
#> 
#> [[27]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3360d98]
#> 
#> [[28]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x33610e0]
#> 
#> [[29]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3361498]
#> 
#> [[30]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3361a48]
#> 
#> [[31]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3361d90]
#> 
#> [[32]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x33620d8]
#> 
#> [[33]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3362810]
#> 
#> [[34]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# ctxt_stack() is called within fn() with intervening frames:
fn <- function(trim) identity(identity(ctxt_stack(trim = trim)))
fn(0)

#> [[1]]
#> <frame 36> (34)
#> expr: identity(ctxt_stack(trim = trim))
#> env:  [local 0x5e0e358]
#> 
#> [[2]]
#> <frame 35> (34)
#> expr: identity(identity(ctxt_stack(trim = trim)))
#> env:  [local 0x5e0e128]
#> 
#> [[3]]
#> <frame 34> (33)
#> expr: fn(0)
#> env:  [local 0x5e0def8]
#> 
#> [[4]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x68b4778]
#> 
#> [[5]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5e0db78]
#> 
#> [[6]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x5e0d718]
#> 
#> [[7]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x5e0cbb8]
#> 
#> [[8]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x5e0ffa0]
#> 
#> [[9]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x5e0fc58]
#> 
#> [[10]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x5e0f910]
#> 
#> [[11]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x5e0f248]
#> 
#> [[12]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x5e0f018]
#> 
#> [[13]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x5e0ee90]
#> 
#> [[14]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x5e0edb0]
#> 
#> [[15]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x5f35bb0]
#> 
#> [[16]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x6ce9c70]
#> 
#> [[17]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7b21098]
#> 
#> [[18]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6328878]
#> 
#> [[19]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x3f036e0]
#> 
#> [[20]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x49d7b50]
#> 
#> [[21]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x60639f8]
#> 
#> [[22]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x60a2350]
#> 
#> [[23]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x3645df8]
#> 
#> [[24]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x363f030]
#> 
#> [[25]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x336f358]
#> 
#> [[26]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3df7338]
#> 
#> [[27]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpO1GfL0/file4f8948c60f87"),  <...>
#> env:  [frame 0x3369470]
#> 
#> [[28]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpO1GfL0/file4f8948c60f87"),  <...>
#> env:  [frame 0x33697f0]
#> 
#> [[29]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3364538]
#> 
#> [[30]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3360d98]
#> 
#> [[31]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x33610e0]
#> 
#> [[32]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3361498]
#> 
#> [[33]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3361a48]
#> 
#> [[34]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3361d90]
#> 
#> [[35]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x33620d8]
#> 
#> [[36]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3362810]
#> 
#> [[37]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# We can trim the first layer of those:
fn(1)

#> [[1]]
#> <frame 34> (33)
#> expr: fn(1)
#> env:  [local 0x526eab8]
#> 
#> [[2]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x68b4778]
#> 
#> [[3]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x526e8f8]
#> 
#> [[4]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x526e6c8]
#> 
#> [[5]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x526e118]
#> 
#> [[6]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x527b500]
#> 
#> [[7]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x527b1b8]
#> 
#> [[8]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x527ae70]
#> 
#> [[9]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x527a620]
#> 
#> [[10]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x527a1c0]
#> 
#> [[11]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x5279eb0]
#> 
#> [[12]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x5279cf0]
#> 
#> [[13]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x52b45d0]
#> 
#> [[14]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x6ce9c70]
#> 
#> [[15]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7b21098]
#> 
#> [[16]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6328878]
#> 
#> [[17]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x3f036e0]
#> 
#> [[18]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x49d7b50]
#> 
#> [[19]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x60639f8]
#> 
#> [[20]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x60a2350]
#> 
#> [[21]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x3645df8]
#> 
#> [[22]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x363f030]
#> 
#> [[23]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x336f358]
#> 
#> [[24]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3df7338]
#> 
#> [[25]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpO1GfL0/file4f8948c60f87"),  <...>
#> env:  [frame 0x3369470]
#> 
#> [[26]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpO1GfL0/file4f8948c60f87"),  <...>
#> env:  [frame 0x33697f0]
#> 
#> [[27]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3364538]
#> 
#> [[28]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3360d98]
#> 
#> [[29]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x33610e0]
#> 
#> [[30]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3361498]
#> 
#> [[31]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3361a48]
#> 
#> [[32]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3361d90]
#> 
#> [[33]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x33620d8]
#> 
#> [[34]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3362810]
#> 
#> [[35]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# The outside intervening frames (at the fn() call site) are still
# returned, but can be trimmed as well:
identity(identity(fn(1)))

#> [[1]]
#> <frame 36> (33)
#> expr: fn(1)
#> env:  [local 0x4ae6638]
#> 
#> [[2]]
#> <frame 35> (33)
#> expr: identity(fn(1))
#> env:  [local 0x4ae6520]
#> 
#> [[3]]
#> <frame 34> (33)
#> expr: identity(identity(fn(1)))
#> env:  [local 0x4ae6408]
#> 
#> [[4]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x68b4778]
#> 
#> [[5]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x4ae6248]
#> 
#> [[6]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x4ae6018]
#> 
#> [[7]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x4ae5a68]
#> 
#> [[8]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x4ae54f0]
#> 
#> [[9]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x4ae8fd8]
#> 
#> [[10]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x4ae8c90]
#> 
#> [[11]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x4ae85c8]
#> 
#> [[12]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x4ae8398]
#> 
#> [[13]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x4ae8210]
#> 
#> [[14]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x4ae8130]
#> 
#> [[15]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x4afcb88]
#> 
#> [[16]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x6ce9c70]
#> 
#> [[17]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7b21098]
#> 
#> [[18]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6328878]
#> 
#> [[19]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x3f036e0]
#> 
#> [[20]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x49d7b50]
#> 
#> [[21]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x60639f8]
#> 
#> [[22]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x60a2350]
#> 
#> [[23]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x3645df8]
#> 
#> [[24]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x363f030]
#> 
#> [[25]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x336f358]
#> 
#> [[26]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3df7338]
#> 
#> [[27]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpO1GfL0/file4f8948c60f87"),  <...>
#> env:  [frame 0x3369470]
#> 
#> [[28]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpO1GfL0/file4f8948c60f87"),  <...>
#> env:  [frame 0x33697f0]
#> 
#> [[29]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3364538]
#> 
#> [[30]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3360d98]
#> 
#> [[31]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x33610e0]
#> 
#> [[32]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3361498]
#> 
#> [[33]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3361a48]
#> 
#> [[34]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3361d90]
#> 
#> [[35]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x33620d8]
#> 
#> [[36]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3362810]
#> 
#> [[37]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(fn(2)))

#> [[1]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x68b4778]
#> 
#> [[2]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x3523b80]
#> 
#> [[3]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x3523950]
#> 
#> [[4]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x35231e0]
#> 
#> [[5]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x3521fb8]
#> 
#> [[6]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x3525aa0]
#> 
#> [[7]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x3525720]
#> 
#> [[8]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x3525058]
#> 
#> [[9]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x3524df0]
#> 
#> [[10]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x3524c68]
#> 
#> [[11]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x3524b88]
#> 
#> [[12]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x3548bf8]
#> 
#> [[13]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x6ce9c70]
#> 
#> [[14]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7b21098]
#> 
#> [[15]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6328878]
#> 
#> [[16]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x3f036e0]
#> 
#> [[17]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x49d7b50]
#> 
#> [[18]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x60639f8]
#> 
#> [[19]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x60a2350]
#> 
#> [[20]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x3645df8]
#> 
#> [[21]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x363f030]
#> 
#> [[22]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x336f358]
#> 
#> [[23]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3df7338]
#> 
#> [[24]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpO1GfL0/file4f8948c60f87"),  <...>
#> env:  [frame 0x3369470]
#> 
#> [[25]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpO1GfL0/file4f8948c60f87"),  <...>
#> env:  [frame 0x33697f0]
#> 
#> [[26]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3364538]
#> 
#> [[27]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3360d98]
#> 
#> [[28]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x33610e0]
#> 
#> [[29]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3361498]
#> 
#> [[30]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3361a48]
#> 
#> [[31]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3361d90]
#> 
#> [[32]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x33620d8]
#> 
#> [[33]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3362810]
#> 
#> [[34]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g <- function(trim) identity(identity(fn(trim)))
g(2)

#> [[1]]
#> <frame 34> (33)
#> expr: g(2)
#> env:  [local 0x542b4d0]
#> 
#> [[2]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x68b4778]
#> 
#> [[3]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x542b690]
#> 
#> [[4]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x542b8c0]
#> 
#> [[5]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x542be70]
#> 
#> [[6]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x542c3e8]
#> 
#> [[7]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x542c730]
#> 
#> [[8]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x542ca78]
#> 
#> [[9]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x51c79f0]
#> 
#> [[10]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x51c7c20]
#> 
#> [[11]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x51c7da8]
#> 
#> [[12]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x51c7e88]
#> 
#> [[13]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x1682b78]
#> 
#> [[14]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x6ce9c70]
#> 
#> [[15]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7b21098]
#> 
#> [[16]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6328878]
#> 
#> [[17]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x3f036e0]
#> 
#> [[18]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x49d7b50]
#> 
#> [[19]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x60639f8]
#> 
#> [[20]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x60a2350]
#> 
#> [[21]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x3645df8]
#> 
#> [[22]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x363f030]
#> 
#> [[23]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x336f358]
#> 
#> [[24]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3df7338]
#> 
#> [[25]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpO1GfL0/file4f8948c60f87"),  <...>
#> env:  [frame 0x3369470]
#> 
#> [[26]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpO1GfL0/file4f8948c60f87"),  <...>
#> env:  [frame 0x33697f0]
#> 
#> [[27]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3364538]
#> 
#> [[28]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3360d98]
#> 
#> [[29]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x33610e0]
#> 
#> [[30]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3361498]
#> 
#> [[31]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3361a48]
#> 
#> [[32]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3361d90]
#> 
#> [[33]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x33620d8]
#> 
#> [[34]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3362810]
#> 
#> [[35]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g(3)

#> [[1]]
#> <frame 33> (32)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x68b4778]
#> 
#> [[2]]
#> <frame 32> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x6e6ac48]
#> 
#> [[3]]
#> <frame 31> (22)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x6e6aeb0]
#> 
#> [[4]]
#> <frame 30> (22)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x6e6b460]
#> 
#> [[5]]
#> <frame 29> (28)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x6e6ba10]
#> 
#> [[6]]
#> <frame 28> (27)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x6075e88]
#> 
#> [[7]]
#> <frame 27> (26)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x60761d0]
#> 
#> [[8]]
#> <frame 26> (25)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x6076898]
#> 
#> [[9]]
#> <frame 25> (24)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x6076ac8]
#> 
#> [[10]]
#> <frame 24> (22)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x6076c50]
#> 
#> [[11]]
#> <frame 23> (22)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x6076d30]
#> 
#> [[12]]
#> <frame 22> (21)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x48ac140]
#> 
#> [[13]]
#> <frame 21> (20)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x6ce9c70]
#> 
#> [[14]]
#> <frame 20> (19)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7b21098]
#> 
#> [[15]]
#> <frame 19> (18)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x6328878]
#> 
#> [[16]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x3f036e0]
#> 
#> [[17]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x49d7b50]
#> 
#> [[18]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x60639f8]
#> 
#> [[19]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x60a2350]
#> 
#> [[20]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x3645df8]
#> 
#> [[21]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x363f030]
#> 
#> [[22]]
#> <frame 12> (0)
#> expr: (function (..., crayon_enabled, crayon_colors, pkgdown_internet)  <...>
#> env:  [frame 0x336f358]
#> 
#> [[23]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x3df7338]
#> 
#> [[24]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/tmp/RtmpO1GfL0/file4f8948c60f87"),  <...>
#> env:  [frame 0x3369470]
#> 
#> [[25]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/tmp/RtmpO1GfL0/file4f8948c60f87"),  <...>
#> env:  [frame 0x33697f0]
#> 
#> [[26]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x3364538]
#> 
#> [[27]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3360d98]
#> 
#> [[28]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x33610e0]
#> 
#> [[29]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x3361498]
#> 
#> [[30]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x3361a48]
#> 
#> [[31]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x3361d90]
#> 
#> [[32]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x33620d8]
#> 
#> [[33]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x3362810]
#> 
#> [[34]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"