Call stack information

The eval_ and call_ families of functions provide a replacement for the base R functions prefixed with sys. (which are all about the context stack), as well as for parent.frame() (which is the only base R function for querying the call stack). The context stack includes all R-level evaluation contexts. It is linear in terms of execution history but due to lazy evaluation it is potentially nonlinear in terms of call history. The call stack history, on the other hand, is homogenous.

global_frame()

current_frame()

ctxt_frame(n = 1)

call_frame(n = 1, clean = TRUE)

ctxt_depth()

call_depth()

ctxt_stack(n = NULL, trim = 0)

call_stack(n = NULL, clean = TRUE)

Arguments

n
The number of frames to go back in the stack.

clean
Whether to post-process the call stack to clean non-standard frames. If TRUE, suboptimal call-stack entries by base::eval() will be cleaned up: the duplicate frame created by eval() is eliminated.

trim
The number of layers of intervening frames to trim off the stack. See stack_trim() and examples.

Details

ctxt_frame() and call_frame() return a frame object containing the following fields: expr and env (call expression and evaluation environment), pos and caller_pos (position of current frame in the context stack and position of the caller), and fun (function of the current frame). ctxt_stack() and call_stack() return a list of all context or call frames on the stack. Finally, ctxt_depth() and call_depth() report the current context position or the number of calling frames on the stack.

The base R functions take two sorts of arguments to indicate which frame to query: which and n. The n argument is straightforward: it's the number of frames to go down the stack, with n = 1 referring to the current context. The which argument is more complicated and changes meaning for values lower than 1. For the sake of consistency, the rlang functions all take the same kind of argument n. This argument has a single meaning (the number of frames to go down the stack) and cannot be lower than 1.

Note finally that parent.frame(1) corresponds to call_frame(2)$env, as n = 1 always refers to the current frame. This makes the _frame() and _stack() functions consistent: ctxt_frame(2) is the same as ctxt_stack()[[2]]. Also, ctxt_depth() returns one more frame than base::sys.nframe() because it counts the global frame. That is consistent with the _stack() functions which return the global frame as well. This way, call_stack(call_depth()) is the same as global_frame().

Life cycle

These functions are soft-deprecated and replaced by trace_back().

Examples

# Expressions within arguments count as contexts
identity(identity(ctxt_depth())) # returns 2

#> Warning: `ctxt_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 25

# But they are not part of the call stack because arguments are
# evaluated within the calling function (or the global environment
# if called at top level)
identity(identity(call_depth())) # returns 0

#> Warning: `call_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 13

# The context stacks includes all intervening execution frames. The
# call stack doesn't:
f <- function(x) identity(x)
f(f(ctxt_stack()))

#> [[1]]
#> <frame 26> (25)
#> expr: identity(x)
#> env:  [local 0x7ff33fad51c8]
#> 
#> [[2]]
#> <frame 25> (22)
#> expr: f(ctxt_stack())
#> env:  [local 0x7ff33fad50b0]
#> 
#> [[3]]
#> <frame 24> (23)
#> expr: identity(x)
#> env:  [local 0x7ff33fad4f98]
#> 
#> [[4]]
#> <frame 23> (22)
#> expr: f(f(ctxt_stack()))
#> env:  [local 0x7ff33fad4e80]
#> 
#> [[5]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff34086f940]
#> 
#> [[6]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff33fad4cc0]
#> 
#> [[7]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7ff33fad4a90]
#> 
#> [[8]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7ff33fad44a8]
#> 
#> [[9]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7ff33fad7e08]
#> 
#> [[10]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7ff33fad7ac0]
#> 
#> [[11]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7ff33fad7778]
#> 
#> [[12]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7ff33fad70b0]
#> 
#> [[13]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7ff33fad6e80]
#> 
#> [[14]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7ff33fad6cf8]
#> 
#> [[15]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7ff33fad6c18]
#> 
#> [[16]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7ff33faeacc0]
#> 
#> [[17]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7ff344559de0]
#> 
#> [[18]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7ff3445532e0]
#> 
#> [[19]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7ff344516240]
#> 
#> [[20]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7ff344415258]
#> 
#> [[21]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7ff3443bf728]
#> 
#> [[22]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7ff33c82cae8]
#> 
#> [[23]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff33c712d48]
#> 
#> [[24]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff33a5357f0]
#> 
#> [[25]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7ff33db30db0]
#> 
#> [[26]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7ff33c529a90]
#> 
#> [[27]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(f(call_stack()))

#> Warning: `call_stack()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [[1]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff34086f940]
#> 
#> [[2]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7ff33f450ee8]
#> 
#> [[3]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7ff344559de0]
#> 
#> [[4]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7ff3445532e0]
#> 
#> [[5]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7ff344516240]
#> 
#> [[6]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7ff344415258]
#> 
#> [[7]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7ff3443bf728]
#> 
#> [[8]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7ff33c82cae8]
#> 
#> [[9]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff33c712d48]
#> 
#> [[10]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff33a5357f0]
#> 
#> [[11]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7ff33db30db0]
#> 
#> [[12]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7ff33c529a90]
#> 
#> [[13]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

g <- function(cmd) cmd()
f(g(ctxt_stack))

#> [[1]]
#> <frame 25> (22)
#> expr: g(ctxt_stack)
#> env:  [local 0x7ff33e49a898]
#> 
#> [[2]]
#> <frame 24> (23)
#> expr: identity(x)
#> env:  [local 0x7ff33e49a780]
#> 
#> [[3]]
#> <frame 23> (22)
#> expr: f(g(ctxt_stack))
#> env:  [local 0x7ff33e49a5f8]
#> 
#> [[4]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff34086f940]
#> 
#> [[5]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff33e49a400]
#> 
#> [[6]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7ff33e4a00e0]
#> 
#> [[7]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7ff33e49fac0]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7ff33e49f510]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7ff33e49f190]
#> 
#> [[10]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7ff33e49eda0]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7ff33e49e6a0]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7ff33e49e470]
#> 
#> [[13]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7ff33e49e2b0]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7ff33e4a40e0]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7ff34050b820]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7ff344559de0]
#> 
#> [[17]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7ff3445532e0]
#> 
#> [[18]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7ff344516240]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7ff344415258]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7ff3443bf728]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7ff33c82cae8]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff33c712d48]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff33a5357f0]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7ff33db30db0]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7ff33c529a90]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(g(call_stack))

#> [[1]]
#> <frame 25> (21)
#> expr: g(call_stack)
#> env:  [local 0x7ff3403fbf58]
#> 
#> [[2]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff34086f940]
#> 
#> [[3]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7ff33f843cb8]
#> 
#> [[4]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7ff344559de0]
#> 
#> [[5]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7ff3445532e0]
#> 
#> [[6]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7ff344516240]
#> 
#> [[7]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7ff344415258]
#> 
#> [[8]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7ff3443bf728]
#> 
#> [[9]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7ff33c82cae8]
#> 
#> [[10]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff33c712d48]
#> 
#> [[11]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff33a5357f0]
#> 
#> [[12]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7ff33db30db0]
#> 
#> [[13]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7ff33c529a90]
#> 
#> [[14]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

# The rlang _stack() functions return a list of frame
# objects. Use purrr::transpose() or index a field with
# purrr::map()'s to extract a particular field from a stack:

# stack <- f(f(call_stack()))
# purrr::map(stack, "env")
# purrr::transpose(stack)$expr

# current_frame() is an alias for ctxt_frame(1)
fn <- function() list(current = current_frame(), first = ctxt_frame(1))
fn()

#> Warning: `current_frame()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> $current
#> <frame 23> (22)
#> expr: fn()
#> env:  [local 0x7ff34008baf8]
#> 
#> $first
#> <frame 23> (22)
#> expr: fn()
#> env:  [local 0x7ff34008baf8]
#> 

# While current_frame() is the top of the stack, global_frame() is
# the bottom:
fn <- function() {
  n <- ctxt_depth()
  ctxt_frame(n)
}
identical(fn(), global_frame())

#> [1] TRUE

# ctxt_stack() returns a stack with all intervening frames. You can
# trim layers of intervening frames with the trim argument:
identity(identity(ctxt_stack()))

#> [[1]]
#> <frame 24> (22)
#> expr: identity(ctxt_stack())
#> env:  [local 0x7ff33f0c5c30]
#> 
#> [[2]]
#> <frame 23> (22)
#> expr: identity(identity(ctxt_stack()))
#> env:  [local 0x7ff33f0c5b18]
#> 
#> [[3]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff34086f940]
#> 
#> [[4]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff33f0c5958]
#> 
#> [[5]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7ff33f0c5728]
#> 
#> [[6]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7ff33f0c7ca0]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7ff33f0c76f0]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7ff33f0c7370]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7ff33f0c6fb8]
#> 
#> [[10]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7ff33f0cc7c8]
#> 
#> [[11]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7ff33f0cc598]
#> 
#> [[12]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7ff33f0cc410]
#> 
#> [[13]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7ff33f0cc330]
#> 
#> [[14]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7ff33f0e1bc0]
#> 
#> [[15]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7ff344559de0]
#> 
#> [[16]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7ff3445532e0]
#> 
#> [[17]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7ff344516240]
#> 
#> [[18]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7ff344415258]
#> 
#> [[19]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7ff3443bf728]
#> 
#> [[20]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7ff33c82cae8]
#> 
#> [[21]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff33c712d48]
#> 
#> [[22]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff33a5357f0]
#> 
#> [[23]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7ff33db30db0]
#> 
#> [[24]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7ff33c529a90]
#> 
#> [[25]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(ctxt_stack(trim = 1)))

#> [[1]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff34086f940]
#> 
#> [[2]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff33b716f48]
#> 
#> [[3]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7ff33b716ce0]
#> 
#> [[4]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7ff33b71a640]
#> 
#> [[5]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7ff33b71a0c8]
#> 
#> [[6]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7ff33b719d48]
#> 
#> [[7]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7ff33b719a00]
#> 
#> [[8]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7ff33b719300]
#> 
#> [[9]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7ff33b719098]
#> 
#> [[10]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7ff33b718f10]
#> 
#> [[11]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7ff33b718e30]
#> 
#> [[12]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7ff33c636af0]
#> 
#> [[13]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7ff344559de0]
#> 
#> [[14]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7ff3445532e0]
#> 
#> [[15]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7ff344516240]
#> 
#> [[16]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7ff344415258]
#> 
#> [[17]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7ff3443bf728]
#> 
#> [[18]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7ff33c82cae8]
#> 
#> [[19]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff33c712d48]
#> 
#> [[20]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff33a5357f0]
#> 
#> [[21]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7ff33db30db0]
#> 
#> [[22]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7ff33c529a90]
#> 
#> [[23]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# ctxt_stack() is called within fn() with intervening frames:
fn <- function(trim) identity(identity(ctxt_stack(trim = trim)))
fn(0)

#> [[1]]
#> <frame 25> (23)
#> expr: identity(ctxt_stack(trim = trim))
#> env:  [local 0x7ff33c59fb70]
#> 
#> [[2]]
#> <frame 24> (23)
#> expr: identity(identity(ctxt_stack(trim = trim)))
#> env:  [local 0x7ff33c59fa58]
#> 
#> [[3]]
#> <frame 23> (22)
#> expr: fn(0)
#> env:  [local 0x7ff33c59f7b8]
#> 
#> [[4]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff34086f940]
#> 
#> [[5]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff33c59f5f8]
#> 
#> [[6]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7ff33c59f390]
#> 
#> [[7]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7ff33c59ebb0]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7ff33c5a30f0]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7ff33c5a29f0]
#> 
#> [[10]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7ff33c5a6200]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7ff33c5a52e8]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7ff33c5a50b8]
#> 
#> [[13]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7ff33c5a7d30]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7ff33c5a7c18]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7ff33dba7748]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7ff344559de0]
#> 
#> [[17]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7ff3445532e0]
#> 
#> [[18]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7ff344516240]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7ff344415258]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7ff3443bf728]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7ff33c82cae8]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff33c712d48]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff33a5357f0]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7ff33db30db0]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7ff33c529a90]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# We can trim the first layer of those:
fn(1)

#> [[1]]
#> <frame 23> (22)
#> expr: fn(1)
#> env:  [local 0x7ff33db196a8]
#> 
#> [[2]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff34086f940]
#> 
#> [[3]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff33db19440]
#> 
#> [[4]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7ff33db190c0]
#> 
#> [[5]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7ff33db189f8]
#> 
#> [[6]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7ff33db1dc20]
#> 
#> [[7]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7ff33db1d718]
#> 
#> [[8]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7ff33db1d168]
#> 
#> [[9]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7ff33db1c918]
#> 
#> [[10]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7ff33db1c608]
#> 
#> [[11]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7ff33db20320]
#> 
#> [[12]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7ff33db201d0]
#> 
#> [[13]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7ff33c9d4dd0]
#> 
#> [[14]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7ff344559de0]
#> 
#> [[15]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7ff3445532e0]
#> 
#> [[16]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7ff344516240]
#> 
#> [[17]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7ff344415258]
#> 
#> [[18]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7ff3443bf728]
#> 
#> [[19]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7ff33c82cae8]
#> 
#> [[20]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff33c712d48]
#> 
#> [[21]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff33a5357f0]
#> 
#> [[22]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7ff33db30db0]
#> 
#> [[23]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7ff33c529a90]
#> 
#> [[24]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# The outside intervening frames (at the fn() call site) are still
# returned, but can be trimmed as well:
identity(identity(fn(1)))

#> [[1]]
#> <frame 25> (22)
#> expr: fn(1)
#> env:  [local 0x7ff33a1cc0e8]
#> 
#> [[2]]
#> <frame 24> (22)
#> expr: identity(fn(1))
#> env:  [local 0x7ff33a1cfe70]
#> 
#> [[3]]
#> <frame 23> (22)
#> expr: identity(identity(fn(1)))
#> env:  [local 0x7ff33a1cfce8]
#> 
#> [[4]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff34086f940]
#> 
#> [[5]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff33a1cfab8]
#> 
#> [[6]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7ff33a1cf4d0]
#> 
#> [[7]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7ff33a1ce3c0]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7ff33a1d1888]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7ff33a1d1038]
#> 
#> [[10]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7ff33a1d0bd8]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7ff33a1d01c8]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7ff33a1d3c08]
#> 
#> [[13]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7ff33a1d3a80]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7ff33a1d3968]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7ff33a216990]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7ff344559de0]
#> 
#> [[17]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7ff3445532e0]
#> 
#> [[18]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7ff344516240]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7ff344415258]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7ff3443bf728]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7ff33c82cae8]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff33c712d48]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff33a5357f0]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7ff33db30db0]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7ff33c529a90]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(fn(2)))

#> [[1]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff34086f940]
#> 
#> [[2]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff3423448f8]
#> 
#> [[3]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7ff342344b28]
#> 
#> [[4]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7ff3423411c8]
#> 
#> [[5]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7ff342341740]
#> 
#> [[6]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7ff342341a88]
#> 
#> [[7]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7ff342341dd0]
#> 
#> [[8]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7ff342342498]
#> 
#> [[9]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7ff3423426c8]
#> 
#> [[10]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7ff342342850]
#> 
#> [[11]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7ff342342930]
#> 
#> [[12]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7ff3423322d8]
#> 
#> [[13]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7ff344559de0]
#> 
#> [[14]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7ff3445532e0]
#> 
#> [[15]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7ff344516240]
#> 
#> [[16]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7ff344415258]
#> 
#> [[17]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7ff3443bf728]
#> 
#> [[18]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7ff33c82cae8]
#> 
#> [[19]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff33c712d48]
#> 
#> [[20]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff33a5357f0]
#> 
#> [[21]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7ff33db30db0]
#> 
#> [[22]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7ff33c529a90]
#> 
#> [[23]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g <- function(trim) identity(identity(fn(trim)))
g(2)

#> [[1]]
#> <frame 23> (22)
#> expr: g(2)
#> env:  [local 0x7ff342d5b638]
#> 
#> [[2]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff34086f940]
#> 
#> [[3]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff342d5b7f8]
#> 
#> [[4]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7ff342d5ba28]
#> 
#> [[5]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7ff342d5bfd8]
#> 
#> [[6]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7ff342d58640]
#> 
#> [[7]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7ff342d58988]
#> 
#> [[8]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7ff342d58cd0]
#> 
#> [[9]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7ff342d59398]
#> 
#> [[10]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7ff342d595c8]
#> 
#> [[11]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7ff342d59750]
#> 
#> [[12]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7ff342d59830]
#> 
#> [[13]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7ff342d453a0]
#> 
#> [[14]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7ff344559de0]
#> 
#> [[15]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7ff3445532e0]
#> 
#> [[16]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7ff344516240]
#> 
#> [[17]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7ff344415258]
#> 
#> [[18]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7ff3443bf728]
#> 
#> [[19]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7ff33c82cae8]
#> 
#> [[20]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff33c712d48]
#> 
#> [[21]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff33a5357f0]
#> 
#> [[22]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7ff33db30db0]
#> 
#> [[23]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7ff33c529a90]
#> 
#> [[24]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g(3)

#> [[1]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff34086f940]
#> 
#> [[2]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7ff3440e6588]
#> 
#> [[3]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7ff3440e67b8]
#> 
#> [[4]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7ff3440e6d68]
#> 
#> [[5]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7ff3440e72e0]
#> 
#> [[6]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7ff3440e7628]
#> 
#> [[7]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7ff3440e7970]
#> 
#> [[8]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7ff3440e8038]
#> 
#> [[9]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7ff3440e4358]
#> 
#> [[10]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7ff3440e44e0]
#> 
#> [[11]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7ff3440e45c0]
#> 
#> [[12]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7ff34416ea40]
#> 
#> [[13]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7ff344559de0]
#> 
#> [[14]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7ff3445532e0]
#> 
#> [[15]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7ff344516240]
#> 
#> [[16]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7ff344415258]
#> 
#> [[17]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7ff3443bf728]
#> 
#> [[18]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7ff33c82cae8]
#> 
#> [[19]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff33c712d48]
#> 
#> [[20]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7ff33a5357f0]
#> 
#> [[21]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7ff33db30db0]
#> 
#> [[22]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7ff33c529a90]
#> 
#> [[23]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"