Call stack information

The eval_ and call_ families of functions provide a replacement for the base R functions prefixed with sys. (which are all about the context stack), as well as for parent.frame() (which is the only base R function for querying the call stack). The context stack includes all R-level evaluation contexts. It is linear in terms of execution history but due to lazy evaluation it is potentially nonlinear in terms of call history. The call stack history, on the other hand, is homogenous.

global_frame()

current_frame()

ctxt_frame(n = 1)

call_frame(n = 1, clean = TRUE)

ctxt_depth()

call_depth()

ctxt_stack(n = NULL, trim = 0)

call_stack(n = NULL, clean = TRUE)

Arguments

n
The number of frames to go back in the stack.

clean
Whether to post-process the call stack to clean non-standard frames. If TRUE, suboptimal call-stack entries by base::eval() will be cleaned up: the duplicate frame created by eval() is eliminated.

trim
The number of layers of intervening frames to trim off the stack. See stack_trim() and examples.

Details

ctxt_frame() and call_frame() return a frame object containing the following fields: expr and env (call expression and evaluation environment), pos and caller_pos (position of current frame in the context stack and position of the caller), and fun (function of the current frame). ctxt_stack() and call_stack() return a list of all context or call frames on the stack. Finally, ctxt_depth() and call_depth() report the current context position or the number of calling frames on the stack.

The base R functions take two sorts of arguments to indicate which frame to query: which and n. The n argument is straightforward: it's the number of frames to go down the stack, with n = 1 referring to the current context. The which argument is more complicated and changes meaning for values lower than 1. For the sake of consistency, the rlang functions all take the same kind of argument n. This argument has a single meaning (the number of frames to go down the stack) and cannot be lower than 1.

Note finally that parent.frame(1) corresponds to call_frame(2)$env, as n = 1 always refers to the current frame. This makes the _frame() and _stack() functions consistent: ctxt_frame(2) is the same as ctxt_stack()[[2]]. Also, ctxt_depth() returns one more frame than base::sys.nframe() because it counts the global frame. That is consistent with the _stack() functions which return the global frame as well. This way, call_stack(call_depth()) is the same as global_frame().

Life cycle

These functions are soft-deprecated and replaced by trace_back().

Examples

# Expressions within arguments count as contexts
identity(identity(ctxt_depth())) # returns 2

#> Warning: `ctxt_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 26

# But they are not part of the call stack because arguments are
# evaluated within the calling function (or the global environment
# if called at top level)
identity(identity(call_depth())) # returns 0

#> Warning: `call_depth()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [1] 13

# The context stacks includes all intervening execution frames. The
# call stack doesn't:
f <- function(x) identity(x)
f(f(ctxt_stack()))

#> [[1]]
#> <frame 27> (26)
#> expr: identity(x)
#> env:  [local 0x7fcc18b4ec30]
#> 
#> [[2]]
#> <frame 26> (23)
#> expr: f(ctxt_stack())
#> env:  [local 0x7fcc18b4ed48]
#> 
#> [[3]]
#> <frame 25> (24)
#> expr: identity(x)
#> env:  [local 0x7fcc18b4ee60]
#> 
#> [[4]]
#> <frame 24> (23)
#> expr: f(f(ctxt_stack()))
#> env:  [local 0x7fcc18b4ef78]
#> 
#> [[5]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fcc1a9cf6b8]
#> 
#> [[6]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fcc18b4f138]
#> 
#> [[7]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fcc18b4f368]
#> 
#> [[8]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fcc18b4f918]
#> 
#> [[9]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fcc18b4fe90]
#> 
#> [[10]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fcc18b501d8]
#> 
#> [[11]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fcc18b50558]
#> 
#> [[12]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fcc18dfb310]
#> 
#> [[13]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fcc18dfb540]
#> 
#> [[14]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fcc18dfb6c8]
#> 
#> [[15]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fcc18dfb7a8]
#> 
#> [[16]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fcc18b07268]
#> 
#> [[17]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fcc13828ff0]
#> 
#> [[18]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fcc13b45720]
#> 
#> [[19]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fcc13c1fb10]
#> 
#> [[20]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fcc17565530]
#> 
#> [[21]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fcc17564df8]
#> 
#> [[22]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fcc17592758]
#> 
#> [[23]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fcc177071b0]
#> 
#> [[24]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fcc178a8d68]
#> 
#> [[25]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fcc172c8390]
#> 
#> [[26]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fcc172c34b0]
#> 
#> [[27]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fcc1597a838]
#> 
#> [[28]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(f(call_stack()))

#> Warning: `call_stack()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> [[1]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fcc1a9cf6b8]
#> 
#> [[2]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fcc18daf280]
#> 
#> [[3]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fcc13828ff0]
#> 
#> [[4]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fcc13b45720]
#> 
#> [[5]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fcc13c1fb10]
#> 
#> [[6]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fcc17565530]
#> 
#> [[7]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fcc17592758]
#> 
#> [[8]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fcc177071b0]
#> 
#> [[9]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fcc178a8d68]
#> 
#> [[10]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fcc172c8390]
#> 
#> [[11]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fcc172c34b0]
#> 
#> [[12]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fcc1597a838]
#> 
#> [[13]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

g <- function(cmd) cmd()
f(g(ctxt_stack))

#> [[1]]
#> <frame 26> (23)
#> expr: g(ctxt_stack)
#> env:  [local 0x7fcc1b2226e0]
#> 
#> [[2]]
#> <frame 25> (24)
#> expr: identity(x)
#> env:  [local 0x7fcc1b2227f8]
#> 
#> [[3]]
#> <frame 24> (23)
#> expr: f(g(ctxt_stack))
#> env:  [local 0x7fcc1b222910]
#> 
#> [[4]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fcc1a9cf6b8]
#> 
#> [[5]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fcc1b222ad0]
#> 
#> [[6]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fcc1b222d00]
#> 
#> [[7]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fcc1b223358]
#> 
#> [[8]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fcc1b2238d0]
#> 
#> [[9]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fcc1b223c18]
#> 
#> [[10]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fcc1b223f60]
#> 
#> [[11]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fcc1b220718]
#> 
#> [[12]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fcc1b220948]
#> 
#> [[13]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fcc1b220b40]
#> 
#> [[14]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fcc1b220c20]
#> 
#> [[15]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fcc1b2108a0]
#> 
#> [[16]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fcc13828ff0]
#> 
#> [[17]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fcc13b45720]
#> 
#> [[18]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fcc13c1fb10]
#> 
#> [[19]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fcc17565530]
#> 
#> [[20]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fcc17564df8]
#> 
#> [[21]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fcc17592758]
#> 
#> [[22]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fcc177071b0]
#> 
#> [[23]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fcc178a8d68]
#> 
#> [[24]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fcc172c8390]
#> 
#> [[25]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fcc172c34b0]
#> 
#> [[26]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fcc1597a838]
#> 
#> [[27]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

f(g(call_stack))

#> [[1]]
#> <frame 26> (22)
#> expr: g(call_stack)
#> env:  [local 0x7fcc1b0d4600]
#> 
#> [[2]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fcc1a9cf6b8]
#> 
#> [[3]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fcc1b0bec30]
#> 
#> [[4]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fcc13828ff0]
#> 
#> [[5]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fcc13b45720]
#> 
#> [[6]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fcc13c1fb10]
#> 
#> [[7]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fcc17565530]
#> 
#> [[8]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fcc17592758]
#> 
#> [[9]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fcc177071b0]
#> 
#> [[10]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fcc178a8d68]
#> 
#> [[11]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fcc172c8390]
#> 
#> [[12]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fcc172c34b0]
#> 
#> [[13]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fcc1597a838]
#> 
#> [[14]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "call_stack" "stack"     

# The rlang _stack() functions return a list of frame
# objects. Use purrr::transpose() or index a field with
# purrr::map()'s to extract a particular field from a stack:

# stack <- f(f(call_stack()))
# purrr::map(stack, "env")
# purrr::transpose(stack)$expr

# current_frame() is an alias for ctxt_frame(1)
fn <- function() list(current = current_frame(), first = ctxt_frame(1))
fn()

#> Warning: `current_frame()` is deprecated as of rlang 0.3.0.
#> This warning is displayed once per session.

#> $current
#> <frame 24> (23)
#> expr: fn()
#> env:  [local 0x7fcc1a97edd0]
#> 
#> $first
#> <frame 24> (23)
#> expr: fn()
#> env:  [local 0x7fcc1a97edd0]
#> 

# While current_frame() is the top of the stack, global_frame() is
# the bottom:
fn <- function() {
  n <- ctxt_depth()
  ctxt_frame(n)
}
identical(fn(), global_frame())

#> [1] TRUE

# ctxt_stack() returns a stack with all intervening frames. You can
# trim layers of intervening frames with the trim argument:
identity(identity(ctxt_stack()))

#> [[1]]
#> <frame 25> (23)
#> expr: identity(ctxt_stack())
#> env:  [local 0x7fcc18f06390]
#> 
#> [[2]]
#> <frame 24> (23)
#> expr: identity(identity(ctxt_stack()))
#> env:  [local 0x7fcc18f06278]
#> 
#> [[3]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fcc1a9cf6b8]
#> 
#> [[4]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fcc1837ddc8]
#> 
#> [[5]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fcc1837db98]
#> 
#> [[6]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fcc1837e4d8]
#> 
#> [[7]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fcc183c9e70]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fcc183c9b28]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fcc183c97e0]
#> 
#> [[10]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fcc183c9118]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fcc183c8ee8]
#> 
#> [[12]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fcc183c8d60]
#> 
#> [[13]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fcc183c8c80]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fcc1852a060]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fcc13828ff0]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fcc13b45720]
#> 
#> [[17]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fcc13c1fb10]
#> 
#> [[18]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fcc17565530]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fcc17564df8]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fcc17592758]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fcc177071b0]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fcc178a8d68]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fcc172c8390]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fcc172c34b0]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fcc1597a838]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(ctxt_stack(trim = 1)))

#> [[1]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fcc1a9cf6b8]
#> 
#> [[2]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fcc1811e070]
#> 
#> [[3]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fcc1811dd60]
#> 
#> [[4]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fcc1811d7b0]
#> 
#> [[5]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fcc1811d238]
#> 
#> [[6]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fcc1811cef0]
#> 
#> [[7]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fcc1811cba8]
#> 
#> [[8]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fcc1811c4e0]
#> 
#> [[9]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fcc1811c2b0]
#> 
#> [[10]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fcc18120038]
#> 
#> [[11]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fcc1811ff58]
#> 
#> [[12]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fcc1813d2e0]
#> 
#> [[13]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fcc13828ff0]
#> 
#> [[14]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fcc13b45720]
#> 
#> [[15]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fcc13c1fb10]
#> 
#> [[16]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fcc17565530]
#> 
#> [[17]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fcc17564df8]
#> 
#> [[18]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fcc17592758]
#> 
#> [[19]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fcc177071b0]
#> 
#> [[20]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fcc178a8d68]
#> 
#> [[21]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fcc172c8390]
#> 
#> [[22]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fcc172c34b0]
#> 
#> [[23]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fcc1597a838]
#> 
#> [[24]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# ctxt_stack() is called within fn() with intervening frames:
fn <- function(trim) identity(identity(ctxt_stack(trim = trim)))
fn(0)

#> [[1]]
#> <frame 26> (24)
#> expr: identity(ctxt_stack(trim = trim))
#> env:  [local 0x7fcc1479de50]
#> 
#> [[2]]
#> <frame 25> (24)
#> expr: identity(identity(ctxt_stack(trim = trim)))
#> env:  [local 0x7fcc1479dd38]
#> 
#> [[3]]
#> <frame 24> (23)
#> expr: fn(0)
#> env:  [local 0x7fcc1479dc20]
#> 
#> [[4]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fcc1a9cf6b8]
#> 
#> [[5]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fcc1479da60]
#> 
#> [[6]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fcc1479d830]
#> 
#> [[7]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fcc1479d248]
#> 
#> [[8]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fcc1479ccd0]
#> 
#> [[9]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fcc1479c988]
#> 
#> [[10]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fcc1479c608]
#> 
#> [[11]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fcc1479fe50]
#> 
#> [[12]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fcc1479fc20]
#> 
#> [[13]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fcc1479fa60]
#> 
#> [[14]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fcc1479f980]
#> 
#> [[15]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fcc147bc400]
#> 
#> [[16]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fcc13828ff0]
#> 
#> [[17]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fcc13b45720]
#> 
#> [[18]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fcc13c1fb10]
#> 
#> [[19]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fcc17565530]
#> 
#> [[20]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fcc17564df8]
#> 
#> [[21]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fcc17592758]
#> 
#> [[22]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fcc177071b0]
#> 
#> [[23]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fcc178a8d68]
#> 
#> [[24]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fcc172c8390]
#> 
#> [[25]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fcc172c34b0]
#> 
#> [[26]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fcc1597a838]
#> 
#> [[27]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# We can trim the first layer of those:
fn(1)

#> [[1]]
#> <frame 24> (23)
#> expr: fn(1)
#> env:  [local 0x7fcc144d53f8]
#> 
#> [[2]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fcc1a9cf6b8]
#> 
#> [[3]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fcc144d5238]
#> 
#> [[4]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fcc144d4fd0]
#> 
#> [[5]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fcc144d49e8]
#> 
#> [[6]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fcc144d43c8]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fcc144d9f58]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fcc144d9c10]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fcc144d9548]
#> 
#> [[10]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fcc144d9318]
#> 
#> [[11]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fcc144d90e8]
#> 
#> [[12]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fcc144d8f98]
#> 
#> [[13]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fcc144efe78]
#> 
#> [[14]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fcc13828ff0]
#> 
#> [[15]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fcc13b45720]
#> 
#> [[16]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fcc13c1fb10]
#> 
#> [[17]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fcc17565530]
#> 
#> [[18]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fcc17564df8]
#> 
#> [[19]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fcc17592758]
#> 
#> [[20]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fcc177071b0]
#> 
#> [[21]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fcc178a8d68]
#> 
#> [[22]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fcc172c8390]
#> 
#> [[23]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fcc172c34b0]
#> 
#> [[24]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fcc1597a838]
#> 
#> [[25]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# The outside intervening frames (at the fn() call site) are still
# returned, but can be trimmed as well:
identity(identity(fn(1)))

#> [[1]]
#> <frame 26> (23)
#> expr: fn(1)
#> env:  [local 0x7fcc17aae4e0]
#> 
#> [[2]]
#> <frame 25> (23)
#> expr: identity(fn(1))
#> env:  [local 0x7fcc17aae390]
#> 
#> [[3]]
#> <frame 24> (23)
#> expr: identity(identity(fn(1)))
#> env:  [local 0x7fcc17aae278]
#> 
#> [[4]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fcc1a9cf6b8]
#> 
#> [[5]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fcc17ab1fc8]
#> 
#> [[6]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fcc17ab1d98]
#> 
#> [[7]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fcc17ab17b0]
#> 
#> [[8]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fcc17ab1238]
#> 
#> [[9]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fcc17ab0ef0]
#> 
#> [[10]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fcc17ab0b70]
#> 
#> [[11]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fcc17ab0400]
#> 
#> [[12]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fcc17ab3e78]
#> 
#> [[13]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fcc17ab3cf0]
#> 
#> [[14]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fcc17ab3c10]
#> 
#> [[15]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fcc1802d9a8]
#> 
#> [[16]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fcc13828ff0]
#> 
#> [[17]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fcc13b45720]
#> 
#> [[18]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fcc13c1fb10]
#> 
#> [[19]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fcc17565530]
#> 
#> [[20]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fcc17564df8]
#> 
#> [[21]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fcc17592758]
#> 
#> [[22]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fcc177071b0]
#> 
#> [[23]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fcc178a8d68]
#> 
#> [[24]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fcc172c8390]
#> 
#> [[25]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fcc172c34b0]
#> 
#> [[26]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fcc1597a838]
#> 
#> [[27]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

identity(identity(fn(2)))

#> [[1]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fcc1a9cf6b8]
#> 
#> [[2]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fcc17859938]
#> 
#> [[3]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fcc17859708]
#> 
#> [[4]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fcc17859158]
#> 
#> [[5]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fcc17858ba8]
#> 
#> [[6]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fcc17858860]
#> 
#> [[7]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fcc17858518]
#> 
#> [[8]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fcc1785bd60]
#> 
#> [[9]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fcc1785b698]
#> 
#> [[10]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fcc1785b510]
#> 
#> [[11]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fcc1785b430]
#> 
#> [[12]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fcc178816d0]
#> 
#> [[13]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fcc13828ff0]
#> 
#> [[14]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fcc13b45720]
#> 
#> [[15]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fcc13c1fb10]
#> 
#> [[16]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fcc17565530]
#> 
#> [[17]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fcc17564df8]
#> 
#> [[18]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fcc17592758]
#> 
#> [[19]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fcc177071b0]
#> 
#> [[20]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fcc178a8d68]
#> 
#> [[21]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fcc172c8390]
#> 
#> [[22]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fcc172c34b0]
#> 
#> [[23]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fcc1597a838]
#> 
#> [[24]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g <- function(trim) identity(identity(fn(trim)))
g(2)

#> [[1]]
#> <frame 24> (23)
#> expr: g(2)
#> env:  [local 0x7fcc172f3fd8]
#> 
#> [[2]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fcc1a9cf6b8]
#> 
#> [[3]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fcc172f3e18]
#> 
#> [[4]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fcc172f3be8]
#> 
#> [[5]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fcc172f3638]
#> 
#> [[6]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fcc172f3088]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fcc172f2d40]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fcc172f29c0]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fcc172f81d0]
#> 
#> [[10]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fcc172f7fa0]
#> 
#> [[11]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fcc172f7e18]
#> 
#> [[12]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fcc172f7d38]
#> 
#> [[13]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fcc1730b7c0]
#> 
#> [[14]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fcc13828ff0]
#> 
#> [[15]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fcc13b45720]
#> 
#> [[16]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fcc13c1fb10]
#> 
#> [[17]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fcc17565530]
#> 
#> [[18]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fcc17564df8]
#> 
#> [[19]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fcc17592758]
#> 
#> [[20]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fcc177071b0]
#> 
#> [[21]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fcc178a8d68]
#> 
#> [[22]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fcc172c8390]
#> 
#> [[23]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fcc172c34b0]
#> 
#> [[24]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fcc1597a838]
#> 
#> [[25]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

g(3)

#> [[1]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fcc1a9cf6b8]
#> 
#> [[2]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fcc167338a8]
#> 
#> [[3]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fcc16733640]
#> 
#> [[4]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fcc16736fa0]
#> 
#> [[5]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fcc16736a28]
#> 
#> [[6]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fcc167366e0]
#> 
#> [[7]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fcc16736398]
#> 
#> [[8]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fcc16735c60]
#> 
#> [[9]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fcc16735a30]
#> 
#> [[10]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fcc167358a8]
#> 
#> [[11]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fcc167357c8]
#> 
#> [[12]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fcc16768018]
#> 
#> [[13]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fcc13828ff0]
#> 
#> [[14]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fcc13b45720]
#> 
#> [[15]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fcc13c1fb10]
#> 
#> [[16]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fcc17565530]
#> 
#> [[17]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fcc17564df8]
#> 
#> [[18]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fcc17592758]
#> 
#> [[19]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fcc177071b0]
#> 
#> [[20]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fcc178a8d68]
#> 
#> [[21]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fcc172c8390]
#> 
#> [[22]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fcc172c34b0]
#> 
#> [[23]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fcc1597a838]
#> 
#> [[24]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"