Trim top call layers from the evaluation stack

ctxt_stack() can be tricky to use in real code because all intervening frames are returned with the stack, including those at ctxt_stack() own call site. stack_trim() makes it easy to remove layers of intervening calls.

Usage

stack_trim(stack, n = 1)

Arguments

stack

An evaluation stack.

n

The number of call frames (not eval frames) to trim off the top of the stack. In other words, the number of layers of intervening frames to trim.

Life cycle

These functions are deprecated and replaced by trace_back().

Examples

# Intervening frames appear on the evaluation stack:
identity(identity(ctxt_stack()))
#> [[1]]
#> <frame 27> (25)
#> expr: identity(ctxt_stack())
#> env:  [local 0x558924692d10]
#> 
#> [[2]]
#> <frame 26> (25)
#> expr: identity(identity(ctxt_stack()))
#> env:  [local 0x558924692bf8]
#> 
#> [[3]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x558924930608]
#> 
#> [[4]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x558924692a38]
#> 
#> [[5]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x558924692808]
#> 
#> [[6]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x558924692258]
#> 
#> [[7]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x558924691ce0]
#> 
#> [[8]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x558924691998]
#> 
#> [[9]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x5589246915e0]
#> 
#> [[10]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x55892469c9c8]
#> 
#> [[11]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x55892469c798]
#> 
#> [[12]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x55892469c610]
#> 
#> [[13]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x55892469c530]
#> 
#> [[14]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x5589246dbc80]
#> 
#> [[15]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x558924a9eb60]
#> 
#> [[16]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x5589255ef840]
#> 
#> [[17]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x55892423f668]
#> 
#> [[18]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x558922e47840]
#> 
#> [[19]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x558923306300]
#> 
#> [[20]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x5589233099f8]
#> 
#> [[21]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x558925c289b0]
#> 
#> [[22]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x558920a29548]
#> 
#> [[23]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x558920067fb8]
#> 
#> [[24]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x55891f9d2cb0]
#> 
#> [[25]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x55891f9cea80]
#> 
#> [[26]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x55891e36e928]
#> 
#> [[27]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x55891cd03698]
#> 
#> [[28]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# stack_trim() will trim the first n layers of calls:
stack_trim(identity(identity(ctxt_stack())))
#> [[1]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x558924930608]
#> 
#> [[2]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x55892138ffa8]
#> 
#> [[3]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x55892138fd78]
#> 
#> [[4]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x55892138f7c8]
#> 
#> [[5]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x55892138f1a8]
#> 
#> [[6]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x55892138ee28]
#> 
#> [[7]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x55892138eae0]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x55892138e3e0]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x55892138e1b0]
#> 
#> [[10]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x558921391e20]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x558921391d40]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x5589213a4df0]
#> 
#> [[13]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x558924a9eb60]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x5589255ef840]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x55892423f668]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x558922e47840]
#> 
#> [[17]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x558923306300]
#> 
#> [[18]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x5589233099f8]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x558925c289b0]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x558920a29548]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x558920067fb8]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x55891f9d2cb0]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x55891f9cea80]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x55891e36e928]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x55891cd03698]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# Note that it also takes care of calls intervening at its own call
# site:
identity(identity(
  stack_trim(identity(identity(ctxt_stack())))
))
#> [[1]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x558924930608]
#> 
#> [[2]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5589210603b8]
#> 
#> [[3]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x5589210600a8]
#> 
#> [[4]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x55892105fac0]
#> 
#> [[5]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x55892105f4d8]
#> 
#> [[6]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x55892105f158]
#> 
#> [[7]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x55892105eda0]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x558921062428]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x5589210621c0]
#> 
#> [[10]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x558921062038]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x558921061f58]
#> 
#> [[12]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x558921083688]
#> 
#> [[13]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x558924a9eb60]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x5589255ef840]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x55892423f668]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x558922e47840]
#> 
#> [[17]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x558923306300]
#> 
#> [[18]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x5589233099f8]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x558925c289b0]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x558920a29548]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x558920067fb8]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x55891f9d2cb0]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x55891f9cea80]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x55891e36e928]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x55891cd03698]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# It is especially useful when used within a function that needs to
# inspect the evaluation stack but should nonetheless be callable
# within nested calls without side effects:
stack_util <- function() {
  # n = 2 means that two layers of intervening calls should be
  # removed: The layer at ctxt_stack()'s call site (including the
  # stack_trim() call), and the layer at stack_util()'s call.
  stack <- stack_trim(ctxt_stack(), n = 2)
  stack
}
user_fn <- function() {
  # A user calls your stack utility with intervening frames:
  identity(identity(stack_util()))
}
# These intervening frames won't appear in the evaluation stack
identity(user_fn())
#> [[1]]
#> <frame 27> (25)
#> expr: user_fn()
#> env:  [local 0x5589209f1228]
#> 
#> [[2]]
#> <frame 26> (25)
#> expr: identity(user_fn())
#> env:  [local 0x5589209f11b8]
#> 
#> [[3]]
#> <frame 25> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x558924930608]
#> 
#> [[4]]
#> <frame 24> (14)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x5589209f0ff8]
#> 
#> [[5]]
#> <frame 23> (14)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x5589209f0d90]
#> 
#> [[6]]
#> <frame 22> (14)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x5589209f4568]
#> 
#> [[7]]
#> <frame 21> (20)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x5589209f3f48]
#> 
#> [[8]]
#> <frame 20> (19)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x5589209f3bc8]
#> 
#> [[9]]
#> <frame 19> (18)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x5589209f3848]
#> 
#> [[10]]
#> <frame 18> (17)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x5589209f30d8]
#> 
#> [[11]]
#> <frame 17> (16)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x5589209f2e00]
#> 
#> [[12]]
#> <frame 16> (14)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x5589209f2c08]
#> 
#> [[13]]
#> <frame 15> (14)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x5589209f2b28]
#> 
#> [[14]]
#> <frame 14> (13)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x558920a24000]
#> 
#> [[15]]
#> <frame 13> (12)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#> env:  [frame 0x558924a9eb60]
#> 
#> [[16]]
#> <frame 12> (11)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x5589255ef840]
#> 
#> [[17]]
#> <frame 11> (10)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x55892423f668]
#> 
#> [[18]]
#> <frame 10> (9)
#> expr: run_examples(tags$tag_examples[[1]], env = if (is.null(examples_env)) NULL else new.env(parent = examples_env),  <...>
#> env:  [frame 0x558922e47840]
#> 
#> [[19]]
#> <frame 9> (7)
#> expr: data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x558923306300]
#> 
#> [[20]]
#> <frame 8> (7)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples_env = examples_env,  <...>
#> env:  [frame 0x5589233099f8]
#> 
#> [[21]]
#> <frame 7> (6)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x558925c289b0]
#> 
#> [[22]]
#> <frame 6> (5)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x558920a29548]
#> 
#> [[23]]
#> <frame 5> (4)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x558920067fb8]
#> 
#> [[24]]
#> <frame 4> (3)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x55891f9d2cb0]
#> 
#> [[25]]
#> <frame 3> (2)
#> expr: build_site(pkg, preview = FALSE, install = install, new_process = new_process,  <...>
#> env:  [frame 0x55891f9cea80]
#> 
#> [[26]]
#> <frame 2> (1)
#> expr: build_site_github_pages(pkg, dest_dir = dest_dir, ..., clean = clean)
#> env:  [frame 0x55891e36e928]
#> 
#> [[27]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x55891cd03698]
#> 
#> [[28]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"