Trim top call layers from the evaluation stack

ctxt_stack() can be tricky to use in real code because all intervening frames are returned with the stack, including those at ctxt_stack() own call site. stack_trim() makes it easy to remove layers of intervening calls.

stack_trim(stack, n = 1)

Arguments

stack
An evaluation stack.

n
The number of call frames (not eval frames) to trim off the top of the stack. In other words, the number of layers of intervening frames to trim.

Life cycle

These functions are soft-deprecated and replaced by trace_back().

Examples

# Intervening frames appear on the evaluation stack:
identity(identity(ctxt_stack()))

#> [[1]]
#> <frame 37> (35)
#> expr: identity(ctxt_stack())
#> env:  [local 0x7f85c399d5e8]
#> 
#> [[2]]
#> <frame 36> (35)
#> expr: identity(identity(ctxt_stack()))
#> env:  [local 0x7f85c399d3f0]
#> 
#> [[3]]
#> <frame 35> (34)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f85c504e858]
#> 
#> [[4]]
#> <frame 34> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f85c399d230]
#> 
#> [[5]]
#> <frame 33> (24)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f85c399cb68]
#> 
#> [[6]]
#> <frame 32> (24)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f85c399c388]
#> 
#> [[7]]
#> <frame 31> (30)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f85c399f7a8]
#> 
#> [[8]]
#> <frame 30> (29)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f85c399f348]
#> 
#> [[9]]
#> <frame 29> (28)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f85c399ebd8]
#> 
#> [[10]]
#> <frame 28> (27)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f85c399e1c8]
#> 
#> [[11]]
#> <frame 27> (26)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f85c3967390]
#> 
#> [[12]]
#> <frame 26> (24)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f85c3967208]
#> 
#> [[13]]
#> <frame 25> (24)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f85c39670f0]
#> 
#> [[14]]
#> <frame 24> (23)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f85c3c48ab0]
#> 
#> [[15]]
#> <frame 23> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x7f85c2fd2500]
#> 
#> [[16]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x7f85c2fd6688]
#> 
#> [[17]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x7f85c2fde340]
#> 
#> [[18]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f85c45a7070]
#> 
#> [[19]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f85c45ab380]
#> 
#> [[20]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f85c8cd1518]
#> 
#> [[21]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f85c427fa00]
#> 
#> [[22]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f85c9e96658]
#> 
#> [[23]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x7f85bf3b51d0]
#> 
#> [[24]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x7f85bfd2f170]
#> 
#> [[25]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x7f85bfd2df38]
#> 
#> [[26]]
#> <frame 12> (0)
#> expr: (function (...)  <...>
#> env:  [frame 0x7f85bd2c6fa0]
#> 
#> [[27]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x7f85bd2bcd00]
#> 
#> [[28]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/var/folders/b9/1vbq6rn93_1fk71sn95dqb8r0000gn/T//Rtmpgg6zPp/fileee10610f48f9"),  <...>
#> env:  [frame 0x7f85bd2b8f58]
#> 
#> [[29]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/var/folders/b9/1vbq6rn93_1fk71sn95dqb8r0000gn/T//Rtmpgg6zPp/fileee10610f48f9"),  <...>
#> env:  [frame 0x7f85bd2b9ce8]
#> 
#> [[30]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x7f85bd2b2010]
#> 
#> [[31]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x7f85bd2ae790]
#> 
#> [[32]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x7f85bd2aead8]
#> 
#> [[33]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x7f85bd2aee90]
#> 
#> [[34]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x7f85bd2af440]
#> 
#> [[35]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x7f85bd2af788]
#> 
#> [[36]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x7f85bd2afad0]
#> 
#> [[37]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x7f85bd2b0240]
#> 
#> [[38]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# stack_trim() will trim the first n layers of calls:
stack_trim(identity(identity(ctxt_stack())))

#> [[1]]
#> <frame 35> (34)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f85c504e858]
#> 
#> [[2]]
#> <frame 34> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f85c4cdd1c0]
#> 
#> [[3]]
#> <frame 33> (24)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f85c4cdcf90]
#> 
#> [[4]]
#> <frame 32> (24)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f85c4cdc858]
#> 
#> [[5]]
#> <frame 31> (30)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f85c4cdc2e0]
#> 
#> [[6]]
#> <frame 30> (29)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f85c4ce1ea8]
#> 
#> [[7]]
#> <frame 29> (28)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f85c4ce1a80]
#> 
#> [[8]]
#> <frame 28> (27)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f85c4ce12d8]
#> 
#> [[9]]
#> <frame 27> (26)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f85c4ce10a8]
#> 
#> [[10]]
#> <frame 26> (24)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f85c4ce0f20]
#> 
#> [[11]]
#> <frame 25> (24)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f85c4ce0e40]
#> 
#> [[12]]
#> <frame 24> (23)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f85c4d7c0e8]
#> 
#> [[13]]
#> <frame 23> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x7f85c2fd2500]
#> 
#> [[14]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x7f85c2fd6688]
#> 
#> [[15]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x7f85c2fde340]
#> 
#> [[16]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f85c45a7070]
#> 
#> [[17]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f85c45ab380]
#> 
#> [[18]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f85c8cd1518]
#> 
#> [[19]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f85c427fa00]
#> 
#> [[20]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f85c9e96658]
#> 
#> [[21]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x7f85bf3b51d0]
#> 
#> [[22]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x7f85bfd2f170]
#> 
#> [[23]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x7f85bfd2df38]
#> 
#> [[24]]
#> <frame 12> (0)
#> expr: (function (...)  <...>
#> env:  [frame 0x7f85bd2c6fa0]
#> 
#> [[25]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x7f85bd2bcd00]
#> 
#> [[26]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/var/folders/b9/1vbq6rn93_1fk71sn95dqb8r0000gn/T//Rtmpgg6zPp/fileee10610f48f9"),  <...>
#> env:  [frame 0x7f85bd2b8f58]
#> 
#> [[27]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/var/folders/b9/1vbq6rn93_1fk71sn95dqb8r0000gn/T//Rtmpgg6zPp/fileee10610f48f9"),  <...>
#> env:  [frame 0x7f85bd2b9ce8]
#> 
#> [[28]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x7f85bd2b2010]
#> 
#> [[29]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x7f85bd2ae790]
#> 
#> [[30]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x7f85bd2aead8]
#> 
#> [[31]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x7f85bd2aee90]
#> 
#> [[32]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x7f85bd2af440]
#> 
#> [[33]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x7f85bd2af788]
#> 
#> [[34]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x7f85bd2afad0]
#> 
#> [[35]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x7f85bd2b0240]
#> 
#> [[36]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# Note that it also takes care of calls intervening at its own call
# site:
identity(identity(
  stack_trim(identity(identity(ctxt_stack())))
))

#> [[1]]
#> <frame 35> (34)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f85c504e858]
#> 
#> [[2]]
#> <frame 34> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f85bf5bee20]
#> 
#> [[3]]
#> <frame 33> (24)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f85bf5be598]
#> 
#> [[4]]
#> <frame 32> (24)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f85bf5bdd80]
#> 
#> [[5]]
#> <frame 31> (30)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f85bf5c1050]
#> 
#> [[6]]
#> <frame 30> (29)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f85bf5c0bb8]
#> 
#> [[7]]
#> <frame 29> (28)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f85bf5c0528]
#> 
#> [[8]]
#> <frame 28> (27)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f85bf5c2b10]
#> 
#> [[9]]
#> <frame 27> (26)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f85bf5c27c8]
#> 
#> [[10]]
#> <frame 26> (24)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f85bf5c2598]
#> 
#> [[11]]
#> <frame 25> (24)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f85bf5c24b8]
#> 
#> [[12]]
#> <frame 24> (23)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f85bf601f40]
#> 
#> [[13]]
#> <frame 23> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x7f85c2fd2500]
#> 
#> [[14]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x7f85c2fd6688]
#> 
#> [[15]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x7f85c2fde340]
#> 
#> [[16]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f85c45a7070]
#> 
#> [[17]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f85c45ab380]
#> 
#> [[18]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f85c8cd1518]
#> 
#> [[19]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f85c427fa00]
#> 
#> [[20]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f85c9e96658]
#> 
#> [[21]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x7f85bf3b51d0]
#> 
#> [[22]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x7f85bfd2f170]
#> 
#> [[23]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x7f85bfd2df38]
#> 
#> [[24]]
#> <frame 12> (0)
#> expr: (function (...)  <...>
#> env:  [frame 0x7f85bd2c6fa0]
#> 
#> [[25]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x7f85bd2bcd00]
#> 
#> [[26]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/var/folders/b9/1vbq6rn93_1fk71sn95dqb8r0000gn/T//Rtmpgg6zPp/fileee10610f48f9"),  <...>
#> env:  [frame 0x7f85bd2b8f58]
#> 
#> [[27]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/var/folders/b9/1vbq6rn93_1fk71sn95dqb8r0000gn/T//Rtmpgg6zPp/fileee10610f48f9"),  <...>
#> env:  [frame 0x7f85bd2b9ce8]
#> 
#> [[28]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x7f85bd2b2010]
#> 
#> [[29]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x7f85bd2ae790]
#> 
#> [[30]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x7f85bd2aead8]
#> 
#> [[31]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x7f85bd2aee90]
#> 
#> [[32]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x7f85bd2af440]
#> 
#> [[33]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x7f85bd2af788]
#> 
#> [[34]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x7f85bd2afad0]
#> 
#> [[35]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x7f85bd2b0240]
#> 
#> [[36]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# It is especially useful when used within a function that needs to
# inspect the evaluation stack but should nonetheless be callable
# within nested calls without side effects:
stack_util <- function() {
  # n = 2 means that two layers of intervening calls should be
  # removed: The layer at ctxt_stack()'s call site (including the
  # stack_trim() call), and the layer at stack_util()'s call.
  stack <- stack_trim(ctxt_stack(), n = 2)
  stack
}
user_fn <- function() {
  # A user calls your stack utility with intervening frames:
  identity(identity(stack_util()))
}
# These intervening frames won't appear in the evaluation stack
identity(user_fn())

#> [[1]]
#> <frame 37> (35)
#> expr: user_fn()
#> env:  [local 0x7f85bc571e30]
#> 
#> [[2]]
#> <frame 36> (35)
#> expr: identity(user_fn())
#> env:  [local 0x7f85bc571dc0]
#> 
#> [[3]]
#> <frame 35> (34)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f85c504e858]
#> 
#> [[4]]
#> <frame 34> (24)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f85bc571c00]
#> 
#> [[5]]
#> <frame 33> (24)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f85bc575838]
#> 
#> [[6]]
#> <frame 32> (24)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f85bc574e28]
#> 
#> [[7]]
#> <frame 31> (30)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f85bc573ff0]
#> 
#> [[8]]
#> <frame 30> (29)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f85bc573b90]
#> 
#> [[9]]
#> <frame 29> (28)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f85bc577560]
#> 
#> [[10]]
#> <frame 28> (27)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f85bc5766f0]
#> 
#> [[11]]
#> <frame 27> (26)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f85bc5763a8]
#> 
#> [[12]]
#> <frame 26> (24)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f85bc5761e8]
#> 
#> [[13]]
#> <frame 25> (24)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f85bc5760d0]
#> 
#> [[14]]
#> <frame 24> (23)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f85bc5f3448]
#> 
#> [[15]]
#> <frame 23> (22)
#> expr: evaluate::evaluate(code, env, new_device = TRUE)
#> env:  [frame 0x7f85c2fd2500]
#> 
#> [[16]]
#> <frame 22> (21)
#> expr: .f(code = .l[[c(1L, 1L)]], run = .l[[c(2L, 1L)]], show = .l[[c(3L,  <...>
#> env:  [frame 0x7f85c2fd6688]
#> 
#> [[17]]
#> <frame 21> (20)
#> expr: purrr::pmap_chr(list(code = code, run = run, show = show), format_example_chunk,  <...>
#> env:  [frame 0x7f85c2fde340]
#> 
#> [[18]]
#> <frame 20> (18)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f85c45a7070]
#> 
#> [[19]]
#> <frame 19> (18)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f85c45ab380]
#> 
#> [[20]]
#> <frame 18> (17)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f85c8cd1518]
#> 
#> [[21]]
#> <frame 17> (16)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f85c427fa00]
#> 
#> [[22]]
#> <frame 16> (15)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f85c9e96658]
#> 
#> [[23]]
#> <frame 15> (14)
#> expr: build_reference(pkg, lazy = lazy, document = document, examples = examples,  <...>
#> env:  [frame 0x7f85bf3b51d0]
#> 
#> [[24]]
#> <frame 14> (13)
#> expr: build_site_local(pkg = pkg, examples = examples, document = document,  <...>
#> env:  [frame 0x7f85bfd2f170]
#> 
#> [[25]]
#> <frame 13> (12)
#> expr: pkgdown::build_site(...)
#> env:  [frame 0x7f85bfd2df38]
#> 
#> [[26]]
#> <frame 12> (0)
#> expr: (function (...)  <...>
#> env:  [frame 0x7f85bd2c6fa0]
#> 
#> [[27]]
#> <frame 11> (0)
#> expr: (function (what, args, quote = FALSE, envir = parent.frame())  <...>
#> env:  [frame 0x7f85bd2bcd00]
#> 
#> [[28]]
#> <frame 10> (0)
#> expr: do.call(do.call, c(readRDS("/var/folders/b9/1vbq6rn93_1fk71sn95dqb8r0000gn/T//Rtmpgg6zPp/fileee10610f48f9"),  <...>
#> env:  [frame 0x7f85bd2b8f58]
#> 
#> [[29]]
#> <frame 9> (0)
#> expr: saveRDS(do.call(do.call, c(readRDS("/var/folders/b9/1vbq6rn93_1fk71sn95dqb8r0000gn/T//Rtmpgg6zPp/fileee10610f48f9"),  <...>
#> env:  [frame 0x7f85bd2b9ce8]
#> 
#> [[30]]
#> <frame 8> (0)
#> expr: withCallingHandlers({ <...>
#> env:  [frame 0x7f85bd2b2010]
#> 
#> [[31]]
#> <frame 7> (6)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x7f85bd2ae790]
#> 
#> [[32]]
#> <frame 6> (5)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [frame 0x7f85bd2aead8]
#> 
#> [[33]]
#> <frame 5> (2)
#> expr: tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#> env:  [frame 0x7f85bd2aee90]
#> 
#> [[34]]
#> <frame 4> (3)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [frame 0x7f85bd2af440]
#> 
#> [[35]]
#> <frame 3> (2)
#> expr: tryCatchOne(tryCatchList(expr, names[-nh], parentenv, handlers[-nh]),  <...>
#> env:  [frame 0x7f85bd2af788]
#> 
#> [[36]]
#> <frame 2> (1)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [frame 0x7f85bd2afad0]
#> 
#> [[37]]
#> <frame 1> (0)
#> expr: tryCatch(withCallingHandlers({ <...>
#> env:  [frame 0x7f85bd2b0240]
#> 
#> [[38]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"