Trim top call layers from the evaluation stack

ctxt_stack() can be tricky to use in real code because all intervening frames are returned with the stack, including those at ctxt_stack() own call site. stack_trim() makes it easy to remove layers of intervening calls.

stack_trim(stack, n = 1)

Arguments

stack
An evaluation stack.

n
The number of call frames (not eval frames) to trim off the top of the stack. In other words, the number of layers of intervening frames to trim.

Life cycle

These functions are deprecated and replaced by trace_back().

Examples

# Intervening frames appear on the evaluation stack:
identity(identity(ctxt_stack()))

#> [[1]]
#> <frame 25> (23)
#> expr: identity(ctxt_stack())
#> env:  [local 0x7f9b34f5d838]
#> 
#> [[2]]
#> <frame 24> (23)
#> expr: identity(identity(ctxt_stack()))
#> env:  [local 0x7f9b34f5d950]
#> 
#> [[3]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f9b34ffb560]
#> 
#> [[4]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f9b34f5db10]
#> 
#> [[5]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f9b34f5dd40]
#> 
#> [[6]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f9b34f5e2f0]
#> 
#> [[7]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f9b34f5a958]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f9b34f5aca0]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f9b34f5afe8]
#> 
#> [[10]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f9b34f5b6b0]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f9b34f5b8e0]
#> 
#> [[12]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f9b34f5ba68]
#> 
#> [[13]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f9b34f5bb48]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f9b34f4b528]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f9b32e25620]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f9b32e1a580]
#> 
#> [[17]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f9b32e2ac10]
#> 
#> [[18]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f9b2e61e398]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7f9b2e61ea98]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f9b32a838b8]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f9b3055c9f0]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f9b31209f80]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f9b304b2718]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f9b304b1940]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f9b2c6b4a50]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# stack_trim() will trim the first n layers of calls:
stack_trim(identity(identity(ctxt_stack())))

#> [[1]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f9b34ffb560]
#> 
#> [[2]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f9b34b56d18]
#> 
#> [[3]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f9b34b59f08]
#> 
#> [[4]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f9b34b598e8]
#> 
#> [[5]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f9b34b5c3d8]
#> 
#> [[6]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f9b34b5c090]
#> 
#> [[7]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f9b34b5bcd8]
#> 
#> [[8]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f9b34b5b610]
#> 
#> [[9]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f9b34b5b3e0]
#> 
#> [[10]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f9b34b5b258]
#> 
#> [[11]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f9b34b5b178]
#> 
#> [[12]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f9b34ba0ea0]
#> 
#> [[13]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f9b32e25620]
#> 
#> [[14]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f9b32e1a580]
#> 
#> [[15]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f9b32e2ac10]
#> 
#> [[16]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f9b2e61e398]
#> 
#> [[17]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7f9b2e61ea98]
#> 
#> [[18]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f9b32a838b8]
#> 
#> [[19]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f9b3055c9f0]
#> 
#> [[20]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f9b31209f80]
#> 
#> [[21]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f9b304b2718]
#> 
#> [[22]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f9b304b1940]
#> 
#> [[23]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f9b2c6b4a50]
#> 
#> [[24]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# Note that it also takes care of calls intervening at its own call
# site:
identity(identity(
  stack_trim(identity(identity(ctxt_stack())))
))

#> [[1]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f9b34ffb560]
#> 
#> [[2]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f9b31b180c8]
#> 
#> [[3]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f9b31b17e98]
#> 
#> [[4]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f9b31b178e8]
#> 
#> [[5]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f9b31b17370]
#> 
#> [[6]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f9b31b16fb8]
#> 
#> [[7]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f9b31b16c70]
#> 
#> [[8]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f9b31b204b8]
#> 
#> [[9]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f9b31b20288]
#> 
#> [[10]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f9b31b20100]
#> 
#> [[11]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f9b31b1ffe8]
#> 
#> [[12]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f9b2e698690]
#> 
#> [[13]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f9b32e25620]
#> 
#> [[14]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f9b32e1a580]
#> 
#> [[15]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f9b32e2ac10]
#> 
#> [[16]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f9b2e61e398]
#> 
#> [[17]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7f9b2e61ea98]
#> 
#> [[18]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f9b32a838b8]
#> 
#> [[19]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f9b3055c9f0]
#> 
#> [[20]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f9b31209f80]
#> 
#> [[21]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f9b304b2718]
#> 
#> [[22]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f9b304b1940]
#> 
#> [[23]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f9b2c6b4a50]
#> 
#> [[24]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# It is especially useful when used within a function that needs to
# inspect the evaluation stack but should nonetheless be callable
# within nested calls without side effects:
stack_util <- function() {
  # n = 2 means that two layers of intervening calls should be
  # removed: The layer at ctxt_stack()'s call site (including the
  # stack_trim() call), and the layer at stack_util()'s call.
  stack <- stack_trim(ctxt_stack(), n = 2)
  stack
}
user_fn <- function() {
  # A user calls your stack utility with intervening frames:
  identity(identity(stack_util()))
}
# These intervening frames won't appear in the evaluation stack
identity(user_fn())

#> [[1]]
#> <frame 25> (23)
#> expr: user_fn()
#> env:  [local 0x7f9b3162a268]
#> 
#> [[2]]
#> <frame 24> (23)
#> expr: identity(user_fn())
#> env:  [local 0x7f9b3162a1f8]
#> 
#> [[3]]
#> <frame 23> (22)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f9b34ffb560]
#> 
#> [[4]]
#> <frame 22> (12)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f9b3162a038]
#> 
#> [[5]]
#> <frame 21> (12)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f9b31629dd0]
#> 
#> [[6]]
#> <frame 20> (12)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f9b31629820]
#> 
#> [[7]]
#> <frame 19> (18)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f9b31629270]
#> 
#> [[8]]
#> <frame 18> (17)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f9b3162cdc8]
#> 
#> [[9]]
#> <frame 17> (16)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f9b3162ca80]
#> 
#> [[10]]
#> <frame 16> (15)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f9b3162c3b8]
#> 
#> [[11]]
#> <frame 15> (14)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f9b3162c188]
#> 
#> [[12]]
#> <frame 14> (12)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f9b3162c000]
#> 
#> [[13]]
#> <frame 13> (12)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f9b3162bf20]
#> 
#> [[14]]
#> <frame 12> (11)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f9b316414d8]
#> 
#> [[15]]
#> <frame 11> (10)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7f9b32e25620]
#> 
#> [[16]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7f9b32e1a580]
#> 
#> [[17]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f9b32e2ac10]
#> 
#> [[18]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7f9b2e61e398]
#> 
#> [[19]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7f9b2e61ea98]
#> 
#> [[20]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f9b32a838b8]
#> 
#> [[21]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f9b3055c9f0]
#> 
#> [[22]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f9b31209f80]
#> 
#> [[23]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f9b304b2718]
#> 
#> [[24]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7f9b304b1940]
#> 
#> [[25]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7f9b2c6b4a50]
#> 
#> [[26]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"