Trim top call layers from the evaluation stack

ctxt_stack() can be tricky to use in real code because all intervening frames are returned with the stack, including those at ctxt_stack() own call site. stack_trim() makes it easy to remove layers of intervening calls.

stack_trim(stack, n = 1)

Arguments

stack
An evaluation stack.

n
The number of call frames (not eval frames) to trim off the top of the stack. In other words, the number of layers of intervening frames to trim.

Examples

# Intervening frames appear on the evaluation stack:
identity(identity(ctxt_stack()))

#> [[1]]
#> <frame 33> (31)
#> expr: identity(ctxt_stack())
#> env:  [local 0x7f848aa803c0]
#> 
#> [[2]]
#> <frame 32> (31)
#> expr: identity(identity(ctxt_stack()))
#> env:  [local 0x7f848aa80270]
#> 
#> [[3]]
#> <frame 31> (30)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f848c1d26d0]
#> 
#> [[4]]
#> <frame 30> (20)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f848aa80040]
#> 
#> [[5]]
#> <frame 29> (20)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f848aa81430]
#> 
#> [[6]]
#> <frame 28> (20)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f848aa82548]
#> 
#> [[7]]
#> <frame 27> (26)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f848aa83660]
#> 
#> [[8]]
#> <frame 26> (25)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f848aa83078]
#> 
#> [[9]]
#> <frame 25> (24)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f848aa843f8]
#> 
#> [[10]]
#> <frame 24> (23)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f848aa848e8]
#> 
#> [[11]]
#> <frame 23> (22)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f848aa85698]
#> 
#> [[12]]
#> <frame 22> (20)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f848aa85008]
#> 
#> [[13]]
#> <frame 21> (20)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f848aa85e60]
#> 
#> [[14]]
#> <frame 20> (19)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f8489c50408]
#> 
#> [[15]]
#> <frame 19> (18)
#> expr: evaluate::evaluate(text, code_env, new_device = TRUE)
#> env:  [frame 0x7f848bcd41c8]
#> 
#> [[16]]
#> <frame 18> (16)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f848d225510]
#> 
#> [[17]]
#> <frame 17> (16)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f848bb42468]
#> 
#> [[18]]
#> <frame 16> (12)
#> expr: data_reference_topic(topic, pkg, path = path, examples = examples,  <...>
#> env:  [frame 0x7f848c255a00]
#> 
#> [[19]]
#> <frame 15> (14)
#> expr: stopifnot(is.list(x), is.list(val))
#> env:  [frame 0x7f848c770aa0]
#> 
#> [[20]]
#> <frame 14> (13)
#> expr: utils::modifyList(data, data_template(pkg, depth = depth))
#> env:  [frame 0x7f848c770870]
#> 
#> [[21]]
#> <frame 13> (12)
#> expr: render_page(pkg, "reference-topic", data = data_reference_topic(topic,  <...>
#> env:  [frame 0x7f848c1ccb88]
#> 
#> [[22]]
#> <frame 12> (11)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f848c56c438]
#> 
#> [[23]]
#> <frame 11> (10)
#> expr: purrr::map(., build_reference_topic, path, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f848c08b1c8]
#> 
#> [[24]]
#> <frame 10> (8)
#> expr: function_list[[k]](value)
#> env:  [frame 0x7f848c092ca0]
#> 
#> [[25]]
#> <frame 9> (8)
#> expr: withVisible(function_list[[k]](value))
#> env:  [frame 0x7f848c092b50]
#> 
#> [[26]]
#> <frame 8> (7)
#> expr: freduce(value, `_function_list`)
#> env:  [frame 0x7f848c0b74a0]
#> 
#> [[27]]
#> <frame 7> (6)
#> expr: `_fseq`(`_lhs`)
#> env:  [frame 0x7f848c0b7270]
#> 
#> [[28]]
#> <frame 6> (5)
#> expr: eval(expr, envir, enclos)
#> env:  [frame 0x7f848baacca0]
#> 
#> [[29]]
#> <frame 5> (3)
#> expr: eval(quote(`_fseq`(`_lhs`)), env, env)
#> env:  [frame 0x7f848c0b70b0]
#> 
#> [[30]]
#> <frame 4> (3)
#> expr: withVisible(eval(quote(`_fseq`(`_lhs`)), env, env))
#> env:  [frame 0x7f848b96abf0]
#> 
#> [[31]]
#> <frame 3> (2)
#> expr: pkg$topics %>% purrr::transpose() %>% purrr::map(build_reference_topic,  <...>
#> env:  [frame 0x7f848baac8b0]
#> 
#> [[32]]
#> <frame 2> (1)
#> expr: build_reference(pkg, lazy = FALSE, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f848b919aa0]
#> 
#> [[33]]
#> <frame 1> (0)
#> expr: pkgdown::build_site("/Users/lionel/Dropbox/R/hadley/rlang")
#> env:  [frame 0x7f8489f5c2a8]
#> 
#> [[34]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# stack_trim() will trim the first n layers of calls:
stack_trim(identity(identity(ctxt_stack())))

#> [[1]]
#> <frame 31> (30)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f848c1d26d0]
#> 
#> [[2]]
#> <frame 30> (20)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f84962c17c8]
#> 
#> [[3]]
#> <frame 29> (20)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f84962c1a30]
#> 
#> [[4]]
#> <frame 28> (20)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f84963dca70]
#> 
#> [[5]]
#> <frame 27> (26)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f84963dc078]
#> 
#> [[6]]
#> <frame 26> (25)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f84963dc3c0]
#> 
#> [[7]]
#> <frame 25> (24)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f84963dc708]
#> 
#> [[8]]
#> <frame 24> (23)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f84963dbe28]
#> 
#> [[9]]
#> <frame 23> (22)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f84963db078]
#> 
#> [[10]]
#> <frame 22> (20)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f84963db238]
#> 
#> [[11]]
#> <frame 21> (20)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f84963db350]
#> 
#> [[12]]
#> <frame 20> (19)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f84962ae3b8]
#> 
#> [[13]]
#> <frame 19> (18)
#> expr: evaluate::evaluate(text, code_env, new_device = TRUE)
#> env:  [frame 0x7f848bcd41c8]
#> 
#> [[14]]
#> <frame 18> (16)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f848d225510]
#> 
#> [[15]]
#> <frame 17> (16)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f848bb42468]
#> 
#> [[16]]
#> <frame 16> (12)
#> expr: data_reference_topic(topic, pkg, path = path, examples = examples,  <...>
#> env:  [frame 0x7f848c255a00]
#> 
#> [[17]]
#> <frame 15> (14)
#> expr: stopifnot(is.list(x), is.list(val))
#> env:  [frame 0x7f848c770aa0]
#> 
#> [[18]]
#> <frame 14> (13)
#> expr: utils::modifyList(data, data_template(pkg, depth = depth))
#> env:  [frame 0x7f848c770870]
#> 
#> [[19]]
#> <frame 13> (12)
#> expr: render_page(pkg, "reference-topic", data = data_reference_topic(topic,  <...>
#> env:  [frame 0x7f848c1ccb88]
#> 
#> [[20]]
#> <frame 12> (11)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f848c56c438]
#> 
#> [[21]]
#> <frame 11> (10)
#> expr: purrr::map(., build_reference_topic, path, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f848c08b1c8]
#> 
#> [[22]]
#> <frame 10> (8)
#> expr: function_list[[k]](value)
#> env:  [frame 0x7f848c092ca0]
#> 
#> [[23]]
#> <frame 9> (8)
#> expr: withVisible(function_list[[k]](value))
#> env:  [frame 0x7f848c092b50]
#> 
#> [[24]]
#> <frame 8> (7)
#> expr: freduce(value, `_function_list`)
#> env:  [frame 0x7f848c0b74a0]
#> 
#> [[25]]
#> <frame 7> (6)
#> expr: `_fseq`(`_lhs`)
#> env:  [frame 0x7f848c0b7270]
#> 
#> [[26]]
#> <frame 6> (5)
#> expr: eval(expr, envir, enclos)
#> env:  [frame 0x7f848baacca0]
#> 
#> [[27]]
#> <frame 5> (3)
#> expr: eval(quote(`_fseq`(`_lhs`)), env, env)
#> env:  [frame 0x7f848c0b70b0]
#> 
#> [[28]]
#> <frame 4> (3)
#> expr: withVisible(eval(quote(`_fseq`(`_lhs`)), env, env))
#> env:  [frame 0x7f848b96abf0]
#> 
#> [[29]]
#> <frame 3> (2)
#> expr: pkg$topics %>% purrr::transpose() %>% purrr::map(build_reference_topic,  <...>
#> env:  [frame 0x7f848baac8b0]
#> 
#> [[30]]
#> <frame 2> (1)
#> expr: build_reference(pkg, lazy = FALSE, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f848b919aa0]
#> 
#> [[31]]
#> <frame 1> (0)
#> expr: pkgdown::build_site("/Users/lionel/Dropbox/R/hadley/rlang")
#> env:  [frame 0x7f8489f5c2a8]
#> 
#> [[32]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# Note that it also takes care of calls intervening at its own call
# site:
identity(identity(
  stack_trim(identity(identity(ctxt_stack())))
))

#> [[1]]
#> <frame 31> (30)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f848c1d26d0]
#> 
#> [[2]]
#> <frame 30> (20)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f8494834f80]
#> 
#> [[3]]
#> <frame 29> (20)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f8494834ca8]
#> 
#> [[4]]
#> <frame 28> (20)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f8494e9b9c0]
#> 
#> [[5]]
#> <frame 27> (26)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f849491bd80]
#> 
#> [[6]]
#> <frame 26> (25)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f849491b9c8]
#> 
#> [[7]]
#> <frame 25> (24)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f8494a1f028]
#> 
#> [[8]]
#> <frame 24> (23)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f848ad02898]
#> 
#> [[9]]
#> <frame 23> (22)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f848ad02668]
#> 
#> [[10]]
#> <frame 22> (20)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f848ad02438]
#> 
#> [[11]]
#> <frame 21> (20)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f848ad02320]
#> 
#> [[12]]
#> <frame 20> (19)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f848c5cfcd8]
#> 
#> [[13]]
#> <frame 19> (18)
#> expr: evaluate::evaluate(text, code_env, new_device = TRUE)
#> env:  [frame 0x7f848bcd41c8]
#> 
#> [[14]]
#> <frame 18> (16)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f848d225510]
#> 
#> [[15]]
#> <frame 17> (16)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f848bb42468]
#> 
#> [[16]]
#> <frame 16> (12)
#> expr: data_reference_topic(topic, pkg, path = path, examples = examples,  <...>
#> env:  [frame 0x7f848c255a00]
#> 
#> [[17]]
#> <frame 15> (14)
#> expr: stopifnot(is.list(x), is.list(val))
#> env:  [frame 0x7f848c770aa0]
#> 
#> [[18]]
#> <frame 14> (13)
#> expr: utils::modifyList(data, data_template(pkg, depth = depth))
#> env:  [frame 0x7f848c770870]
#> 
#> [[19]]
#> <frame 13> (12)
#> expr: render_page(pkg, "reference-topic", data = data_reference_topic(topic,  <...>
#> env:  [frame 0x7f848c1ccb88]
#> 
#> [[20]]
#> <frame 12> (11)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f848c56c438]
#> 
#> [[21]]
#> <frame 11> (10)
#> expr: purrr::map(., build_reference_topic, path, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f848c08b1c8]
#> 
#> [[22]]
#> <frame 10> (8)
#> expr: function_list[[k]](value)
#> env:  [frame 0x7f848c092ca0]
#> 
#> [[23]]
#> <frame 9> (8)
#> expr: withVisible(function_list[[k]](value))
#> env:  [frame 0x7f848c092b50]
#> 
#> [[24]]
#> <frame 8> (7)
#> expr: freduce(value, `_function_list`)
#> env:  [frame 0x7f848c0b74a0]
#> 
#> [[25]]
#> <frame 7> (6)
#> expr: `_fseq`(`_lhs`)
#> env:  [frame 0x7f848c0b7270]
#> 
#> [[26]]
#> <frame 6> (5)
#> expr: eval(expr, envir, enclos)
#> env:  [frame 0x7f848baacca0]
#> 
#> [[27]]
#> <frame 5> (3)
#> expr: eval(quote(`_fseq`(`_lhs`)), env, env)
#> env:  [frame 0x7f848c0b70b0]
#> 
#> [[28]]
#> <frame 4> (3)
#> expr: withVisible(eval(quote(`_fseq`(`_lhs`)), env, env))
#> env:  [frame 0x7f848b96abf0]
#> 
#> [[29]]
#> <frame 3> (2)
#> expr: pkg$topics %>% purrr::transpose() %>% purrr::map(build_reference_topic,  <...>
#> env:  [frame 0x7f848baac8b0]
#> 
#> [[30]]
#> <frame 2> (1)
#> expr: build_reference(pkg, lazy = FALSE, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f848b919aa0]
#> 
#> [[31]]
#> <frame 1> (0)
#> expr: pkgdown::build_site("/Users/lionel/Dropbox/R/hadley/rlang")
#> env:  [frame 0x7f8489f5c2a8]
#> 
#> [[32]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# It is especially useful when used within a function that needs to
# inspect the evaluation stack but should nonetheless be callable
# within nested calls without side effects:
stack_util <- function() {
  # n = 2 means that two layers of intervening calls should be
  # removed: The layer at ctxt_stack()'s call site (including the
  # stack_trim() call), and the layer at stack_util()'s call.
  stack <- stack_trim(ctxt_stack(), n = 2)
  stack
}
user_fn <- function() {
  # A user calls your stack utility with intervening frames:
  identity(identity(stack_util()))
}
# These intervening frames won't appear in the evaluation stack
identity(user_fn())

#> [[1]]
#> <frame 33> (31)
#> expr: user_fn()
#> env:  [local 0x7f84919322d8]
#> 
#> [[2]]
#> <frame 32> (31)
#> expr: identity(user_fn())
#> env:  [local 0x7f8491932268]
#> 
#> [[3]]
#> <frame 31> (30)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f848c1d26d0]
#> 
#> [[4]]
#> <frame 30> (20)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7f84919320a8]
#> 
#> [[5]]
#> <frame 29> (20)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7f8491931e08]
#> 
#> [[6]]
#> <frame 28> (20)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7f8491930f58]
#> 
#> [[7]]
#> <frame 27> (26)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7f8491fa6238]
#> 
#> [[8]]
#> <frame 26> (25)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7f848e3b2828]
#> 
#> [[9]]
#> <frame 25> (24)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7f848e3b24a8]
#> 
#> [[10]]
#> <frame 24> (23)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7f848e3b0d88]
#> 
#> [[11]]
#> <frame 23> (22)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7f848e3b0b58]
#> 
#> [[12]]
#> <frame 22> (20)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7f848d7278d0]
#> 
#> [[13]]
#> <frame 21> (20)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7f848d7277b8]
#> 
#> [[14]]
#> <frame 20> (19)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7f8491f55b88]
#> 
#> [[15]]
#> <frame 19> (18)
#> expr: evaluate::evaluate(text, code_env, new_device = TRUE)
#> env:  [frame 0x7f848bcd41c8]
#> 
#> [[16]]
#> <frame 18> (16)
#> expr: as_data.tag_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f848d225510]
#> 
#> [[17]]
#> <frame 17> (16)
#> expr: as_data(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7f848bb42468]
#> 
#> [[18]]
#> <frame 16> (12)
#> expr: data_reference_topic(topic, pkg, path = path, examples = examples,  <...>
#> env:  [frame 0x7f848c255a00]
#> 
#> [[19]]
#> <frame 15> (14)
#> expr: stopifnot(is.list(x), is.list(val))
#> env:  [frame 0x7f848c770aa0]
#> 
#> [[20]]
#> <frame 14> (13)
#> expr: utils::modifyList(data, data_template(pkg, depth = depth))
#> env:  [frame 0x7f848c770870]
#> 
#> [[21]]
#> <frame 13> (12)
#> expr: render_page(pkg, "reference-topic", data = data_reference_topic(topic,  <...>
#> env:  [frame 0x7f848c1ccb88]
#> 
#> [[22]]
#> <frame 12> (11)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7f848c56c438]
#> 
#> [[23]]
#> <frame 11> (10)
#> expr: purrr::map(., build_reference_topic, path, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7f848c08b1c8]
#> 
#> [[24]]
#> <frame 10> (8)
#> expr: function_list[[k]](value)
#> env:  [frame 0x7f848c092ca0]
#> 
#> [[25]]
#> <frame 9> (8)
#> expr: withVisible(function_list[[k]](value))
#> env:  [frame 0x7f848c092b50]
#> 
#> [[26]]
#> <frame 8> (7)
#> expr: freduce(value, `_function_list`)
#> env:  [frame 0x7f848c0b74a0]
#> 
#> [[27]]
#> <frame 7> (6)
#> expr: `_fseq`(`_lhs`)
#> env:  [frame 0x7f848c0b7270]
#> 
#> [[28]]
#> <frame 6> (5)
#> expr: eval(expr, envir, enclos)
#> env:  [frame 0x7f848baacca0]
#> 
#> [[29]]
#> <frame 5> (3)
#> expr: eval(quote(`_fseq`(`_lhs`)), env, env)
#> env:  [frame 0x7f848c0b70b0]
#> 
#> [[30]]
#> <frame 4> (3)
#> expr: withVisible(eval(quote(`_fseq`(`_lhs`)), env, env))
#> env:  [frame 0x7f848b96abf0]
#> 
#> [[31]]
#> <frame 3> (2)
#> expr: pkg$topics %>% purrr::transpose() %>% purrr::map(build_reference_topic,  <...>
#> env:  [frame 0x7f848baac8b0]
#> 
#> [[32]]
#> <frame 2> (1)
#> expr: build_reference(pkg, lazy = FALSE, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7f848b919aa0]
#> 
#> [[33]]
#> <frame 1> (0)
#> expr: pkgdown::build_site("/Users/lionel/Dropbox/R/hadley/rlang")
#> env:  [frame 0x7f8489f5c2a8]
#> 
#> [[34]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"