Trim top call layers from the evaluation stack

ctxt_stack() can be tricky to use in real code because all intervening frames are returned with the stack, including those at ctxt_stack() own call site. stack_trim() makes it easy to remove layers of intervening calls.

stack_trim(stack, n = 1)

Arguments

stack
An evaluation stack.

n
The number of call frames (not eval frames) to trim off the top of the stack. In other words, the number of layers of intervening frames to trim.

Life cycle

These functions are deprecated and replaced by trace_back().

Examples

# Intervening frames appear on the evaluation stack:
identity(identity(ctxt_stack()))

#> [[1]]
#> <frame 26> (24)
#> expr: identity(ctxt_stack())
#> env:  [local 0x7fd7a140efe0]
#> 
#> [[2]]
#> <frame 25> (24)
#> expr: identity(identity(ctxt_stack()))
#> env:  [local 0x7fd7a140f0f8]
#> 
#> [[3]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fd79f7d0890]
#> 
#> [[4]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fd7a140f2b8]
#> 
#> [[5]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fd7a140f4e8]
#> 
#> [[6]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fd7a140bb88]
#> 
#> [[7]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fd7a140c100]
#> 
#> [[8]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fd7a140c448]
#> 
#> [[9]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fd7a140c790]
#> 
#> [[10]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fd7a140ce58]
#> 
#> [[11]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fd7a140d088]
#> 
#> [[12]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fd7a140d210]
#> 
#> [[13]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fd7a140d2f0]
#> 
#> [[14]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fd79f7dc740]
#> 
#> [[15]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fd79f670808]
#> 
#> [[16]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7fd7a0a4aa28]
#> 
#> [[17]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fd79a13e588]
#> 
#> [[18]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fd79a31a520]
#> 
#> [[19]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fd79be2da38]
#> 
#> [[20]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fd79be311a0]
#> 
#> [[21]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fd79be66218]
#> 
#> [[22]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fd7a0313e70]
#> 
#> [[23]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fd79d506738]
#> 
#> [[24]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fd79b5d7798]
#> 
#> [[25]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fd79b5de490]
#> 
#> [[26]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fd79d3498a8]
#> 
#> [[27]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# stack_trim() will trim the first n layers of calls:
stack_trim(identity(identity(ctxt_stack())))

#> [[1]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fd79f7d0890]
#> 
#> [[2]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fd7a303a4c0]
#> 
#> [[3]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fd7a303a6f0]
#> 
#> [[4]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fd7a303aca0]
#> 
#> [[5]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fd7a303b218]
#> 
#> [[6]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fd7a303b560]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fd7a303b8a8]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fd7a3038060]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fd7a3038290]
#> 
#> [[10]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fd7a3038418]
#> 
#> [[11]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fd7a30384f8]
#> 
#> [[12]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fd7a3027f10]
#> 
#> [[13]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fd79f670808]
#> 
#> [[14]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7fd7a0a4aa28]
#> 
#> [[15]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fd79a13e588]
#> 
#> [[16]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fd79a31a520]
#> 
#> [[17]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fd79be2da38]
#> 
#> [[18]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fd79be311a0]
#> 
#> [[19]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fd79be66218]
#> 
#> [[20]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fd7a0313e70]
#> 
#> [[21]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fd79d506738]
#> 
#> [[22]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fd79b5d7798]
#> 
#> [[23]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fd79b5de490]
#> 
#> [[24]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fd79d3498a8]
#> 
#> [[25]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# Note that it also takes care of calls intervening at its own call
# site:
identity(identity(
  stack_trim(identity(identity(ctxt_stack())))
))

#> [[1]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fd79f7d0890]
#> 
#> [[2]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fd7a31b0928]
#> 
#> [[3]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fd7a31b0b58]
#> 
#> [[4]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fd7a31ad1f8]
#> 
#> [[5]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fd7a31ad770]
#> 
#> [[6]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fd7a31adab8]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fd7a31ade00]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fd7a31ae4c8]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fd7a31ae6f8]
#> 
#> [[10]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fd7a31ae880]
#> 
#> [[11]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fd7a31ae960]
#> 
#> [[12]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fd7a319e2d0]
#> 
#> [[13]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fd79f670808]
#> 
#> [[14]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7fd7a0a4aa28]
#> 
#> [[15]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fd79a13e588]
#> 
#> [[16]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fd79a31a520]
#> 
#> [[17]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fd79be2da38]
#> 
#> [[18]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fd79be311a0]
#> 
#> [[19]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fd79be66218]
#> 
#> [[20]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fd7a0313e70]
#> 
#> [[21]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fd79d506738]
#> 
#> [[22]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fd79b5d7798]
#> 
#> [[23]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fd79b5de490]
#> 
#> [[24]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fd79d3498a8]
#> 
#> [[25]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# It is especially useful when used within a function that needs to
# inspect the evaluation stack but should nonetheless be callable
# within nested calls without side effects:
stack_util <- function() {
  # n = 2 means that two layers of intervening calls should be
  # removed: The layer at ctxt_stack()'s call site (including the
  # stack_trim() call), and the layer at stack_util()'s call.
  stack <- stack_trim(ctxt_stack(), n = 2)
  stack
}
user_fn <- function() {
  # A user calls your stack utility with intervening frames:
  identity(identity(stack_util()))
}
# These intervening frames won't appear in the evaluation stack
identity(user_fn())

#> [[1]]
#> <frame 26> (24)
#> expr: user_fn()
#> env:  [local 0x7fd7a3408ae8]
#> 
#> [[2]]
#> <frame 25> (24)
#> expr: identity(user_fn())
#> env:  [local 0x7fd7a3408b58]
#> 
#> [[3]]
#> <frame 24> (23)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fd79f7d0890]
#> 
#> [[4]]
#> <frame 23> (13)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fd7a3404e08]
#> 
#> [[5]]
#> <frame 22> (13)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fd7a3405038]
#> 
#> [[6]]
#> <frame 21> (13)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fd7a34055e8]
#> 
#> [[7]]
#> <frame 20> (19)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fd7a3405b60]
#> 
#> [[8]]
#> <frame 19> (18)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fd7a3405ea8]
#> 
#> [[9]]
#> <frame 18> (17)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fd7a34061f0]
#> 
#> [[10]]
#> <frame 17> (16)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fd7a34068b8]
#> 
#> [[11]]
#> <frame 16> (15)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fd7a3406ae8]
#> 
#> [[12]]
#> <frame 15> (13)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fd7a3406c70]
#> 
#> [[13]]
#> <frame 14> (13)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fd7a3402e40]
#> 
#> [[14]]
#> <frame 13> (12)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fd7a33f6ae8]
#> 
#> [[15]]
#> <frame 12> (11)
#> expr: evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fd79f670808]
#> 
#> [[16]]
#> <frame 11> (10)
#> expr: downlit::evaluate_and_highlight(code, fig_save = fig_save_topic,  <...>
#> env:  [frame 0x7fd7a0a4aa28]
#> 
#> [[17]]
#> <frame 10> (9)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fd79a13e588]
#> 
#> [[18]]
#> <frame 9> (8)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fd79a31a520]
#> 
#> [[19]]
#> <frame 8> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fd79be2da38]
#> 
#> [[20]]
#> <frame 7> (6)
#> expr: withCallingHandlers(data_reference_topic(topic, pkg, examples = examples,  <...>
#> env:  [frame 0x7fd79be311a0]
#> 
#> [[21]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fd79be66218]
#> 
#> [[22]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fd7a0313e70]
#> 
#> [[23]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fd79d506738]
#> 
#> [[24]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fd79b5d7798]
#> 
#> [[25]]
#> <frame 2> (1)
#> expr: build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE,  <...>
#> env:  [frame 0x7fd79b5de490]
#> 
#> [[26]]
#> <frame 1> (0)
#> expr: pkgdown::deploy_to_branch(new_process = FALSE)
#> env:  [frame 0x7fd79d3498a8]
#> 
#> [[27]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"