Trim top call layers from the evaluation stack

ctxt_stack() can be tricky to use in real code because all intervening frames are returned with the stack, including those at ctxt_stack() own call site. stack_trim() makes it easy to remove layers of intervening calls.

stack_trim(stack, n = 1)

Arguments

stack
An evaluation stack.

n
The number of call frames (not eval frames) to trim off the top of the stack. In other words, the number of layers of intervening frames to trim.

Life cycle

These functions are deprecated and replaced by trace_back().

Examples

# Intervening frames appear on the evaluation stack:
identity(identity(ctxt_stack()))

#> [[1]]
#> <frame 24> (22)
#> expr: identity(ctxt_stack())
#> env:  [local 0x7fa65fb6d048]
#> 
#> [[2]]
#> <frame 23> (22)
#> expr: identity(identity(ctxt_stack()))
#> env:  [local 0x7fa65fb6c4b0]
#> 
#> [[3]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fa65f2d2758]
#> 
#> [[4]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fa65fb6d8e0]
#> 
#> [[5]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fa65fb72f68]
#> 
#> [[6]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fa65fb72868]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fa65fb722b8]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fa65fb71f70]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fa65fb74ad0]
#> 
#> [[10]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fa65fb74398]
#> 
#> [[11]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fa65fb74168]
#> 
#> [[12]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fa65fb73fe0]
#> 
#> [[13]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fa65fb73f00]
#> 
#> [[14]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fa65fbaa6a8]
#> 
#> [[15]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fa6603785d8]
#> 
#> [[16]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fa660bcc388]
#> 
#> [[17]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fa6606fc158]
#> 
#> [[18]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fa662f26680]
#> 
#> [[19]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fa662eb4f50]
#> 
#> [[20]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fa65a09e4f8]
#> 
#> [[21]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fa65e9d0910]
#> 
#> [[22]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fa65b6abf30]
#> 
#> [[23]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7fa65d322948]
#> 
#> [[24]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7fa65d091540]
#> 
#> [[25]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# stack_trim() will trim the first n layers of calls:
stack_trim(identity(identity(ctxt_stack())))

#> [[1]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fa65f2d2758]
#> 
#> [[2]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fa65db3ae98]
#> 
#> [[3]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fa65db3ac68]
#> 
#> [[4]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fa65db3a648]
#> 
#> [[5]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fa65db3a098]
#> 
#> [[6]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fa65db39d18]
#> 
#> [[7]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fa65dedfee0]
#> 
#> [[8]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fa65dedf770]
#> 
#> [[9]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fa65dedf540]
#> 
#> [[10]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fa65dedf3b8]
#> 
#> [[11]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fa65dedf2d8]
#> 
#> [[12]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fa65dd71410]
#> 
#> [[13]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fa6603785d8]
#> 
#> [[14]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fa660bcc388]
#> 
#> [[15]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fa6606fc158]
#> 
#> [[16]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fa662f26680]
#> 
#> [[17]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fa662eb4f50]
#> 
#> [[18]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fa65a09e4f8]
#> 
#> [[19]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fa65e9d0910]
#> 
#> [[20]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fa65b6abf30]
#> 
#> [[21]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7fa65d322948]
#> 
#> [[22]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7fa65d091540]
#> 
#> [[23]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# Note that it also takes care of calls intervening at its own call
# site:
identity(identity(
  stack_trim(identity(identity(ctxt_stack())))
))

#> [[1]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fa65f2d2758]
#> 
#> [[2]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fa65f87eb30]
#> 
#> [[3]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fa65f87e900]
#> 
#> [[4]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fa65f87e318]
#> 
#> [[5]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fa65f881c40]
#> 
#> [[6]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fa65f881888]
#> 
#> [[7]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fa65f881508]
#> 
#> [[8]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fa65f880d98]
#> 
#> [[9]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fa65f880b30]
#> 
#> [[10]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fa65f8809a8]
#> 
#> [[11]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fa65f8808c8]
#> 
#> [[12]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fa65f899b28]
#> 
#> [[13]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fa6603785d8]
#> 
#> [[14]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fa660bcc388]
#> 
#> [[15]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fa6606fc158]
#> 
#> [[16]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fa662f26680]
#> 
#> [[17]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fa662eb4f50]
#> 
#> [[18]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fa65a09e4f8]
#> 
#> [[19]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fa65e9d0910]
#> 
#> [[20]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fa65b6abf30]
#> 
#> [[21]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7fa65d322948]
#> 
#> [[22]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7fa65d091540]
#> 
#> [[23]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"     

# It is especially useful when used within a function that needs to
# inspect the evaluation stack but should nonetheless be callable
# within nested calls without side effects:
stack_util <- function() {
  # n = 2 means that two layers of intervening calls should be
  # removed: The layer at ctxt_stack()'s call site (including the
  # stack_trim() call), and the layer at stack_util()'s call.
  stack <- stack_trim(ctxt_stack(), n = 2)
  stack
}
user_fn <- function() {
  # A user calls your stack utility with intervening frames:
  identity(identity(stack_util()))
}
# These intervening frames won't appear in the evaluation stack
identity(user_fn())

#> [[1]]
#> <frame 24> (22)
#> expr: user_fn()
#> env:  [local 0x7fa65f23da80]
#> 
#> [[2]]
#> <frame 23> (22)
#> expr: identity(user_fn())
#> env:  [local 0x7fa65f23da10]
#> 
#> [[3]]
#> <frame 22> (21)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fa65f2d2758]
#> 
#> [[4]]
#> <frame 21> (11)
#> expr: eval(expr, envir, enclos)
#> env:  [local 0x7fa65f23d7e0]
#> 
#> [[5]]
#> <frame 20> (11)
#> expr: withVisible(eval(expr, envir, enclos))
#> env:  [local 0x7fa65f23d540]
#> 
#> [[6]]
#> <frame 19> (11)
#> expr: withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = wHandler,  <...>
#> env:  [local 0x7fa65f23cee8]
#> 
#> [[7]]
#> <frame 18> (17)
#> expr: doTryCatch(return(expr), name, parentenv, handler)
#> env:  [local 0x7fa65f23c970]
#> 
#> [[8]]
#> <frame 17> (16)
#> expr: tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> env:  [local 0x7fa65f23c5f0]
#> 
#> [[9]]
#> <frame 16> (15)
#> expr: tryCatchList(expr, classes, parentenv, handlers)
#> env:  [local 0x7fa65f23c190]
#> 
#> [[10]]
#> <frame 15> (14)
#> expr: tryCatch(expr, error = function(e) { <...>
#> env:  [local 0x7fa65f23f888]
#> 
#> [[11]]
#> <frame 14> (13)
#> expr: try(f, silent = TRUE)
#> env:  [local 0x7fa65f23f658]
#> 
#> [[12]]
#> <frame 13> (11)
#> expr: handle(ev <- withCallingHandlers(withVisible(eval(expr, envir,  <...>
#> env:  [local 0x7fa65f23f498]
#> 
#> [[13]]
#> <frame 12> (11)
#> expr: timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr,  <...>
#> env:  [local 0x7fa65f23f3b8]
#> 
#> [[14]]
#> <frame 11> (10)
#> expr: evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos,  <...>
#> env:  [frame 0x7fa65f267ff8]
#> 
#> [[15]]
#> <frame 10> (9)
#> expr: evaluate::evaluate(x, child_env(env), new_device = TRUE)
#> env:  [frame 0x7fa6603785d8]
#> 
#> [[16]]
#> <frame 9> (8)
#> expr: highlight_examples(code, topic, env = env)
#> env:  [frame 0x7fa660bcc388]
#> 
#> [[17]]
#> <frame 8> (7)
#> expr: run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv()),  <...>
#> env:  [frame 0x7fa6606fc158]
#> 
#> [[18]]
#> <frame 7> (6)
#> expr: data_reference_topic(topic, pkg, examples = examples, run_dont_run = run_dont_run)
#> env:  [frame 0x7fa662f26680]
#> 
#> [[19]]
#> <frame 6> (5)
#> expr: .f(.x[[i]], ...)
#> env:  [frame 0x7fa662eb4f50]
#> 
#> [[20]]
#> <frame 5> (4)
#> expr: purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy,  <...>
#> env:  [frame 0x7fa65a09e4f8]
#> 
#> [[21]]
#> <frame 4> (3)
#> expr: build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fa65e9d0910]
#> 
#> [[22]]
#> <frame 3> (2)
#> expr: build_site_local(pkg = pkg, examples = examples, run_dont_run = run_dont_run,  <...>
#> env:  [frame 0x7fa65b6abf30]
#> 
#> [[23]]
#> <frame 2> (1)
#> expr: build_site(".", override = list(destination = dest_dir), devel = FALSE,  <...>
#> env:  [frame 0x7fa65d322948]
#> 
#> [[24]]
#> <frame 1> (0)
#> expr: pkgdown:::deploy_local(new_process = FALSE, remote_url = url)
#> env:  [frame 0x7fa65d091540]
#> 
#> [[25]]
#> <frame 0> [global]
#> expr: NULL
#> env:  [global]
#> 
#> attr(,"class")
#> [1] "ctxt_stack" "stack"