tidyeval-data {rlang}	R Documentation

Data pronoun for tidy evaluation

Description

This pronoun is installed by functions performing tidy evaluation. It allows you to refer to overscoped data explicitly.

Usage

.data

Format

An object of class dictionary of length 0.

Details

You can import this object in your package namespace to avoid R CMD check errors when referring to overscoped objects.

Examples

quo <- quo(.data$foo)
eval_tidy(quo, list(foo = "bar"))

[Package rlang version 0.1.6 Index]