Capture a call trace.

A call trace captures the sequence of calls that lead to the current function, sometimes called the call stack. Because of lazy evaluation, the call stack in R is actually a tree, which the print method of this object will reveal.

trace_back(top = NULL, bottom = NULL)

Arguments

top
The first frame environment to be included in the backtrace. This becomes the top of the backtrace tree and represents the oldest call in the backtrace.

This is needed in particular when you call trace_back() indirectly or from a larger context, for example in tests or inside an RMarkdown document where you don't want all of the knitr evaluation mechanisms to appear in the backtrace.

bottom
The last frame environment to be included in the backtrace. This becomes the rightmost leaf of the backtrace tree and represents the youngest call in the backtrace.

Set this when you would like to capture a backtrace without the capture context.

Can also be an integer that will be passed to caller_env().

Examples

# Trim backtraces automatically (this improves the generated
# documentation for the rlang website and the same trick can be
# useful within knitr documents):
options(rlang_trace_top_env = current_env())

f <- function() g()
g <- function() h()
h <- function() trace_back()

# When no lazy evaluation is involved the backtrace is linear
# (i.e. every call has only one child)
f()

#> █
#> └─global::f()
#>   └─global::g()
#>     └─global::h()

# Lazy evaluation introduces a tree like structure
identity(identity(f()))

#> █
#> ├─[ base::identity(...) ]
#> ├─[ base::identity(...) ]
#> └─global::f()
#>   └─global::g()
#>     └─global::h()

identity(try(f()))

#> █
#> ├─[ base::identity(...) ]
#> ├─[ base::try(...) ] with 4 more calls
#> └─global::f()
#>   └─global::g()
#>     └─global::h()

try(identity(f()))

#> █
#> ├─[ base::try(...) ] with 4 more calls
#> ├─[ base::identity(...) ]
#> └─global::f()
#>   └─global::g()
#>     └─global::h()

# When printing, you can request to simplify this tree to only show
# the direct sequence of calls that lead to `trace_back()`
x <- try(identity(f()))
x

#> █
#> ├─[ base::try(...) ] with 4 more calls
#> ├─[ base::identity(...) ]
#> └─global::f()
#>   └─global::g()
#>     └─global::h()

print(x, simplify = "branch")

#>  ─base::try(identity(f()))
#>  ─global::f()
#>  ─global::g()
#>  ─global::h()

# With a little cunning you can also use it to capture the
# tree from within a base NSE function
x <- NULL
with(mtcars, {x <<- f(); 10})

#> [1] 10

x

#> █
#> ├─[ base::with(...) ]
#> └─base::with.default(...)
#>   └─[ base::eval(...) ] with 1 more call
#>     └─global::f()
#>       └─global::g()
#>         └─global::h()

# When code is executed indirectly, i.e. via source or within an
# RMarkdown document, you'll tend to get a lot of guff at the beginning
# related to the execution environment:
conn <- textConnection("f()")
source(conn, echo = TRUE)

#> 
#> > f()

#> Error in f(): could not find function "f"

close(conn)

# To automatically strip this off, pass `top = globalenv()`.
# This will automatically trim off calls prior to the last appearance
# of the global environment on the stack
h <- function() trace_back(globalenv())

conn <- textConnection("f()")
source(conn, echo = TRUE)

#> 
#> > f()

#> Error in f(): could not find function "f"

close(conn)

# Restore defaults
options(rlang_trace_top_env = NULL)